Agda.Compiler.Backend

Description

Interface for compiler backend writers.

Documentation

module Agda.Compiler.Backend.Base

data Recompile menv mod

Constructors

Recompile menv

Skip mod

data IsMain

Constructors

IsMain

NotMain

type Flag opts = opts -> OptM opts

f :: Flag opts is an action on the option record that results from parsing an option. f opts produces either an error message or an updated options record

toTreeless :: EvaluationStrategy -> QName -> TCM (Maybe TTerm)

Convert to treeless using default pipeline

module Agda.Syntax.Treeless

module Agda.TypeChecking.Monad

crInterface :: CheckResult -> Interface

crMode :: CheckResult -> ModuleCheckMode

crWarnings :: CheckResult -> Set TCWarning

data CheckResult where

The result and associated parameters of a type-checked file, when invoked directly via interaction or a backend. Note that the constructor is not exported.

Bundled Patterns

pattern CheckResult :: Interface -> Set TCWarning -> ModuleCheckMode -> Source -> CheckResult
Flattened unidirectional pattern for CheckResult for destructuring inside the ModuleInfo field.

backendInteraction :: AbsolutePath -> [Backend] -> TCM () -> (AbsolutePath -> TCM CheckResult) -> TCM ()

parseBackendOptions :: [Backend] -> [String] -> CommandLineOptions -> OptM ([Backend], CommandLineOptions)

callBackend :: BackendName -> IsMain -> CheckResult -> TCM ()

Call the compilerMain function of the given backend.

callBackendInteractTop :: BackendName -> String -> CommandM ()

Call the backendInteractTop function of the given backend.

callBackendInteractHole :: BackendName -> String -> InteractionId -> Range -> String -> CommandM ()

Call the backendInteractHole function of the given backend.