Agda.Compiler.JS.Compiler

Description

Main module for JS backend.

Documentation

jsBackend :: Backend

jsBackend' :: Backend' JSOptions JSOptions JSModuleEnv Module (Maybe Export)

data JSOptions

Constructors

JSOptions

Fields

optJSCompile :: Bool

optJSOptimize :: Bool

optJSMinify :: Bool

Remove spaces etc. See https://en.wikipedia.org/wiki/Minification_(programming).

optJSVerify :: Bool

Run generated code through interpreter.

optJSModuleStyle :: JSModuleStyle

defaultJSOptions :: JSOptions

jsCommandLineFlags :: [OptDescr (Flag JSOptions)]

jsPreCompile :: JSOptions -> TCM JSOptions

jsPostCompile :: JSOptions -> IsMain -> Map TopLevelModuleName Module -> TCM ()

After all modules have been compiled, copy RTE modules and verify compiled modules.

data JSModuleEnv

Constructors

JSModuleEnv

Fields

jsCoinductionKit :: Maybe CoinductionKit

jsCompile :: Bool

Should this module be compiled?

jsPreModule :: JSOptions -> IsMain -> TopLevelModuleName -> Maybe FilePath -> TCM (Recompile JSModuleEnv Module)

jsPostModule :: JSOptions -> JSModuleEnv -> IsMain -> TopLevelModuleName -> [Maybe Export] -> TCM Module

jsCompileDef :: JSOptions -> JSModuleEnv -> IsMain -> Definition -> TCM (Maybe Export)

prefix :: [Char]

jsMod :: TopLevelModuleName -> GlobalId

jsFileName :: GlobalId -> String

jsMember :: Name -> MemberId

global' :: QName -> TCM (Exp, JSQName)

global :: QName -> TCM (Exp, JSQName)

reorder :: [Export] -> [Export]

reorder' :: Set JSQName -> [Export] -> [Export]

isTopLevelValue :: Export -> Bool

isEmptyObject :: Export -> Bool

insertAfter :: Set JSQName -> Export -> [Export] -> [Export]

type EnvWithOpts = (JSOptions, JSModuleEnv)

definition :: EnvWithOpts -> (QName, Definition) -> TCM (Maybe Export)

checkCompilerPragmas :: QName -> TCM ()

Ensure that there is at most one pragma for a name.

defJSDef :: Definition -> Maybe String

definition' :: EnvWithOpts -> QName -> Definition -> Type -> JSQName -> TCM (Maybe Export)

compileTerm :: EnvWithOpts -> TTerm -> TCM Exp

compilePrim :: TPrim -> Exp

compileAlt :: EnvWithOpts -> TAlt -> TCM ((QName, MemberId), Exp)

eraseLocalVars :: [Bool] -> TTerm -> TTerm

visitorName :: QName -> TCM MemberId

flatName :: MemberId

local :: Nat -> Exp

qname :: QName -> TCM Exp

literal :: Literal -> Exp

litqname :: QName -> Exp

litmeta :: MetaId -> Exp

writeModule :: Bool -> JSModuleStyle -> Module -> TCM ()

outFile :: GlobalId -> TCM FilePath

outFile_ :: TCM FilePath

primitives :: Set PrimitiveId

Primitives implemented in the JS Agda RTS.

TODO: Primitives that are not part of this set, and for which defJSDef does not return anything, are silently compiled to Undefined. A better approach might be to list exactly those primitives which should be compiled to Undefined.

Orphan instances

NFData JSModuleStyle

Instance details

Methods

rnf :: JSModuleStyle -> ()