Agda.Compiler.JS.Pretty

Documentation

data JSModuleStyle

Constructors

JSES6

JSCJS

JSAMD

data Doc

Constructors

Doc String

Indent Int Doc

Group Doc

Beside Doc Doc

Above Doc Doc

Enclose Doc Doc Doc

Space

Empty

minifiedCodeLinesLength :: Int

render :: Bool -> Doc -> String

($+$) :: Doc -> Doc -> Doc infixr 5

($++$) :: Doc -> Doc -> Doc infixr 5

Separate by blank line.

(<+>) :: Doc -> Doc -> Doc infixr 6

Separate by space that will be removed by minify.

For non-removable space, use d <> " " <> d'.

text :: String -> Doc

group :: Doc -> Doc

indentBy :: Int -> Doc -> Doc

enclose :: Doc -> Doc -> Doc -> Doc

space :: Doc

indent :: Doc -> Doc

hcat :: [Doc] -> Doc

vcat :: [Doc] -> Doc

vsep :: [Doc] -> Doc

Concatenate vertically, separated by blank lines.

punctuate :: Doc -> [Doc] -> Doc

parens :: Doc -> Doc

brackets :: Doc -> Doc

braces :: Doc -> Doc

mparens :: Bool -> Doc -> Doc

Apply parens to Doc if boolean is true.

unescape :: Char -> String

unescapes :: String -> Doc

class Pretty a where

Minimal complete definition

pretty | prettyPrec

Methods

pretty :: (Nat, Bool, JSModuleStyle) -> a -> Doc

prettyPrec :: Int -> (Nat, Bool, JSModuleStyle) -> a -> Doc

prettyShow :: Pretty a => Bool -> JSModuleStyle -> a -> String

class Pretties a where

Methods

pretties :: (Nat, Bool, JSModuleStyle) -> a -> [Doc]

block :: (Nat, Bool, JSModuleStyle) -> Exp -> Doc

modname :: GlobalId -> Doc

exports :: (Nat, Bool, JSModuleStyle) -> Set JSQName -> [Export] -> Doc

variableName :: String -> String

isValidJSIdent :: String -> Bool

Check if a string is a valid JS identifier. The check ignores keywords as we prepend z_ to our identifiers. The check is conservative and may not admit all valid JS identifiers.