Agda.Compiler.MAlonzo.Compiler

Documentation

ghcBackend :: Backend

ghcBackend' :: Backend' GHCFlags GHCEnv GHCModuleEnv GHCModule GHCDefinition

data GHCFlags

Constructors

GHCFlags

Fields

flagGhcCompile :: Bool

flagGhcCallGhc :: Bool

flagGhcBin :: Maybe FilePath

Use the compiler at PATH instead of "ghc"

flagGhcFlags :: [String]

flagGhcStrictData :: Bool

Make inductive constructors strict?

flagGhcStrict :: Bool

Make functions strict?

defaultGHCFlags :: GHCFlags

ghcCommandLineFlags :: [OptDescr (Flag GHCFlags)]

withCompilerFlag :: FilePath -> Flag GHCFlags

class Monad m => ReadGHCOpts m where

Monads that can read GHCOptions

Methods

askGhcOpts :: m GHCOptions

data GHCModule

Constructors

GHCModule

Fields

ghcModModuleEnv :: GHCModuleEnv

ghcModMainFuncs :: [MainFunctionDef]

The main function definition(s), if both the module is the IsMain module (root/focused) and a suitable main function was defined.

data GHCDefinition

Constructors

GHCDefinition

Fields

ghcDefUsesFloat :: UsesFloat

ghcDefDecls :: [Decl]

ghcDefDefinition :: Definition

ghcDefMainDef :: Maybe MainFunctionDef

ghcDefImports :: Set ModuleName

ghcPreCompile :: GHCFlags -> TCM GHCEnv

ghcPostCompile :: GHCEnv -> IsMain -> Map ModuleName GHCModule -> TCM ()

ghcPreModule

Arguments

:: GHCEnv

-> IsMain
Are we looking at the main module?

-> ModuleName

-> Maybe FilePath
Path to the .agdai file.

-> TCM (Recompile GHCModuleEnv GHCModule)
Could we confirm the existence of a main function?

ghcPostModule

Arguments

:: GHCEnv

-> GHCModuleEnv

-> IsMain
Are we looking at the main module?

-> ModuleName

-> [GHCDefinition]
Compiled module content.

-> TCM GHCModule

ghcCompileDef :: GHCEnv -> GHCModuleEnv -> IsMain -> Definition -> TCM GHCDefinition

ghcMayEraseType :: QName -> TCM Bool

We do not erase types that have a HsData pragma. This is to ensure a stable interface to third-party code.

imports :: BuiltinThings PrimFun -> Set ModuleName -> [Definition] -> [ImportDecl]

newtype UsesFloat

Constructors

UsesFloat Bool

pattern YesFloat :: UsesFloat

pattern NoFloat :: UsesFloat

mazRTEFloatImport :: UsesFloat -> [ImportDecl]

definition :: Definition -> HsCompileM (UsesFloat, [Decl], Maybe CheckedMainFunctionDef)

constructorCoverageCode :: QName -> Int -> [QName] -> HaskellType -> [HaskellCode] -> HsCompileM [Decl]

data CCEnv

Environment for naming of local variables. Invariant: reverse ccCxt ++ ccNameSupply

Constructors

CCEnv

Fields

_ccNameSupply :: NameSupply

Supply of fresh names

_ccContext :: CCContext

Names currently in scope

type NameSupply = [Name]

type CCContext = [Name]

ccNameSupply :: Lens' NameSupply CCEnv

ccContext :: Lens' CCContext CCEnv

initCCEnv :: CCEnv

Initial environment for expression generation.

lookupIndex :: Int -> CCContext -> Name

Term variables are de Bruijn indices.

type CCT m = ReaderT CCEnv (WriterT UsesFloat (HsCompileT m))

Constructor coverage monad transformer

type CC = CCT TCM

Constructor coverage monad

liftCC :: Monad m => HsCompileT m a -> CCT m a

freshNames :: Monad m => Int -> ([Name] -> CCT m a) -> CCT m a

intros :: Monad m => Int -> ([Name] -> CCT m a) -> CCT m a

Introduce n variables into the context.

checkConstructorType :: QName -> HaskellCode -> HsCompileM [Decl]

checkCover :: HasConstInfo m => QName -> HaskellType -> Nat -> [QName] -> [HaskellCode] -> m [Decl]

closedTerm_ :: TTerm -> HsCompileM Exp

closedTerm :: TTerm -> HsCompileM (Exp, UsesFloat)

mkIf :: TTerm -> CC TTerm

term :: TTerm -> CC Exp

Extract Agda term to Haskell expression. Erased arguments are extracted as (). Types are extracted as ().

noApplication :: TTerm -> CC Exp

Translate a non-application, non-coercion, non-constructor, non-definition term.

hsCoerce :: Exp -> Exp

compilePrim :: TPrim -> Exp

alt :: Int -> TAlt -> CC Alt

literal :: forall m. Monad m => Literal -> CCT m Exp

hslit :: Literal -> Literal

litString :: Text -> Exp

litqname :: QName -> Exp

litqnamepat :: QName -> Pat

condecl :: QName -> Induction -> HsCompileM ConDecl

compiledcondecl

Arguments

:: Maybe Nat
The constructor's arity (after erasure).

-> QName

-> HsCompileM Decl

compiledTypeSynonym :: QName -> String -> Nat -> Decl

tvaldecl

Arguments

:: QName

-> Induction
Is the type inductive or coinductive?

-> Nat

-> [ConDecl]

-> Maybe Clause

-> [Decl]

infodecl :: QName -> [Decl] -> [Decl]

type MonadGHCIO m = (MonadIO m, ReadGHCOpts m)

copyRTEModules :: MonadGHCIO m => m ()

writeModule :: MonadGHCIO m => Module -> m ()

outFileAndDir :: MonadGHCIO m => ModuleName -> m (FilePath, FilePath)

curOutFileAndDir :: (MonadGHCIO m, ReadGHCModuleEnv m) => m (FilePath, FilePath)

curOutFile :: (MonadGHCIO m, ReadGHCModuleEnv m) => m FilePath

callGHC :: ReaderT GHCModule TCM ()