Agda.Compiler.MAlonzo.Pragmas

Documentation

type GHCPragma = String

Raw Haskell pragma (something enclosed in {-# ... #-}).

type GHCImport = String

Raw Haskell import (something starting with import ).

type HaskellCode = String

type HaskellType = String

data HaskellPragma

GHC backend translation pragmas.

Constructors

HsDefn Range HaskellCode

HsType Range HaskellType

HsData Range HaskellType [HaskellCode]
@COMPILE GHC X = data D (c₁ | ... | cₙ)

HsExport Range HaskellCode
COMPILE GHC x as f

getHaskellPragma :: QName -> TCM (Maybe HaskellPragma)

Retrieve and parse a COMPILE GHC pragma stored for a name.

parsePragma :: CompilerPragma -> Maybe HaskellPragma

parseHaskellPragma :: CompilerPragma -> TCM (Maybe HaskellPragma)

sanityCheckPragma :: QName -> HaskellPragma -> TCM (Maybe HaskellPragma)

Check whether the parsed COMPILE GHC pragma matches the kind of identifier it attaches to.

getHaskellConstructor :: QName -> HsCompileM (Maybe HaskellCode)

foreignHaskell :: Interface -> (Set GHCPragma, Set GHCImport, [HaskellCode])

Get content of FOREIGN GHC pragmas, sorted by KindOfForeignCode: file header pragmas, import statements, rest.

data KindOfForeignCode

Classify FOREIGN Haskell code.

Constructors

ForeignFileHeaderPragma
A pragma that must appear before the module header.

ForeignImport
An import statement. Must appear right after the module header.

ForeignOther
The rest. To appear after the import statements.

classifyForeign :: String -> KindOfForeignCode

Classify a FOREIGN GHC declaration.

classifyPragma :: String -> KindOfForeignCode

Classify a Haskell pragma into whether it is a file header pragma or not.

partitionByKindOfForeignCode :: (a -> KindOfForeignCode) -> [a] -> ([a], [a], [a])

Partition a list by KindOfForeignCode attribute.