Agda.Compiler.MAlonzo.Strict

Description

Strictification of Haskell code

Documentation

class MakeStrict a where

The function makeStrict makes every function argument, case and generator pattern, and LocalBind binding strict (except for those patterns that are marked as irrefutable, and anything in a FakeDecl or FakeExp). Note that only the outermost patterns are made strict.

Methods

makeStrict :: a -> a