Agda.Compiler.ToTreeless

Documentation

toTreeless :: EvaluationStrategy -> QName -> TCM (Maybe TTerm)

toTreelessWith :: BuildPipeline -> CCConfig -> QName -> TCM (Maybe TTerm)

Converts compiled clauses to treeless syntax.

Note: Do not use any of the concrete names in the returned term for identification purposes! If you wish to do so, first apply the Agda.Compiler.Treeless.NormalizeNames transformation.

closedTermToTreeless :: CCConfig -> Term -> TCM TTerm

data Pipeline

Constructors

FixedPoint Int Pipeline

Sequential [Pipeline]

SinglePass CompilerPass

data CompilerPass

Constructors

CompilerPass

Fields

passTag :: String

passVerbosity :: Int

passName :: String

passCode :: EvaluationStrategy -> TTerm -> TCM TTerm

compilerPass :: String -> Int -> String -> (EvaluationStrategy -> TTerm -> TCM TTerm) -> Pipeline

compilerPipeline :: BuildPipeline

type CCConfig = (EvaluationStrategy, CCSubst)

data CCSubst

Constructors

EraseUnused

IgnoreUnused