Agda.Interaction.EmacsCommand

Description

Code for instructing Emacs to do things

Documentation

displayInfo :: String -> DocTree -> Bool -> Maybe ModuleToSource -> Lisp String

displayInfo header content append m displays content with header header in some suitable way in the Agda information buffer. If append is True, then the content is appended to previous content (if any), otherwise any previous content is deleted. If m is Just ModuleToSource then content is rendered with annotations, otherwise annotations in content are discarded and just the text is displayed.

clearRunningInfo :: Lisp String

Clear the running info buffer.

displayRunningInfo :: DocTree -> Maybe ModuleToSource -> Lisp String

Display running information about what the type-checker is up to.