Agda.Interaction.EmacsTop

Documentation

mimicGHCi :: TCM () -> TCM ()

mimicGHCi is a fake ghci interpreter for the Emacs frontend and for interaction tests.

mimicGHCi reads the Emacs frontend commands from stdin, interprets them and print the result into stdout.

namedMetaOf :: OutputConstraint Expr a -> a

showGoals :: Goals -> TCM String

Print open metas nicely.

showInfoError :: Info_Error -> TCM String

Serializing Info_Error

explainWhyInScope :: MonadPretty m => WhyInScopeData -> m Doc

prettyResponseContext

Arguments

:: InteractionId
Context of this meta-variable.

-> Bool
Print the elements in reverse order?

-> [ResponseContextEntry]

-> TCM Doc

Pretty-prints the context of the given meta-variable.

prettyTypeOfMeta :: Rewrite -> InteractionId -> TCM Doc

Pretty-prints the type of the meta-variable.