Agda.Interaction.FindFile

Description

Functions which map between module names and file names.

Note that file name lookups are cached in the TCState. The code assumes that no Agda source files are added or removed from the include directories while the code is being type checked.

Documentation

newtype SourceFile

SourceFiles must exist and be registered in our file dictionary.

Constructors

SourceFile

Fields

srcFileId :: FileId

data InterfaceFile

File must exist.

toIFile :: HasCallStack => SourceFile -> TCM AbsolutePath

Converts an Agda file name to the corresponding interface file name. Note that we do not guarantee that the file exists.

mkInterfaceFile

Arguments

:: AbsolutePath
Path to the candidate interface file

-> IO (Maybe InterfaceFile)
Interface file iff it exists

Makes an interface file from an AbsolutePath candidate. If the file does not exist, then fail by returning Nothing.

data FindError

Errors which can arise when trying to find a source file.

Invariant: All paths are absolute.

Constructors

NotFound (List1 AbsolutePath)
An Agda module was not found. We were looking for it in one of the given include directories.

Ambiguous (List2 AbsolutePath)
Several matching files were found.

findErrorToTypeError :: TopLevelModuleName -> FindError -> TypeError

Given the module name which the error applies to this function converts a FindError to a TypeError.

findFile :: TopLevelModuleName -> TCM SourceFile

Finds the source file corresponding to a given top-level module name. The returned paths are absolute.

Raises an error if the file cannot be found.

findFile' :: TopLevelModuleName -> TCM (Either FindError SourceFile)

Tries to find the source file corresponding to a given top-level module name. The returned paths are absolute.

SIDE EFFECT: Updates stModuleToSource.

findFile'_

Arguments

:: List1 AbsolutePath
Include paths.

-> TopLevelModuleName

-> StateT ModuleToSourceId TCM (Either FindError SourceFile)

A variant of findFile' which manipulates an extra ModuleToSourceId

findFile''

Arguments

:: List1 AbsolutePath
Include paths.

-> TopLevelModuleName

-> StateT ModuleToSource IO (Either FindError SourceFile)

A variant of findFile' which does not require TCM.

findInterfaceFile'

Arguments

:: HasCallStack

=> SourceFile
Path to the source file

-> TCM (Maybe InterfaceFile)
Maybe path to the interface file

Finds the interface file corresponding to a given top-level module file. The returned paths are absolute.

Raises Nothing if the interface file cannot be found.

findInterfaceFile :: HasCallStack => TopLevelModuleName -> TCM (Maybe InterfaceFile)

Finds the interface file corresponding to a given top-level module file. The returned paths are absolute.

Raises an error if the source file cannot be found, and returns Nothing if the source file can be found but not the interface file.

checkModuleName

Arguments

:: TopLevelModuleName
The name of the module.

-> SourceFile
The file from which it was loaded.

-> Maybe TopLevelModuleName
The expected name, coming from an import statement.

-> TCM ()

Ensures that the module name matches the file name. The file corresponding to the module name (according to the include path) has to be the same as the given file name.

rootNameModule :: AbsolutePath -> String

replaceModuleExtension :: String -> AbsolutePath -> AbsolutePath

dropAgdaExtension :: FilePath -> FilePath

Remove an existing Agda file extension from a file path.

hasAgdaExtension :: FilePath -> Bool

Check if a file path has an Agda extension.

stripAgdaExtension :: FilePath -> Maybe FilePath

Remove an Agda file extension from a filepath, if possible.