Agda.Interaction.Highlighting.Emacs

Description

Functions which give precise syntax highlighting info to Emacs.

Documentation

lispifyHighlightingInfo

Arguments

:: HighlightingInfo

-> RemoveTokenBasedHighlighting

-> HighlightingMethod

-> ModuleToSource
Must contain a mapping for every definition site's module.

-> IO (Lisp String)

Turns syntax highlighting information into a list of S-expressions.

lispifyTokenBased :: TokenBased -> Lisp String

Formats the TokenBased tag for the Emacs backend. No quotes are added.