Agda.Interaction.Highlighting.Precise

Description

Types used for precise syntax highlighting.

Files

data Aspect

Syntactic aspects of the code. (These cannot overlap.)

Constructors

Comment

Keyword

String

Number

Hole

Symbol
Symbols like forall, =, ->, etc.

PrimitiveType
Things like Set and Prop.

Name (Maybe NameKind) Bool
Is the name an operator part?

Pragma
Text occurring in pragmas that does not have a more specific aspect.

Background
Non-code contents in literate Agda

Markup
Delimiters used to separate the Agda code blocks from the other contents in literate Agda

data NameKind

NameKinds are figured out during scope checking.

Constructors

Bound
Bound variable.

Generalizable
Generalizable variable. (This includes generalizable variables that have been generalized).

Constructor Induction
Inductive or coinductive constructor.

Datatype

Field
Record field.

Function

Module
Module name.

Postulate

Primitive
Primitive.

Record
Record type.

Argument
Named argument, like x in {x = v}

Macro
Macro.

data OtherAspect

Other aspects, generated by type checking. (These can overlap with each other and with Aspects.)

Constructors

Error

ErrorWarning
A warning that is considered fatal in the end.

DottedPattern

UnsolvedMeta

UnsolvedConstraint
Unsolved constraint not connected to meta-variable. This could for instance be an emptyness constraint.

TerminationProblem

PositivityProblem

Deadcode
Used for highlighting unreachable clauses, unreachable RHS (because of an absurd pattern), etc.

ShadowingInTelescope
Used for shadowed repeated variable names in telescopes.

CoverageProblem

IncompletePattern
When this constructor is used it is probably a good idea to include a note explaining why the pattern is incomplete.

TypeChecks
Code which is being type-checked.

MissingDefinition
Function declaration without matching definition NB: We put CatchallClause last so that it is overwritten by other, more important, aspects in the emacs mode.

CatchallClause

ConfluenceProblem

data Aspects

Meta information which can be associated with a character/character range.

Constructors

Aspects

Fields

aspect :: Maybe Aspect

otherAspects :: Set OtherAspect

note :: String

This note, if not null, can be displayed as a tool-tip or something like that. It should contain useful information about the range (like the module containing a certain identifier, or the fixity of an operator).

definitionSite :: Maybe DefinitionSite

The definition site of the annotated thing, if applicable and known.

tokenBased :: !TokenBased

Is this entry token-based?

data DefinitionSite

Constructors

DefinitionSite

Fields

defSiteModule :: TopLevelModuleName

The defining module.

defSitePos :: Int

The file position in that module. File positions are counted from 1.

defSiteHere :: Bool

Has this DefinitionSite been created at the defining site of the name?

defSiteAnchor :: Maybe String

A pretty name for the HTML linking.

data TokenBased

Is the highlighting "token-based", i.e. based only on information from the lexer?

Constructors

TokenBased

NotOnlyTokenBased

newtype File

A File is a mapping from file positions to meta information.

The first position in the file has number 1.

Constructors

File

Fields

mapping :: IntMap Aspects

type HighlightingInfo = CompressedFile

Syntax highlighting information.

Creation

parserBased :: Aspects

A variant of mempty with tokenBased set to NotOnlyTokenBased.

singleton :: Ranges -> Aspects -> File

singleton rs m is a file whose positions are those in rs, and in which every position is associated with m.

several :: [Ranges] -> Aspects -> File

Like singleton, but with several Ranges instead of only one.

kindOfNameToNameKind :: KindOfName -> NameKind

Conversion from classification of the scope checker.

Merging

merge :: File -> File -> File

Merges files.

Inspection

smallestPos :: File -> Maybe Int

Returns the smallest position, if any, in the File.

toMap :: File -> IntMap Aspects

Convert the File to a map from file positions (counting from 1) to meta information.

Compressed files

newtype CompressedFile

A compressed File, in which consecutive positions with the same Aspects are stored together.

Constructors

CompressedFile

Fields

ranges :: [(Range, Aspects)]

compressedFileInvariant :: CompressedFile -> Bool

Invariant for compressed files.

Note that these files are not required to be maximally compressed, because ranges are allowed to be empty, and the Aspectss in adjacent ranges are allowed to be equal.

compress :: File -> CompressedFile

Compresses a file by merging consecutive positions with equal meta information into longer ranges.

decompress :: CompressedFile -> File

Decompresses a compressed file.

noHighlightingInRange :: Ranges -> CompressedFile -> CompressedFile

Clear any highlighting info for the given ranges. Used to make sure unsolved meta highlighting overrides error highlighting.

Creation

singletonC :: Ranges -> Aspects -> CompressedFile

singletonC rs m is a file whose positions are those in rs, and in which every position is associated with m.

severalC :: [Ranges] -> Aspects -> CompressedFile

Like singletonR, but with a list of Ranges instead of a single one.

splitAtC :: Int -> CompressedFile -> (CompressedFile, CompressedFile)

splitAtC p f splits the compressed file f into (f1, f2), where all the positions in f1 are < p, and all the positions in f2 are >= p.

selectC :: Range -> CompressedFile -> CompressedFile

Inspection

smallestPosC :: CompressedFile -> Maybe Int

Returns the smallest position, if any, in the CompressedFile.

Merge

mergeC :: CompressedFile -> CompressedFile -> CompressedFile

Merges compressed files.