Agda.Interaction.InteractionTop

Documentation

atTopLevel :: CommandM a -> CommandM a

cmd_goal_type_context_and :: GoalTypeAux -> Rewrite -> InteractionId -> Range -> String -> CommandM ()

Displays the current goal, the given document, and the current context.

Should not modify the state.

cmd_load'

Arguments

:: FilePath
File to load into interaction.

-> [String]
Arguments to Agda for loading this file

-> Bool
Allow unsolved meta-variables?

-> Mode
Full type-checking, or only scope-checking?

-> (CheckResult -> CommandM a)
Continuation after successful loading.

-> CommandM a

cmd_load' file argv unsolvedOk cmd loads the module in file file, using argv as the command-line options.

If type checking completes without any exceptions having been encountered then the command cmd r is executed, where r is the result of typeCheckMain.

commandMToIO :: (forall x. (CommandM a -> IO x) -> IO x) -> CommandM a

Opposite of liftIO for CommandM.

This function should only be applied to computations that are guaranteed not to raise any errors (except for IOExceptions).

decorate :: Doc -> String

displayStatus :: CommandM ()

Displays or updates status information.

Does not change the state.

display_info :: DisplayInfo -> CommandM ()

display_info does what display_info' False does, but additionally displays some status information (see status and displayStatus).

extlam_dropName :: UnicodeOrAscii -> CaseContext -> String -> String

getBackendName :: CompilerBackend -> BackendName

Interpret an interaction

getOldInteractionScope :: InteractionId -> CommandM ScopeInfo

give_gen

Arguments

:: UseForce
Should safety checks be skipped?

-> InteractionId

-> Range

-> String

-> GiveRefine

-> CommandM ()

A "give"-like action (give, refine, etc).

give_gen force ii rng s give_ref mk_newtxt acts on interaction point ii occupying range rng, placing the new content given by string s, and replacing ii by the newly created interaction points in the state if safety checks pass (unless force is applied).

handleCommand :: (forall a. CommandM a -> CommandM a) -> CommandM () -> CommandM () -> CommandM ()

handleCommand_ :: CommandM () -> CommandM ()

Do setup and error handling for a command.

highlightExpr :: Expr -> TCM ()

independent :: Interaction -> Bool

Can the command run even if the relevant file has not been loaded into the state?

initialiseCommandQueue

Arguments

:: IO Command
Returns the next command.

-> IO CommandQueue

Creates a command queue, and forks a thread that writes commands to the queue. The queue is returned.

insertOldInteractionScope :: InteractionId -> ScopeInfo -> CommandM ()

interpret :: Interaction -> CommandM ()

liftCommandMT :: (forall x. TCM x -> TCM x) -> CommandM a -> CommandM a

Lift a TCM action transformer to a CommandM action transformer.

liftCommandMTLocalState :: (forall x. TCM x -> TCM x) -> CommandM a -> CommandM a

Ditto, but restore state.

makeCaseVariant :: CaseContext -> MakeCaseVariant

maybeAbort :: (IOTCM -> CommandM a) -> CommandM (Command' (Maybe a))

If the next command from the command queue is anything but an actual command, then the command is returned.

If the command is an IOTCM command, then the following happens: The given computation is applied to the command and executed. If an abort command is encountered (and acted upon), then the computation is interrupted, the persistent state and all options are restored, and some commands are sent to the frontend. If the computation was not interrupted, then its result is returned.

maybeTimed :: CommandM a -> CommandM (Maybe CPUTime, a)

modifyOldInteractionScopes :: (OldInteractionScopes -> OldInteractionScopes) -> CommandM ()

A Lens for oldInteractionScopes.

modifyTheInteractionPoints :: ([InteractionId] -> [InteractionId]) -> CommandM ()

A Lens for theInteractionPoints.

parseAndDoAtToplevel

Arguments

:: (Expr -> TCM a)
The command to perform.

-> String
The expression to parse.

-> CommandM (Maybe CPUTime, a)

Parses and scope checks an expression (using the "inside scope" as the scope), performs the given command with the expression as input, and returns the result and the time it takes.

putResponse :: Response -> CommandM ()

Put a response by the callback function given by stInteractionOutputCallback.

removeOldInteractionScope :: InteractionId -> CommandM ()

runInteraction :: IOTCM -> CommandM ()

Run an IOTCM value, catch the exceptions, emit output

If an error happens the state of CommandM does not change, but stPersistent may change (which contains successfully loaded interfaces for example).

searchAbout :: Rewrite -> Range -> String -> CommandM ()

Shows all the top-level names in scope which mention all the given identifiers in their type.

setCommandLineOpts :: CommandLineOptions -> CommandM ()

Sets the command line options and updates the status information.

showModuleContents :: Rewrite -> Range -> String -> CommandM ()

Shows all the top-level names in the given module, along with their types.

solveInstantiatedGoals :: Rewrite -> Maybe InteractionId -> CommandM ()

Solved goals already instantiated internally The second argument potentially limits it to one specific goal.

sortInteractionPoints :: (MonadInteractionPoints m, MonadError TCErr m, MonadDebug m) => [InteractionId] -> m [InteractionId]

Sorts interaction points based on their ranges.

status :: CommandM Status

Computes some status information.

Does not change the state.

tellEmacsToJumpToError :: Range -> [Response]

Tells the Emacs mode to go to the first error position (if any).

tellToUpdateHighlighting :: Maybe (HighlightingInfo, HighlightingMethod, ModuleToSource) -> IO [Response]

Tell to highlight the code using the given highlighting info (unless it is Nothing).

updateInteractionPointsAfter :: Interaction -> Bool

Should Resp_InteractionPoints be issued after the command has run?

whyInScope :: String -> CommandM ()

Explain why something is in scope.

withCurrentFile :: CommandM a -> CommandM a

Set envCurrentPath to theCurrentFile, if any.

data GiveRefine

Constructors

Give

Refine

Intro

ElaborateGive Rewrite