Agda.Interaction.Library

Description

Library management.

Sample use:

    -- Get libraries as listed in .agda/libraries file.
    libs <- getInstalledLibraries Nothing

    -- Get the libraries (and immediate paths) relevant for projectRoot.
    -- This involves locating and processing the .agda-lib file for the project.
    (libNames, includePaths) <-  getDefaultLibraries projectRoot True

    -- Get include paths of depended-on libraries.
    resolvedPaths <- libraryIncludePaths Nothing libs libNames

    let allPaths = includePaths ++ resolvedPaths
  

Documentation

findProjectRoot :: FilePath -> LibM (Maybe FilePath)

Get project root

getAgdaAppDir :: IO FilePath

Get the path to ~/.agda (system-specific). Can be overwritten by the AGDA_DIR environment variable.

(This is not to be confused with the directory for the data files that Agda needs (e.g. the primitive modules).)

getDefaultLibraries

Arguments

:: FilePath
Project root.

-> Bool
Use defaults if no .agda-lib file exists for this project?

-> LibM ([LibName], [FilePath])
The returned LibNames are all non-empty strings.

Get dependencies and include paths for given project root:

Look for .agda-lib files according to findAgdaLibFiles. If none are found, use default dependencies (according to defaults file) and current directory (project root).

getInstalledLibraries

Arguments

:: Maybe FilePath
Override the default libraries file?

-> LibM [AgdaLibFile]
Content of library files. (Might have empty LibNames.)

Parse the descriptions of the libraries Agda knows about.

Returns none if there is no libraries file.

getTrustedExecutables

Arguments

:: LibM (Map ExeName FilePath)
Content of executables files.

Return the trusted executables Agda knows about.

Returns none if there is no executables file.

libraryIncludePaths

Arguments

:: Maybe FilePath
libraries file (error reporting only).

-> [AgdaLibFile]
Libraries Agda knows about.

-> [LibName]
(Non-empty) library names to be resolved to (lists of) pathes.

-> LibM [FilePath]
Resolved pathes (no duplicates). Contains "." if [LibName] does.

Get all include pathes for a list of libraries to use.

getAgdaLibFiles' :: FilePath -> LibErrorIO [AgdaLibFile]

Get the contents of .agda-lib files in the given project root.

getPrimitiveLibDir :: IO FilePath

Returns the absolute default lib dir. This directory is used to store the Primitive.agda file.

type LibName = String

A symbolic library name.

data OptionsPragma

The options from an OPTIONS pragma (or a .agda-lib file).

In the future it might be nice to switch to a more structured representation. Note that, currently, there is not a one-to-one correspondence between list elements and options.

Constructors

OptionsPragma

Fields

pragmaStrings :: [String]

The options.

pragmaRange :: Range

The range of the options in the pragma (not including things like an OPTIONS keyword).

data AgdaLibFile

Content of a .agda-lib file.

Constructors

AgdaLibFile

Fields

_libName :: LibName

The symbolic name of the library.

_libFile :: FilePath

Path to this .agda-lib file (not content of the file).

_libAbove :: !Int

How many directories above the Agda file is the .agda-lib file located?

_libIncludes :: [FilePath]

Roots where to look for the modules of the library.

_libDepends :: [LibName]

Dependencies.

_libPragmas :: OptionsPragma

Default pragma options for all files in the library.

type ExeName = Text

A symbolic executable name.

type LibM = ExceptT LibErrors (WriterT [LibWarning] (StateT LibState IO))

Throws LibErrors exceptions, still collects LibWarnings.

mkLibM :: [AgdaLibFile] -> LibErrorIO a -> LibM a

Raise collected LibErrors as exception.

data LibWarning

Constructors

LibWarning (Maybe LibPositionInfo) LibWarning'

data LibPositionInfo

Information about which .agda-lib file we are reading and from where in the libraries file it came from.

Constructors

LibPositionInfo

Fields

libFilePos :: Maybe FilePath

Name of libraries file.

lineNumPos :: LineNumber

Line number in libraries file.

filePos :: FilePath

Library file.

libraryWarningName :: LibWarning -> WarningName

data ProjectConfig

A file can either belong to a project located at a given root containing one or more .agda-lib files, or be part of the default project.

Constructors

ProjectConfig

Fields

configRoot :: FilePath

configAgdaLibFiles :: [FilePath]

configAbove :: !Int

How many directories above the Agda file is the .agda-lib file located?

DefaultProjectConfig

Exported for testing

data VersionView

Library names are structured into the base name and a suffix of version numbers, e.g. mylib-1.2.3. The version suffix is optional.

Constructors

VersionView

Fields

vvBase :: LibName

Actual library name.

vvNumbers :: [Integer]

Major version, minor version, subminor version, etc., all non-negative. Note: a priori, there is no reason why the version numbers should be Ints.

versionView :: LibName -> VersionView

Split a library name into basename and a list of version numbers.

versionView "foo-1.2.3"    == VersionView "foo" [1, 2, 3]
versionView "foo-01.002.3" == VersionView "foo" [1, 2, 3]

Note that because of leading zeros, versionView is not injective. (unVersionView . versionView would produce a normal form.)

unVersionView :: VersionView -> LibName

Print a VersionView, inverse of versionView (modulo leading zeros).

findLib' :: (a -> LibName) -> LibName -> [a] -> [a]

Generalized version of findLib for testing.

findLib' id "a"   [ "a-1", "a-02", "a-2", "b" ] == [ "a-02", "a-2" ]

findLib' id "a"   [ "a", "a-1", "a-01", "a-2", "b" ] == [ "a" ]
findLib' id "a-1" [ "a", "a-1", "a-01", "a-2", "b" ] == [ "a-1", "a-01" ]
findLib' id "a-2" [ "a", "a-1", "a-01", "a-2", "b" ] == [ "a-2" ]
findLib' id "c"   [ "a", "a-1", "a-01", "a-2", "b" ] == []