Agda.Interaction.Library.Base

Description

Basic data types for library management.

Documentation

type LibName = String

A symbolic library name.

data LibrariesFile

Constructors

LibrariesFile

Fields

lfPath :: FilePath

E.g. ~.agdalibraries.

lfExists :: Bool

The libraries file might not exist, but we may print its assumed location in error messages.

type ExeName = Text

A symbolic executable name.

data ExecutablesFile

Constructors

ExecutablesFile

Fields

efPath :: FilePath

E.g. ~.agdaexecutables.

efExists :: Bool

The executables file might not exist, but we may print its assumed location in error messages.

libNameForCurrentDir :: LibName

The special name "." is used to indicated that the current directory should count as a project root.

data ProjectConfig

A file can either belong to a project located at a given root containing one or more .agda-lib files, or be part of the default project.

Constructors

ProjectConfig

Fields

configRoot :: FilePath

configAgdaLibFiles :: [FilePath]

DefaultProjectConfig

data AgdaLibFile

Content of a .agda-lib file.

Constructors

AgdaLibFile

Fields

_libName :: LibName

The symbolic name of the library.

_libFile :: FilePath

Path to this .agda-lib file (not content of the file).

_libIncludes :: [FilePath]

Roots where to look for the modules of the library.

_libDepends :: [LibName]

Dependencies.

_libPragmas :: [String]

Default pragma options for all files in the library.

emptyLibFile :: AgdaLibFile

libName :: Lens' LibName AgdaLibFile

Lenses for AgdaLibFile

libFile :: Lens' FilePath AgdaLibFile

libIncludes :: Lens' [FilePath] AgdaLibFile

libDepends :: Lens' [LibName] AgdaLibFile

libPragmas :: Lens' [String] AgdaLibFile

Library warnings and errors

type LineNumber = Int

data LibPositionInfo

Constructors

LibPositionInfo

Fields

libFilePos :: Maybe FilePath

Name of libraries file

lineNumPos :: LineNumber

Line number in libraries file.

filePos :: FilePath

Library file

data LibWarning

Constructors

LibWarning (Maybe LibPositionInfo) LibWarning'

data LibWarning'

Library Warnings.

Constructors

UnknownField String

ExeNotFound ExecutablesFile FilePath
Raised when a trusted executable can not be found.

ExeNotExecutable ExecutablesFile FilePath
Raised when a trusted executable does not have the executable permission.

data LibError

Constructors

LibError (Maybe LibPositionInfo) LibError'

libraryWarningName :: LibWarning -> WarningName

data LibError'

Collected errors while processing library files.

Constructors

LibNotFound LibrariesFile LibName
Raised when a library name could not successfully be resolved to an .agda-lib file.

AmbiguousLib LibName [AgdaLibFile]
Raised when a library name is defined in several .agda-lib files.

OtherError String
Generic error.

type LibState = (Map FilePath ProjectConfig, Map FilePath AgdaLibFile)

Cache locations of project configurations and parsed .agda-lib files

type LibErrorIO = WriterT [Either LibError LibWarning] (StateT LibState IO)

Collects LibErrors and LibWarnings.

type LibM = ExceptT Doc (WriterT [LibWarning] (StateT LibState IO))

Throws Doc exceptions, still collects LibWarnings.

warnings :: MonadWriter [Either LibError LibWarning] m => [LibWarning] -> m ()

warnings' :: MonadWriter [Either LibError LibWarning] m => [LibWarning'] -> m ()

raiseErrors' :: MonadWriter [Either LibError LibWarning] m => [LibError'] -> m ()

raiseErrors :: MonadWriter [Either LibError LibWarning] m => [LibError] -> m ()

getCachedProjectConfig :: (MonadState LibState m, MonadIO m) => FilePath -> m (Maybe ProjectConfig)

storeCachedProjectConfig :: (MonadState LibState m, MonadIO m) => FilePath -> ProjectConfig -> m ()

getCachedAgdaLibFile :: (MonadState LibState m, MonadIO m) => FilePath -> m (Maybe AgdaLibFile)

storeCachedAgdaLibFile :: (MonadState LibState m, MonadIO m) => FilePath -> AgdaLibFile -> m ()

Prettyprinting errors and warnings

formatLibPositionInfo :: LibPositionInfo -> String -> Doc

formatLibError :: [AgdaLibFile] -> LibError -> Doc

Pretty-print LibError.