Agda.Interaction.Options.Arguments

Description

Definitions and tools for arguments of command-line options.

Names of the arguments of options that have a finite enumeration.

optionValues :: Map String [String]

Print the possible values of an argument of the given type.

--color

colorArg :: String

Argument to --color.

colorValues :: [String]

Possible values for --color.

--emacs-mode

emacsModeArg :: String

Argument to --emacs-mode.

emacsModeValues :: [String]

Possible values for --emacs-mode.

--help

helpArg :: String

Argument to --help.

helpValues :: [String]

Possible values for --help.

--profile

profileArg :: String

Argument to --profile.

profileValues :: [String]

Possible values for --profile.

--trace-imports

traceImportsArg :: String

Argument to --trace-imports.

traceImportsValues :: [String]

Possible values for --trace-imports.

--warning

warningArg :: String

Argument to --warning.

warningValues :: [String]

Possible values for --warning.