Agda.Interaction.Options.Warnings

Documentation

data WarningMode

A WarningMode has two components: a set of warnings to be displayed and a flag stating whether warnings should be turned into fatal errors.

Constructors

WarningMode

Fields

_warningSet :: Set WarningName

Invariant: When UnusedImportsAll_ is in the warning set, so must be UnusedImports_.

_warn2Error :: Bool

warningSet :: Lens' WarningMode (Set WarningName)

warn2Error :: Lens' WarningMode Bool

lensSingleWarning :: WarningName -> Lens' WarningMode Bool

defaultWarningSet :: String

The defaultWarningMode is a curated set of warnings covering non-fatal errors and disabling style-related ones

allWarnings :: Set WarningName

usualWarnings :: Set WarningName

noWarnings :: Set WarningName

unsolvedWarnings :: Set WarningName

unusedImportsWarnings :: Set WarningName

Both of these warnings are disabled by -WnoUnusedImports.

incompleteMatchWarnings :: Set WarningName

errorWarnings :: Set WarningName

exactSplitWarnings :: Set WarningName

Warnings enabled by --exact-split.

defaultWarningMode :: WarningMode

data WarningModeError

Some warnings are errors and cannot be turned off.

Constructors

Unknown Text
Unknown warning.

NoNoError Text
Warning that cannot be disabled.

NoUnusedImportsAll
-WnoUnusedImports=all is not supported, use -WnoUnusedImports instead.

prettyWarningModeError :: WarningModeError -> Text

warningModeUpdate :: String -> Either WarningModeError WarningModeUpdate

warningModeUpdate str computes the action of str over the current WarningMode: it may reset the set of warnings, add or remove a specific flag or demand that any warning be turned into an error

warningSets :: [(String, (Set WarningName, String))]

Common sets of warnings

data WarningName

The WarningName data enumeration is meant to have a one-to-one correspondance to existing warnings in the codebase.

Constructors

OptionRenamed_

WarningProblem_
Some warning could not be set or unset. Parser Warnings

OverlappingTokensWarning_

MisplacedAttributes_

UnknownPolarity_

UnknownAttribute_

UnsupportedAttribute_

MultipleAttributes_

MismatchedBrackets_

LibUnknownField_

DivergentModalityInClause_

EmptyAbstract_

EmptyConstructor_

EmptyField_

EmptyGeneralize_

EmptyInstance_

EmptyMacro_

EmptyMutual_

EmptyPostulate_

EmptyPrimitive_

EmptyPrivate_

EmptyRewritePragma_

EmptyWhere_

EmptyPolarityPragma_

HiddenGeneralize_

InvalidCatchallPragma_

InvalidConstructorBlock_

InvalidCoverageCheckPragma_

InvalidDataOrRecDefParameter_

InvalidNoPositivityCheckPragma_

InvalidNoUniverseCheckPragma_

DuplicateRecordDirective_

InvalidTerminationCheckPragma_

InvalidTacticAttribute_

MissingDataDeclaration_

MissingDefinitions_

NotAllowedInMutual_

OpenImportAbstract_

OpenImportPrivate_

PolarityPragmasButNotPostulates_

PragmaCompiled_

PragmaNoTerminationCheck_

ShadowingInTelescope_

UnknownFixityInMixfixDecl_

UnknownNamesInFixityDecl_

UnknownNamesInPolarityPragmas_

UselessAbstract_

UselessImport_

UselessInstance_

UselessMacro_

UselessPrivate_

AbsurdPatternRequiresAbsentRHS_

AsPatternShadowsConstructorOrPatternSynonym_

PatternShadowsConstructor_

CantGeneralizeOverSorts_

ClashesViaRenaming_

CoverageIssue_

CoverageNoExactSplit_

InlineNoExactSplit_

DeprecationWarning_

DuplicateUsing_

FixingCohesion_

FixingPolarity_

FixingRelevance_

FixityInRenamingModule_

InvalidCharacterLiteral_

UnusedImports_

UnusedImportsAll_

UselessPragma_

IllegalDeclarationInDataDefinition_

IllformedAsClause_

InstanceArgWithExplicitArg_

InstanceWithExplicitArg_

InstanceNoOutputTypeName_

InteractionMetaBoundaries_

InversionDepthReached_

ModuleDoesntExport_

NotInScope_

NotStrictlyPositive_

ConstructorDoesNotFitInData_

CoinductiveEtaRecord_

UnsupportedIndexedMatch_

OldBuiltin_

BuiltinDeclaresIdentifier_

PlentyInHardCompileTimeMode_

PragmaCompileErased_

PragmaCompileList_

PragmaCompileMaybe_

PragmaCompileUnparsable_

PragmaCompileWrong_

PragmaCompileWrongName_

PragmaExpectsDefinedSymbol_

PragmaExpectsUnambiguousConstructorOrFunction_

PragmaExpectsUnambiguousProjectionOrFunction_

NoMain_

NotARewriteRule_

RewriteLHSNotDefinitionOrConstructor_

RewriteVariablesNotBoundByLHS_

RewriteVariablesBoundMoreThanOnce_

RewriteVariablesBoundInSingleton_

RewriteLHSReduces_

RewriteHeadSymbolIsProjectionLikeFunction_

RewriteHeadSymbolIsTypeConstructor_

RewriteHeadSymbolContainsMetas_

RewriteConstructorParametersNotGeneral_

RewriteContainsUnsolvedMetaVariables_

RewriteBlockedOnProblems_

RewriteRequiresDefinitions_

RewriteDoesNotTargetRewriteRelation_

RewriteBeforeFunctionDefinition_

RewriteBeforeMutualFunctionDefinition_

ConfluenceCheckingIncompleteBecauseOfMeta_

ConfluenceForCubicalNotSupported_

RewriteMaybeNonConfluent_

RewriteNonConfluent_

RewriteAmbiguousRules_

RewriteMissingRule_

DuplicateRewriteRule_

SafeFlagEta_

SafeFlagInjective_

SafeFlagNoCoverageCheck_

SafeFlagNonTerminating_

SafeFlagNoPositivityCheck_

SafeFlagNoUniverseCheck_

SafeFlagPolarity_

SafeFlagPostulate_

SafeFlagPragma_

SafeFlagTerminating_

SafeFlagWithoutKFlagPrimEraseEquality_

TerminationIssue_

TooManyArgumentsToSort_

UnreachableClauses_

UnsolvedConstraints_

UnsolvedInteractionMetas_

UnsolvedMetaVariables_

UselessHiding_

UselessInline_

UselessPatternDeclarationForRecord_

UselessPublic_

UselessTactic_

UserWarning_

InvalidDisplayForm_

UnusedVariablesInDisplayForm_

RewritesNothing_

WithClauseProjectionFixityMismatch_

WithoutKFlagPrimEraseEquality_

ConflictingPragmaOptions_

WrongInstanceDeclaration_

TooManyPolarities_

TopLevelPolarity_

CoInfectiveImport_

InfectiveImport_

DuplicateFields_

TooManyFields_

MissingTypeSignatureForOpaque_

NotAffectedByOpaque_

UnfoldingWrongName_

UnfoldTransparentName_

UselessOpaque_

HiddenNotInArgumentPosition_

InstanceNotInArgumentPosition_

MacroInLetBindings_

AbstractInLetBindings_

FaceConstraintCannotBeHidden_

FaceConstraintCannotBeNamed_

CustomBackendWarning_

warningNameToString :: WarningName -> String

string2WarningName :: String -> Maybe WarningName

The flag corresponding to a warning is precisely the name of the constructor minus the trailing underscore.

usageWarning :: String

warningUsage generated using warningNameDescription