Agda.Main

Description

Agda main module.

Documentation

runAgda :: [Backend] -> IO ()

The main function

runAgda' :: [Backend] -> IO ()

The main function without importing built-in backends

runAgdaArgs :: [Backend] -> [String] -> IO ()

type Interactor a = TCM () -> (AbsolutePath -> TCM CheckResult) -> TCM a

data FrontendType

Major mode of operation, not including the standard mode (checking the given main module).

Constructors

FrontEndInteraction InteractionFormat
--interaction or --interaction-json.

FrontEndRepl
--interactive.

FrontEndBuildLibrary
--build-library.

data InteractionFormat

Constructors

InteractionEmacs
--interaction.

InteractionJson
--interaction-json.

pattern FrontEndEmacs :: FrontendType

pattern FrontEndJson :: FrontendType

buildLibraryInteractor :: Interactor ()

interactionInteractor :: InteractionFormat -> Interactor ()

Emacs/JSON mode. Note that it ignores the "check" action because it calls typeCheck directly.

replInteractor :: Maybe AbsolutePath -> Interactor ()

The (deprecated) repl mode.

defaultInteractor :: AbsolutePath -> Interactor ()

The interactor to use when there are no frontends or backends specified.

getInteractor :: MonadError String m => [Backend] -> Maybe AbsolutePath -> CommandLineOptions -> m (Maybe (Interactor ()))

runAgdaWithOptions

Arguments

:: Interactor a
Backend interaction

-> String
program name

-> CommandLineOptions
parsed command line options

-> TCM a

Run Agda with parsed command line options

printUsage :: [Backend] -> Help -> IO ()

Print usage information.

printVersion :: [Backend] -> PrintAgdaVersion -> IO ()

Print version information.

printAgdaDataDir :: IO ()

printAgdaAppDir :: IO ()

printOptions :: IO ()

optionError :: String -> IO ()

What to do for bad options.

runTCMPrettyErrors :: TCM () -> IO ()

Run a TCM action in IO; catch and pretty print errors.

helpForLocaleError :: TCErr -> IO ()

If the error is an IO error, and the error message suggests that the problem is related to locales or code pages, print out some extra information.