Agda.Mimer.Types

Results

data MimerResult

Constructors

MimerExpr String
Returns String rather than Expr because the give action expects a string.

MimerList [(Int, String)]

MimerNoResult

isNoResult :: MimerResult -> Bool

data SearchStepResult

Constructors

ResultExpr Expr

OpenBranch SearchBranch

NoSolution

Search branches

data SearchBranch

Constructors

SearchBranch

Fields

sbTCState :: TCState

sbGoals :: [Goal]

sbCost :: Int

sbCache :: Map CheckpointId ComponentCache

sbComponentsUsed :: Map Name Int

Number of times each component has been used

data Goal

Constructors

Goal

Fields

goalMeta :: MetaId

nextGoal :: SearchBranch -> (Goal, SearchBranch)

Take the first goal off a search branch. Precondition: the set of goals is non-empty.

Components

type ComponentCache = Map Component (Maybe [Component])

data BaseComponents

Components that are not changed during search. Components that do change (local variables and let bindings) are stored in each SearchBranch.

Constructors

BaseComponents

Fields

hintFns :: [Component]

hintDataTypes :: [Component]

hintRecordTypes :: [Component]

hintAxioms :: [Component]

Excluding those producing Level

hintLevel :: [Component]

A definition in a where clause

hintProjections :: [Component]

Variables that are candidates for arguments to recursive calls

hintThisFn :: Maybe Component

hintLetVars :: [Open Component]

hintRecVars :: Open [(Term, NoSubst Term Int)]

Variable terms and which argument they come from

type CompId = Int

data Component

Constructors

Component

Fields

compId :: CompId

Unique id for the component. Used for the cache.

compName :: Maybe Name

Used for keeping track of how many times a component has been used

compPars :: Nat

How many arguments should be dropped (e.g. constructor parameters)

compTerm :: Term

compType :: Type

compRec :: Bool

Is this a recursive call

compMetas :: [MetaId]

compCost :: Cost

mkComponent :: CompId -> [MetaId] -> Cost -> Maybe Name -> Nat -> Term -> Type -> Component

mkComponentQ :: CompId -> Cost -> QName -> Nat -> Term -> Type -> Component

noName :: Maybe Name

addCost :: Cost -> Component -> Component

deleteCompMeta :: MetaId -> Component -> Component

SearchOptions

data SearchOptions

Constructors

SearchOptions

Fields

searchBaseComponents :: BaseComponents

searchHintMode :: HintMode

searchTimeout :: MilliSeconds

searchGenProjectionsLocal :: Bool

searchGenProjectionsLet :: Bool

searchGenProjectionsExternal :: Bool

searchGenProjectionsRec :: Bool

searchSpeculateProjections :: Bool

searchTopMeta :: MetaId

searchTopEnv :: TCEnv

searchTopCheckpoint :: CheckpointId

searchInteractionId :: InteractionId

searchFnName :: Maybe QName

searchCosts :: Costs

searchStats :: IORef MimerStats

searchRewrite :: Rewrite

searchBuiltinFlat :: Maybe QName

type Cost = Int

data Costs

Constructors

Costs

Fields

costLocal :: Cost

costFn :: Cost

costDataCon :: Cost

costRecordCon :: Cost

costSpeculateProj :: Cost

costProj :: Cost

costAxiom :: Cost

costLet :: Cost

costLevel :: Cost

costSet :: Cost

costRecCall :: Cost

costNewMeta :: Cost

Cost of a new meta-variable appearing in a non-implicit position

costNewHiddenMeta :: Cost

Cost of a new meta-variable appearing in an implicit position

costCompReuse :: Nat -> Cost

Cost of reusing a component n times. Only counted when n>1.

noCost :: Cost

defaultCosts :: Costs

Measure performance

data MimerStats

Constructors

MimerStats

Fields

statCompHit :: !Int

Could make use of an already generated component

statCompGen :: !Int

Could use a generator for a component

statCompRegen :: !Int

Had to regenerate the cache (new context)

statCompNoRegen :: !Int

Did not have to regenerate the cache

statMetasCreated :: !Int

Total number of meta-variables created explicitly (not through unification)

statTypeEqChecks :: !Int

Number of times type equality is tested (with unification)

statRefineSuccess :: !Int

Number of times a refinement has been successful

statRefineFail :: !Int

Number of times a refinement has failed

emptyMimerStats :: MimerStats

incCompHit :: MimerStats -> MimerStats

incCompGen :: MimerStats -> MimerStats

incCompRegen :: MimerStats -> MimerStats

incCompNoRegen :: MimerStats -> MimerStats

incMetasCreated :: MimerStats -> MimerStats

incTypeEqChecks :: MimerStats -> MimerStats

incRefineSuccess :: MimerStats -> MimerStats

incRefineFail :: MimerStats -> MimerStats

Pretty printing

haskellRecord :: Doc -> [(Doc, Doc)] -> Doc

keyValueList :: [(Doc, Doc)] -> Doc