Agda.Setup

Description

Agda's self-setup.

Documentation

getAgdaAppDir :: IO FilePath

Get the path to ~/.agda (system-specific). Can be overwritten by the AGDA_DIR environment variable.

(This is not to be confused with the directory getDataDir for the data files that Agda needs (e.g. the primitive modules).)

getDataDir :: IO FilePath

getDataFileName :: FilePath -> IO FilePath

This overrides the getDataFileName from 'Paths_Agda'.

setup :: Bool -> IO ()

False: Check whether we need to setup Agda. This function can be called when starting up Agda.

True: force a setup e.g. when passing Agda option --setup.

Copies the embedded data files to the designated data directory.