Agda.Setup.DataFiles

Description

The list of data files Agda uses.

Because of TemplateHaskell state restrictions, this cannot be define in 'Setup'.

Documentation

dataPath :: FilePath -> FilePath

A data file path relative to the project root.

dataFiles :: [FilePath]

Agda's embedded data files.

emacsModeDir :: FilePath

The subdirectory in the Agda data directory containing the emacs mode.

emacsLispFiles :: [FilePath]

The Agda mode's Emacs Lisp files, given in the order in which they should be compiled.