Agda.Syntax.Abstract

Description

The abstract syntax. This is what you get after desugaring and scope analysis of the concrete syntax. The type checker works on abstract syntax, producing internal syntax (Agda.Syntax.Internal).

Documentation

class SubstExpr a where

Minimal complete definition

Nothing

Methods

substExpr :: [(Name, Expr)] -> a -> a

default substExpr :: (Functor t, SubstExpr b, t b ~ a) => [(Name, Expr)] -> a -> a

type PatternSynDefns = Map QName PatternSynDefn

type PatternSynDefn = ([Arg Name], Pattern' Void)

class NameToExpr a where

Turn a name into an expression.

Methods

nameToExpr :: a -> Expr

class AnyAbstract a where

Are we in an abstract block?

In that case some definition is abstract.

Methods

anyAbstract :: a -> Bool

type HoleContent = HoleContent' () BindName Pattern Expr

type Patterns = [NamedArg Pattern]

type Pattern = Pattern' Expr

type NAPs1 e = List1 (NamedArg (Pattern' e))

type NAPs e = [NamedArg (Pattern' e)]

data Pattern' e

Parameterised over the type of dot patterns.

Constructors

VarP BindName

ConP ConPatInfo AmbiguousQName (NAPs e)

ProjP PatInfo ProjOrigin AmbiguousQName
Destructor pattern d.

DefP PatInfo AmbiguousQName (NAPs e)
Defined pattern: function definition f ps. It is also abused to convert destructor patterns into concrete syntax thus, we put AmbiguousQName here as well.

WildP PatInfo
Underscore pattern entered by user. Or generated at type checking for implicit arguments.

AsP PatInfo BindName (Pattern' e)

DotP PatInfo e
Dot pattern .e

AbsurdP PatInfo

LitP PatInfo Literal

PatternSynP PatInfo AmbiguousQName (NAPs e)

RecP PatInfo [FieldAssignment' (Pattern' e)]

EqualP PatInfo [(e, e)]

WithP PatInfo (Pattern' e)
| p, for with-patterns.

AnnP PatInfo e (Pattern' e)
Pattern with type annotation

type LHSCore = LHSCore' Expr

data LHSCore' e

The lhs in projection-application and with-pattern view. Parameterised over the type e of dot patterns.

Constructors

LHSHead
The head applied to ordinary patterns.

Fields

lhsDefName :: QName

Head f.

lhsPats :: [NamedArg (Pattern' e)]

Applied to patterns ps.

LHSProj
Projection.

Fields

lhsDestructor :: AmbiguousQName

Record projection identifier.

lhsFocus :: NamedArg (LHSCore' e)

Main argument of projection.

lhsPats :: [NamedArg (Pattern' e)]

Applied to patterns ps.

LHSWith
With patterns.

Fields

lhsHead :: LHSCore' e

E.g. the LHSHead.

lhsWithPatterns :: [Arg (Pattern' e)]

Applied to with patterns | p1 | ... | pn. These patterns are not prefixed with WithP!

lhsPats :: [NamedArg (Pattern' e)]

Applied to patterns ps.

data LHS

The lhs of a clause in focused (projection-application) view (outside-in). Projection patters are represented as LHSProjs.

Constructors

LHS

Fields

lhsInfo :: LHSInfo

Range.

lhsCore :: LHSCore

Copatterns.

data SpineLHS

The lhs of a clause in spine view (inside-out). Projection patterns are contained in spLhsPats, represented as ProjP d.

Constructors

SpineLHS

Fields

spLhsInfo :: LHSInfo

Range.

spLhsDefName :: QName

Name of function we are defining.

spLhsPats :: [NamedArg Pattern]

Elimination by pattern, projections, with-patterns.

data RHS

Constructors

RHS

Fields

rhsExpr :: Expr

rhsConcrete :: Maybe Expr

We store the original concrete expression in case we have to reproduce it during interactive case splitting. Nothing for internally generated rhss.

AbsurdRHS

WithRHS QName [WithExpr] [Clause]
The QName is the name of the with function.

RewriteRHS

Fields

rewriteExprs :: [RewriteEqn]

The QNames are the names of the generated with functions, one for each Expr.

rewriteStrippedPats :: [ProblemEq]

The patterns stripped by with-desugaring. These are only present if this rewrite follows a with.

rewriteRHS :: RHS

The RHS should not be another RewriteRHS.

rewriteWhereDecls :: WhereDeclarations

The where clauses are attached to the RewriteRHS by

type WithExpr = WithExpr' Expr

type WithExpr' e = Named BindName (Arg e)

type RewriteEqn = RewriteEqn' QName BindName Pattern Expr

type SpineClause = Clause' SpineLHS

type Clause = Clause' LHS

data WhereDeclarations

Constructors

WhereDecls

Fields

whereModule :: Maybe ModuleName

whereDecls :: Maybe Declaration

The declaration is a Section.

data Clause' lhs

We could throw away where clauses at this point and translate them to let. It's not obvious how to remember that the let was really a where clause though, so for the time being we keep it here.

Constructors

Clause

Fields

clauseLHS :: lhs

clauseStrippedPats :: [ProblemEq]

Only in with-clauses where we inherit some already checked patterns from the parent. These live in the context of the parent clause left-hand side.

clauseRHS :: RHS

clauseWhereDecls :: WhereDeclarations

clauseCatchall :: Bool

data ProblemEq

A user pattern together with an internal term that it should be equal to after splitting is complete. Special cases: * User pattern is a variable but internal term isn't: this will be turned into an as pattern. * User pattern is a dot pattern: this pattern won't trigger any splitting but will be checked for equality after all splitting is complete and as patterns have been bound. * User pattern is an absurd pattern: emptiness of the type will be checked after splitting is complete. * User pattern is an annotated wildcard: type annotation will be checked after splitting is complete.

Constructors

ProblemEq

Fields

problemInPat :: Pattern

problemInst :: Term

problemType :: Dom Type

data DataDefParams

Constructors

DataDefParams

Fields

dataDefGeneralizedParams :: Set Name

We don't yet know the position of generalized parameters from the data sig, so we keep these in a set on the side.

dataDefParams :: [LamBinding]

data GeneralizeTelescope

Constructors

GeneralizeTel

Fields

generalizeTelVars :: Map QName Name

Maps generalize variables to the corresponding bound variable (to be introduced by the generalisation).

generalizeTel :: Telescope

type Telescope = [TypedBinding]

type Telescope1 = List1 TypedBinding

data TypedBinding

A typed binding. Appears in dependent function spaces, typed lambdas, and telescopes. It might be tempting to simplify this to only bind a single name at a time, and translate, say, (x y : A) to (x : A)(y : A) before type-checking. However, this would be slightly problematic:

We would have to typecheck the type A several times.

If A contains a meta variable or hole, it would be duplicated by such a translation.

While 1. is only slightly inefficient, 2. would be an outright bug. Duplicating A could not be done naively, we would have to make sure that the metas of the copy are aliases of the metas of the original.

Constructors

TBind Range TacticAttr (List1 (NamedArg Binder)) Expr
As in telescope (x y z : A) or type (x y z : A) -> B.

TLet Range (List1 LetBinding)
E.g. (let x = e) or (let open M).

data LamBinding

A lambda binding is either domain free or typed.

Constructors

DomainFree TacticAttr (NamedArg Binder)
. x or {x} or .x or {x = y} or x@p or (p)

DomainFull TypedBinding
. (xs:e) or {xs:e} or (let Ds)

type Binder = Binder' BindName

data Binder' a

Constructors

Binder

Fields

binderPattern :: Maybe Pattern

binderName :: a

type TacticAttr = Maybe Expr

type Field = TypeSignature

type Constructor = TypeSignature

type TypeSignature = Declaration

Only Axioms.

data LetBinding

Bindings that are valid in a let.

Constructors

LetBind LetInfo ArgInfo BindName Expr Expr
LetBind info rel name type defn

LetPatBind LetInfo Pattern Expr
Irrefutable pattern binding.

LetApply ModuleInfo ModuleName ModuleApplication ScopeCopyInfo ImportDirective
LetApply mi newM (oldM args) renamings dir. The ImportDirective is for highlighting purposes.

LetOpen ModuleInfo ModuleName ImportDirective
only for highlighting and abstractToConcrete

LetDeclaredVariable BindName
Only used for highlighting. Refers to the first occurrence of x in let x : A; x = e. | LetGeneralize DefInfo ArgInfo Expr

data Pragma

Constructors

OptionsPragma [String]

BuiltinPragma RString ResolvedName
ResolvedName is not UnknownName. Name can be ambiguous e.g. for built-in constructors.

BuiltinNoDefPragma RString KindOfName QName
Builtins that do not come with a definition, but declare a name for an Agda concept.

RewritePragma Range [QName]
Range is range of REWRITE keyword.

CompilePragma RString QName String

StaticPragma QName

EtaPragma QName
For coinductive records, use pragma instead of regular eta-equality definition (as it is might make Agda loop).

InjectivePragma QName

InlinePragma Bool QName

DisplayPragma QName [NamedArg Pattern] Expr

data ModuleApplication

Constructors

SectionApp Telescope ModuleName [NamedArg Expr]
tel. M args: applies M to args and abstracts tel.

RecordModuleInstance ModuleName
M {{...}}

type ImportedName = ImportedName' QName ModuleName

type Renaming = Renaming' QName ModuleName

type ImportDirective = ImportDirective' QName ModuleName

type DefInfo = DefInfo' Expr

data Declaration

Constructors

Axiom KindOfName DefInfo ArgInfo (Maybe [Occurrence]) QName Expr
Type signature (can be irrelevant, but not hidden).

The fourth argument contains an optional assignment of polarities to arguments.

Generalize (Set QName) DefInfo ArgInfo QName Expr
First argument is set of generalizable variables used in the type.

Field DefInfo QName (Arg Expr)
record field

Primitive DefInfo QName (Arg Expr)
primitive function

Mutual MutualInfo [Declaration]
a bunch of mutually recursive definitions

Section Range ModuleName GeneralizeTelescope [Declaration]

Apply ModuleInfo ModuleName ModuleApplication ScopeCopyInfo ImportDirective
The ImportDirective is for highlighting purposes.

Import ModuleInfo ModuleName ImportDirective
The ImportDirective is for highlighting purposes.

Pragma Range Pragma

Open ModuleInfo ModuleName ImportDirective
only retained for highlighting purposes

FunDef DefInfo QName Delayed [Clause]
sequence of function clauses

DataSig DefInfo QName GeneralizeTelescope Expr
lone data signature

DataDef DefInfo QName UniverseCheck DataDefParams [Constructor]

RecSig DefInfo QName GeneralizeTelescope Expr
lone record signature

RecDef DefInfo QName UniverseCheck RecordDirectives DataDefParams Expr [Declaration]
The Expr gives the constructor type telescope, (x1 : A1)..(xn : An) -> Prop, and the optional name is the constructor's name. The optional Range is for the pattern attribute.

PatternSynDef QName [Arg BindName] (Pattern' Void)
Only for highlighting purposes

UnquoteDecl MutualInfo [DefInfo] [QName] Expr

UnquoteDef [DefInfo] [QName] Expr

ScopedDecl ScopeInfo [Declaration]
scope annotation

type RecordDirectives = RecordDirectives' QName

data ScopeCopyInfo

Constructors

ScopeCopyInfo

Fields

renModules :: Ren ModuleName

renNames :: Ren QName

type Ren a = Map a (List1 a)

Renaming (generic).

type RecordAssigns = [RecordAssign]

type RecordAssign = Either Assign ModuleName

type Assigns = [Assign]

type Assign = FieldAssignment' Expr

Record field assignment f = e.

data Expr

Expressions after scope checking (operators parsed, names resolved).

Constructors

Var Name
Bound variable.

Def' QName Suffix
Constant: axiom, function, data or record type, with a possible suffix.

Proj ProjOrigin AmbiguousQName
Projection (overloaded).

Con AmbiguousQName
Constructor (overloaded).

PatternSyn AmbiguousQName
Pattern synonym.

Macro QName
Macro.

Lit ExprInfo Literal
Literal.

QuestionMark MetaInfo InteractionId
Meta variable for interaction. The InteractionId is usually identical with the metaNumber of MetaInfo. However, if you want to print an interaction meta as just ? instead of ?n, you should set the metaNumber to Nothing while keeping the InteractionId.

Underscore MetaInfo
Meta variable for hidden argument (must be inferred locally).

Dot ExprInfo Expr
.e, for postfix projection.

App AppInfo Expr (NamedArg Expr)
Ordinary (binary) application.

WithApp ExprInfo Expr [Expr]
With application.

Lam ExprInfo LamBinding Expr
λ bs → e.

AbsurdLam ExprInfo Hiding
λ() or λ{}.

ExtendedLam ExprInfo DefInfo Erased QName (List1 Clause)

Pi ExprInfo Telescope1 Expr
Dependent function space Γ → A.

Generalized (Set QName) Expr
Like a Pi, but the ordering is not known

Fun ExprInfo (Arg Expr) Expr
Non-dependent function space.

Let ExprInfo (List1 LetBinding) Expr
let bs in e.

ETel Telescope
Only used when printing telescopes.

Rec ExprInfo RecordAssigns
Record construction.

RecUpdate ExprInfo Expr Assigns
Record update.

ScopedExpr ScopeInfo Expr
Scope annotation.

Quote ExprInfo
Quote an identifier QName.

QuoteTerm ExprInfo
Quote a term.

Unquote ExprInfo
The splicing construct: unquote ...

Tactic ExprInfo Expr [NamedArg Expr]
tactic e x1 .. xn

DontCare Expr
For printing DontCare from Syntax.Internal.

type Args = [NamedArg Expr]

newtype BindName

A name in a binding position: we also compare the nameConcrete when comparing the binders for equality.

With --caching on we compare abstract syntax to determine if we can reuse previous typechecking results: during that comparison two names can have the same nameId but be semantically different, e.g. in {_ : A} -> .. vs. {r : A} -> ...

Constructors

BindName

Fields

unBind :: Name

pattern Def :: QName -> Expr

Pattern synonym for regular Def

mkBindName :: Name -> BindName

generalized :: Set QName -> Expr -> Expr

Smart constructor for Generalized

initCopyInfo :: ScopeCopyInfo

mkBinder :: a -> Binder' a

mkBinder_ :: Name -> Binder

extractPattern :: Binder' a -> Maybe (Pattern, a)

mkDomainFree :: NamedArg Binder -> LamBinding

mkTBind :: Range -> List1 (NamedArg Binder) -> Expr -> TypedBinding

mkTLet :: Range -> [LetBinding] -> Maybe TypedBinding

mkPi :: ExprInfo -> Telescope -> Expr -> Expr

noDataDefParams :: DataDefParams

noWhereDecls :: WhereDeclarations

axiomName :: Declaration -> QName

The name defined by the given axiom.

Precondition: The declaration has to be a (scoped) Axiom.

app :: Expr -> [NamedArg Expr] -> Expr

mkLet :: ExprInfo -> [LetBinding] -> Expr -> Expr

patternToExpr :: Pattern -> Expr

lambdaLiftExpr :: [Name] -> Expr -> Expr

insertImplicitPatSynArgs :: HasRange a => (Range -> a) -> Range -> [Arg Name] -> [NamedArg a] -> Maybe ([(Name, a)], [Arg Name])

module Agda.Syntax.Abstract.Name

Orphan instances

KillRange Suffix

Instance details

Methods

killRange :: KillRangeT Suffix