Agda.Syntax.Abstract.Pattern

Description

Auxiliary functions to handle patterns in the abstract syntax.

Generic and specific traversals.

Generic traversals

type NAP = NamedArg Pattern

class MapNamedArgPattern a where

Minimal complete definition

Nothing

Methods

mapNamedArgPattern :: (NAP -> NAP) -> a -> a

default mapNamedArgPattern :: forall (f :: Type -> Type) a'. (Functor f, MapNamedArgPattern a', a ~ f a') => (NAP -> NAP) -> a -> a

class APatternLike p where

Generic pattern traversal.

Minimal complete definition

Nothing

Associated Types

type ADotT p

Methods

foldrAPattern

Arguments

:: Monoid m

=> (Pattern' (ADotT p) -> m -> m)
Combine a pattern and the value computed from its subpatterns.

-> p

-> m

Fold pattern.

default foldrAPattern :: forall m (f :: Type -> Type) b. (Monoid m, Foldable f, APatternLike b, ADotT p ~ ADotT b, f b ~ p) => (Pattern' (ADotT p) -> m -> m) -> p -> m

traverseAPatternM

Arguments

:: Monad m

=> (Pattern' (ADotT p) -> m (Pattern' (ADotT p)))
pre: Modification before recursion.

-> (Pattern' (ADotT p) -> m (Pattern' (ADotT p)))
post: Modification after recursion.

-> p

-> m p

Traverse pattern.

default traverseAPatternM :: forall (f :: Type -> Type) q m. (Traversable f, APatternLike q, ADotT p ~ ADotT q, f q ~ p, Monad m) => (Pattern' (ADotT p) -> m (Pattern' (ADotT p))) -> (Pattern' (ADotT p) -> m (Pattern' (ADotT p))) -> p -> m p

foldAPattern :: (APatternLike p, Monoid m) => (Pattern' (ADotT p) -> m) -> p -> m

Compute from each subpattern a value and collect them all in a monoid.

preTraverseAPatternM

Arguments

:: (APatternLike p, Monad m)

=> (Pattern' (ADotT p) -> m (Pattern' (ADotT p)))
pre: Modification before recursion.

-> p

-> m p

Traverse pattern(s) with a modification before the recursive descent.

postTraverseAPatternM

Arguments

:: (APatternLike p, Monad m)

=> (Pattern' (ADotT p) -> m (Pattern' (ADotT p)))
post: Modification after recursion.

-> p

-> m p

Traverse pattern(s) with a modification after the recursive descent.

mapAPattern :: APatternLike p => (Pattern' (ADotT p) -> Pattern' (ADotT p)) -> p -> p

Map pattern(s) with a modification after the recursive descent.

Specific folds

patternVars :: APatternLike p => p -> [Name]

Collect pattern variables in left-to-right textual order.

containsAPattern :: APatternLike p => (Pattern' (ADotT p) -> Bool) -> p -> Bool

Check if a pattern contains a specific (sub)pattern.

containsAbsurdPattern :: APatternLike p => p -> Bool

Check if a pattern contains an absurd pattern. For instance, suc (), does so.

Precondition: contains no pattern synonyms.

containsAsPattern :: APatternLike p => p -> Bool

Check if a pattern contains an @-pattern.

checkPatternLinearity :: (Monad m, APatternLike p) => p -> ([Name] -> m ()) -> m ()

Check if any user-written pattern variables occur more than once, and throw the given error if they do.

Specific traversals

substPattern :: [(Name, Pattern)] -> Pattern -> Pattern

Pattern substitution.

For the embedded expression, the given pattern substitution is turned into an expression substitution.

substPattern'

Arguments

:: (e -> e)
Substitution function for expressions.

-> [(Name, Pattern' e)]
(Parallel) substitution.

-> Pattern' e
Input pattern.

-> Pattern' e

Pattern substitution, parametrized by substitution function for embedded expressions.

patternToExpr :: Pattern -> Expr

Convert a pattern to an expression.

Does not support all cases of patterns. Result has no Range info, except in identifiers.

This function is only used in expanding pattern synonyms and in Agda.Syntax.Translation.InternalToAbstract, so we can cut some corners.

class PatternToExpr p e where

Converting a pattern to an expression.

The Hiding context is remembered to create instance metas when translating absurd patterns in instance position.

Minimal complete definition

Nothing

Methods

patToExpr :: p -> Reader Hiding e

default patToExpr :: forall (t :: Type -> Type) p' e'. (Traversable t, PatternToExpr p' e', p ~ t p', e ~ t e') => p -> Reader Hiding e

Other pattern utilities

splitOffTrailingWithPatterns :: Patterns -> (Patterns, Patterns)

Split patterns into (patterns, trailing with-patterns).

trailingWithPatterns :: Patterns -> Patterns

Get the tail of with-patterns of a pattern spine.

data LHSPatternView e

The next patterns are ...

(This view discards PatInfo.)

Constructors

LHSAppP (NAPs e)
Application patterns (non-empty list).

LHSProjP ProjOrigin AmbiguousQName (NamedArg (Pattern' e))
A projection pattern. Is also stored unmodified here.

LHSWithP [Pattern' e]
With patterns (non-empty list). These patterns are not prefixed with WithP.

lhsPatternView :: IsProjP e => NAPs e -> Maybe (LHSPatternView e, NAPs e)

Construct the LHSPatternView of the given list (if not empty).

Return the view and the remaining patterns.

Left-hand-side manipulation

class LHSToSpine a b where

Convert a focused lhs to spine view and back.

Methods

lhsToSpine :: a -> b

spineToLhs :: b -> a

lhsCoreToSpine :: LHSCore' e -> QNamed [NamedArg (Pattern' e)]

spineToLhsCore :: IsProjP e => QNamed [NamedArg (Pattern' e)] -> LHSCore' e

lhsCoreApp :: LHSCore' e -> [NamedArg (Pattern' e)] -> LHSCore' e

Add applicative patterns (non-projection / non-with patterns) to the right.

lhsCoreWith :: LHSCore' e -> [Arg (Pattern' e)] -> LHSCore' e

Add with-patterns to the right.

lhsCoreAddChunk :: IsProjP e => LHSCore' e -> LHSPatternView e -> LHSCore' e

lhsCoreAddSpine :: IsProjP e => LHSCore' e -> [NamedArg (Pattern' e)] -> LHSCore' e

Add projection, with, and applicative patterns to the right.

lhsCoreAllPatterns :: LHSCore' e -> [Pattern' e]

Used for checking pattern linearity.

lhsCoreToPattern :: LHSCore -> Pattern

Used in 'AbstractToConcrete'. Returns a DefP.

mapLHSHead :: (QName -> [NamedArg Pattern] -> LHSCore) -> LHSCore -> LHSCore

Orphan instances

IsWithP (Pattern' e)
Check for with-pattern.

Instance details

Methods

isWithP :: Pattern' e -> Maybe (Pattern' e)