Agda.Syntax.Abstract.PatternSynonyms

Description

Pattern synonym utilities: folding pattern synonym definitions for printing and merging pattern synonym definitions to handle overloaded pattern synonyms.

Documentation

matchPatternSyn :: PatternSynDefn -> Expr -> Maybe [WithHiding Expr]

Match an expression against a pattern synonym.

matchPatternSynP :: PatternSynDefn -> Pattern' e -> Maybe [WithHiding (Pattern' e)]

Match a pattern against a pattern synonym.

mergePatternSynDefs :: List1 PatternSynDefn -> Maybe PatternSynDefn

Merge a list of pattern synonym definitions. Fails unless all definitions have the same shape (i.e. equal up to renaming of variables and constructor names).