Agda.Syntax.Abstract.Views

Documentation

data AppView' arg

Constructors

Application Expr [NamedArg arg]

type AppView = AppView' Expr

appView :: Expr -> AppView

Gather applications to expose head and spine.

Note: everything is an application, possibly of itself to 0 arguments

appView' :: Expr -> AppView' (AppInfo, Expr)

maybeProjTurnPostfix :: Expr -> Maybe Expr

unAppView :: AppView -> Expr

data LamView

Collects plain lambdas.

Constructors

LamView [LamBinding] Expr

lamView :: Expr -> LamView

data PiView

Collect A.Pis.

Constructors

PiView [(ExprInfo, Telescope1)] Type

piView :: Expr -> PiView

unPiView :: PiView -> Expr

asView :: Pattern -> ([Name], Pattern)

Gather top-level AsPatterns to expose underlying pattern.

unScope :: Expr -> Expr

Remove top ScopedExpr wrappers.

deepUnscope :: ExprLike a => a -> a

Remove ScopedExpr wrappers everywhere.

NB: Unless the implementation of ExprLike for clauses has been finished, this does not work for clauses yet.

deepUnscopeDecls :: [Declaration] -> [Declaration]

deepUnscopeDecl :: Declaration -> [Declaration]

Traversal

type RecurseExprFn (m :: Type -> Type) a = Applicative m => (Expr -> m Expr -> m Expr) -> a -> m a

type RecurseExprRecFn (m :: Type -> Type) = forall a. ExprLike a => a -> m a

type FoldExprFn m a = Monoid m => (Expr -> m) -> a -> m

type FoldExprRecFn m = forall a. ExprLike a => a -> m

type TraverseExprFn (m :: Type -> Type) a = (Applicative m, Monad m) => (Expr -> m Expr) -> a -> m a

type TraverseExprRecFn (m :: Type -> Type) = forall a. ExprLike a => a -> m a

class ExprLike a where

Apply an expression rewriting to every subexpression, inside-out. See Agda.Syntax.Internal.Generic.

Minimal complete definition

Nothing

Methods

recurseExpr :: RecurseExprFn m a

The first expression is pre-traversal, the second one post-traversal.

default recurseExpr :: forall (f :: Type -> Type) a' m. (Traversable f, ExprLike a', a ~ f a', Applicative m) => (Expr -> m Expr -> m Expr) -> a -> m a

foldExpr :: FoldExprFn m a

traverseExpr :: TraverseExprFn m a

mapExpr :: (Expr -> Expr) -> a -> a

Getting all declared names

type KName = WithKind QName

class DeclaredNames a where

Extracts "all" names which are declared in a Declaration.

Includes: local modules and where clauses. Excludes: open public, let, with function names, extended lambdas.

Minimal complete definition

Nothing

Methods

declaredNames :: Collection KName m => a -> m

default declaredNames :: forall (t :: Type -> Type) b m. (Foldable t, DeclaredNames b, t b ~ a, Collection KName m) => a -> m