Agda.Syntax.Common.Aspect

Documentation

data Induction

Constructors

Inductive

CoInductive

data Aspect

Constructors

Comment

Keyword

String

Number

Hole

Symbol
Symbols like forall, =, ->, etc.

PrimitiveType
Things like Set and Prop.

Name (Maybe NameKind) Bool
Bool: Is the name an operator part?

Pragma
Text occurring in pragmas that does not have a more specific aspect.

Background
Non-code contents in literate Agda.

Markup
Delimiters used to separate the Agda code blocks from the other contents in literate Agda.

URL Text
Uniform Resource Locator aka clickable link.

data NameKind

NameKinds are figured out during scope checking.

Constructors

Bound
Bound variable.

Generalizable
Generalizable variable. (This includes generalizable variables that have been generalized).

Constructor Induction
Inductive or coinductive constructor.

Datatype

Field
Record field.

Function

Module
Module name.

Postulate

Primitive
Primitive.

Record
Record type.

Argument
Named argument, like x in {x = v}

Macro
Macro.

data OtherAspect

Other aspects, generated by type checking. (These can overlap with each other and with Aspects.)

Constructors

Error

ErrorWarning
A warning that is considered fatal in the end.

DottedPattern

UnsolvedMeta

UnsolvedConstraint
Unsolved constraint not connected to meta-variable. This could for instance be an emptyness constraint.

TerminationProblem

PositivityProblem

Deadcode
Used for highlighting unreachable clauses, unreachable RHS (because of an absurd pattern), etc.

ShadowingInTelescope
Used for shadowed repeated variable names in telescopes.

CoverageProblem

IncompletePattern
When this constructor is used it is probably a good idea to include a note explaining why the pattern is incomplete.

TypeChecks
Code which is being type-checked.

MissingDefinition
Function declaration without matching definition.

InstanceProblem
Unusable instance etc.

CosmeticProblem
Nothing serious, just a beauty flaw.

CatchallClause

ConfluenceProblem

data Aspects

Syntactic aspects of the code. (These cannot overlap.)

Meta information which can be associated with a character/character range.

Constructors

Aspects

Fields

aspect :: Maybe Aspect

otherAspects :: Set OtherAspect

note :: String

This note, if not null, can be displayed as a tool-tip or something like that. It should contain useful information about the range (like the module containing a certain identifier, or the fixity of an operator).

definitionSite :: Maybe DefinitionSite

The definition site of the annotated thing, if applicable and known.

tokenBased :: !TokenBased

Is this entry token-based?

data DefinitionSite

Constructors

DefinitionSite

Fields

defSiteModule :: TopLevelModuleName' Range

The defining module.

defSitePos :: Int

The file position in that module. File positions are counted from 1.

defSiteHere :: Bool

Has this DefinitionSite been created at the defining site of the name?

defSiteAnchor :: Maybe String

A pretty name for the HTML linking.

data TokenBased

Is the highlighting "token-based", i.e. based only on information from the lexer?

Constructors

TokenBased

NotOnlyTokenBased

rangeDefinitionSite :: HasRange a => a -> Aspects

Compute the Aspects which (possibly) contain a DefinitionSite for the given thing-with-range.