Agda.Syntax.Common.KeywordRange

Description

A abstract Range type dedicated to keyword occurrences in the source.

Documentation

data KwRange

Range dedicated to a keyword or fixed token sequence.

Motivation: by lacking a SetRange instance we indicate that it cannot be updated.

kwRange :: HasRange a => a -> KwRange

Create a keyword range.