Agda.Syntax.Common.Pretty

Description

Pretty printing functions.

Documentation

(<?>) :: Doc -> Doc -> Doc infixl 6

a ? b = hang a 2 b

align :: Int -> [(String, Doc)] -> Doc

align max rows lays out the elements of rows in two columns, with the second components aligned. The alignment column of the second components is at most max characters to the right of the left-most column.

Precondition: max > 0.

annotate :: Aspects -> Doc -> Doc

Only adds an annotation node if the Aspects is non-null.

annotateAspect :: Aspect -> Doc -> Doc

Attach a simple Aspect, rather than a full set of Aspects, to a document.

braces

Arguments

:: Doc

-> Doc
Wrap document in {...}

brackets

Arguments

:: Doc

-> Doc
Wrap document in [...]

colon :: Doc

comma :: Doc

definedAt :: HasRange a => Doc -> a -> Doc

Attach the position of something as the "binding site" of a Doc.

dot :: Doc

doubleQuotes

Arguments

:: Doc

-> Doc
Wrap document in "..."

equals :: Doc

fsep :: Foldable t => t Doc -> Doc

fwords :: String -> Doc

githubIssue :: Int -> Doc

Link to an issue on the Agda bug tracker.

hcat :: Foldable t => t Doc -> Doc

hlArgument :: Doc -> Doc

hlBound :: Doc -> Doc

hlComment :: Doc -> Doc

hlConstructor :: Induction -> Doc -> Doc

hlDatatype :: Doc -> Doc

hlField :: Doc -> Doc

hlFunction :: Doc -> Doc

hlGeneralizable :: Doc -> Doc

hlHole :: Doc -> Doc

hlKeyword :: Doc -> Doc

hlMacro :: Doc -> Doc

hlModule :: Doc -> Doc

hlNameKind :: NameKind -> Doc -> Doc

hlNumber :: Doc -> Doc

hlPostulate :: Doc -> Doc

hlPragma :: Doc -> Doc

hlPrimitive :: Doc -> Doc

hlPrimitiveType :: Doc -> Doc

hlRecord :: Doc -> Doc

hlString :: Doc -> Doc

hlSymbol :: Doc -> Doc

href :: Text -> Doc -> Doc

Attach a link to a document (cf. LaTeX href).

hsep :: Foldable t => t Doc -> Doc

hsepWith :: Doc -> Doc -> Doc -> Doc

Separate, but only if both separees are not null.

lbrace :: Doc

lbrack :: Doc

lparen :: Doc

mparens :: Bool -> Doc -> Doc

Apply parens to Doc if boolean is true.

multiLineText :: String -> Doc

Handles strings with newlines properly (preserving indentation)

parens

Arguments

:: Doc

-> Doc
Wrap document in (...)

parensNonEmpty :: Doc -> Doc

Only wrap in parens if not empty

pipe :: Doc

pluralS :: Sized a => a -> Doc -> Doc

prefixedThings :: Doc -> [Doc] -> Doc

Used for with-like telescopes

prettyAssign :: (Pretty k, Pretty v) => (k, v) -> Doc

Pretty print a single association.

prettyInterval :: Eq a => Position' a -> Position' a -> Doc

prettyLineColumn :: Position' a -> Doc

prettyList_ :: Pretty a => [a] -> Doc

Comma separated list, without the brackets.

prettyMap :: (Pretty k, Pretty v) => [(k, v)] -> Doc

Pretty print an association list.

prettySet :: Pretty a => [a] -> Doc

Pretty print a set.

prettyShow :: Pretty a => a -> String

Use instead of show when printing to world.

pshow :: Show a => a -> Doc

pshow = text . show

punctuate :: Foldable t => Doc -> t Doc -> [Doc]

pwords :: String -> [Doc]

quotes

Arguments

:: Doc

-> Doc
Wrap document in '...'

rbrace :: Doc

rbrack :: Doc

rparen :: Doc

semi :: Doc

sep :: Foldable t => t Doc -> Doc

singPlural :: Sized a => a -> c -> c -> c

textNonEmpty :: String -> Doc

Return empty for empty strings.

twords :: Text -> [Doc]

url :: String -> Doc

A URL formatted as link (cf. LaTeX url).

vcat :: Foldable t => t Doc -> Doc

vsep :: [Doc] -> Doc

type Doc = Doc Aspects

The type of documents. We use documents annotated by Aspects to record syntactic highlighting information that is generated during pretty-printing.

type DocTree = DocTree Aspects

class Pretty a where

While Show is for rendering data in Haskell syntax, Pretty is for displaying data to the world, i.e., the user and the environment.

Atomic data has no inner document structure, so just implement pretty as pretty a = text $ ... a ....

Minimal complete definition

pretty | prettyPrec

Methods

pretty :: a -> Doc

prettyPrec :: Int -> a -> Doc

prettyList :: [a] -> Doc

($$) :: Doc a -> Doc a -> Doc a infixl 5

Above, except that if the last line of the first argument stops at least one position before the first line of the second begins, these two lines are overlapped. For example:

   text "hi" $$ nest 5 (text "there")

lays out as

   hi   there

rather than

   hi
        there

$$ is associative, with identity empty, and also satisfies

(x $$ y) <> z = x $$ (y <> z), if y non-empty.

($+$) :: Doc a -> Doc a -> Doc a infixl 5

Above, with no overlapping. $+$ is associative, with identity empty.

(<+>) :: Doc a -> Doc a -> Doc a infixl 6

Beside, separated by space, unless one of the arguments is empty. <+> is associative, with identity empty.

cat :: [Doc a] -> Doc a

Either hcat or vcat.

char :: Char -> Doc a

A document of height and width 1, containing a literal character.

double

Arguments

:: Double

-> Doc a
double n = text (show n)

fcat :: [Doc a] -> Doc a

"Paragraph fill" version of cat.

float

Arguments

:: Float

-> Doc a
float n = text (show n)

fullRender

Arguments

:: Mode
Rendering mode.

-> Int
Line length.

-> Float
Ribbons per line.

-> (TextDetails -> a -> a)
What to do with text.

-> a
What to do at the end.

-> Doc b
The document.

-> a
Result.

The general rendering interface. Please refer to the Style and Mode types for a description of rendering mode, line length and ribbons.

fullRenderAnn

Arguments

:: Mode
Rendering mode.

-> Int
Line length.

-> Float
Ribbons per line.

-> (AnnotDetails b -> a -> a)
What to do with text.

-> a
What to do at the end.

-> Doc b
The document.

-> a
Result.

The general rendering interface, supporting annotations. Please refer to the Style and Mode types for a description of rendering mode, line length and ribbons.

hang :: Doc a -> Int -> Doc a -> Doc a

hang d1 n d2 = sep [d1, nest n d2]

int

Arguments

:: Int

-> Doc a
int n = text (show n)

integer

Arguments

:: Integer

-> Doc a
integer n = text (show n)

isEmpty :: Doc a -> Bool

Returns True if the document is empty

nest :: Int -> Doc a -> Doc a

Nest (or indent) a document by a given number of positions (which may also be negative). nest satisfies the laws:

nest 0 x = x

nest k (nest k' x) = nest (k+k') x

nest k (x <> y) = nest k z <> nest k y

nest k (x $$ y) = nest k x $$ nest k y

nest k empty = empty

x <> nest k y = x <> y, if x non-empty

The side condition on the last law is needed because empty is a left identity for <>.

ptext :: String -> Doc a

Same as text. Used to be used for Bytestrings.

rational

Arguments

:: Rational

-> Doc a
rational n = text (show n)

render :: Doc a -> String

Render the Doc to a String using the default Style (see style).

renderSpans :: Doc ann -> (String, [Span ann])

Render an annotated Doc to a String and list of annotations (see Span) using the default Style (see style).

renderStyle :: Style -> Doc a -> String

Render the Doc to a String using the given Style.

sizedText :: Int -> String -> Doc a

Some text with any width. (text s = sizedText (length s) s)

style :: Style

The default style (mode=PageMode, lineLength=100, ribbonsPerLine=1.5).

text :: String -> Doc a

A document of height 1 containing a literal string. text satisfies the following laws:

text s <> text t = text (s++t)

text "" <> x = x, if x non-empty

The side condition on the last law is necessary because text "" has height 1, while empty has no height.

zeroWidthText :: String -> Doc a

Some text, but without any width. Use for non-printing text such as a HTML or Latex tags

data Mode

Rendering mode.

Constructors

PageMode
Normal rendering (lineLength and ribbonsPerLine respected').

ZigZagMode
With zig-zag cuts.

LeftMode
No indentation, infinitely long lines (lineLength ignored), but explicit new lines, i.e., text "one" $$ text "two", are respected.

OneLineMode
All on one line, lineLength ignored and explicit new lines ($$) are turned into spaces.

data Span a

A Span represents the result of an annotation after a Doc has been rendered, capturing where the annotation now starts and ends in the rendered output.

Constructors

Span

Fields

spanStart :: !Int

spanLength :: !Int

spanAnnotation :: a

data Style

A rendering style. Allows us to specify constraints to choose among the many different rendering options.

Constructors

Style

Fields

mode :: Mode

The rendering mode.

lineLength :: Int

Maximum length of a line, in characters.

ribbonsPerLine :: Float

Ratio of line length to ribbon length. A ribbon refers to the characters on a line excluding indentation. So a lineLength of 100, with a ribbonsPerLine of 2.0 would only allow up to 50 characters of ribbon to be displayed on a line, while allowing it to be indented up to 50 characters.

pattern Chr :: !Char -> TextDetails

A single Char fragment

pattern PStr :: String -> TextDetails

Used to represent a Fast String fragment but now deprecated and identical to the Str constructor.