Agda.Syntax.Concrete

Description

The concrete syntax is a raw representation of the program text without any desugaring at all. This is what the parser produces. The idea is that if we figure out how to keep the concrete syntax around, it can be printed exactly as the user wrote it.

Expressions

data Expr

Concrete expressions. Should represent exactly what the user wrote.

Constructors

Ident QName
ex: x

Lit Range Literal
ex: 1 or "foo"

QuestionMark Range (Maybe Nat)
ex: ? or {! ... !}

Underscore Range (Maybe String)
ex: _ or _A_5

RawApp Range (List2 Expr)
before parsing operators

App Range Expr (NamedArg Expr)
ex: e e, e {e}, or e {x = e}

OpApp Range QName (Set1 Name) OpAppArgs
ex: e + e The QName is possibly ambiguous, but it must correspond to one of the names in the set.

WithApp Range Expr (List1 Expr)
ex: e | e1 | .. | en

HiddenArg Range (Named_ Expr)
ex: {e} or {x=e}

InstanceArg Range (Named_ Expr)
ex: {{e}} or {{x=e}}

Lam Range (List1 LamBinding) Expr
ex: \x {y} -> e or \(x:A){y:B} -> e

AbsurdLam Range Hiding
ex: \ ()

ExtendedLam Range Erased (List1 LamClause)
ex: \ { p11 .. p1a -> e1 ; .. ; pn1 .. pnz -> en }

Fun Range (Arg Expr) Expr
ex: e -> e or .e -> e (NYI: {e} -> e)

Pi Telescope1 Expr
ex: (xs:e) -> e or {xs:e} -> e

Rec Range RecordAssignments
ex: record {x = a; y = b}, or record { x = a; M1; M2 }

RecUpdate Range Expr [FieldAssignment]
ex: record e {x = a; y = b}

Let Range (List1 Declaration) (Maybe Expr)
ex: let Ds in e, missing body when parsing do-notation let

Paren Range Expr
ex: (e)

IdiomBrackets Range [Expr]
ex: (| e1 | e2 | .. | en |) or (|)

DoBlock Range (List1 DoStmt)
ex: do x <- m1; m2

Absurd Range
ex: () or {}, only in patterns

As Range Name Expr
ex: x@p, only in patterns

Dot KwRange Expr
ex: .p, only in patterns

DoubleDot KwRange Expr
ex: ..A, used for parsing ..A -> B

Quote Range
ex: quote, should be applied to a name

QuoteTerm Range
ex: quoteTerm, should be applied to a term

Tactic Range Expr
ex: @(tactic t), used to declare tactic arguments

Unquote Range
ex: unquote, should be applied to a term of type Term

DontCare Expr
to print irrelevant things

Equal Range Expr Expr
ex: a = b, used internally in the parser

Ellipsis Range
..., used internally to parse patterns.

KnownIdent NameKind QName
An identifier coming from abstract syntax, for which we know a precise syntactic highlighting class (used in printing).

KnownOpApp NameKind Range QName (Set1 Name) OpAppArgs
An operator application coming from abstract syntax, for which we know a precise syntactic highlighting class (used in printing).

Generalized Expr

data OpApp e

Constructors

SyntaxBindingLambda Range (List1 LamBinding) e
An abstraction inside a special syntax declaration (see Issue 358 why we introduce this).

Ordinary e

fromOrdinary :: e -> OpApp e -> e

type OpAppArgs = OpAppArgs' Expr

type OpAppArgs' e = List1 (NamedArg (MaybePlaceholder (OpApp e)))

type OpAppArgs0 e = [NamedArg (MaybePlaceholder (OpApp e))]

module Agda.Syntax.Concrete.Name

data AppView

The Expr is not an application.

Constructors

AppView Expr [NamedArg Expr]

appView :: Expr -> AppView

unAppView :: AppView -> Expr

toNamedArg :: Expr -> NamedArg Expr

Parse outermost hiding information.

unNamedArg :: NamedArg Expr -> Expr

Unparse hiding information.

rawApp :: List1 Expr -> Expr

rawAppP :: List1 Pattern -> Pattern

isSingleIdentifierP :: Pattern -> Maybe Name

removeParenP :: Pattern -> Pattern

isPattern :: Expr -> Maybe Pattern

Turn an expression into a pattern. Fails if the expression is not a valid pattern.

isAbsurdP :: Pattern -> Maybe (Range, Hiding)

isBinderP :: Pattern -> Maybe Binder

observeHiding :: Expr -> WithHiding Expr

Observe the hiding status of an expression

observeRelevance :: Expr -> (Relevance, Expr)

Observe the relevance status of an expression

observeModifiers :: Expr -> Arg Expr

Observe various modifiers applied to an expression

exprToPatternWithHoles :: Expr -> Pattern

Turn an expression into a pattern, turning non-pattern subexpressions into WildP.

returnExpr :: Expr -> Maybe Expr

Bindings

data Binder' a

A Binder x@p, the pattern is optional

Constructors

Binder

Fields

binderPattern :: Maybe Pattern

binderNameOrigin :: BinderNameOrigin

binderName :: a

type Binder = Binder' BoundName

mkBinder_ :: Name -> Binder

mkBinder :: a -> Binder' a

type LamBinding = LamBinding' TypedBinding

A lambda binding is either domain free or typed.

data LamBinding' a

Constructors

DomainFree (NamedArg Binder)
. x or {x} or .x or .{x} or {.x} or x@p or (p)

DomainFull a
. (xs : e) or {xs : e}

dropTypeAndModality :: LamBinding -> [LamBinding]

Drop type annotations and lets from bindings.

type TypedBinding = TypedBinding' Expr

A typed binding.

data TypedBinding' e

Constructors

TBind Range (List1 (NamedArg Binder)) e
Binding (x1@p1 ... xn@pn : A).

TLet Range (List1 Declaration)
Let binding (let Ds) or (open M args).

type RecordAssignment = Either FieldAssignment ModuleAssignment

type RecordAssignments = [RecordAssignment]

type FieldAssignment = FieldAssignment' Expr

data FieldAssignment' a

Constructors

FieldAssignment

Fields

_nameFieldA :: Name

_exprFieldA :: a

nameFieldA :: forall a f. Functor f => (Name -> f Name) -> FieldAssignment' a -> f (FieldAssignment' a)

exprFieldA :: forall a f. Functor f => (a -> f a) -> FieldAssignment' a -> f (FieldAssignment' a)

data ModuleAssignment

Constructors

ModuleAssignment

Fields

_qnameModA :: QName

_exprModA :: [Expr]

_importDirModA :: ImportDirective

data BoundName

Constructors

BName

Fields

boundName :: Name

bnameFixity :: Fixity'

bnameTactic :: TacticAttribute

From @tactic attribute.

bnameIsFinite :: Bool

The @finite cannot be parsed, it comes from the builtin Partial only.

mkBoundName_ :: Name -> BoundName

mkBoundName :: Name -> Fixity' -> BoundName

type TacticAttribute = TacticAttribute' Expr

newtype TacticAttribute' a

Constructors

TacticAttribute

Fields

theTacticAttribute :: Maybe (Ranged a)

type Telescope = [TypedBinding]

type Telescope1 = List1 TypedBinding

A telescope is a sequence of typed bindings. Bound variables are in scope in later types.

lamBindingsToTelescope :: Range -> [LamBinding] -> Telescope

We can try to get a Telescope from a [LamBinding]. If we have a type annotation already, we're happy. Otherwise we manufacture a binder with an underscore for the type.

makePi :: Telescope -> Expr -> Expr

Smart constructor for Pi: check whether the Telescope is empty

mkLam :: Range -> [LamBinding] -> Expr -> Expr

Smart constructor for Lam: check for non-zero bindings.

mkLet :: Range -> [Declaration] -> Expr -> Expr

Smart constructor for Let: check for non-zero let bindings.

mkTLet :: Range -> [Declaration] -> Maybe (TypedBinding' e)

Smart constructor for TLet: check for non-zero let bindings.

Declarations

data Declaration

The representation type of a declaration. The comments indicate which type in the intended family the constructor targets.

Constructors

TypeSig ArgInfo TacticAttribute Name Expr
Axioms and functions can be irrelevant. (Hiding should be NotHidden)

FieldSig IsInstance TacticAttribute Name (Arg Expr)

Generalize KwRange [TypeSignature]
Variables to be generalized, can be hidden and/or irrelevant.

Field KwRange [FieldSignature]

FunClause LHS RHS WhereClause Catchall

DataSig Range Erased Name [LamBinding] Expr
lone data signature in mutual block

Data Range Erased Name [LamBinding] Expr [TypeSignatureOrInstanceBlock]

DataDef Range Name [LamBinding] [TypeSignatureOrInstanceBlock]

RecordSig Range Erased Name [LamBinding] Expr
lone record signature in mutual block

RecordDef Range Name [RecordDirective] [LamBinding] [Declaration]

Record Range Erased Name [RecordDirective] [LamBinding] Expr [Declaration]

Infix Fixity (List1 Name)

Syntax Name Notation
notation declaration for a name

PatternSyn Range Name [WithHiding Name] Pattern

Mutual KwRange [Declaration]

InterleavedMutual KwRange [Declaration]

Abstract KwRange [Declaration]

Private KwRange Origin [Declaration]
In Agda.Syntax.Concrete.Definitions we generate private blocks temporarily, which should be treated different that user-declared private blocks. Thus the Origin.

InstanceB KwRange [Declaration]
The KwRange here only refers to the range of the instance keyword. The range of the whole block InstanceB r ds is fuseRange r ds.

LoneConstructor KwRange [Declaration]

Macro KwRange [Declaration]

Postulate KwRange [TypeSignatureOrInstanceBlock]

Primitive KwRange [TypeSignature]

Open Range QName ImportDirective

Import Range QName (Maybe AsName) !OpenShortHand ImportDirective

ModuleMacro Range Erased Name ModuleApplication !OpenShortHand ImportDirective

Module Range Erased QName Telescope [Declaration]

UnquoteDecl Range [Name] Expr
unquoteDecl xs = e

UnquoteDef Range [Name] Expr
unquoteDef xs = e

UnquoteData Range Name [Name] Expr
unquoteDecl data d constructor xs = e

Pragma Pragma

Opaque KwRange [Declaration]
opaque ...

Unfolding KwRange [QName]
unfolding ...

isPragma :: CMaybe Pragma m => Declaration -> m

Return Pragma if Declaration is Pragma.

data RecordDirective

Isolated record directives parsed as Declarations

Constructors

Induction (Ranged Induction)
Range of keyword [co]inductive.

Constructor Name IsInstance

Eta (Ranged HasEta0)
Range of [no-]eta-equality keyword.

PatternOrCopattern Range
If declaration pattern is present, give its range.

type RecordDirectives = RecordDirectives' (Maybe (Name, IsInstance))

data ModuleApplication

Constructors

SectionApp Range Telescope QName [Expr]
tel M exprs where M exprs is a RawApp just after parsing.

RecordModuleInstance Range QName
M {{...}}

type TypeSignature = Declaration

Just type signatures.

type TypeSignatureOrInstanceBlock = Declaration

Just type signatures or instance blocks.

type ImportDirective = ImportDirective' Name Name

The things you are allowed to say when you shuffle names between name spaces (i.e. in import, namespace, or open declarations).

type Using = Using' Name Name

type ImportedName = ImportedName' Name Name

An imported name can be a module or a defined name.

type Renaming = Renaming' Name Name

type RenamingDirective = RenamingDirective' Name Name

type HidingDirective = HidingDirective' Name Name

data AsName' a

The content of the as-clause of the import statement.

Constructors

AsName

Fields

asName :: a

The "as" name.

asRange :: Range

The range of the "as" keyword. Retained for highlighting purposes.

type AsName = AsName' (Either Expr Name)

From the parser, we get an expression for the as-Name, which we have to parse into a Name.

data OpenShortHand

Constructors

DoOpen

DontOpen

type RewriteEqn = RewriteEqn' () Name Pattern Expr

type WithExpr = Named Name (Arg Expr)

data LHS

Left hand sides can be written in infix style. For example:

n + suc m = suc (n + m)
(f ∘ g) x = f (g x)

We use fixity information to see which name is actually defined.

Constructors

LHS
Original pattern (including with-patterns), rewrite equations and with-expressions.

Fields

lhsOriginalPattern :: Pattern

e.g. f ps | wps

lhsRewriteEqn :: [RewriteEqn]

(rewrite e | with p <- e in eq) (many)

lhsWithExpr :: [WithExpr]

with e1 in eq | {e2} | ... (many)

data Pattern

Concrete patterns. No literals in patterns at the moment.

Constructors

IdentP Bool QName
c or x

If the boolean is False, then the QName must not refer to a constructor or a pattern synonym. The value False is used when a hidden argument pun is expanded.

QuoteP Range
quote

AppP Pattern (NamedArg Pattern)
p p' or p {x = p'}

RawAppP Range (List2 Pattern)
p1..pn before parsing operators

OpAppP Range QName (Set1 Name) (List1 (NamedArg Pattern))
eg: p => p' for operator _=>_ The QName is possibly ambiguous, but it must correspond to one of the names in the set.

HiddenP Range (Named_ Pattern)
{p} or {x = p}

InstanceP Range (Named_ Pattern)
{{p}} or {{x = p}}

ParenP Range Pattern
(p)

WildP Range
_

AbsurdP Range
()

AsP Range Name Pattern
x@p

DotP KwRange Range Expr
.e, the KwRange is for the dot, the Range for the whole thing (including the dot).

LitP Range Literal
0, 1, etc.

RecP Range [FieldAssignment' Pattern]
record {x = p; y = q}

EqualP Range (List1 (Expr, Expr))
i = i1 i.e. cubical face lattice generator

EllipsisP Range (Maybe Pattern)
..., only as left-most pattern. Second arg is Nothing before expansion, and Just p after expanding ellipsis to p.

WithP Range Pattern
| p, for with-patterns.

data LHSCore

Processed (operator-parsed) intermediate form of the core f ps of LHS. Corresponds to lhsOriginalPattern.

Constructors

LHSHead

Fields

lhsDefName :: QName

f

lhsPats :: [NamedArg Pattern]

ps

LHSProj

Fields

lhsDestructor :: QName

Record projection.

lhsPatsLeft :: [NamedArg Pattern]

Patterns for record indices (currently none).

lhsFocus :: NamedArg LHSCore

Main argument.

lhsPats :: [NamedArg Pattern]

ps

LHSWith

Fields

lhsHead :: LHSCore

lhsWithPatterns :: List1 Pattern

At least one (| p).

lhsPats :: [NamedArg Pattern]

ps

LHSEllipsis

Fields

lhsEllipsisRange :: Range

lhsEllipsisPat :: LHSCore

Pattern that was expanded from an ellipsis ....

data LamClause

Constructors

LamClause

Fields

lamLHS :: [Pattern]

Possibly empty sequence.

lamRHS :: RHS

lamCatchall :: Catchall

type RHS = RHS' Expr

data RHS' e

Constructors

AbsurdRHS
No right hand side because of absurd match.

RHS e

type WhereClause = WhereClause' [Declaration]

where block following a clause.

data WhereClause' decls

The generalization WhereClause' is for the sake of Agda.Concrete.Generic.

Constructors

NoWhere
No where clauses.

AnyWhere Range decls
Ordinary where. Range of the where keyword. List of declarations can be empty.

SomeWhere Range Erased Name Access decls
Named where: module M where ds. Range of the keywords module and where. The Access flag applies to the Name (not the module contents!) and is propagated from the parent function. List of declarations can be empty.

data WhereClause_

Type of where block following a clause.

Constructors

NoWhere_
No where clauses.

AnyWhere_
Ordinary where.

SomeWhere_
Named where: module M where ....

data ExprWhere

An expression followed by a where clause. Currently only used to give better a better error message in interaction.

Constructors

ExprWhere Expr WhereClause

whereClause_ :: WhereClause' a -> WhereClause_

data DoStmt

Constructors

DoBind Range Pattern Expr [LamClause]
p ← e where cs

DoThen Expr

DoLet Range (List1 Declaration)

data Pragma

Constructors

OptionsPragma Range [String]

BuiltinPragma Range RString QName

RewritePragma Range Range [QName]
Second Range is for REWRITE keyword.

ForeignPragma Range (Ranged BackendName) String

CompilePragma Range (Ranged BackendName) QName String

StaticPragma Range QName

InlinePragma Range Bool QName
INLINE or NOINLINE

ImpossiblePragma Range [String]
Throws an internal error in the scope checker. The Strings are words to be displayed with the error.

EtaPragma Range QName
For coinductive records, use pragma instead of regular eta-equality definition (as it is might make Agda loop).

WarningOnUsage Range QName Text
Applies to the named function

WarningOnImport Range Text
Applies to the current module

InjectivePragma Range QName
Mark a definition as injective for the pattern matching unifier.

InjectiveForInferencePragma Range QName
Mark a definition as injective for the conversion checker

DisplayPragma Range Pattern Expr
Display lhs as rhs (modifies the printer).

CatchallPragma Range
Applies to the following function clause.

TerminationCheckPragma Range (TerminationCheck Name)
Applies to the following function (and all that are mutually recursive with it) or to the functions in the following mutual block.

NoCoverageCheckPragma Range
Applies to the following function (and all that are mutually recursive with it) or to the functions in the following mutual block.

NoPositivityCheckPragma Range
Applies to the following data/record type or mutual block.

PolarityPragma Range Name [Ranged Occurrence]

NoUniverseCheckPragma Range
Applies to the following data/record type.

NotProjectionLikePragma Range QName
Applies to the stated function

OverlapPragma Range [QName] OverlapMode
Applies to the given name(s), which must be instance names (checked by the type checker).

data Module

Modules: Top-level pragmas plus other top-level declarations.

Constructors

Mod

Fields

modPragmas :: [Pragma]

modDecls :: [Declaration]

data ThingWithFixity x

Decorating something with Fixity'.

Constructors

ThingWithFixity x Fixity'

type HoleContent = HoleContent' () Name Pattern Expr

data HoleContent' qn nm p e

Extended content of an interaction hole.

Constructors

HoleContentExpr e
e

HoleContentRewrite [RewriteEqn' qn nm p e]
(rewrite | invert) e0 | ... | en

spanAllowedBeforeModule :: [Declaration] -> ([Declaration], [Declaration])

Splits off allowed (= import) declarations before the first non-allowed declaration. After successful parsing, the first non-allowed declaration should be a module declaration.

ungatherRecordDirectives :: RecordDirectives -> [RecordDirective]