Agda.Syntax.Concrete.Definitions

Description

Preprocess Declarations, producing NiceDeclarations.

Attach fixity and syntax declarations to the definition they refer to.

Distribute the following attributes to the individual definitions: abstract, instance, postulate, primitive, private, termination pragmas.

Gather the function clauses belonging to one function definition.

Expand ellipsis ... in function clauses following with.

Infer mutual blocks. A block starts when a lone signature is encountered, and ends when all lone signatures have seen their definition.

Handle interleaved mutual blocks. In an `interleaved mutual' block we:

leave the data and fun sigs in place

classify signatures in constructor block based on their return type and group them all as a data def at the position in the block where the first constructor for the data sig in question occured

classify fun clauses based on the declared function used and group them all as a fundef at the position in the block where the first such fun clause appeared

Report basic well-formedness error, when one of the above transformation fails. When possible, errors should be deferred to the scope checking phase (ConcreteToAbstract), where we are in the TCM and can produce more informative error messages.

Documentation

data NiceDeclaration

The nice declarations. No fixity declarations and function definitions are contained in a single constructor instead of spread out between type signatures and clauses. The private, postulate, abstract and instance modifiers have been distributed to the individual declarations.

Observe the order of components:

Range Fixity' Access IsAbstract IsInstance TerminationCheck PositivityCheck

further attributes

(Q)Name

content (Expr, Declaration ...)

Constructors

Axiom Range Access IsAbstract IsInstance ArgInfo Name Expr
IsAbstract argument: We record whether a declaration was made in an abstract block.

ArgInfo argument: Axioms and functions can be declared irrelevant. (Hiding should be NotHidden.)

NiceField Range Access IsAbstract IsInstance TacticAttribute Name (Arg Expr)

PrimitiveFunction Range Access IsAbstract Name (Arg Expr)

NiceMutual Range TerminationCheck CoverageCheck PositivityCheck [NiceDeclaration]

NiceModule Range Access IsAbstract QName Telescope [Declaration]

NiceModuleMacro Range Access Name ModuleApplication OpenShortHand ImportDirective

NiceOpen Range QName ImportDirective

NiceImport Range QName (Maybe AsName) OpenShortHand ImportDirective

NicePragma Range Pragma

NiceRecSig Range Access IsAbstract PositivityCheck UniverseCheck Name [LamBinding] Expr

NiceDataSig Range Access IsAbstract PositivityCheck UniverseCheck Name [LamBinding] Expr

NiceFunClause Range Access IsAbstract TerminationCheck CoverageCheck Catchall Declaration
An uncategorized function clause, could be a function clause without type signature or a pattern lhs (e.g. for irrefutable let). The Declaration is the actual FunClause.

FunSig Range Access IsAbstract IsInstance IsMacro ArgInfo TerminationCheck CoverageCheck Name Expr

FunDef Range [Declaration] IsAbstract IsInstance TerminationCheck CoverageCheck Name [Clause]
Block of function clauses (we have seen the type signature before). The Declarations are the original declarations that were processed into this FunDef and are only used in notSoNiceDeclaration. Andreas, 2017-01-01: Because of issue #2372, we add IsInstance here. An alias should know that it is an instance.

NiceDataDef Range Origin IsAbstract PositivityCheck UniverseCheck Name [LamBinding] [NiceConstructor]

NiceLoneConstructor Range [NiceConstructor]

NiceRecDef Range Origin IsAbstract PositivityCheck UniverseCheck Name RecordDirectives [LamBinding] [Declaration]
(Maybe Range) gives range of the 'pattern' declaration.

NicePatternSyn Range Access Name [Arg Name] Pattern

NiceGeneralize Range Access ArgInfo TacticAttribute Name Expr

NiceUnquoteDecl Range Access IsAbstract IsInstance TerminationCheck CoverageCheck [Name] Expr

NiceUnquoteDef Range Access IsAbstract TerminationCheck CoverageCheck [Name] Expr

type NiceConstructor = NiceTypeSignature

Only Axioms.

type NiceTypeSignature = NiceDeclaration

Only Axioms.

data Clause

One clause in a function definition. There is no guarantee that the LHS actually declares the Name. We will have to check that later.

Constructors

Clause Name Catchall LHS RHS WhereClause [Clause]

data DeclarationException

Exception with internal source code callstack

Constructors

DeclarationException

Fields

deLocation :: CallStack

deException :: DeclarationException'

data DeclarationWarning

Constructors

DeclarationWarning

Fields

dwLocation :: CallStack

dwWarning :: DeclarationWarning'

data DeclarationWarning'

Non-fatal errors encountered in the Nicifier.

Constructors

EmptyAbstract Range
Empty abstract block.

EmptyConstructor Range
Empty constructor block.

EmptyField Range
Empty field block.

EmptyGeneralize Range
Empty variable block.

EmptyInstance Range
Empty instance block

EmptyMacro Range
Empty macro block.

EmptyMutual Range
Empty mutual block.

EmptyPostulate Range
Empty postulate block.

EmptyPrivate Range
Empty private block.

EmptyPrimitive Range
Empty primitive block.

InvalidCatchallPragma Range
A {-# CATCHALL #-} pragma that does not precede a function clause.

InvalidConstructor Range
Invalid definition in a constructor block

InvalidConstructorBlock Range
Invalid constructor block (not inside an interleaved mutual block)

InvalidCoverageCheckPragma Range
A {-# NON_COVERING #-} pragma that does not apply to any function.

InvalidNoPositivityCheckPragma Range
A {-# NO_POSITIVITY_CHECK #-} pragma that does not apply to any data or record type.

InvalidNoUniverseCheckPragma Range
A {-# NO_UNIVERSE_CHECK #-} pragma that does not apply to a data or record type.

InvalidRecordDirective Range
A record directive outside of a record / below existing fields.

InvalidTerminationCheckPragma Range
A {-# TERMINATING #-} and {-# NON_TERMINATING #-} pragma that does not apply to any function.

MissingDefinitions [(Name, Range)]
Declarations (e.g. type signatures) without a definition.

NotAllowedInMutual Range String

OpenPublicPrivate Range
private has no effect on open public. (But the user might think so.)

OpenPublicAbstract Range
abstract has no effect on open public. (But the user might think so.)

PolarityPragmasButNotPostulates [Name]

PragmaNoTerminationCheck Range
Pragma {-# NO_TERMINATION_CHECK #-} has been replaced by {-# TERMINATING #-} and {-# NON_TERMINATING #-}.

PragmaCompiled Range
COMPILE pragmas are not allowed in safe mode

ShadowingInTelescope (List1 (Name, List2 Range))

UnknownFixityInMixfixDecl [Name]

UnknownNamesInFixityDecl [Name]

UnknownNamesInPolarityPragmas [Name]

UselessAbstract Range
abstract block with nothing that can (newly) be made abstract.

UselessInstance Range
instance block with nothing that can (newly) become an instance.

UselessPrivate Range
private block with nothing that can (newly) be made private.

unsafeDeclarationWarning :: DeclarationWarning -> Bool

Nicifier warnings turned into errors in --safe mode.

data Nice a

Nicifier monad. Preserve the state when throwing an exception.

runNice :: Nice a -> (Either DeclarationException a, NiceWarnings)

Run a Nicifier computation, return result and warnings (in chronological order).

niceDeclarations :: Fixities -> [Declaration] -> Nice [NiceDeclaration]

Main. Fixities (or more precisely syntax declarations) are needed when grouping function clauses.

notSoNiceDeclarations :: NiceDeclaration -> [Declaration]

(Approximately) convert a NiceDeclaration back to a list of Declarations.

niceHasAbstract :: NiceDeclaration -> Maybe IsAbstract

Has the NiceDeclaration a field of type IsAbstract?

type Measure = Name

Termination measure is, for now, a variable name.

declarationWarningName :: DeclarationWarning -> WarningName