Agda.Syntax.Concrete.Definitions.Types

Documentation

data NiceDeclaration

The nice declarations. No fixity declarations and function definitions are contained in a single constructor instead of spread out between type signatures and clauses. The private, postulate, abstract and instance modifiers have been distributed to the individual declarations.

Observe the order of components:

Range Fixity' Access IsAbstract IsInstance TerminationCheck PositivityCheck

further attributes

(Q)Name

content (Expr, Declaration ...)

Constructors

Axiom Range Access IsAbstract IsInstance ArgInfo Name Expr
IsAbstract argument: We record whether a declaration was made in an abstract block.

ArgInfo argument: Axioms and functions can be declared irrelevant. (Hiding should be NotHidden.)

NiceField Range Access IsAbstract IsInstance TacticAttribute Name (Arg Expr)

PrimitiveFunction Range Access IsAbstract Name (Arg Expr)

NiceMutual KwRange TerminationCheck CoverageCheck PositivityCheck [NiceDeclaration]

NiceModule Range Access IsAbstract Erased QName Telescope [Declaration]

NiceModuleMacro Range Access Erased Name ModuleApplication OpenShortHand ImportDirective

NiceOpen Range QName ImportDirective

NiceImport Range QName (Maybe AsName) OpenShortHand ImportDirective

NicePragma Range Pragma

NiceRecSig Range Erased Access IsAbstract PositivityCheck UniverseCheck Name [LamBinding] Expr

NiceDataSig Range Erased Access IsAbstract PositivityCheck UniverseCheck Name [LamBinding] Expr

NiceFunClause Range Access IsAbstract TerminationCheck CoverageCheck Catchall Declaration
An uncategorized function clause, could be a function clause without type signature or a pattern lhs (e.g. for irrefutable let). The Declaration is the actual FunClause.

FunSig Range Access IsAbstract IsInstance IsMacro ArgInfo TerminationCheck CoverageCheck Name Expr

FunDef Range (List1 Declaration) IsAbstract IsInstance TerminationCheck CoverageCheck Name (List1 Clause)
Block of function clauses (we have seen the type signature before). The Declarations are the original declarations that were processed into this FunDef and are only used in notSoNiceDeclaration. Andreas, 2017-01-01: Because of issue #2372, we add IsInstance here. An alias should know that it is an instance.

NiceDataDef Range Origin IsAbstract PositivityCheck UniverseCheck Name [LamBinding] [NiceConstructor]

NiceLoneConstructor KwRange [NiceConstructor]

NiceRecDef Range Origin IsAbstract PositivityCheck UniverseCheck Name [RecordDirective] [LamBinding] [Declaration]
(Maybe Range) gives range of the 'pattern' declaration.

NicePatternSyn Range Access Name [WithHiding Name] Pattern

NiceGeneralize Range Access ArgInfo TacticAttribute Name Expr

NiceUnquoteDecl Range Access IsAbstract IsInstance TerminationCheck CoverageCheck [Name] Expr

NiceUnquoteDef Range Access IsAbstract TerminationCheck CoverageCheck [Name] Expr

NiceUnquoteData Range Access IsAbstract PositivityCheck UniverseCheck Name [Name] Expr

NiceOpaque KwRange [QName] [NiceDeclaration]

type TerminationCheck = TerminationCheck Measure

type Measure = Name

Termination measure is, for now, a variable name.

type NiceConstructor = NiceTypeSignature

Only Axioms.

type NiceTypeSignature = NiceDeclaration

Only Axioms.

data Clause

One clause in a function definition. There is no guarantee that the LHS actually declares the Name. We will have to check that later.

Constructors

Clause Name Catchall ArgInfo LHS RHS WhereClause [Clause]

data MutualChecks

When processing a mutual block we collect the various checks present in the block before combining them.

Constructors

MutualChecks

Fields

mutualTermination :: [TerminationCheck]

mutualCoverage :: [CoverageCheck]

mutualPositivity :: [PositivityCheck]

data InferredMutual

In an inferred mutual block we keep accumulating nice declarations until all of the lone signatures have an attached definition. The type is therefore a bit span-like: we return an initial segment (the inferred mutual block) together with leftovers.

Constructors

InferredMutual

Fields

inferredChecks :: MutualChecks

inferredBlock :: [NiceDeclaration]

inferredLeftovers :: [NiceDeclaration]

extendInferredBlock :: NiceDeclaration -> InferredMutual -> InferredMutual

type InterleavedMutual = Map Name InterleavedDecl

In an `interleaved mutual' block we collect the data signatures, function signatures, as well as their associated constructors and function clauses respectively. Each signature is given a position in the block (from 0 onwards) and each set of constructor / clauses is given a *distinct* one. This allows for interleaved forward declarations similar to what one gets in a new-style mutual block.

data InterleavedDecl

Constructors

InterleavedData

Fields

interleavedDeclNum :: DeclNum

Internal number of the data signature.

interleavedDeclSig :: NiceDeclaration

The data signature.

interleavedDataCons :: Maybe (DeclNum, List1 [NiceConstructor])

Constructors associated to the data signature.

InterleavedFun

Fields

interleavedDeclNum :: DeclNum

Internal number of the data signature.

interleavedDeclSig :: NiceDeclaration

The data signature.

interleavedFunClauses :: Maybe (DeclNum, List1 (List1 Declaration, List1 Clause))

Function clauses associated to the function signature.

data InterleavedState

Extra state for the Nice monad to process interleaved mutual blocks.

Constructors

ISt

Fields

interleavedMutual :: InterleavedMutual

interleavedMutualChecks :: MutualChecks

interleavedCurrentDeclNum :: DeclNum

type DeclNum = Int

Numbering declarations in an interleaved mutual block.

isInterleavedFun :: InterleavedDecl -> Maybe ()

isInterleavedData :: InterleavedDecl -> Maybe ()

interleavedDecl :: Name -> InterleavedDecl -> [(DeclNum, NiceDeclaration)]

data KindOfBlock

Several declarations expect only type signatures as sub-declarations. These are:

Constructors

PostulateBlock
postulate.

PrimitiveBlock
primitive.

FieldBlock
field. Ensured by parser.

DataBlock
data ... where. Here we got a bad error message for Agda-2.5 (Issue 1698).

ConstructorBlock
constructor, in interleaved mutual.

declName :: NiceDeclaration -> String

data InMutual

Constructors

InMutual
we are nicifying a mutual block

NotInMutual
we are nicifying decls not in a mutual block

data DataRecOrFun

The kind of the forward declaration.

Constructors

DataName
Name of a data type

Fields

_kindPosCheck :: PositivityCheck

_kindUniCheck :: UniverseCheck

RecName
Name of a record type

Fields

_kindPosCheck :: PositivityCheck

_kindUniCheck :: UniverseCheck

FunName ArgInfo TerminationCheck CoverageCheck
Name of a function.

isFunName :: DataRecOrFun -> Bool

sameKind :: DataRecOrFun -> DataRecOrFun -> Bool

terminationCheck :: DataRecOrFun -> TerminationCheck

coverageCheck :: DataRecOrFun -> CoverageCheck

positivityCheck :: DataRecOrFun -> PositivityCheck

mutualChecks :: DataRecOrFun -> MutualChecks

universeCheck :: DataRecOrFun -> UniverseCheck