Agda.Syntax.Concrete.Fixity

Description

Collecting fixity declarations (and polarity pragmas) for concrete declarations.

Documentation

type Fixities = Map Name Fixity'

type Polarities = Map Name PragmaPolarities

class Monad m => MonadFixityError (m :: Type -> Type) where

Methods

throwMultipleFixityDecls :: List1 (Name, Pair Fixity') -> m a

throwMultiplePolarityPragmas :: List1 Name -> m a

warnUnknownNamesInFixityDecl :: Set1 Name -> m ()

warnUnknownNamesInPolarityPragmas :: Set1 Name -> m ()

warnUnknownFixityInMixfixDecl :: Set1 Name -> m ()

warnPolarityPragmasButNotPostulates :: Set1 Name -> m ()

warnEmptyPolarityPragma :: Range -> m ()

data DoWarn

Constructors

NoWarn

DoWarn

fixitiesAndPolarities :: MonadFixityError m => DoWarn -> [Declaration] -> m (Fixities, Polarities)

Get the fixities and polarity pragmas from the current block. Doesn't go inside modules and where blocks. The reason for this is that these declarations have to appear at the same level (or possibly outside an abstract or mutual block) as their target declaration.