Agda.Syntax.Concrete.Glyph

Description

Choice of Unicode or ASCII glyphs.

Documentation

data UnicodeOrAscii

We want to know whether we are allowed to insert unicode characters or not.

Constructors

UnicodeOk
true: Unicode characters are allowed.

AsciiOnly
'false: Stick to ASCII.

unsafeSetUnicodeOrAscii :: UnicodeOrAscii -> IO ()

specialCharactersForGlyphs :: UnicodeOrAscii -> SpecialCharacters

Return the glyph set based on a given (unicode or ascii) glyph mode

braces' :: Doc -> Doc

dbraces :: Doc -> Doc

forallQ :: Doc

leftIdiomBrkt :: Doc

rightIdiomBrkt :: Doc

emptyIdiomBrkt :: Doc

arrow :: Doc

lambda :: Doc

fromArrow :: Doc

data SpecialCharacters

Picking the appropriate set of special characters depending on whether we are allowed to use unicode or have to limit ourselves to ascii.

Constructors

SpecialCharacters

Fields

_dbraces :: Doc -> Doc

_lambda :: Doc

_arrow :: Doc

_forallQ :: Doc

_leftIdiomBrkt :: Doc

_rightIdiomBrkt :: Doc

_emptyIdiomBrkt :: Doc

_fromArrow :: Doc