Agda.Syntax.Concrete.Name

Description

Names in the concrete syntax are just strings (or lists of strings for qualified names).

Documentation

data Name

A name is a non-empty list of alternating Ids and Holes. A normal name is represented by a singleton list, and operators are represented by a list with Holes where the arguments should go. For instance: [Hole,Id "+",Hole] is infix addition.

Equality and ordering on Names are defined to ignore range so same names in different locations are equal.

Constructors

Name
A (mixfix) identifier.

Fields

nameRange :: Range

nameInScope :: NameInScope

nameNameParts :: NameParts

NoName
_.

Fields

nameRange :: Range

nameId :: NameId

type NameParts = List1 NamePart

isOpenMixfix :: Name -> Bool

An open mixfix identifier is either prefix, infix, or suffix. That is to say: at least one of its extremities is a Hole

data NamePart

Mixfix identifiers are composed of words and holes, e.g. _+_ or if_then_else_ or [_/_].

Constructors

Hole
_ part.

Id RawName
Identifier part.

data QName

QName is a list of namespaces and the name of the constant. For the moment assumes namespaces are just Names and not explicitly applied modules. Also assumes namespaces are generative by just using derived equality. We will have to define an equality instance to non-generative namespaces (as well as having some sort of lookup table for namespace names).

Constructors

Qual Name QName
A.rest.

QName Name
x.

Constructing simple Names.

simpleName :: RawName -> Name

Create an ordinary InScope name.

simpleBinaryOperator :: RawName -> Name

Create a binary operator name in scope.

simpleHole :: Name

Create an ordinary InScope name containing a single Hole.

Operations on Name and NamePart

lensNameParts :: Lens' Name NameParts

Don't use on 'NoName{}'.

nameToRawName :: Name -> RawName

nameParts :: Name -> NameParts

nameStringParts :: Name -> [RawName]

stringNameParts :: String -> NameParts

Parse a string to parts of a concrete name.

Note: stringNameParts "_" == [Id "_"] == nameParts NoName{}

class NumHoles a where

Number of holes in a Name (i.e., arity of a mixfix-operator).

Methods

numHoles :: a -> Int

isOperator :: Name -> Bool

Is the name an operator? Needs at least 2 NameParts.

isHole :: NamePart -> Bool

isPrefix :: Name -> Bool

isPostfix :: Name -> Bool

isInfix :: Name -> Bool

isNonfix :: Name -> Bool

Keeping track of which names are (not) in scope

data NameInScope

Constructors

InScope

NotInScope

class LensInScope a where

Minimal complete definition

lensInScope

Methods

lensInScope :: Lens' a NameInScope

isInScope :: a -> NameInScope

mapInScope :: (NameInScope -> NameInScope) -> a -> a

setInScope :: a -> a

setNotInScope :: a -> a

Generating fresh names

data FreshNameMode

Method by which to generate fresh unshadowed names.

Constructors

UnicodeSubscript
Append an integer Unicode subscript: x, x₁, x₂, …

AsciiCounter
Append an integer ASCII counter: x, x1, x2, …

nextRawName :: FreshNameMode -> RawName -> RawName

nextName :: FreshNameMode -> Name -> Name

Get the next version of the concrete name. For instance, nextName "x" = "x₁". The name must not be a NoName.

lastIdPart :: Lens' NameParts RawName

Zoom on the last non-hole in a name.

firstNonTakenName :: FreshNameMode -> (Name -> Bool) -> Name -> Name

Get the first version of the concrete name that does not satisfy the given predicate.

nameSuffix :: Lens' Name (Maybe Suffix)

Lens for accessing and modifying the suffix of a name. The suffix of a NoName is always Nothing, and should not be changed.

nameSuffixView :: Name -> (Maybe Suffix, Name)

Split a name into a base name plus a suffix.

setNameSuffix :: Maybe Suffix -> Name -> Name

Replaces the suffix of a name. Unless the suffix is Nothing, the name should not be NoName.

nameRoot :: Name -> RawName

Get a raw version of the name with all suffixes removed. For instance, nameRoot "x₁₂₃" = "x".

sameRoot :: Name -> Name -> Bool

Operations on qualified names

lensQNameName :: Lens' QName Name

Lens for the unqualified part of a QName

qualify :: QName -> Name -> QName

qualify A.B x == A.B.x

unqualify :: QName -> Name

unqualify A.B.x == x

The range is preserved.

qnameParts :: QName -> List1 Name

qnameParts A.B.x = [A, B, x]

isQualified :: QName -> Bool

Is the name (un)qualified?

isUnqualified :: QName -> Maybe Name

No name stuff

noName_ :: Name

noName_ = noName noRange

noName :: Range -> Name

class IsNoName a where

Check whether a name is the empty name "_".

Minimal complete definition

Nothing

Methods

isNoName :: a -> Bool

default isNoName :: forall (t :: Type -> Type) b. (Foldable t, IsNoName b, t b ~ a) => a -> Bool

Showing names

Printing names

Range instances

NFData instances