Agda.Syntax.Concrete.Operators.Parser

Documentation

placeholder :: PositionInName -> Parser e (MaybePlaceholder e)

maybePlaceholder :: Maybe PositionInName -> Parser e e -> Parser e (MaybePlaceholder e)

satNoPlaceholder :: (e -> Maybe a) -> Parser e a

data ExprView e

Constructors

LocalV QName

WildV e

OtherV e

AppV e (NamedArg e)

OpAppV QName (Set1 Name) (OpAppArgs' e)
The QName is possibly ambiguous, but it must correspond to one of the names in the set.

HiddenArgV (Named_ e)

InstanceArgV (Named_ e)

LamV (List1 LamBinding) e

ParenV e

class HasRange e => IsExpr e where

Methods

exprView :: e -> ExprView e

unExprView :: ExprView e -> e

patternView :: e -> Maybe Pattern

data ParseSections

Should sections be parsed?

Constructors

ParseSections

DoNotParseSections

parse :: IsExpr e => (ParseSections, Parser e a) -> [e] -> [a]

Runs a parser. If sections should be parsed, then identifiers with at least two name parts are split up into multiple tokens, using PositionInName to record the tokens' original positions within their respective identifiers.

Parser combinators

partP :: IsExpr e => [Name] -> RawName -> Parser e Range

Parse a specific identifier as a NamePart

atLeastTwoParts :: IsExpr e => Parser e Name

Parses a split-up, unqualified name consisting of at least two name parts.

The parser does not check that underscores and other name parts alternate. The range of the resulting name is the range of the first name part that is not an underscore.

patternBinder :: IsExpr e => Parser e Binder

Parses a potentially pattern-matching binder

type family OperatorType (k :: NotationKind) e

Used to define the return type of opP.

data NK (k :: NotationKind) where

A singleton type for NotationKind (except for the constructor NoNotation).

Constructors

In :: NK 'InfixNotation

Pre :: NK 'PrefixNotation

Post :: NK 'PostfixNotation

Non :: NK 'NonfixNotation

opP :: forall e (k :: NotationKind). IsExpr e => ParseSections -> Parser e e -> NewNotation -> NK k -> Parser e (OperatorType k e)

Parse the "operator part" of the given notation.

Normal holes (but not binders) at the beginning and end are ignored.

If the notation does not contain any binders, then a section notation is allowed.

argsP :: IsExpr e => Parser e e -> Parser e [NamedArg e]

appP :: IsExpr e => Parser e e -> Parser e [NamedArg e] -> Parser e e

atomP :: IsExpr e => (QName -> Bool) -> Parser e e