Agda.Syntax.Concrete.Pattern

Description

Tools for patterns in concrete syntax.

Documentation

class IsEllipsis a where

Check for ellipsis ....

Methods

isEllipsis :: a -> Bool

class HasEllipsis a where

Has the lhs an occurrence of the ellipsis ...?

Methods

hasEllipsis :: a -> Bool

class IsWithP p where

Check for with-pattern | p.

Minimal complete definition

Nothing

Methods

isWithP :: p -> Maybe p

default isWithP :: forall q (f :: Type -> Type). (IsWithP q, Decoration f, f q ~ p) => p -> Maybe p

LHS manipulation (see also 'Pattern')

data LHSPatternView

The next patterns are ...

(This view discards PatInfo.)

Constructors

LHSAppP (List1 (NamedArg Pattern))
Application patterns (non-empty list).

LHSWithP (List1 Pattern)
With patterns (non-empty list). These patterns are not prefixed with WithP.

lhsPatternView :: [NamedArg Pattern] -> Maybe (LHSPatternView, [NamedArg Pattern])

Construct the LHSPatternView of the given list (if not empty).

Return the view and the remaining patterns.

lhsCoreApp :: LHSCore -> List1 (NamedArg Pattern) -> LHSCore

Add applicative patterns (non-projection / non-with patterns) to the right.

lhsCoreWith :: LHSCore -> List1 Pattern -> LHSCore

Add with-patterns to the right.

lhsCoreAddSpine :: LHSCore -> [NamedArg Pattern] -> LHSCore

Append patterns to LHSCore, separating with patterns from the rest.

mapLhsOriginalPattern :: (Pattern -> Pattern) -> LHS -> LHS

Modify the Pattern component in LHS.

mapLhsOriginalPatternM :: (Functor m, Applicative m) => (Pattern -> m Pattern) -> LHS -> m LHS

Effectfully modify the Pattern component in LHS.

hasCopatterns :: LHSCore -> Bool

Does the LHS contain projection patterns?

Generic fold

class CPatternLike p where

Generic pattern traversal.

See APatternLike.

Minimal complete definition

Nothing

Methods

foldrCPattern

Arguments

:: Monoid m

=> (Pattern -> m -> m)
Combine a pattern and the value computed from its subpatterns.

-> p

-> m

Fold pattern.

default foldrCPattern :: forall m (f :: Type -> Type) q. (Monoid m, Foldable f, CPatternLike q, f q ~ p) => (Pattern -> m -> m) -> p -> m

traverseCPatternA

Arguments

:: (Applicative m, Functor m)

=> (Pattern -> m Pattern -> m Pattern)
Combine a pattern and the its recursively computed version.

-> p

-> m p

Traverse pattern with option of post-traversal modification.

default traverseCPatternA :: forall (f :: Type -> Type) q m. (Traversable f, CPatternLike q, f q ~ p, Applicative m, Functor m) => (Pattern -> m Pattern -> m Pattern) -> p -> m p

traverseCPatternM

Arguments

:: Monad m

=> (Pattern -> m Pattern)
pre: Modification before recursion.

-> (Pattern -> m Pattern)
post: Modification after recursion.

-> p

-> m p

Traverse pattern.

default traverseCPatternM :: forall (f :: Type -> Type) q m. (Traversable f, CPatternLike q, f q ~ p, Monad m) => (Pattern -> m Pattern) -> (Pattern -> m Pattern) -> p -> m p

foldCPattern :: (CPatternLike p, Monoid m) => (Pattern -> m) -> p -> m

Compute a value from each subpattern and collect all values in a monoid.

preTraverseCPatternM

Arguments

:: (CPatternLike p, Monad m)

=> (Pattern -> m Pattern)
pre: Modification before recursion.

-> p

-> m p

Traverse pattern(s) with a modification before the recursive descent.

postTraverseCPatternM

Arguments

:: (CPatternLike p, Monad m)

=> (Pattern -> m Pattern)
post: Modification after recursion.

-> p

-> m p

Traverse pattern(s) with a modification after the recursive descent.

mapCPattern :: CPatternLike p => (Pattern -> Pattern) -> p -> p

Map pattern(s) with a modification after the recursive descent.

Specific folds.

patternQNames :: CPatternLike p => p -> [QName]

Get all the identifiers in a pattern in left-to-right order.

Implemented using difference lists.

patternNames :: Pattern -> [Name]

Get all the identifiers in a pattern in left-to-right order.

hasWithPatterns :: CPatternLike p => p -> Bool

Does the pattern contain a with-pattern? (Shortcutting.)

isWithPattern :: Pattern -> Bool

Is WithP?

numberOfWithPatterns :: CPatternLike p => p -> Int

Count the number of with-subpatterns in a pattern?

hasEllipsis' :: CPatternLike p => p -> AffineHole Pattern p

Compute the context in which the ellipsis occurs, if at all. If there are several occurrences, this is an error. This only counts ellipsis that haven't already been expanded.

reintroduceEllipsis :: ExpandedEllipsis -> Pattern -> Pattern

splitEllipsis :: IsWithP p => Int -> [p] -> ([p], [p])

Helpers for pattern and lhs parsing

patternAppView :: Pattern -> List1 (NamedArg Pattern)

View a pattern p as a list p0 .. pn where p0 is the identifier (in most cases a constructor).

Pattern needs to be parsed already (operators resolved).