Agda.Syntax.Fixity

Description

Definitions for fixity, precedence levels, and declared syntax.

Documentation

data ThingWithFixity x

Decorating something with Fixity'.

Constructors

ThingWithFixity x Fixity'

data ParenPreference

Do we prefer parens around arguments like λ x → x or not? See lamBrackets.

Constructors

PreferParen

PreferParenless

preferParen :: ParenPreference -> Bool

preferParenless :: ParenPreference -> Bool

Precendence

data Precedence

Precedence is associated with a context.

Constructors

TopCtx

FunctionSpaceDomainCtx

LeftOperandCtx Fixity

RightOperandCtx Fixity ParenPreference

FunctionCtx

ArgumentCtx ParenPreference

InsideOperandCtx

WithFunCtx

WithArgCtx

DotPatternCtx

type PrecedenceStack = [Precedence]

When printing we keep track of a stack of precedences in order to be able to decide whether it's safe to leave out parens around lambdas. An empty stack is equivalent to TopCtx. Invariant: `notElem TopCtx`.

pushPrecedence :: Precedence -> PrecedenceStack -> PrecedenceStack

headPrecedence :: PrecedenceStack -> Precedence

argumentCtx_ :: Precedence

Argument context preferring parens.

opBrackets :: Fixity -> PrecedenceStack -> Bool

Do we need to bracket an operator application of the given fixity in a context with the given precedence.

opBrackets' :: Bool -> Fixity -> PrecedenceStack -> Bool

Do we need to bracket an operator application of the given fixity in a context with the given precedence.

lamBrackets :: PrecedenceStack -> Bool

Does a lambda-like thing (lambda, let or pi) need brackets in the given context? A peculiar thing with lambdas is that they don't need brackets in certain right operand contexts. To decide we need to look at the stack of precedences and not just the current precedence. Example: m₁ >>= (λ x → x) >>= m₂ (for _>>=_ left associative).

appBrackets :: PrecedenceStack -> Bool

Does a function application need brackets?

appBrackets' :: Bool -> PrecedenceStack -> Bool

Does a function application need brackets?

withAppBrackets :: PrecedenceStack -> Bool

Does a with application need brackets?

piBrackets :: PrecedenceStack -> Bool

Does a function space need brackets?

roundFixBrackets :: PrecedenceStack -> Bool