Agda.Syntax.Internal

Documentation

class TermSize a where

The size of a term is roughly the number of nodes in its syntax tree. This number need not be precise for logical correctness of Agda, it is only used for reporting (and maybe decisions regarding performance).

Not counting towards the term size are:

sort and color annotations,

projections.

Minimal complete definition

tsize

Methods

termSize :: a -> Int

tsize :: a -> Sum Int

data Suggestion

Constructors

forall a.Suggest a => Suggestion a

class Suggest a where

Suggest a name if available (i.e. name is not "_")

Methods

suggestName :: a -> Maybe String

class SgTel a where

Constructing a singleton telescope.

Methods

sgTel :: a -> Telescope

class TelToArgs a where

Drop the types from a telescope.

Methods

telToArgs :: a -> [Arg ArgName]

type ListTel = ListTel' ArgName

type ListTel' a = [Dom (a, Type)]

Telescope as list.

type DummyTermKind = String

data IntervalView

Constructors

IZero

IOne

IMin (Arg Term) (Arg Term)

IMax (Arg Term) (Arg Term)

INeg (Arg Term)

OTerm Term

data PathView

View type as path type.

Constructors

PathType

Fields

pathSort :: Sort

Sort of this type.

pathName :: QName

Builtin PATH.

pathLevel :: Arg Term

Hidden

pathType :: Arg Term

Hidden

pathLhs :: Arg Term

NotHidden

pathRhs :: Arg Term

NotHidden

OType Type
reduced

data EqualityView

View type as equality type.

Constructors

EqualityType

Fields

eqtSort :: Sort

Sort of this type.

eqtName :: QName

Builtin EQUALITY.

eqtParams :: [Arg Term]

Hidden. Empty or Level.

eqtType :: Arg Term

Hidden

eqtLhs :: Arg Term

NotHidden

eqtRhs :: Arg Term

NotHidden

OtherType Type
reduced

IdiomType Type
reduced

type PatternSubstitution = Substitution' DeBruijnPattern

type Substitution = Substitution' Term

data Substitution' a

Substitutions.

Constructors

IdS
Identity substitution. Γ ⊢ IdS : Γ

EmptyS Impossible
Empty substitution, lifts from the empty context. First argument is IMPOSSIBLE. Apply this to closed terms you want to use in a non-empty context. Γ ⊢ EmptyS : ()

a :# (Substitution' a) infixr 4
Substitution extension, `cons'. Γ ⊢ u : Aρ Γ ⊢ ρ : Δ ---------------------- Γ ⊢ u :# ρ : Δ, A

Strengthen Impossible (Substitution' a)
Strengthening substitution. First argument is IMPOSSIBLE. Apply this to a term which does not contain variable 0 to lower all de Bruijn indices by one. Γ ⊢ ρ : Δ --------------------------- Γ ⊢ Strengthen ρ : Δ, A

Wk !Int (Substitution' a)
Weakening substitution, lifts to an extended context. Γ ⊢ ρ : Δ ------------------- Γ, Ψ ⊢ Wk |Ψ| ρ : Δ

Lift !Int (Substitution' a)
Lifting substitution. Use this to go under a binder. Lift 1 ρ == var 0 :# Wk 1 ρ. Γ ⊢ ρ : Δ ------------------------- Γ, Ψρ ⊢ Lift |Ψ| ρ : Δ, Ψ

class PatternVars a where

Extract pattern variables in left-to-right order. A DotP is also treated as variable (see docu for Clause).

Associated Types

type PatternVarOut a

Methods

patternVars :: a -> [Arg (Either (PatternVarOut a) Term)]

data ConPatternInfo

The ConPatternInfo states whether the constructor belongs to a record type (True) or data type (False). In the former case, the PatOrigin of the conPInfo says whether the record pattern orginates from the expansion of an implicit pattern. The Type is the type of the whole record pattern. The scope used for the type is given by any outer scope plus the clause's telescope (clauseTel).

Constructors

ConPatternInfo

Fields

conPInfo :: PatternInfo

Information on the origin of the pattern.

conPRecord :: Bool

False if data constructor. True if record constructor.

conPFallThrough :: Bool

Should the match block on non-canonical terms or can it proceed to the catch-all clause?

conPType :: Maybe (Arg Type)

The type of the whole constructor pattern. Should be present (Just) if constructor pattern is is generated ordinarily by type-checking. Could be absent (Nothing) if pattern comes from some plugin (like Agsy). Needed e.g. for with-clause stripping.

conPLazy :: Bool

Lazy patterns are generated by the forcing translation in the unifier (unifyStep) and are dropped by the clause compiler (TODO: not yet) (compileClauses) when the variables they bind are unused. The GHC backend compiles lazy matches to lazy patterns in Haskell (TODO: not yet).

type DeBruijnPattern = Pattern' DBPatVar

data DBPatVar

Type used when numbering pattern variables.

Constructors

DBPatVar

Fields

dbPatVarName :: PatVarName

dbPatVarIndex :: Int

type Pattern

Arguments

= Pattern' PatVarName
The PatVarName is a name suggestion.

data Pattern' x

Patterns are variables, constructors, or wildcards. QName is used in ConP rather than Name since a constructor might come from a particular namespace. This also meshes well with the fact that values (i.e. the arguments we are matching with) use QName.

Constructors

VarP PatternInfo x
x

DotP PatternInfo Term
.t

ConP ConHead ConPatternInfo [NamedArg (Pattern' x)]
c ps The subpatterns do not contain any projection copatterns.

LitP PatternInfo Literal
E.g. 5, "hello".

ProjP ProjOrigin QName
Projection copattern. Can only appear by itself.

IApplyP PatternInfo Term Term x
Path elimination pattern, like VarP but keeps track of endpoints.

DefP PatternInfo QName [NamedArg (Pattern' x)]
Used for HITs, the QName should be the one from primHComp.

data PatOrigin

Origin of the pattern: what did the user write in this position?

Constructors

PatOSystem
Pattern inserted by the system

PatOSplit
Pattern generated by case split

PatOVar Name
User wrote a variable pattern

PatODot
User wrote a dot pattern

PatOWild
User wrote a wildcard pattern

PatOCon
User wrote a constructor pattern

PatORec
User wrote a record pattern

PatOLit
User wrote a literal pattern

PatOAbsurd
User wrote an absurd pattern

data PatternInfo

Constructors

PatternInfo

Fields

patOrigin :: PatOrigin

patAsNames :: [Name]

type PatVarName = ArgName

Pattern variables.

data Clause

A clause is a list of patterns and the clause body.

The telescope contains the types of the pattern variables and the de Bruijn indices say how to get from the order the variables occur in the patterns to the order they occur in the telescope. The body binds the variables in the order they appear in the telescope.

clauseTel ~ permute clausePerm (patternVars namedClausePats)

Terms in dot patterns are valid in the clause telescope.

For the purpose of the permutation and the body dot patterns count as variables. TODO: Change this!

Constructors

Clause

Fields

clauseLHSRange :: Range

clauseFullRange :: Range

clauseTel :: Telescope

Δ: The types of the pattern variables in dependency order.

namedClausePats :: NAPs

Δ ⊢ ps. The de Bruijn indices refer to Δ.

clauseBody :: Maybe Term

Just v with Δ ⊢ v for a regular clause, or Nothing for an absurd one.

clauseType :: Maybe (Arg Type)

Δ ⊢ t. The type of the rhs under clauseTel. Used, e.g., by TermCheck. Can be Irrelevant if we encountered an irrelevant projection pattern on the lhs.

clauseCatchall :: Bool

Clause has been labelled as CATCHALL.

clauseExact :: Maybe Bool

Pattern matching of this clause is exact, no catch-all case. Computed by the coverage checker. Nothing means coverage checker has not run yet (clause may be inexact). Just False means clause is not exact. Just True means clause is exact.

clauseRecursive :: Maybe Bool

clauseBody contains recursive calls; computed by termination checker. Nothing means that termination checker has not run yet, or that clauseBody contains meta-variables; these could be filled with recursive calls later! Just False means definitely no recursive call. Just True means definitely a recursive call.

clauseUnreachable :: Maybe Bool

Clause has been labelled as unreachable by the coverage checker. Nothing means coverage checker has not run yet (clause may be unreachable). Just False means clause is not unreachable. Just True means clause is unreachable.

clauseEllipsis :: ExpandedEllipsis

Was this clause created by expansion of an ellipsis?

type NAPs = [NamedArg DeBruijnPattern]

Named pattern arguments.

type Blocked_ = Blocked ()

'Blocked a without the a.

type NotBlocked = NotBlocked' Term

type Blocked = Blocked' Term

newtype BraveTerm

Newtypes for terms that produce a dummy, rather than crash, when applied to incompatible eliminations.

Constructors

BraveTerm

Fields

unBrave :: Term

type LevelAtom = Term

type PlusLevel = PlusLevel' Term

data PlusLevel' t

Constructors

Plus Integer t

type Level = Level' Term

data Level' t

A level is a maximum expression of a closed level and 0..n PlusLevel expressions each of which is an atom plus a number.

Constructors

Max Integer [PlusLevel' t]

type Sort = Sort' Term

data Sort' t

Sorts.

Constructors

Type (Level' t)
Set ℓ.

Prop (Level' t)
Prop ℓ.

Inf IsFibrant Integer
Setωᵢ.

SSet (Level' t)
SSet ℓ.

SizeUniv
SizeUniv, a sort inhabited by type Size.

LockUniv
LockUniv, a sort for locks.

IntervalUniv
IntervalUniv, a sort inhabited by the cubical interval.

PiSort (Dom' t t) (Sort' t) (Abs (Sort' t))
Sort of the pi type.

FunSort (Sort' t) (Sort' t)
Sort of a (non-dependent) function type.

UnivSort (Sort' t)
Sort of another sort.

MetaS !MetaId [Elim' t]

DefS QName [Elim' t]
A postulated sort.

DummyS String
A (part of a) term or type which is only used for internal purposes. Replaces the abuse of Prop for a dummy sort. The String typically describes the location where we create this dummy, but can contain other information as well.

data IsFibrant

Constructors

IsFibrant

IsStrict

type Telescope = Tele (Dom Type)

data Tele a

Sequence of types. An argument of the first type is bound in later types and so on.

Constructors

EmptyTel

ExtendTel a (Abs (Tele a))
Abs is never NoAbs.

class LensSort a where

Minimal complete definition

lensSort

Methods

lensSort :: Lens' Sort a

getSort :: a -> Sort

type Type = Type' Term

type Type' a = Type'' Term a

data Type'' t a

Types are terms with a sort annotation.

Constructors

El

Fields

_getSort :: Sort' t

unEl :: a

data Abs a

Binder.

Abs: The bound variable might appear in the body. NoAbs is pseudo-binder, it does not introduce a fresh variable, similar to the const of Haskell.

Constructors

Abs
The body has (at least) one free variable. Danger: unAbs doesn't shift variables properly

Fields

absName :: ArgName

unAbs :: a

NoAbs

Fields

absName :: ArgName

unAbs :: a

type Elims

Arguments

= [Elim]
eliminations ordered left-to-right.

type Elim = Elim' Term

type ConInfo = ConOrigin

data Term

Raw values.

Def is used for both defined and undefined constants. Assume there is a type declaration and a definition for every constant, even if the definition is an empty list of clauses.

Constructors

Var !Int Elims
x es neutral

Lam ArgInfo (Abs Term)
Terms are beta normal. Relevance is ignored

Lit Literal

Def QName Elims
f es, possibly a delta/iota-redex

Con ConHead ConInfo Elims
c es or record { fs = es } es allows only Apply and IApply eliminations, and IApply only for data constructors.

Pi (Dom Type) (Abs Type)
dependent or non-dependent function space

Sort Sort

Level Level

MetaV !MetaId Elims

DontCare Term
Irrelevant stuff in relevant position, but created in an irrelevant context. Basically, an internal version of the irrelevance axiom .irrAx : .A -> A.

Dummy String Elims
A (part of a) term or type which is only used for internal purposes. Replaces the Sort Prop hack. The String typically describes the location where we create this dummy, but can contain other information as well. The second field accumulates eliminations in case we apply a dummy term to more of them.

class LensConName a where

Minimal complete definition

getConName

Methods

getConName :: a -> QName

setConName :: QName -> a -> a

mapConName :: (QName -> QName) -> a -> a

data ConHead

Store the names of the record fields in the constructor. This allows reduction of projection redexes outside of TCM. For instance, during substitution and application.

Constructors

ConHead

Fields

conName :: QName

The name of the constructor.

conDataRecord :: DataOrRecord

Data or record constructor?

conInductive :: Induction

Record constructors can be coinductive.

conFields :: [Arg QName]

The name of the record fields. Arg is stored since the info in the constructor args might not be accurate because of subtyping (issue #2170).

data DataOrRecord

Constructors

IsData

IsRecord PatternOrCopattern

type NamedArgs = [NamedArg Term]

type Args = [Arg Term]

Type of argument lists.

type Dom = Dom' Term

data Dom' t e

Similar to Arg, but we need to distinguish an irrelevance annotation in a function domain (the domain itself is not irrelevant!) from an irrelevant argument.

Dom is used in Pi of internal syntax, in Context and Telescope. Arg is used for actual arguments (Var, Con, Def etc.) and in Abstract syntax and other situations.

cubical

When domFinite = True for the domain of a Pi type, the elements should be compared by tabulating the domain type. Only supported in case the domain type is primIsOne, to obtain the correct equality for partial elements.

Constructors

Dom

Fields

domInfo :: ArgInfo

domFinite :: !Bool

domName :: Maybe NamedName

e.g. x in {x = y : A} -> B.

domTactic :: Maybe t

"@tactic e".

unDom :: e

pattern ClosedLevel :: Integer -> Level

Constant level n

argFromDom :: Dom' t a -> Arg a

namedArgFromDom :: Dom' t a -> NamedArg a

domFromArg :: Arg a -> Dom a

domFromNamedArg :: NamedArg a -> Dom a

defaultDom :: a -> Dom a

defaultArgDom :: ArgInfo -> a -> Dom a

defaultNamedArgDom :: ArgInfo -> String -> a -> Dom a

clausePats :: Clause -> [Arg DeBruijnPattern]

patVarNameToString :: PatVarName -> String

nameToPatVarName :: Name -> PatVarName

defaultPatternInfo :: PatternInfo

varP :: a -> Pattern' a

dotP :: Term -> Pattern' a

litP :: Literal -> Pattern' a

namedVarP :: PatVarName -> Named_ Pattern

namedDBVarP :: Int -> PatVarName -> Named_ DeBruijnPattern

absurdP :: Int -> DeBruijnPattern

Make an absurd pattern with the given de Bruijn index.

noConPatternInfo :: ConPatternInfo

toConPatternInfo :: ConInfo -> ConPatternInfo

Build partial ConPatternInfo from ConInfo

fromConPatternInfo :: ConPatternInfo -> ConInfo

Build ConInfo from ConPatternInfo.

patternInfo :: Pattern' x -> Maybe PatternInfo

Retrieve the PatternInfo from a pattern

patternOrigin :: Pattern' x -> Maybe PatOrigin

Retrieve the origin of a pattern

properlyMatching :: Pattern' a -> Bool

Does the pattern perform a match that could fail?

properlyMatching'

Arguments

:: Bool
Should absurd patterns count as proper match?

-> Bool
Should projection patterns count as proper match?

-> Pattern' a
The pattern.

-> Bool

isEqualityType :: EqualityView -> Bool

isPathType :: PathView -> Bool

isIOne :: IntervalView -> Bool

absurdBody :: Abs Term

Absurd lambdas are internally represented as identity with variable name "()".

isAbsurdBody :: Abs Term -> Bool

absurdPatternName :: PatVarName

isAbsurdPatternName :: PatVarName -> Bool

var :: Nat -> Term

An unapplied variable.

dontCare :: Term -> Term

Add DontCare is it is not already a DontCare.

dummyLocName :: CallStack -> String

Construct a string representing the call-site that created the dummy thing.

dummyTermWith :: DummyTermKind -> CallStack -> Term

Aux: A dummy term to constitute a dummy termlevelsort/type.

dummyLevel :: CallStack -> Level

A dummy level to constitute a level/sort created at location. Note: use macro DUMMY_LEVEL !

dummyTerm :: CallStack -> Term

A dummy term created at location. Note: use macro DUMMY_TERM !

__DUMMY_TERM__ :: HasCallStack => Term

__DUMMY_LEVEL__ :: HasCallStack => Level

dummySort :: CallStack -> Sort

A dummy sort created at location. Note: use macro DUMMY_SORT !

__DUMMY_SORT__ :: HasCallStack => Sort

dummyType :: CallStack -> Type

A dummy type created at location. Note: use macro DUMMY_TYPE !

__DUMMY_TYPE__ :: HasCallStack => Type

dummyDom :: CallStack -> Dom Type

Context entries without a type have this dummy type. Note: use macro DUMMY_DOM !

__DUMMY_DOM__ :: HasCallStack => Dom Type

atomicLevel :: t -> Level' t

varSort :: Int -> Sort

tmSort :: Term -> Sort

tmSSort :: Term -> Sort

levelPlus :: Integer -> Level -> Level

Given a constant m and level l, compute m + l

levelSuc :: Level -> Level

mkType :: Integer -> Sort

mkProp :: Integer -> Sort

mkSSet :: Integer -> Sort

isSort :: Term -> Maybe Sort

impossibleTerm :: CallStack -> Term

mapAbsNamesM :: Applicative m => (ArgName -> m ArgName) -> Tele a -> m (Tele a)

A traversal for the names in a telescope.

mapAbsNames :: (ArgName -> ArgName) -> Tele a -> Tele a

replaceEmptyName :: ArgName -> Tele a -> Tele a

telFromList' :: (a -> ArgName) -> ListTel' a -> Telescope

telFromList :: ListTel -> Telescope

Convert a list telescope to a telescope.

telToList :: Tele (Dom t) -> [Dom (ArgName, t)]

Convert a telescope to its list form.

listTel :: Lens' ListTel Telescope

Lens to edit a Telescope as a list.

stripDontCare :: Term -> Term

Removing a topmost DontCare constructor.

arity :: Type -> Nat

Doesn't do any reduction.

suggests :: [Suggestion] -> String

unSpine :: Term -> Term

Convert top-level postfix projections into prefix projections.

unSpine' :: (ProjOrigin -> Bool) -> Term -> Term

Convert Proj projection eliminations according to their ProjOrigin into Def projection applications.

hasElims :: Term -> Maybe (Elims -> Term, Elims)

A view distinguishing the neutrals Var, Def, and MetaV which can be projected.

pDom :: LensHiding a => a -> Doc -> Doc

prettyPrecLevelSucs :: Int -> Integer -> (Int -> Doc) -> Doc

module Agda.Syntax.Internal.Blockers

module Agda.Syntax.Internal.Elim

module Agda.Syntax.Abstract.Name

newtype MetaId

A meta variable identifier is just a natural number.

Constructors

MetaId

Fields

metaId :: Nat

newtype ProblemId

A "problem" consists of a set of constraints and the same constraint can be part of multiple problems.

Constructors

ProblemId Nat