Agda.Syntax.Internal.Blockers

Blocked Terms

data NotBlocked' t

Even if we are not stuck on a meta during reduction we can fail to reduce a definition by pattern matching for another reason.

Constructors

StuckOn (Elim' t)
The Elim is neutral and blocks a pattern match.

Underapplied
Not enough arguments were supplied to complete the matching.

AbsurdMatch
We matched an absurd clause, results in a neutral Def.

MissingClauses QName
We ran out of clauses for QName, all considered clauses produced an actual mismatch. This can happen when try to reduce a function application but we are still missing some function clauses. See Agda.TypeChecking.Patterns.Match.

ReallyNotBlocked
Reduction was not blocked, we reached a whnf which can be anything but a stuck Def.

data Blocker

What is causing the blocking? Or in other words which metas or problems need to be solved to unblock the blocked computation/constraint.

Constructors

UnblockOnAll (Set Blocker)

UnblockOnAny (Set Blocker)

UnblockOnMeta MetaId
Unblock if meta is instantiated

UnblockOnProblem ProblemId

UnblockOnDef QName
Unblock when function is defined

alwaysUnblock :: Blocker

neverUnblock :: Blocker

unblockOnAll :: Set Blocker -> Blocker

unblockOnAny :: Set Blocker -> Blocker

unblockOnEither :: Blocker -> Blocker -> Blocker

unblockOnBoth :: Blocker -> Blocker -> Blocker

unblockOnMeta :: MetaId -> Blocker

unblockOnProblem :: ProblemId -> Blocker

unblockOnDef :: QName -> Blocker

unblockOnAllMetas :: Set MetaId -> Blocker

unblockOnAnyMeta :: Set MetaId -> Blocker

onBlockingMetasM :: Monad m => (MetaId -> m Blocker) -> Blocker -> m Blocker

allBlockingMetas :: Blocker -> Set MetaId

allBlockingProblems :: Blocker -> Set ProblemId

allBlockingDefs :: Blocker -> Set QName

data Blocked' t a

Something where a meta variable may block reduction. Notably a top-level meta is considered blocking. This did not use to be the case (pre Aug 2020).

Constructors

Blocked

Fields

theBlocker :: Blocker

ignoreBlocking :: a

NotBlocked

Fields

blockingStatus :: NotBlocked' t

ignoreBlocking :: a

stuckOn :: Elim' t -> NotBlocked' t -> NotBlocked' t

When trying to reduce f es, on match failed on one elimination e ∈ es that came with info r :: NotBlocked. stuckOn e r produces the new NotBlocked info.

MissingClauses must be propagated, as this is blockage that can be lifted in the future (as more clauses are added).

StuckOn e0 is also propagated, since it provides more precise information as StuckOn e (as e0 is the original reason why reduction got stuck and usually a subterm of e). An information like StuckOn (Apply (Arg info (Var i []))) (stuck on a variable) could be used by the lhs/coverage checker to trigger a split on that (pattern) variable.

In the remaining cases for r, we are terminally stuck due to StuckOn e. Propagating AbsurdMatch does not seem useful.

Underapplied must not be propagated, as this would mean that f es is underapplied, which is not the case (it is stuck). Note that Underapplied can only arise when projection patterns were missing to complete the original match (in e). (Missing ordinary pattern would mean the e is of function type, but we cannot match against something of function type.)

Handling blocked terms.

blockedOn :: Blocker -> a -> Blocked' t a

blocked :: MetaId -> a -> Blocked' t a

notBlocked :: a -> Blocked' t a

blocked_ :: MetaId -> Blocked' t ()

notBlocked_ :: Blocked' t ()

getBlocker :: Blocked' t a -> Blocker

Waking up logic

data WakeUp

Should a constraint wake up or not? If not, we might refine the unblocker.

Constructors

WakeUp

DontWakeUp (Maybe Blocker)

wakeUpWhen :: (constr -> Bool) -> (constr -> WakeUp) -> constr -> WakeUp

wakeUpWhen_ :: (constr -> Bool) -> constr -> WakeUp

wakeIfBlockedOnProblem :: ProblemId -> Blocker -> WakeUp

wakeIfBlockedOnMeta :: MetaId -> Blocker -> WakeUp

wakeIfBlockedOnDef :: QName -> Blocker -> WakeUp

unblockMeta :: MetaId -> Blocker -> Blocker

unblockProblem :: ProblemId -> Blocker -> Blocker

unblockDef :: QName -> Blocker -> Blocker