Agda.Syntax.Internal.Defs

Description

Extract used definitions from terms.

Documentation

getDefListNoMetas :: Term -> [QName]

Get list of definitions, throw an error if any metavariable is encountered.

class GetDefs a where

Getting the used definitions.

Note: in contrast to foldTerm getDefs also collects from sorts in terms. Thus, this is not an instance of foldTerm.

Methods

getDefs :: (Monoid m, ExpandCase m) => (MetaId -> Maybe Term) -> (QName -> m) -> a -> m

getDefs lookup emb a extracts all used definitions (functions, data/record types) from a, embedded into a monoid via emb. Instantiations of meta variables are obtained via lookup.

Typical monoid instances would be Endo [QName] or Set QName. Do not use [QName] because of obvious quadratic slowdown. Note that emb can also choose to discard a used definition by mapping to the unit of the monoid.