Agda.Syntax.Internal.MetaVars

Documentation

class AllMetas t where

Returns every meta-variable occurrence in the given type, except for those in sort annotations on types.

Minimal complete definition

Nothing

Methods

allMetas :: Monoid m => (MetaId -> m) -> t -> m

default allMetas :: (TermLike t, Monoid m) => (MetaId -> m) -> t -> m

allMetas' :: (TermLike a, Monoid m) => (MetaId -> m) -> a -> m

allMetasList :: AllMetas a => a -> [MetaId]

Returns allMetas in a list. allMetasList = allMetas (:[]).

Note: this resulting list is computed via difference lists. Thus, use this function if you actually need the whole list of metas. Otherwise, use allMetas with a suitable monoid.

noMetas :: AllMetas a => a -> Bool

True if thing contains no metas. noMetas = null . allMetasList.

firstMeta :: AllMetas a => a -> Maybe MetaId

Returns the first meta it find in the thing, if any. firstMeta == listToMaybe . allMetasList.

unblockOnAnyMetaIn :: AllMetas t => t -> Blocker

A blocker that unblocks if any of the metas in a term are solved.

unblockOnAllMetasIn :: AllMetas t => t -> Blocker

A blocker that unblocks if any of the metas in a term are solved.