Agda.Syntax.Internal.Pattern

Tools for clauses

clauseArgs :: Clause -> Args

Translate the clause patterns to terms with free variables bound by the clause telescope.

Precondition: no projection patterns.

clauseElims :: Clause -> Elims

Translate the clause patterns to an elimination spine with free variables bound by the clause telescope.

class FunArity a where

Arity of a function, computed from clauses.

Methods

funArity :: a -> Int

Tools for patterns

class LabelPatVars a b where

Label the pattern variables from left to right using one label for each variable pattern and one for each dot pattern.

Minimal complete definition

Nothing

Associated Types

type PatVarLabel b

Methods

labelPatVars :: a -> State [PatVarLabel b] b

default labelPatVars :: forall (f :: Type -> Type) a' b'. (Traversable f, LabelPatVars a' b', PatVarLabel b ~ PatVarLabel b', f a' ~ a, f b' ~ b) => a -> State [PatVarLabel b] b

unlabelPatVars :: b -> a

Intended, but unpractical due to the absence of type-level lambda, is: labelPatVars :: f (Pattern' x) -> State [i] (f (Pattern' (i,x)))

default unlabelPatVars :: forall (f :: Type -> Type) a' b'. (Traversable f, LabelPatVars a' b', f a' ~ a, f b' ~ b) => b -> a

numberPatVars :: (LabelPatVars a b, PatVarLabel b ~ Int) => Int -> Permutation -> a -> b

Augment pattern variables with their de Bruijn index.

unnumberPatVars :: LabelPatVars a b => b -> a

dbPatPerm :: [NamedArg DeBruijnPattern] -> Maybe Permutation

dbPatPerm' :: Bool -> [NamedArg DeBruijnPattern] -> Maybe Permutation

Computes the permutation from the clause telescope to the pattern variables.

Use as fromMaybe IMPOSSIBLE . dbPatPerm to crash in a controlled way if a de Bruijn index is out of scope here.

The first argument controls whether dot patterns counts as variables or not.

clausePerm :: Clause -> Maybe Permutation

Computes the permutation from the clause telescope to the pattern variables.

Use as fromMaybe IMPOSSIBLE . clausePerm to crash in a controlled way if a de Bruijn index is out of scope here.

patternToElim :: Arg DeBruijnPattern -> Elim

Turn a pattern into a term. Projection patterns are turned into projection eliminations, other patterns into apply elimination.

patternsToElims :: [NamedArg DeBruijnPattern] -> [Elim]

patternToTerm :: DeBruijnPattern -> Term

class MapNamedArgPattern a p where

Minimal complete definition

Nothing

Methods

mapNamedArgPattern :: (NamedArg (Pattern' a) -> NamedArg (Pattern' a)) -> p -> p

default mapNamedArgPattern :: forall (f :: Type -> Type) p'. (Functor f, MapNamedArgPattern a p', p ~ f p') => (NamedArg (Pattern' a) -> NamedArg (Pattern' a)) -> p -> p

class PatternLike a b where

Generic pattern traversal.

Pre-applies a pattern modification, recurses, and post-applies another one.

Minimal complete definition

Nothing

Methods

foldrPattern

Arguments

:: Monoid m

=> (Pattern' a -> m -> m)
Combine a pattern and the value computed from its subpatterns.

-> b

-> m

Fold pattern.

default foldrPattern :: forall m (f :: Type -> Type) p. (Monoid m, Foldable f, PatternLike a p, f p ~ b) => (Pattern' a -> m -> m) -> b -> m

traversePatternM

Arguments

:: Monad m

=> (Pattern' a -> m (Pattern' a))
pre: Modification before recursion.

-> (Pattern' a -> m (Pattern' a))
post: Modification after recursion.

-> b

-> m b

Traverse pattern.

default traversePatternM :: forall (f :: Type -> Type) p m. (Traversable f, PatternLike a p, f p ~ b, Monad m) => (Pattern' a -> m (Pattern' a)) -> (Pattern' a -> m (Pattern' a)) -> b -> m b

foldPattern :: (PatternLike a b, Monoid m) => (Pattern' a -> m) -> b -> m

Compute from each subpattern a value and collect them all in a monoid.

preTraversePatternM

Arguments

:: (PatternLike a b, Monad m)

=> (Pattern' a -> m (Pattern' a))
pre: Modification before recursion.

-> b

-> m b

Traverse pattern(s) with a modification before the recursive descent.

postTraversePatternM

Arguments

:: (PatternLike a b, Monad m)

=> (Pattern' a -> m (Pattern' a))
post: Modification after recursion.

-> b

-> m b

Traverse pattern(s) with a modification after the recursive descent.

class CountPatternVars a where

Minimal complete definition

Nothing

Methods

countPatternVars :: a -> Int

default countPatternVars :: forall (f :: Type -> Type) b. (Foldable f, CountPatternVars b, f b ~ a) => a -> Int

class PatternVarModalities p where

Associated Types

type PatVar p

Methods

patternVarModalities :: p -> [(PatVar p, Modality)]

Get the list of pattern variables annotated with modalities.

hasDefP :: [NamedArg DeBruijnPattern] -> Bool