Agda.Syntax.Internal.Univ

Description

Kinds of standard universes: Prop, Type, SSet.

Types

data Univ

Flavor of standard universe (Prop < Type < SSet,).

Constructors

UProp
Fibrant universe of propositions.

UType
Fibrant universe.

USSet
Non-fibrant universe.

data IsFibrant

We have IsFibrant < IsStrict.

Constructors

IsFibrant
Fibrant universe.

IsStrict
Non-fibrant universe.

Universe kind arithmetic

univUniv :: Univ -> Univ

The successor universe type of a universe.

funUniv :: Univ -> Univ -> Univ

Compute the universe type of a function space from the universe types of domain and codomain.

Inverting funUniv

domainUniv

Arguments

:: Bool
Have UProp?

-> Univ
Univ kind of the funSort.

-> Univ
Univ kind of the codomain.

-> Maybe Univ
Univ kind of the domain, if unique.

Conclude u1 from funUniv u1 u2 and u2.

codomainUniv

Arguments

:: Univ
Univ kind of the funSort.

-> Univ
Univ kind of the domain.

-> Maybe Univ
Univ kind of the codomain, if uniquely exists.

Conclude u2 from funUniv u1 u2 and u1.

Fibrancy

univFibrancy :: Univ -> IsFibrant

Fibrancy of standard universes.

Printing

showUniv :: Univ -> String

Hacky showing of standard universes, does not take actual names into account.