Agda.Syntax.Parser

Types

data Parser a

Wrapped Parser type.

Parse functions

parse :: Parser a -> String -> PM a

Parse without top-level layout.

parsePosString :: Parser a -> Position -> String -> PM a

parseFile

Arguments

:: Show a

=> Parser a

-> AbsolutePath
The path to the file.

-> String
The file contents. Note that the file is not read from disk.

-> PM (a, FileType)

Parsers

moduleParser :: Parser Module

Parses a module.

moduleNameParser :: Parser QName

Parses a module name.

acceptableFileExts :: [String]

Extensions supported by parseFile.

exprParser :: Parser Expr

Parses an expression.

exprWhereParser :: Parser ExprWhere

Parses an expression followed by a where clause.

holeContentParser :: Parser HoleContent

Parses an expression or some other content of an interaction hole.

tokensParser :: Parser [Token]

Gives the parsed token stream (including comments).

Reading files.

readFilePM :: AbsolutePath -> PM Text

Returns the contents of the given file.

Parse errors

data ParseError

Parse errors: what you get if parsing fails.

Constructors

ParseError
Errors that arise at a specific position in the file

Fields

errSrcFile :: !SrcFile

The file in which the error occurred.

errPos :: !PositionWithoutFile

Where the error occurred.

errInput :: String

The remaining input.

errPrevToken :: String

The previous token.

errMsg :: String

Hopefully an explanation of what happened.

OverlappingTokensError
Parse errors that concern a range in a file.

Fields

errRange :: !(Range' SrcFile)

The range of the bigger overlapping token

InvalidExtensionError
Parse errors that concern a whole file.

Fields

errPath :: !AbsolutePath

The file which the error concerns.

errValidExts :: [String]

ReadFileError

Fields

errPath :: !AbsolutePath

The file which the error concerns.

errIOError :: IOError

data ParseWarning

Warnings for parsing.

Constructors

OverlappingTokensWarning
Parse errors that concern a range in a file.

Fields

warnRange :: !(Range' SrcFile)

The range of the bigger overlapping token

UnsupportedAttribute Range !(Maybe String)
Unsupported attribute.

MultipleAttributes Range !(Maybe String)
Multiple attributes.

newtype PM a

A monad for handling parse errors and warnings.

Constructors

PM

Fields

unPM :: ExceptT ParseError (StateT [ParseWarning] IO) a

runPMIO :: MonadIO m => PM a -> m (Either ParseError a, [ParseWarning])

Run a PM computation, returning a list of warnings in first-to-last order and either a parse error or the parsed thing.