Agda.Syntax.Parser.LexActions

Description

This module contains the building blocks used to construct the lexer.

Main function

lexToken :: Parser Token

Scan the input to find the next token. Calls alexScanUser. This is the main lexing function where all the work happens. The function lexer, used by the parser is the continuation version of this function.

Lex actions

General actions

token :: (String -> Parser tok) -> LexAction tok

The most general way of parsing a token.

withInterval :: ((Interval, String) -> tok) -> LexAction tok

Parse a token from an Interval and the lexed string.

withInterval' :: (String -> a) -> ((Interval, a) -> tok) -> LexAction tok

Like withInterval, but applies a function to the string.

withInterval_ :: (Interval -> r) -> LexAction r

Return a token without looking at the lexed string.

withLayout :: Keyword -> LexAction r -> LexAction r

Executed for layout keywords. Enters the layout state and performs the given action.

andThen :: Parser () -> LexAction r -> LexAction r infixr 1

Prepend some parser manipulation to an action.

skip :: LexAction Token

Throw away the current lexeme.

begin :: LexState -> LexAction Token

Enter a new state without consuming any input.

end :: LexAction Token

Exit the current state without consuming any input.

beginWith :: LexState -> LexAction a -> LexAction a

Enter a new state and perform the given action.

endWith :: LexAction a -> LexAction a

Exit the current state and perform the given action.

begin_ :: LexState -> LexAction Token

Enter a new state throwing away the current lexeme.

end_ :: LexAction Token

Exit the current state throwing away the current lexeme.

lexError :: String -> Parser a

For lexical errors we want to report the current position as the site of the error, whereas for parse errors the previous position is the one we're interested in (since this will be the position of the token we just lexed). This function does parseErrorAt the current position.

Specialized actions

keyword :: Keyword -> LexAction Token

Parse a Keyword token, triggers layout for layoutKeywords.

symbol :: Symbol -> LexAction Token

Parse a Symbol token.

identifier :: LexAction Token

Parse an identifier. Identifiers can be qualified (see Name). Example: Foo.Bar.f

literal :: Read a => (a -> Literal) -> LexAction Token

literal' :: (String -> a) -> (a -> Literal) -> LexAction Token

Parse a literal.

integer :: String -> Integer

Lex predicates

followedBy :: Char -> LexPredicate

True when the given character is the next character of the input string.

eof :: LexPredicate

True if we are at the end of the file.

inState :: LexState -> LexPredicate

True if the given state appears somewhere on the state stack