Agda.Syntax.Parser.Literate

Description

Preprocessors for literate code formats.

Documentation

literateProcessors :: [(String, (Processor, FileType))]

List of valid extensions for literate Agda files, and their corresponding preprocessors.

If you add new extensions, remember to update test/Utils.hs so that test cases ending in the new extensions are found.

literateExtsShortList :: List1 String

Short list of extensions for literate Agda files. For display purposes.

literateTeX :: Processor

Preprocessor for literate TeX.

literateRsT :: Processor

Preprocessor for reStructuredText.

literateMd :: Processor

Preprocessor for Markdown.

literateOrg :: Processor

Preprocessor for Org mode documents.

illiterate :: [Layer] -> String

Blanks the non-code parts of a given file, preserving positions of characters corresponding to code. This way, there is a direct correspondence between source positions and positions in the processed result.

atomizeLayers :: Layers -> [(LayerRole, Char)]

type Processor = PositionWithoutFile -> String -> [Layer]

Type of a literate preprocessor: Invariants:

f : Processor

proposition> f pos s /= []

proposition> f pos s >>= layerContent == s

type Layers = [Layer]

A list of contiguous layers.

data Layer

A sequence of characters in a file playing the same role.

Constructors

Layer

Fields

layerRole :: LayerRole

interval :: IntervalWithoutFile

layerContent :: String

data LayerRole

Role of a character in the file.

Constructors

Markup

Comment

Code

isCode :: LayerRole -> Bool

Returns True if the role corresponds to Agda code.

isCodeLayer :: Layer -> Bool

Returns True if the layer contains Agda code.