Agda.Syntax.Position

Description

Position information for syntax. Crucial for giving good error messages.

Positions

type Position = Position' SrcFile

type PositionWithoutFile = Position' ()

data Position' a

Represents a point in the input.

If two positions have the same srcFile and posPos components, then the final two components should be the same as well, but since this can be hard to enforce the program should not rely too much on the last two components; they are mainly there to improve error messages for the user.

Note the invariant which positions have to satisfy: positionInvariant.

Constructors

Pn'

Fields

srcFile :: !a

File.

posPos :: !Word32

Position, counting from 1.

posLineCol :: !Word64

Line and position numbers as Word32, both counting from 1.

pattern Pn :: a -> Word32 -> Word32 -> Word32 -> Position' a

posLine :: Position' a -> Word32

posCol :: Position' a -> Word32

type SrcFile = Maybe RangeFile

data RangeFile

File information used in the Position, Interval and Range types.

Constructors

RangeFile

Fields

rangeFilePath :: !AbsolutePath

The file's path.

rangeFileName :: !(Maybe (TopLevelModuleName' Range))

The file's top-level module name (if applicable).

This field is optional, but some things may break if the field is not instantiated with an actual top-level module name. For instance, the Eq and Ord instances only make use of this field.

The field uses Maybe rather than Maybe because it should be possible to instantiate it with something that is not yet defined (see parseSource).

This '(TopLevelModuleName' Range)' should not contain a range.

mkRangeFile :: AbsolutePath -> Maybe (TopLevelModuleName' Range) -> RangeFile

A smart constructor for RangeFile.

positionInvariant :: Position' a -> Bool

startPos :: Maybe RangeFile -> Position

The first position in a file: position 1, line 1, column 1.

startPos' :: a -> Position' a

The first position in a file: position 1, line 1, column 1.

movePos :: Position' a -> Char -> Position' a

Advance the position by one character. A newline character ('n') moves the position to the first character in the next line. Any other character moves the position to the next column.

movePosByString :: Foldable t => Position' a -> t Char -> Position' a

Advance the position by a string.

movePosByString = foldl' movePos

backupPos :: Position' a -> Position' a

Backup the position by one character.

Precondition: The character must not be 'n'.

Intervals

type Interval = Interval' SrcFile

type IntervalWithoutFile = Interval' ()

data Interval' a

An interval. The iEnd position is not included in the interval.

Note the invariant which intervals have to satisfy: intervalInvariant.

Constructors

Interval a !PositionWithoutFile !PositionWithoutFile

intervalInvariant :: Ord a => Interval' a -> Bool

iStart :: Interval' a -> Position' a

iEnd :: Interval' a -> Position' a

posToInterval :: a -> PositionWithoutFile -> PositionWithoutFile -> Interval' a

Converts a file name and two positions to an interval. Sort the positions ascendingly.

getIntervalFile :: Interval' a -> a

iLength :: Interval' a -> Word32

The length of an interval.

fuseIntervals :: IntervalWithoutFile -> IntervalWithoutFile -> IntervalWithoutFile

Finds the least interval which covers the arguments.

Ranges

type Range = Range' SrcFile

type RangeWithoutFile = Range' ()

data Range' a

A range is a file name, plus a sequence of intervals, assumed to point to the given file. The intervals should be consecutive and separated.

Note the invariant which ranges have to satisfy: rangeInvariant.

Constructors

NoRange

Range !a (Seq IntervalWithoutFile)

rangeInvariant :: Ord a => Range' a -> Bool

Range invariant.

consecutiveAndSeparated :: Ord a => [Interval' a] -> Bool

Are the intervals consecutive and separated, do they all point to the same file, and do they satisfy the interval invariant?

intervalsToRange :: a -> [IntervalWithoutFile] -> Range' a

Turns a file name plus a list of intervals into a range.

Precondition: consecutiveAndSeparated.

intervalToRange :: a -> IntervalWithoutFile -> Range' a

Converts a file name and an interval to a range.

rangeFromAbsolutePath :: AbsolutePath -> Range

Range pointing to the first position in the given file.

rangeIntervals :: Range' a -> [IntervalWithoutFile]

The intervals that make up the range. The intervals are consecutive and separated (consecutiveAndSeparated).

rangeFile :: Range -> SrcFile

The file the range is pointing to.

rangeModule' :: Range -> Maybe (Maybe (TopLevelModuleName' Range))

The range's top-level module name, if any.

If there is no range, then Nothing is returned. If there is a range without a module name, then Just Nothing is returned.

rangeModule :: Range -> Maybe (TopLevelModuleName' Range)

The range's top-level module name, if any.

rightMargin :: Range -> Range

Conflate a range to its right margin.

noRange :: Range' a

Ranges between two unknown positions

posToRange :: Position' a -> Position' b -> Range' a

Converts two positions to a range.

Precondition: The positions have to point to the same file.

posToRange' :: a -> PositionWithoutFile -> PositionWithoutFile -> Range' a

Converts a file name and two positions to a range.

rStart :: Range' a -> Maybe (Position' a)

The initial position in the range, if any.

rStart' :: Range' a -> Maybe PositionWithoutFile

The initial position in the range, if any.

rEnd :: Range' a -> Maybe (Position' a)

The position after the final position in the range, if any.

rEnd' :: Range' a -> Maybe PositionWithoutFile

The position after the final position in the range, if any.

rangeToInterval :: Range' a -> Maybe IntervalWithoutFile

Converts a range to an interval, if possible. Note that the information about the source file is lost.

rangeToIntervalWithFile :: Range' a -> Maybe (Interval' a)

Converts a range to an interval, if possible.

continuous :: Range' a -> Range' a

Returns the shortest continuous range containing the given one.

continuousPerLine :: Ord a => Range' a -> Range' a

Removes gaps between intervals on the same line.

newtype PrintRange a

Wrapper to indicate that range should be printed.

Constructors

PrintRange a

class HasRange a where

Things that have a range are instances of this class.

Minimal complete definition

Nothing

Methods

getRange :: a -> Range

default getRange :: forall (t :: Type -> Type) b. (Foldable t, HasRange b, t b ~ a) => a -> Range

class HasRangeWithoutFile a where

Things that have a RangeWithoutFile are instances of this class.

Minimal complete definition

Nothing

Methods

getRangeWithoutFile :: a -> RangeWithoutFile

default getRangeWithoutFile :: forall (t :: Type -> Type) b. (Foldable t, HasRangeWithoutFile b, t b ~ a) => a -> RangeWithoutFile

class HasRange a => SetRange a where

If it is also possible to set the range, this is the class.

Instances should satisfy getRange (setRange r x) == r.

Minimal complete definition

Nothing

Methods

setRange :: Range -> a -> a

default setRange :: forall (f :: Type -> Type) b. (Functor f, SetRange b, f b ~ a) => Range -> a -> a

class KillRange a where

Killing the range of an object sets all range information to noRange.

Minimal complete definition

Nothing

Methods

killRange :: KillRangeT a

default killRange :: forall (f :: Type -> Type) b. (Functor f, KillRange b, f b ~ a) => KillRangeT a

type KillRangeT a = a -> a

killRangeMap :: (KillRange k, KillRange v) => KillRangeT (Map k v)

Remove ranges in keys and values of a map.

class KILLRANGE t (b :: Bool) where

Methods

killRangeN :: t -> t

withRangeOf :: (SetRange t, HasRange u) => t -> u -> t

x `withRangeOf` y sets the range of x to the range of y.

fuseRange :: (HasRange u, HasRange t) => u -> t -> Range

Precondition: The ranges must point to the same file (or be empty).

fuseRanges :: Ord a => Range' a -> Range' a -> Range' a

fuseRanges r r' unions the ranges r and r'.

Meaning it finds the least range r0 that covers r and r'.

Precondition: The ranges must point to the same file (or be empty).

beginningOf :: Range -> Range

beginningOf r is an empty range (a single, empty interval) positioned at the beginning of r. If r does not have a beginning, then noRange is returned.

beginningOfFile :: BeginningOfFile a => a -> Range

Return an empty range (a single, empty interval) at the beginning of the file.

interleaveRanges :: HasRangeWithoutFile a => [a] -> [a] -> ([a], [(a, a)])

Interleaves two streams of ranged elements

It will report the conflicts as a list of conflicting pairs. In case of conflict, the element with the earliest start position is placed first. In case of a tie, the element with the earliest ending position is placed first. If both tie, the element from the first list is placed first.