Agda.Syntax.Scope.Base

Description

This module defines the notion of a scope and operations on scopes.

Scope representation

data Scope

A scope is a named collection of names partitioned into public and private names.

Constructors

Scope

Fields

scopeName :: ModuleName

scopeParents :: [ModuleName]

scopeNameSpaces :: ScopeNameSpaces

scopeImports :: Map QName ModuleName

scopeDatatypeModule :: Maybe DataOrRecordModule

data DataOrRecordModule

Constructors

IsDataModule

IsRecordModule

data NameSpaceId

See Access.

Constructors

PrivateNS
Things not exported by this module.

PublicNS
Things defined and exported by this module.

ImportedNS
Things from open public, exported by this module.

allNameSpaces :: [NameSpaceId]

type ScopeNameSpaces = [(NameSpaceId, NameSpace)]

localNameSpace :: Access -> NameSpaceId

nameSpaceAccess :: NameSpaceId -> Access

scopeNameSpace :: NameSpaceId -> Scope -> NameSpace

Get a NameSpace from Scope.

updateScopeNameSpaces :: (ScopeNameSpaces -> ScopeNameSpaces) -> Scope -> Scope

A lens for scopeNameSpaces

updateScopeNameSpacesM :: Functor m => (ScopeNameSpaces -> m ScopeNameSpaces) -> Scope -> m Scope

`Monadic' lens (Functor sufficient).

data ScopeInfo

The complete information about the scope at a particular program point includes the scope stack, the local variables, and the context precedence.

Constructors

ScopeInfo

Fields

_scopeCurrent :: ModuleName

_scopeModules :: Map ModuleName Scope

_scopeVarsToBind :: LocalVars

The variables that will be bound at the end of the current block of variables (i.e. clause). We collect them here instead of binding them immediately so we can avoid shadowing between variables in the same variable block.

_scopeLocals :: LocalVars

_scopePrecedence :: !PrecedenceStack

_scopeInverseName :: NameMap

_scopeInverseModule :: ModuleMap

_scopeInScope :: InScopeSet

_scopeFixities :: Fixities

Maps concrete names C.Name to fixities

_scopePolarities :: Polarities

Maps concrete names C.Name to polarities

_scopeRecords :: Map QName (QName, Maybe Induction)

Maps the name of a record to the name of its (co)constructor.

data NameMapEntry

For the sake of highlighting, the _scopeInverseName map also stores the KindOfName of an A.QName.

Constructors

NameMapEntry

Fields

qnameKind :: !KindOfName

The anameKind.

qnameConcrete :: !(List1 QName)

Possible renderings of the abstract name.

type NameMap = HashMap NameId NameMapEntry

type ModuleMap = HashMap ModuleName [QName]

type LocalVars = AssocList Name LocalVar

Local variables.

data BindingSource

For each bound variable, we want to know whether it was bound by a λ, Π, module telescope, pattern, or let.

Constructors

LambdaBound
λ (currently also used for Π and module parameters)

PatternBound Hiding
f ... =. Remember Hiding for pattern variables {x} and {{x}}. This information is only used for checking pattern synonyms. It is not serialized.

LetBound
let ... in

WithBound
| ... in q

MacroBound
Binding added to scope by one of context-manipulating reflection primitives

data LocalVar

A local variable can be shadowed by an import. In case of reference to a shadowed variable, we want to report a scope error.

Constructors

LocalVar

Fields

localVar :: Name

Unique ID of local variable.

localBindingSource :: BindingSource

Kind of binder used to introduce the variable (λ, let, ...).

localShadowedBy :: [AbstractName]

If this list is not empty, the local variable is shadowed by one or more imports.

shadowLocal :: List1 AbstractName -> LocalVar -> LocalVar

Shadow a local name by a non-empty list of imports.

patternToModuleBound :: LocalVar -> LocalVar

Treat patternBound variable as a module parameter

notShadowedLocal :: LocalVar -> Maybe Name

Project name of unshadowed local variable.

notShadowedLocals :: LocalVars -> AssocList Name Name

Get all locals that are not shadowed by imports.

scopeCurrent :: Lens' ScopeInfo ModuleName

Lenses for ScopeInfo components

scopeModules :: Lens' ScopeInfo (Map ModuleName Scope)

scopeVarsToBind :: Lens' ScopeInfo LocalVars

scopeLocals :: Lens' ScopeInfo LocalVars

scopePrecedence :: Lens' ScopeInfo PrecedenceStack

scopeInverseName :: Lens' ScopeInfo NameMap

scopeInverseModule :: Lens' ScopeInfo ModuleMap

scopeInScope :: Lens' ScopeInfo InScopeSet

scopeFixities :: Lens' ScopeInfo Fixities

scopePolarities :: Lens' ScopeInfo Polarities

scopeRecords :: Lens' ScopeInfo (Map QName (QName, Maybe Induction))

scopeFixitiesAndPolarities :: Lens' ScopeInfo (Fixities, Polarities)

updateVarsToBind :: (LocalVars -> LocalVars) -> ScopeInfo -> ScopeInfo

Lens for scopeVarsToBind.

setVarsToBind :: LocalVars -> ScopeInfo -> ScopeInfo

updateScopeLocals :: (LocalVars -> LocalVars) -> ScopeInfo -> ScopeInfo

Lens for scopeLocals.

setScopeLocals :: LocalVars -> ScopeInfo -> ScopeInfo

Name spaces

data NameSpace

A NameSpace contains the mappings from concrete names that the user can write to the abstract fully qualified names that the type checker wants to read.

Constructors

NameSpace

Fields

nsNames :: NamesInScope

Maps concrete names to a list of abstract names. If this is part of a scope produced by an ImportDirective, the concrete names should (where possible) match those of the directive. E.g., if the directive mentions using (x), then the concrete x in the map should be this x rather then the x from the original scope that is restricted by the import directive. This is important for the highlighting of the UnusedImports warning.

nsNameParts :: NamePartsInScope

Maps name parts to a list of abstract names in which the name part occurs.

nsModules :: ModulesInScope

Maps concrete module names to a list of abstract module names.

nsInScope :: InScopeSet

All abstract names targeted by a concrete name in scope. Computed by recomputeInScopeSets.

type ThingsInScope a = Map Name (List1 a)

type NamesInScope = ThingsInScope AbstractName

type NamePartsInScope = Map RawName NamesInScope

type ModulesInScope = ThingsInScope AbstractModule

type InScopeSet = Set QName

data InScopeTag a where

Set of types consisting of exactly AbstractName and AbstractModule.

A GADT just for some dependent-types trickery.

Constructors

NameTag :: InScopeTag AbstractName

ModuleTag :: InScopeTag AbstractModule

class Ord a => InScope a where

Type class for some dependent-types trickery.

Methods

inScopeTag :: InScopeTag a

inNameSpace :: InScope a => NameSpace -> ThingsInScope a

inNameSpace selects either the name map or the module name map from a NameSpace. What is selected is determined by result type (using the dependent-type trickery).

data NameOrModule

Non-dependent tag for name or module.

Constructors

NameNotModule

ModuleNotName

Operator parsing

data OperatorScope

Names that are relevant to operator parsing.

Constructors

OpScope

Fields

osHasOps :: !Bool

Are there any operators or notations in scope?

osScope :: Map QName (List1 AbstractName)

All relevant names in global scope.

osLocals :: !(AssocList Name Name)

Relevant names in local scope.

Live names and copy trimming

data LiveNames

A set of module and definition names that were explicitly referred to during scope checking.

Constructors

SomeLiveNames
We encountered these names during scope-checking, so they should be preserved.

Fields

liveModules :: !(Set ModuleName)

liveNames :: !(Set QName)

AllLiveNames
An interaction point can see this copy, so we should preserve every name.

isNameAlive :: QName -> LiveNames -> Bool

Is the given name alive in this set?

Note: a qualified name can be kept alive by liveness of any of the modules it belongs to.

isModuleAlive :: ModuleName -> LiveNames -> Bool

Is the given module (or one of its parents) a member of the set?

Unused imports

data OpenedModule

Information gathered during scope checking for the unused-imports warning when a module is opened.

Constructors

OpenedModule

Fields

openedRange :: KwRange

Range of the open keyword.

openedParent :: ModuleName

Parent module into which we merged the opened module.

openedModule :: ModuleName

Module we opened.

openedHasDir :: Bool

Whether the open statement comes with using or renaming.

openedScope :: NamesInScope

The scope imported by the open statement. The keys (Name) of this map should be the concrete names of the opening directive, if such was given.

data UnusedImportsState

Information gathered during scope checking for the unused-imports warning.

Constructors

UnusedImportsState

Fields

unambiguousLookups :: ![AbstractName]

Names that were unambiguously resolved from a concrete name in the source.

ambiguousLookups :: !(IntMap (List2 AbstractName))

Names that were ambiguously resolved from a concrete name in the source. They are stored with their position in the file that is matched with disambiguation information produced by the type checker.

openedModules :: !(IntMap OpenedModule)

Log of module opens with the names they brought into scope.

lensUnambiguousLookups :: Lens' UnusedImportsState [AbstractName]

lensAmbiguousLookups :: Lens' UnusedImportsState (IntMap (List2 AbstractName))

lensOpenedModules :: Lens' UnusedImportsState (IntMap OpenedModule)

Decorated names

data KindOfName

For the sake of parsing left-hand sides, we distinguish constructor and record field names from defined names.

Constructors

ConName
Constructor name (Inductive or don't know).

CoConName
Constructor name (definitely CoInductive).

FldName
Record field name.

PatternSynName
Name of a pattern synonym.

GeneralizeName
Name to be generalized

DisallowedGeneralizeName
Generalizable variable from a let open

MacroName
Name of a macro

QuotableName
A name that can only be quoted. Previous category DefName: (Refined in a flat manner as Enum and Bounded are not hereditary.)

DataName
Name of a data.

RecName
Name of a record.

FunName
Name of a defined function.

AxiomName
Name of a postulate.

PrimName
Name of a primitive.

OtherDefName
A DefName, but either other kind or don't know which kind. End DefName. Keep these together in sequence, for sake of isDefName!

defNameKinds :: [KindOfName]

All kinds of regular definitions.

isDefName :: KindOfName -> Bool

conLikeNameKinds :: [KindOfName]

Constructor and pattern synonyms.

isConName :: KindOfName -> Maybe Induction

conKindOfName :: Induction -> KindOfName

conKindOfName' :: Foldable t => t Induction -> KindOfName

For ambiguous constructors, we might have both alternatives of Induction. In this case, we default to ConName.

approxConInduction :: Foldable t => t Induction -> Induction

For ambiguous constructors, we might have both alternatives of Induction. In this case, we default to Inductive.

exactConInduction :: Foldable t => t Induction -> Maybe Induction

exactConName :: Foldable t => t Induction -> Maybe KindOfName

Only return [Co]ConName if no ambiguity.

data KindsOfNames

A set of KindOfName, for the sake of elemKindsOfNames.

Constructors

AllKindsOfNames

SomeKindsOfNames (Set KindOfName)
Only these kinds.

ExceptKindsOfNames (Set KindOfName)
All but these Kinds.

elemKindsOfNames :: KindOfName -> KindsOfNames -> Bool

allKindsOfNames :: KindsOfNames

someKindsOfNames :: [KindOfName] -> KindsOfNames

exceptKindsOfNames :: [KindOfName] -> KindsOfNames

data WithKind a

Decorate something with KindOfName

Constructors

WithKind

Fields

theKind :: KindOfName

kindedThing :: a

data WhyInScope

Where does a name come from?

This information is solely for reporting to the user, see whyInScope.

Constructors

Defined
Defined in this module.

Opened QName WhyInScope
Imported from another module.

Applied QName WhyInScope
Imported by a module application.

data AbstractName

A decoration of QName.

Constructors

AbsName

Fields

anameName :: QName

The resolved qualified name.

anameKind :: KindOfName

The kind (definition, constructor, record field etc.).

anameLineage :: WhyInScope

Explanation where this name came from.

anameMetadata :: NameMetadata

Additional information needed during scope checking. Currently used for generalized data/record params.

data NameMetadata

Constructors

NameMetadata

Fields

nameDataGeneralizedVars :: Map QName Name

nameDataIsInstance :: IsInstance

noMetadata :: NameMetadata

generalizedVarsMetadata :: Map QName Name -> NameMetadata

instanceMetadata :: IsInstance -> NameMetadata

data AbstractModule

A decoration of abstract syntax module names.

Constructors

AbsModule

Fields

amodName :: ModuleName

The resolved module name.

amodLineage :: WhyInScope

Explanation where this name came from.

lensAnameName :: Lens' AbstractName QName

Van Laarhoven lens on anameName.

lensAmodName :: Lens' AbstractModule ModuleName

Van Laarhoven lens on amodName.

data ResolvedName

Constructors

VarName
Local variable bound by λ, Π, module telescope, pattern, let.

Fields

resolvedVar :: Name

resolvedBindingSource :: BindingSource

What kind of binder?

DefinedName
Function, data/record type, postulate.

Fields

Access

AbstractName

Suffix

anameKind can be DefName, MacroName, QuotableName.

FieldName
Record field name. Needs to be distinguished to parse copatterns.

Fields

(List1 AbstractName)

(FldName ==) . anameKind for all names.

ConstructorName
Data or record constructor name.

Fields

(Set1 Induction)

Inductive or CoInductive or both.

(List1 AbstractName)

isJust . isConName . anameKind for all names.

PatternSynResName
Name of pattern synonym.

Fields

(List1 AbstractName)

(PatternSynName ==) . anameKind for all names.

UnknownName
Unbound name.

data AmbiguousNameReason

Why is a resolved name ambiguous? What did it resolve to?

Invariant (statically enforced): At least two resolvents in total.

Constructors

AmbiguousLocalVar LocalVar (List1 AbstractName)
The name resolves both to a local variable and some declared names.

AmbiguousDeclName (List2 AbstractName)
The name resolves to at least 2 declared names.

data NameResolutionError

A failure in name resolution, indicating the reason that a name which is in scope could not be returned from tryResolveName.

Constructors

IllegalAmbiguity AmbiguousNameReason
Ambiguous names are not supported in this situation.

ConstrOfNonRecord QName ResolvedName
The name was Foo.constructor, and Foo is in scope, but it is not a record.

ambiguousNamesInReason :: AmbiguousNameReason -> List2 QName

The flat list of ambiguous names in AmbiguousNameReason.

data WhyInScopeData

Constructors

WhyInScopeData

Fields

QName

The name x this explanation is about.

FilePath

The directory in which the current module resides.

(Maybe LocalVar)

The local variable that x could denote, if any.

[AbstractName]

The defined names that x could denote.

[AbstractModule]

The modules that x could denote.

whyInScopeDataFromAbstractName :: QName -> AbstractName -> WhyInScopeData

whyInScopeDataFromAmbiguousNameReason :: QName -> AmbiguousNameReason -> WhyInScopeData

Operations on name and module maps.

mergeNames :: Eq a => ThingsInScope a -> ThingsInScope a -> ThingsInScope a

mergeNamesMany :: Eq a => [ThingsInScope a] -> ThingsInScope a

Operations on name spaces

emptyNameSpace :: NameSpace

The empty name space.

mapNameSpace :: (NamesInScope -> NamesInScope) -> (ModulesInScope -> ModulesInScope) -> (InScopeSet -> InScopeSet) -> NameSpace -> NameSpace

Map functions over the names and modules in a name space.

zipNameSpace :: (NamesInScope -> NamesInScope -> NamesInScope) -> (NamePartsInScope -> NamePartsInScope -> NamePartsInScope) -> (ModulesInScope -> ModulesInScope -> ModulesInScope) -> (InScopeSet -> InScopeSet -> InScopeSet) -> NameSpace -> NameSpace -> NameSpace

Zip together two name spaces.

mapNameSpaceM :: Applicative m => (NamesInScope -> m NamesInScope) -> (ModulesInScope -> m ModulesInScope) -> (InScopeSet -> m InScopeSet) -> NameSpace -> m NameSpace

Map monadic function over a namespace.

General operations on scopes

emptyScope :: Scope

The empty scope.

emptyScopeInfo :: ScopeInfo

The empty scope info.

mapScope :: (NameSpaceId -> NamesInScope -> NamesInScope) -> (NameSpaceId -> ModulesInScope -> ModulesInScope) -> (NameSpaceId -> InScopeSet -> InScopeSet) -> Scope -> Scope

Map functions over the names and modules in a scope.

mapScope_ :: (NamesInScope -> NamesInScope) -> (ModulesInScope -> ModulesInScope) -> (InScopeSet -> InScopeSet) -> Scope -> Scope

Same as mapScope but applies the same function to all name spaces.

mapScopeNS :: NameSpaceId -> (NamesInScope -> NamesInScope) -> (ModulesInScope -> ModulesInScope) -> (InScopeSet -> InScopeSet) -> Scope -> Scope

Same as mapScope but applies the function only on the given name space.

mapScopeM :: Applicative m => (NameSpaceId -> NamesInScope -> m NamesInScope) -> (NameSpaceId -> ModulesInScope -> m ModulesInScope) -> (NameSpaceId -> InScopeSet -> m InScopeSet) -> Scope -> m Scope

Map monadic functions over the names and modules in a scope.

mapScopeM_ :: Applicative m => (NamesInScope -> m NamesInScope) -> (ModulesInScope -> m ModulesInScope) -> (InScopeSet -> m InScopeSet) -> Scope -> m Scope

Same as mapScopeM but applies the same function to both the public and private name spaces.

zipScope :: (NameSpaceId -> NamesInScope -> NamesInScope -> NamesInScope) -> (NameSpaceId -> NamePartsInScope -> NamePartsInScope -> NamePartsInScope) -> (NameSpaceId -> ModulesInScope -> ModulesInScope -> ModulesInScope) -> (NameSpaceId -> InScopeSet -> InScopeSet -> InScopeSet) -> Scope -> Scope -> Scope

Zip together two scopes. The resulting scope has the same name as the first scope.

zipScope_ :: (NamesInScope -> NamesInScope -> NamesInScope) -> (NamePartsInScope -> NamePartsInScope -> NamePartsInScope) -> (ModulesInScope -> ModulesInScope -> ModulesInScope) -> (InScopeSet -> InScopeSet -> InScopeSet) -> Scope -> Scope -> Scope

Same as zipScope but applies the same function to both the public and private name spaces.

recomputeInScopeSets :: Scope -> Scope

Recompute the inScope sets of a scope.

filterScope :: (Name -> Bool) -> (Name -> Bool) -> Scope -> Scope

Filter a scope keeping only concrete names matching the predicates. The first predicate is applied to the names and the second to the modules.

mapScopeMaybe :: (Name -> Maybe Name) -> (Name -> Maybe Name) -> Scope -> Scope

Filter a scope keeping only concrete names matching the predicates. The first predicate is applied to the names and the second to the modules.

allNamesInScope :: InScope a => Scope -> ThingsInScope a

Return all names in a scope.

allNamesInScope' :: InScope a => Scope -> ThingsInScope (a, Access)

findNameInScope :: InScope a => Name -> Scope -> [(a, Access)]

Look up a single name in the current scope.

This is equivalent to Map.lookup n . allNamesInScope', but more efficient when only a single name needs to be looked up.

exportedNamesInScope :: InScope a => Scope -> ThingsInScope a

Returns the scope's non-private names.

namesInScope :: InScope a => [NameSpaceId] -> Scope -> ThingsInScope a

allThingsInScope :: Scope -> NameSpace

thingsInScope :: [NameSpaceId] -> Scope -> NameSpace

mergeScope :: Scope -> Scope -> Scope

Merge two scopes. The result has the name of the first scope.

mergeScopes :: [Scope] -> Scope

Merge a non-empty list of scopes. The result has the name of the first scope in the list.

Specific operations on scopes

setScopeAccess :: NameSpaceId -> Scope -> Scope

Move all names in a scope to the given name space (except never move from Imported to Public).

setNameSpace :: NameSpaceId -> NameSpace -> Scope -> Scope

Update a particular name space.

modifyNameSpace :: NameSpaceId -> (NameSpace -> NameSpace) -> Scope -> Scope

Modify a particular name space.

addNameToNameParts :: Name -> AbstractName -> NamePartsInScope -> NamePartsInScope

addNamePartsToScope :: NameSpaceId -> Name -> AbstractName -> Scope -> Scope

Add all associated name parts of a name (operator parts, notation parts) to the scope.

recomputeNameParts :: Scope -> Scope

Recompute name parts in the namespaces.

addNameToScope :: NameSpaceId -> Name -> AbstractName -> Scope -> Scope

Add a name to a scope.

removeNameFromScope :: NameSpaceId -> Name -> Scope -> Scope

Remove a name from a scope. Caution: does not update the nsInScope set. This is only used by rebindName and in that case we add the name right back (but with a different kind).

addModuleToScope :: NameSpaceId -> Name -> AbstractModule -> Scope -> Scope

Add a module to a scope.

data UsingOrHiding

When we get here we cannot have both using and hiding.

Constructors

UsingOnly [ImportedName]

HidingOnly [ImportedName]

usingOrHiding :: ImportDirective -> UsingOrHiding

applyImportDirective :: ImportDirective -> Scope -> Scope

Apply an ImportDirective to a scope:

rename keys (C.Name) according to renaming;

for untouched keys, either of

a) remove keys according to hiding, or b) filter keys according to using.

Both steps could be done in one pass, by first preparing key-filtering functions C.Name -> Maybe C.Name for defined names and module names. However, the penalty of doing it in two passes should not be too high. (Doubling the run time.)

applyImportDirective_

Arguments

:: ImportDirective

-> Scope

-> (Scope, (Set Name, Set Name))
Merged scope, clashing names, clashing module names.

Version of applyImportDirective that also returns sets of name and module name clashes introduced by renaming to identifiers that are already imported by using or lack of hiding.

renameCanonicalNames :: Map QName QName -> Map ModuleName ModuleName -> Scope -> Scope

Rename the abstract names in a scope.

restrictPrivate :: Scope -> Scope

Remove private name space of a scope.

Should be a right identity for exportedNamesInScope. exportedNamesInScope . restrictPrivate == exportedNamesInScope.

restrictLocalPrivate :: ModuleName -> Scope -> Scope

Remove private things from the given module from a scope.

withoutPrivates :: ScopeInfo -> ScopeInfo

Filter privates out of a ScopeInfo

disallowGeneralizedVars :: Scope -> Scope

Disallow using generalized variables from the scope

inScopeBecause :: (WhyInScope -> WhyInScope) -> Scope -> Scope

Add an explanation to why things are in scope.

publicModules :: ScopeInfo -> Map ModuleName Scope

Get the public parts of the public modules of a scope

publicNames :: ScopeInfo -> Set AbstractName

publicNamesOfModules :: Map ModuleName Scope -> [AbstractName]

everythingInScope :: ScopeInfo -> NameSpace

concreteNamesInScope :: ScopeInfo -> Set QName

Get all concrete names in scope. Includes bound variables.

scopeLookup :: InScope a => QName -> ScopeInfo -> [a]

Look up a name in the scope

scopeLookup' :: InScope a => QName -> ScopeInfo -> [(a, Access)]

Inverse look-up

data AllowAmbiguousNames

Constructors

AmbiguousAnything
Used for instance arguments to check whether a name is in scope, but we do not care whether is is ambiguous

AmbiguousConProjs
Ambiguous constructors, projections, or pattern synonyms.

AmbiguousNothing

isNameInScope :: QName -> ScopeInfo -> Bool

isNameInScopeUnqualified :: QName -> ScopeInfo -> Bool

inverseScopeLookupName :: QName -> ScopeInfo -> [QName]

Find the concrete names that map (uniquely) to a given abstract qualified name. Sort by number of modules in the qualified name, unqualified names first.

inverseScopeLookupName' :: AllowAmbiguousNames -> QName -> ScopeInfo -> [QName]

inverseScopeLookupName'' :: AllowAmbiguousNames -> QName -> ScopeInfo -> Maybe NameMapEntry

A version of inverseScopeLookupName that also delivers the KindOfName. Used in highlighting.

inverseScopeLookupModule :: ModuleName -> ScopeInfo -> [QName]

Find the concrete names that map (uniquely) to a given abstract module name. Sort by length, shortest first.

inverseScopeLookupModule' :: AllowAmbiguousNames -> ModuleName -> ScopeInfo -> [QName]

recomputeInScopeSet :: ScopeInfo -> Set QName

recomputeInverseNamesAndModules :: ScopeInfo -> (NameMap, ModuleMap)

recomputeInverseScope' :: ScopeInfo -> ScopeInfo

Update binding site

class SetBindingSite a where

Set the nameBindingSite in an abstract name.

Minimal complete definition

Nothing

Methods

setBindingSite :: Range -> a -> a

default setBindingSite :: forall b (t :: Type -> Type). (SetBindingSite b, Functor t, t b ~ a) => Range -> a -> a

(Debug) printing

prettyNameSpace :: NameSpace -> [Doc]

newtype PrettyWithInstance a

Constructors

PrettyWithInstance a

blockOfLines :: Doc -> [Doc] -> [Doc]

Add first string only if list is non-empty.

Boring instances

Orphan instances

Pretty Suffix

Instance details

Methods

pretty :: Suffix -> Doc

prettyPrec :: Int -> Suffix -> Doc

prettyList :: [Suffix] -> Doc