Agda.Syntax.Scope.Flat

Description

Flattened scopes.

Documentation

data FlatScope

Flattened scopes.

flattenScope :: [[Name]] -> ScopeInfo -> FlatScope

Compute a flattened scope. Only include unqualified names or names qualified by modules in the first argument.

getDefinedNames :: KindsOfNames -> FlatScope -> [List1 NewNotation]

Compute all defined names in scope and their fixities/notations. Note that overloaded names (constructors) can have several fixities/notations. Then we mergeNotations. (See issue 1194.)

localNames :: FlatScope -> ScopeM ([QName], [NewNotation])

Compute all names (first component) and operators/notations (second component) in scope.