Agda.Syntax.TopLevelModuleName

Documentation

type TopLevelModuleNameParts = List1 Text

A top-level module name has one or more name parts.

data RawTopLevelModuleName

Raw top-level module names (with linear-time comparisons).

Constructors

RawTopLevelModuleName

Fields

rawModuleNameRange :: Range

rawModuleNameParts :: TopLevelModuleNameParts

rawTopLevelModuleNameToString :: RawTopLevelModuleName -> String

Turns a raw top-level module name into a string.

hashRawTopLevelModuleName :: RawTopLevelModuleName -> ModuleNameHash

Hashes a raw top-level module name.

rawTopLevelModuleNameForQName :: QName -> RawTopLevelModuleName

Turns a qualified name into a RawTopLevelModuleName. The qualified name is assumed to represent a top-level module name.

rawTopLevelModuleNameForModuleName :: ModuleName -> RawTopLevelModuleName

Computes the RawTopLevelModuleName corresponding to the given module name, which is assumed to represent a top-level module name.

Precondition: The module name must be well-formed.

rawTopLevelModuleNameForModule :: Module -> RawTopLevelModuleName

Computes the top-level module name.

Precondition: The Declaration has to be well-formed. This means that there are only allowed declarations before the first module declaration, typically import declarations. See spanAllowedBeforeModule.

data TopLevelModuleName

Top-level module names (with constant-time comparisons).

Constructors

TopLevelModuleName

Fields

moduleNameRange :: Range

moduleNameId :: !ModuleNameHash

moduleNameParts :: TopLevelModuleNameParts

lensTopLevelModuleNameParts :: Lens' TopLevelModuleNameParts TopLevelModuleName

A lens focusing on the moduleNameParts.

rawTopLevelModuleName :: TopLevelModuleName -> RawTopLevelModuleName

Converts a top-level module name to a raw top-level module name.

unsafeTopLevelModuleName :: RawTopLevelModuleName -> ModuleNameHash -> TopLevelModuleName

Converts a raw top-level module name and a hash to a top-level module name.

This function does not ensure that there are no hash collisions, that is taken care of by topLevelModuleName.

topLevelModuleNameToQName :: TopLevelModuleName -> QName

A corresponding QName. The range of each Name part is the whole range of the TopLevelModuleName.

moduleNameToFileName :: TopLevelModuleName -> String -> FilePath

Turns a top-level module name into a file name with the given suffix.

projectRoot :: AbsolutePath -> TopLevelModuleName -> AbsolutePath

Finds the current project's "root" directory, given a project file and the corresponding top-level module name.

Example: If the module "A.B.C" is located in the file "fooABC.agda", then the root is "foo".

Precondition: The module name must be well-formed.