Agda.Syntax.Translation.AbstractToConcrete

Description

The translation of abstract syntax to concrete syntax has two purposes. First it allows us to pretty print abstract syntax values without having to write a dedicated pretty printer, and second it serves as a sanity check for the concrete to abstract translation: translating from concrete to abstract and then back again should be (more or less) the identity.

Documentation

class ToConcrete a where

Minimal complete definition

Nothing

Associated Types

type ConOfAbs a

Methods

toConcrete :: a -> AbsToCon (ConOfAbs a)

bindToConcrete :: a -> (ConOfAbs a -> AbsToCon b) -> AbsToCon b

toConcreteCtx :: ToConcrete a => Precedence -> a -> AbsToCon (ConOfAbs a)

Translate something in a context of the given precedence.

abstractToConcrete_ :: (ToConcrete a, MonadAbsToCon m) => a -> m (ConOfAbs a)

abstractToConcreteScope :: (ToConcrete a, MonadAbsToCon m) => ScopeInfo -> a -> m (ConOfAbs a)

abstractToConcreteHiding :: (LensHiding i, ToConcrete a, MonadAbsToCon m) => i -> a -> m (ConOfAbs a)

runAbsToCon :: MonadAbsToCon m => AbsToCon c -> m c

data RangeAndPragma

Constructors

RangeAndPragma Range Pragma

abstractToConcreteCtx :: (ToConcrete a, MonadAbsToCon m) => Precedence -> a -> m (ConOfAbs a)

withScope :: ScopeInfo -> AbsToCon a -> AbsToCon a

preserveInteractionIds :: AbsToCon a -> AbsToCon a

type MonadAbsToCon m = (MonadTCEnv m, ReadTCState m, MonadStConcreteNames m, HasOptions m, HasBuiltins m, MonadDebug m)

We need: - Read access to the AbsToCon environment - Read access to the TC environment - Read access to the TC state - Read and write access to the stConcreteNames part of the TC state - Read access to the options - Permission to print debug messages

data AbsToCon a

data Env

noTakenNames :: AbsToCon a -> AbsToCon a

lookupQName :: AllowAmbiguousNames -> QName -> AbsToCon QName