Agda.Syntax.Translation.ConcreteToAbstract

Description

Translation from Agda.Syntax.Concrete to Agda.Syntax.Abstract. Involves scope analysis, figuring out infix operator precedences and tidying up definitions.

Documentation

class ToAbstract c where

Things that can be translated to abstract syntax are instances of this class.

Associated Types

type AbsOfCon c

Methods

toAbstract :: c -> ScopeM (AbsOfCon c)

localToAbstract :: ToAbstract c => c -> (AbsOfCon c -> ScopeM b) -> ScopeM b

This operation does not affect the scope, i.e. the original scope is restored upon completion.

concreteToAbstract_ :: ToAbstract c => c -> ScopeM (AbsOfCon c)

concreteToAbstract :: ToAbstract c => ScopeInfo -> c -> ScopeM (AbsOfCon c)

data TopLevel a

Temporary data type to scope check a file.

Constructors

TopLevel

Fields

topLevelSourceFile :: SourceFile

The file from which we loaded this module.

topLevelExpectedName :: TopLevelModuleName

The expected module name (coming from the import statement that triggered scope checking this file).

topLevelTheThing :: a

The file content.

data TopLevelInfo

Constructors

TopLevelInfo

Fields

topLevelDecls :: [Declaration]

topLevelScope :: ScopeInfo

as seen from inside the module

topLevelModuleName :: TopLevelInfo -> ModuleName

The top-level module name.

importPrimitives :: ScopeM [Declaration]

Declaration open import Agda.Primitive using (Set) when optImportSorts. Prop is added when optProp, and SSet when optTwoLevel.

checkAttributes :: Attributes -> ScopeM ()

Raises an error if the list of attributes contains an unsupported attribute.