Agda.Syntax.Translation.InternalToAbstract

Description

Translating from internal syntax to abstract syntax. Enables nice pretty printing of internal syntax.

TODO

numbers on metas

fake dependent functions to independent functions

meta parameters

shadowing

Documentation

class Reify i where

Minimal complete definition

reify

Associated Types

type ReifiesTo i

Methods

reify :: MonadReify m => i -> m (ReifiesTo i)

reifyWhen :: MonadReify m => Bool -> i -> m (ReifiesTo i)

type MonadReify (m :: Type -> Type) = (PureTCM m, MonadInteractionPoints m, MonadFresh NameId m)

data NamedClause

Constructors

NamedClause QName Bool Clause
Also tracks whether module parameters should be dropped from the patterns.

reifyPatterns :: MonadReify m => [NamedArg DeBruijnPattern] -> m [NamedArg Pattern]

Assumes that pattern variables have been added to the context already. Picks pattern variable names from context.

reifyUnblocked :: Reify i => i -> TCM (ReifiesTo i)

Like reify but instantiates blocking metas, useful for reporting.

blankNotInScope :: (MonadTCEnv m, MonadDebug m, BlankVars a) => a -> m a

blankNotInScope e replaces variables in expression e with _ if they are currently not in scope.

reifyDisplayFormP

Arguments

:: MonadReify m

=> QName
LHS head symbol

-> Patterns
Patterns to be taken into account to find display form.

-> Patterns
Remaining trailing patterns ("with patterns").

-> m (QName, Patterns)
New head symbol and new patterns.

reifyDisplayFormP tries to recursively rewrite a lhs with a display form.

Note: we are not necessarily in the empty context upon entry!