Agda.Termination.CallMatrix

Documentation

type ArgumentIndex = Int

Call matrix indices = function argument indices.

Machine integer Int is sufficient, since we cannot index more arguments than we have addresses on our machine.

newtype CallMatrix' a

Call matrices.

A call matrix for a call f --> g has dimensions ar(g) × ar(f).

Each column corresponds to one formal argument of caller f. Each row corresponds to one argument in the call to g.

In the presence of dot patterns, a call argument can be related to several different formal arguments of f.

See e.g. testsucceedDotPatternTermination.agda:

    data D : Nat -> Set where
      cz : D zero
      c1 : forall n -> D n -> D (suc n)
      c2 : forall n -> D n -> D n

    f : forall n -> D n -> Nat
    f .zero    cz        = zero
    f .(suc n) (c1  n d) = f n (c2 n d)
    f n        (c2 .n d) = f n d
  

Call matrices (without guardedness) are

          -1 -1   n < suc n  and       n <  c1 n d
           ?  =                   c2 n d <= c1 n d

           = -1   n <= n     and  n < c2 n d
           ? -1                   d < c2 n d
  

Here is a part of the original documentation for call matrices (kept for historical reasons):

This datatype encodes information about a single recursive function application. The columns of the call matrix stand for source function arguments (patterns). The rows of the matrix stand for target function arguments. Element (i, j) in the matrix should be computed as follows:

lt (less than) if the j-th argument to the target function is structurally strictly smaller than the i-th pattern.

le (less than or equal) if the j-th argument to the target function is structurally smaller than the i-th pattern.

unknown otherwise.

Constructors

CallMatrix

Fields

mat :: Matrix ArgumentIndex a

type CallMatrix = CallMatrix' Order

mapCallMatrix :: (Matrix ArgumentIndex a -> Matrix ArgumentIndex a) -> CallMatrix' a -> CallMatrix' a

class CallComb a where

Call matrix multiplication and call combination.

Methods

(>*<) :: a -> a -> a

Call matrix augmented with path information.

data CallMatrixAug cinfo

Call matrix augmented with path information.

Constructors

CallMatrixAug

Fields

augCallMatrix :: CallMatrix

The matrix of the (composed call).

augCallInfo :: cinfo

Meta info, like call path.

noAug :: Monoid cinfo => CallMatrix -> CallMatrixAug cinfo

Non-augmented call matrix.

mapCallMatrixAug :: (CallMatrix -> CallMatrix) -> CallMatrixAug cinfo -> CallMatrixAug cinfo

Sets of incomparable call matrices augmented with path information.

newtype CMSet cinfo

Sets of incomparable call matrices augmented with path information. Use overloaded null, empty, singleton, mappend.

Constructors

CMSet

Fields

cmSet :: Favorites (CallMatrixAug cinfo)

mapCMSet :: (Favorites (CallMatrixAug cinfo) -> Favorites (CallMatrixAug cinfo)) -> CMSet cinfo -> CMSet cinfo

insert :: CallMatrixAug cinfo -> CMSet cinfo -> CMSet cinfo

Insert into a call matrix set.

union :: CMSet cinfo -> CMSet cinfo -> CMSet cinfo

Union two call matrix sets.

toList :: CMSet cinfo -> [CallMatrixAug cinfo]

Convert into a list of augmented call matrices.

Printing