Agda.Termination.CutOff

Description

Defines CutOff type which is used in Agda.Interaction.Options. This module's purpose is to eliminate the dependency of Agda.TypeChecking.Monad.Base on the termination checker and everything it imports.

Documentation

data CutOff

Cut off structural order comparison at some depth in termination checker?

Constructors

CutOff !Int
c >= 0 means: record decrease up to including c+1.

DontCutOff

defaultCutOff :: CutOff

The default termination depth.