Agda.Termination.Monad

Description

The monad for the termination checker.

The termination monad TerM is an extension of the type checking monad TCM by an environment with information needed by the termination checker.

Documentation

data Target

The target of the function we are checking.

Constructors

TargetDef QName
The target of recursion is a record, data, or unreducible Def.

TargetRecord
We are termination-checking a record.

TargetOther
None of the above two or unknown.

type Guarded = Order

The current guardedness level.

data TerEnv

The termination environment.

Constructors

TerEnv

Fields

terUseDotPatterns :: Bool

Are we mining dot patterns to find evindence of structal descent?

terSizeSuc :: Maybe QName

The name of size successor, if any.

terSharp :: Maybe QName

The name of the delay constructor (sharp), if any.

terCutOff :: CutOff

Depth at which to cut off the structural order.

terCurrent :: QName

The name of the function we are currently checking.

terMutual :: MutualNames

The names of the functions in the mutual block we are checking. This includes the internally generated functions (with, extendedlambda, coinduction).

terUserNames :: Set QName

The list of name actually appearing in the file (abstract syntax). Excludes the internally generated functions.

terHaveInlinedWith :: Bool

Does the actual clause result from with-inlining? (If yes, it may be ill-typed.)

terTarget :: Target

Target type of the function we are currently termination checking. Only the constructors of Target are considered guarding.

terMaskArgs :: ListInf Bool

Only consider the notMasked False arguments for establishing termination. See issue #1023.

terMaskResult :: Bool

Only consider guardedness if False (not masked).

_terSizeDepth :: Int

How many SIZELT relations do we have in the context (= clause telescope). Used to approximate termination for metas in call args.

terPatterns :: MaskedDeBruijnPatterns

The patterns of the clause we are checking.

terPatternsRaise :: !Int

Number of additional binders we have gone under (and consequently need to raise the patterns to compare to terms). Updated during call graph extraction, hence strict.

terGuarded :: !Guarded

The current guardedness status. Changes as we go deeper into the term. Updated during call graph extraction, hence strict.

terUseSizeLt :: Bool

When extracting usable size variables during construction of the call matrix, can we take the variable for use with SIZELT constraints from the context? Yes, if we are under an inductive constructor. No, if we are under a record constructor. (See issue #1015).

terUsableVars :: VarSet

Pattern variables that can be compared to argument variables using SIZELT.

defaultTerEnv :: TerEnv

An empty termination environment.

Values are set to a safe default meaning that with these initial values the termination checker will not miss termination errors it would have seen with better settings of these values.

Values that do not have a safe default are set to IMPOSSIBLE.

class (Functor m, Monad m) => MonadTer (m :: Type -> Type) where

Termination monad service class.

Minimal complete definition

terAsk, terLocal

Methods

terAsk :: m TerEnv

terLocal :: (TerEnv -> TerEnv) -> m a -> m a

terAsks :: (TerEnv -> a) -> m a

newtype TerM a

Termination monad.

Constructors

TerM

Fields

terM :: ReaderT TerEnv TCM a

runTer :: TerEnv -> TerM a -> TCM a

Generic run method for termination monad.

runTerDefault :: TerM a -> TCM a

Run TerM computation in default environment (created from options).

Modifiers and accessors for the termination environment in the monad.

terGetUseDotPatterns :: TerM Bool

terSetUseDotPatterns :: Bool -> TerM a -> TerM a

terGetSizeSuc :: TerM (Maybe QName)

terGetCurrent :: TerM QName

terSetCurrent :: QName -> TerM a -> TerM a

terGetSharp :: TerM (Maybe QName)

terGetCutOff :: TerM CutOff

terGetMutual :: TerM MutualNames

terGetUserNames :: TerM (Set QName)

terGetTarget :: TerM Target

terSetTarget :: Target -> TerM a -> TerM a

terGetHaveInlinedWith :: TerM Bool

terSetHaveInlinedWith :: TerM a -> TerM a

terGetMaskArgs :: TerM (ListInf Bool)

terSetMaskArgs :: ListInf Bool -> TerM a -> TerM a

terGetMaskResult :: TerM Bool

terSetMaskResult :: Bool -> TerM a -> TerM a

terGetPatterns :: TerM MaskedDeBruijnPatterns

terSetPatterns :: MaskedDeBruijnPatterns -> TerM a -> TerM a

terRaise :: TerM a -> TerM a

terGetGuarded :: TerM Guarded

terModifyGuarded :: (Order -> Order) -> TerM a -> TerM a

terSetGuarded :: Order -> TerM a -> TerM a

terUnguarded :: TerM a -> TerM a

terSizeDepth :: Lens' TerEnv Int

Lens for _terSizeDepth.

terGetUsableVars :: TerM VarSet

Lens for terUsableVars.

terModifyUsableVars :: (VarSet -> VarSet) -> TerM a -> TerM a

terSetUsableVars :: VarSet -> TerM a -> TerM a

terGetUseSizeLt :: TerM Bool

Lens for terUseSizeLt.

terModifyUseSizeLt :: (Bool -> Bool) -> TerM a -> TerM a

terSetUseSizeLt :: Bool -> TerM a -> TerM a

withUsableVars :: UsableSizeVars a => a -> TerM b -> TerM b

Compute usable vars from patterns and run subcomputation.

conUseSizeLt :: QName -> TerM a -> TerM a

Set terUseSizeLt when going under constructor c.

projUseSizeLt :: QName -> TerM a -> TerM a

Set terUseSizeLt for arguments following projection q. We disregard j<i after a non-coinductive projection. However, the projection need not be recursive (Issue 1470).

isProjectionButNotCoinductive :: MonadTCM tcm => QName -> tcm Bool

For termination checking purposes flat should not be considered a projection. That is, it flat doesn't preserve either structural order or guardedness like other projections do. Andreas, 2012-06-09: the same applies to projections of recursive records.

isCoinductiveProjection :: MonadTCM tcm => Bool -> QName -> tcm Bool

Check whether a projection belongs to a coinductive record and is actually recursive. E.g. @ isCoinductiveProjection (Stream.head) = return False

isCoinductiveProjection (Stream.tail) = return True @

De Bruijn pattern stuff

patternDepth :: Pattern' a -> Int

How long is the path to the deepest atomic pattern?

unusedVar :: DeBruijnPattern

A dummy pattern used to mask a pattern that cannot be used for structural descent.

class UsableSizeVars a where

Extract variables from DeBruijnPatterns that could witness a decrease via a SIZELT constraint.

These variables must be under an inductive constructor (with no record constructor in the way), or after a coinductive projection (with no inductive one in the way).

Methods

usableSizeVars :: a -> TerM VarSet

Masked patterns (which are not eligible for structural descent, only for size descent)

type MaskedDeBruijnPatterns = [Masked DeBruijnPattern]

data Masked a

Constructors

Masked

Fields

getMask :: Bool

True if thing not eligible for structural descent.

getMasked :: a

Thing.

masked :: a -> Masked a

notMasked :: a -> Masked a

Call pathes

newtype CallPath

Call paths.

Constructors

CallPath (DList CallInfo)

callInfos :: CallPath -> [CallInfo]

The calls making up the call path.

Size depth estimation

class TerSetSizeDepth b where

A very crude way of estimating the SIZELT chains i > j > k in context. Returns 3 in this case. Overapproximates.

Methods

terSetSizeDepth :: b -> TerM a -> TerM a