Agda.Termination.Order

Description

An Abstract domain of relative sizes, i.e., differences between size of formal function parameter and function argument in recursive call; used in the termination checker.

Structural orderings

data Order

In the paper referred to above, there is an order R with Unknown <= Le <= Lt.

This is generalized to Unknown <= 'Decr k' where Decr 1 replaces Lt and Decr 0 replaces Le. A negative decrease means an increase. The generalization allows the termination checker to record an increase by 1 which can be compensated by a following decrease by 2 which results in an overall decrease.

However, the termination checker of the paper itself terminates because there are only finitely many different call-matrices. To maintain termination of the terminator we set a cutoff point which determines how high the termination checker can count. This value should be set by a global or file-wise option.

See Call for more information.

TODO: document orders which are call-matrices themselves.

Constructors

Decr !Bool !Int
Decrease of callee argument wrt. caller parameter.

The Bool indicates whether the decrease (if any) is usable. In any chain, there needs to be one usable decrease. Unusable decreases come from SIZELT constraints which are not in inductive pattern match or a coinductive copattern match. See issue #2331.

UPDATE: Andreas, 2017-07-26: Feature #2331 is unsound due to size quantification in terms. While the infrastructure for usable/unusable decrease remains in place, no unusable decreases are generated by TermCheck.

Unknown
No relation, infinite increase, or increase beyond termination depth.

Mat !(Matrix Int Order)
Matrix-shaped order, currently UNUSED.

decr :: (?cutoff :: CutOff) => Bool -> Int -> Order

Smart constructor for Decr k :: Order which cuts off too big values.

Possible values for k: - ?cutoff <= k <= ?cutoff + 1.

increase :: Int -> Order -> Order

Raw increase which does not cut off.

decrease :: Int -> Order -> Order

Raw decrease which does not cut off.

setUsability :: Bool -> Order -> Order

(.*.) :: (?cutoff :: CutOff) => Order -> Order -> Order

Multiplication of Orders. (Corresponds to sequential composition.)

supremum :: (?cutoff :: CutOff) => [Order] -> Order

The supremum of a (possibly empty) list of Orders. More information (i.e., more decrease) is bigger. Unknown is no information, thus, smallest.

infimum :: (?cutoff :: CutOff) => [Order] -> Order

The infimum of a (non empty) list of Orders. Gets the worst information. Unknown is the least element, thus, dominant.

orderSemiring :: (?cutoff :: CutOff) => Semiring Order

We use a record for semiring instead of a type class since implicit arguments cannot occur in instance constraints, like instance (?cutoff :: Int) => SemiRing Order.

le :: Order

le, lt, decreasing, unknown: for backwards compatibility, and for external use.

lt :: Order

Usable decrease.

unknown :: Order

orderMat :: Matrix Int Order -> Order

Smart constructor for matrix shaped orders, avoiding empty and singleton matrices.

collapseO :: (?cutoff :: CutOff) => Order -> Order

nonIncreasing :: Order -> Bool

decreasing :: Order -> Bool

Decreasing and usable?

isDecr :: Order -> Bool

Matrix-shaped order is decreasing if any diagonal element is decreasing.

class NotWorse a where

A partial order, aimed at deciding whether a call graph gets worse during the completion.

Methods

notWorse :: a -> a -> Bool

isOrder :: (?cutoff :: CutOff) => Order -> Bool