Agda.TypeChecking.Abstract

Description

Functions for abstracting terms over other terms.

Documentation

abstractType

Arguments

:: Range
Range of the rewrite expression, if any, otherwise empty.

-> Type
Type of the term to abstract.

-> Term
Term to abstract.

-> Type
Type to abstract in.

-> TCM Type
Type with hole (de Bruijn index 0) for the abstracted term.

abstractType r a v b[v] = b where v : a.

piAbstractTerm :: ArgInfo -> Term -> Type -> Type -> TCM Type

piAbstractTerm NotHidden v a b[v] = (w : a) -> b[w] piAbstractTerm Hidden v a b[v] = {w : a} -> b[w]

piAbstract :: Arg (Term, EqualityView) -> Type -> TCM Type

piAbstract (v, a) b[v] = (w : a) -> b[w]

For the inspect idiom, it does something special: piAbstract (v, a) b[v, refl v] = (w : a) (w' : Eq a w v) -> b[w, w'] (Not all refl v@ subterms are abstracted: see caveat below.)

For rewrite, it does something special: piAbstract (prf, Eq a v v') b[v,prf] = (w : a) (w' : Eq a w v') -> b[w, w']

class IsPrefixOf a where

isPrefixOf u v = Just es if v == u applyE es.

Methods

isPrefixOf :: a -> a -> Maybe Elims

abstractTerm :: Type -> Term -> Type -> Term -> TCM Term

class AbsTerm a where

Methods

absTerm

Arguments

:: Nat
De Bruijn index that should be the placeholder for the abstracted term.

-> Term
Term to abstract.

-> a
Where to abstract.

-> a
If the given de Bruijn index is free in the result, abstraction actually happened.

subst j u . absTerm j u == id

swap01 :: TermSubst a => a -> a

This swaps var 0 and var 1.

Equality of terms for the sake of with-abstraction.

class EqualSy a where

Methods

equalSy :: a -> a -> Bool