Agda.TypeChecking.CompiledClause

Description

Case trees.

After coverage checking, pattern matching is translated to case trees, i.e., a tree of successive case splits on one variable at a time.

Documentation

data WithArity c

Constructors

WithArity

Fields

arity :: Int

content :: c

data Case c

Branches in a case tree.

Constructors

Branches

Fields

projPatterns :: Bool

We are constructing a record here (copatterns). conBranches lists projections.

conBranches :: Map QName (WithArity c)

Map from constructor (or projection) names to their arity and the case subtree. (Projections have arity 0.)

etaBranch :: Maybe (ConHead, WithArity c)

Eta-expand with the given (eta record) constructor. If this is present, there should not be any conBranches or litBranches.

litBranches :: Map Literal c

Map from literal to case subtree.

catchAllBranch :: Maybe c

(Possibly additional) catch-all clause.

fallThrough :: Maybe Bool

(if True) In case of non-canonical argument use catchAllBranch.

lazyMatch :: Bool

Lazy pattern match. Requires single (non-copattern) branch with no lit branches and no catch-all.

data CompiledClauses' a

Case tree with bodies.

Constructors

Case (Arg Int) (Case (CompiledClauses' a))
Case n bs stands for a match on the n-th argument (counting from zero) with bs as the case branches. If the n-th argument is a projection, we have only conBranches with arity 0.

Done [Arg ArgName] a
Done xs b stands for the body b where the xs contains hiding and name suggestions for the free variables. This is needed to build lambdas on the right hand side for partial applications which can still reduce.

Fail [Arg ArgName]
Absurd case. Add the free variables here as well so we can build correct number of lambdas for strict backends. (#4280)

type CompiledClauses = CompiledClauses' Term

litCase :: Literal -> c -> Case c

conCase :: QName -> Bool -> WithArity c -> Case c

etaCase :: ConHead -> WithArity c -> Case c

projCase :: QName -> c -> Case c

catchAll :: c -> Case c

checkLazyMatch :: Case c -> Case c

Check that the requirements on lazy matching (single inductive case) are met, and set lazy to False otherwise.

hasCatchAll :: CompiledClauses -> Bool

Check whether a case tree has a catch-all clause.

hasProjectionPatterns :: CompiledClauses -> Bool

Check whether a case tree has any projection patterns

Pretty instances.

prettyMap_ :: (Pretty k, Pretty v) => Map k v -> [Doc]

KillRange instances.

TermLike instances