Agda.TypeChecking.Conversion

Documentation

antiUnify :: MonadConversion m => ProblemId -> Type -> Term -> Term -> m Term

When comparing argument spines (in compareElims) where the first arguments don't match, we keep going, substituting the anti-unification of the two terms in the telescope. More precisely:

@ (u = v : A)[pid] w = antiUnify pid A u v us = vs : Δ[w/x] ------------------------------------------------------------- u us = v vs : (x : A) Δ @

The simplest case of anti-unification is to return a fresh metavariable (created by blockTermOnProblem), but if there's shared structure between the two terms we can expose that.

This is really a crutch that lets us get away with things that otherwise would require heterogenous conversion checking. See for instance issue #2384.

antiUnifyArgs :: MonadConversion m => ProblemId -> Dom Type -> Arg Term -> Arg Term -> m (Arg Term)

antiUnifyElims :: MonadConversion m => ProblemId -> Type -> Term -> Elims -> Elims -> m Term

antiUnifyType :: MonadConversion m => ProblemId -> Type -> Type -> m Type

assignE :: MonadConversion m => CompareDirection -> MetaId -> Elims -> Term -> CompareAs -> (Term -> Term -> m ()) -> m ()

Try to assign meta. If meta is projected, try to eta-expand and run conversion check again.

bothAbsurd :: MonadConversion m => QName -> QName -> m Bool

coerce :: (MonadConversion m, MonadTCM m) => Comparison -> Term -> Type -> Type -> m Term

coerce v a b coerces v : a to type b, returning a v' : b with maybe extra hidden applications or hidden abstractions.

In principle, this function can host coercive subtyping, but currently it only tries to fix problems with hidden function types.

coerceSize :: MonadConversion m => (Type -> Type -> m ()) -> Term -> Type -> Type -> m ()

Account for situations like k : (Size< j) <= (Size< k + 1)

Actually, the semantics is (Size<= k) ∩ (Size< j) ⊆ rhs which gives a disjunctive constraint. Mmmh, looks like stuff TODO.

For now, we do a cheap heuristics.

compareArgs :: MonadConversion m => [Polarity] -> [IsForced] -> Type -> Term -> Args -> Args -> m ()

Type-directed equality on argument lists

compareAs :: MonadConversion m => Comparison -> CompareAs -> Term -> Term -> m ()

Type directed equality on terms or types.

compareAs' :: MonadConversion m => Comparison -> CompareAs -> Term -> Term -> m ()

compareAsDir :: MonadConversion m => CompareDirection -> CompareAs -> Term -> Term -> m ()

compareAtom :: MonadConversion m => Comparison -> CompareAs -> Term -> Term -> m ()

Syntax directed equality on atomic values

compareAtomDir :: MonadConversion m => CompareDirection -> CompareAs -> Term -> Term -> m ()

compareDom

Arguments

:: MonadConversion m

=> Comparison
cmp The comparison direction

-> Dom Type
a1 The smaller domain.

-> Dom Type
a2 The other domain.

-> Abs Type
b1 The smaller codomain.

-> Abs Type
b2 The bigger codomain.

-> m ()
Error continuation

-> m ()
Success continuation

-> m ()

Check whether a1 cmp a2 and continue in context extended by a1.

compareElims :: MonadConversion m => [Polarity] -> [IsForced] -> Type -> Term -> [Elim] -> [Elim] -> m ()

compareElims pols a v els1 els2 performs type-directed equality on eliminator spines. t is the type of the head v.

compareInterval :: MonadConversion m => Comparison -> Type -> Term -> Term -> m ()

compareIrrelevant :: MonadConversion m => Type -> Term -> Term -> m ()

Compare two terms in irrelevant position. This always succeeds. However, we can dig for solutions of irrelevant metas in the terms we compare. (Certainly not the systematic solution, that'd be proof search...)

compareLevel :: MonadConversion m => Comparison -> Level -> Level -> m ()

compareMetas :: MonadConversion m => Comparison -> CompareAs -> MetaId -> Elims -> MetaId -> Elims -> m ()

Check whether x xArgs cmp y yArgs

compareSort :: MonadConversion m => Comparison -> Sort -> Sort -> m ()

compareTerm :: MonadConversion m => Comparison -> Type -> Term -> Term -> m ()

Type directed equality on values.

compareTerm' :: MonadConversion m => Comparison -> Type -> Term -> Term -> m ()

compareTermOnFace :: MonadConversion m => Comparison -> Term -> Type -> Term -> Term -> m ()

compareTermOnFace' :: MonadConversion m => (Substitution -> Comparison -> Type -> Term -> Term -> m ()) -> Comparison -> Term -> Type -> Term -> Term -> m ()

compareType :: MonadConversion m => Comparison -> Type -> Type -> m ()

Equality on Types

compareWithPol :: MonadConversion m => Polarity -> (Comparison -> a -> a -> m ()) -> a -> a -> m ()

computeElimHeadType :: MonadConversion m => QName -> Elims -> Elims -> m Type

Compute the head type of an elimination. For projection-like functions this requires inferring the type of the principal argument.

convError :: TypeError -> TCM ()

Ignore errors in irrelevant context.

equalAtom :: MonadConversion m => CompareAs -> Term -> Term -> m ()

equalLevel :: MonadConversion m => Level -> Level -> m ()

equalSort :: MonadConversion m => Sort -> Sort -> m ()

Check that the first sort equal to the second.

equalTerm :: MonadConversion m => Type -> Term -> Term -> m ()

equalTermOnFace :: MonadConversion m => Term -> Type -> Term -> Term -> m ()

equalTermOnFace φ A u v = _ , φ ⊢ u = v : A

equalType :: MonadConversion m => Type -> Type -> m ()

forallFaceMaps :: MonadConversion m => Term -> (IntMap Bool -> Blocker -> Term -> m a) -> (IntMap Bool -> Substitution -> m a) -> m [a]

guardPointerEquality :: MonadConversion m => a -> a -> String -> m () -> m ()

guardPointerEquality x y s m behaves as m if x and y are equal as pointers, or does nothing otherwise. Use with care, see the documentation for unsafeComparePointers

intersectVars :: Elims -> Elims -> Maybe [Bool]

intersectVars us vs checks whether all relevant elements in us and vs are variables, and if yes, returns a prune list which says True for arguments which are different and can be pruned.

isCanonical :: [Conj] -> Bool

leqConj :: MonadConversion m => Conj -> Conj -> m Bool

leqConj r q = r ≤ q in the I lattice, when r and q are conjuctions. ' (∧ r_i) ≤ (∧ q_j) iff ' (∧ r_i) ∧ (∧ q_j) = (∧ r_i) iff ' {r_i | i} ∪ {q_j | j} = {r_i | i} iff ' {q_j | j} ⊆ {r_i | i}

leqInterval :: MonadConversion m => [Conj] -> [Conj] -> m Bool

leqInterval r q = r ≤ q in the I lattice. (∨ r_i) ≤ (∨ q_j) iff ∀ i. ∃ j. r_i ≤ q_j

leqLevel :: MonadConversion m => Level -> Level -> m ()

leqSort :: MonadConversion m => Sort -> Sort -> m ()

Check that the first sort is less or equal to the second.

We can put SizeUniv below Inf, but otherwise, it is unrelated to the other universes.

leqType :: MonadConversion m => Type -> Type -> m ()

polFromCmp :: Comparison -> Polarity

sameVars :: Elims -> Elims -> Bool

Check if to lists of arguments are the same (and all variables). Precondition: the lists have the same length.

tryConversion :: (MonadConstraint m, MonadWarning m, MonadError TCErr m, MonadFresh ProblemId m) => m () -> m Bool

Try whether a computation runs without errors or new constraints (may create new metas, though). Restores state upon failure.

tryConversion' :: (MonadConstraint m, MonadWarning m, MonadError TCErr m, MonadFresh ProblemId m) => m a -> m (Maybe a)

Try whether a computation runs without errors or new constraints (may create new metas, though). Return Just the result upon success. Return Nothing and restore state upon failure.

type Conj = (IntMap BoolSet, [Term])

type MonadConversion (m :: Type -> Type) = (PureTCM m, MonadConstraint m, MonadMetaSolver m, MonadError TCErr m, MonadWarning m, MonadStatistics m, MonadFresh ProblemId m, MonadFresh Int m)

failConversion

Arguments

:: (MonadTCError m, HasBuiltins m)

=> Comparison
Equality or subtyping?

-> Term
See convErrLhs

-> Term
See convErrRhs

-> CompareAs
Will be converted to a FailedCompareAs.

-> m a

Throw a floating ConversionError, given the arguments of compareAs.