Agda.TypeChecking.Coverage.SplitTree

Description

Split tree for transforming pattern clauses into case trees.

The coverage checker generates a split tree from the clauses. The clause compiler uses it to transform clauses to case trees.

The initial problem is a set of clauses. The root node designates on which argument to split and has subtrees for all the constructors. Splitting continues until there is only a single clause left at each leaf of the split tree.

Documentation

type SplitTree = SplitTree' SplitTag

type SplitTrees = SplitTrees' SplitTag

data SplitTree' a

Abstract case tree shape.

Constructors

SplittingDone
No more splits coming. We are at a single, all-variable clause.

Fields

splitBindings :: Int

The number of variables bound in the clause

SplitAt
A split is necessary.

Fields

splitArg :: Arg Int

Arg. no to split at.

splitLazy :: LazySplit

splitTrees :: SplitTrees' a

Sub split trees.

data LazySplit

Constructors

LazySplit

StrictSplit

type SplitTrees' a = [(a, SplitTree' a)]

Split tree branching. A finite map from constructor names to splittrees A list representation seems appropriate, since we are expecting not so many constructors per data type, and there is no need for random access.

data SplitTag

Tag for labeling branches of a split tree. Each branch is associated to either a constructor or a literal, or is a catchall branch (currently only used for splitting on a literal type).

Constructors

SplitCon QName

SplitLit Literal

SplitCatchall

Printing a split tree

data SplitTreeLabel a

Constructors

SplitTreeLabel

Fields

lblConstructorName :: Maybe a

Nothing for root of split tree

lblSplitArg :: Maybe (Arg Int)

lblLazy :: LazySplit

lblBindings :: Maybe Int

toTree :: SplitTree' a -> Tree (SplitTreeLabel a)

Convert a split tree into a Tree (for printing).

toTrees :: SplitTrees' a -> Forest (SplitTreeLabel a)