Agda.TypeChecking.Errors

Documentation

renderError :: MonadTCM tcm => TCErr -> tcm String

prettyError :: MonadTCM tcm => TCErr -> tcm Doc

tcErrString :: TCErr -> String

prettyTCWarnings' :: Set TCWarning -> TCM [Doc]

prettyTCWarnings :: Set TCWarning -> TCM String

tcWarningsToError :: [TCWarning] -> TCM ()

Turns warnings, if any, into errors.

applyFlagsToTCWarningsPreserving :: HasOptions m => Set WarningName -> Set TCWarning -> m (Set TCWarning)

Depending which flags are set, one may happily ignore some warnings.

applyFlagsToTCWarnings :: HasOptions m => Set TCWarning -> m (Set TCWarning)

getAllUnsolvedWarnings :: (ReadTCState m, MonadWarning m, MonadTCM m) => m [TCWarning]

getAllWarningsPreserving :: (ReadTCState m, MonadWarning m, MonadTCM m) => Set WarningName -> WhichWarnings -> m (Set TCWarning)

getAllWarnings :: (ReadTCState m, MonadWarning m, MonadTCM m) => WhichWarnings -> m (Set TCWarning)

Collect all warnings that have accumulated in the state.

getAllWarningsOfTCErr :: TCErr -> TCM (Set TCWarning)

dropTopLevelModule :: MonadPretty m => QName -> m QName

Drops the filename component of the qualified name.

topLevelModuleDropper :: (MonadDebug m, MonadTCEnv m, ReadTCState m) => m (QName -> QName)

Produces a function which drops the filename component of the qualified name.

explainWhyInScope :: MonadPretty m => WhyInScopeData -> m Doc

class Verbalize a where

Methods

verbalize :: a -> String

Orphan instances

PrettyTCM ExecError

Instance details

Methods

prettyTCM :: MonadPretty m => ExecError -> m Doc

PrettyTCM GHCBackendError

Instance details

Methods

prettyTCM :: MonadPretty m => GHCBackendError -> m Doc

PrettyTCM InteractionError

Instance details

Methods

prettyTCM :: MonadPretty m => InteractionError -> m Doc

PrettyTCM JSBackendError

Instance details

Methods

prettyTCM :: MonadPretty m => JSBackendError -> m Doc

PrettyTCM MissingTypeSignatureInfo

Instance details

Methods

prettyTCM :: MonadPretty m => MissingTypeSignatureInfo -> m Doc

PrettyTCM NegativeUnification

Instance details

Methods

prettyTCM :: MonadPretty m => NegativeUnification -> m Doc

PrettyTCM SplitError

Instance details

Methods

prettyTCM :: MonadPretty m => SplitError -> m Doc

PrettyTCM TCErr

Instance details

Methods

prettyTCM :: MonadPretty m => TCErr -> m Doc

PrettyTCM TypeError

Instance details

Methods

prettyTCM :: MonadPretty m => TypeError -> m Doc

PrettyTCM UnificationFailure

Instance details

Methods

prettyTCM :: MonadPretty m => UnificationFailure -> m Doc

PrettyTCM UnquoteError

Instance details

Methods

prettyTCM :: MonadPretty m => UnquoteError -> m Doc