Agda.TypeChecking.Free

Description

Computing the free variables of a term.

The distinction between rigid and strongly rigid occurrences comes from: Jason C. Reed, PhD thesis, 2009, page 96 (see also his LFMTP 2009 paper)

The main idea is that x = t(x) is unsolvable if x occurs strongly rigidly in t. It might have a solution if the occurrence is not strongly rigid, e.g.

x = f -> suc (f (x ( y -> k))) has x = f -> suc (f (suc k))

Jason C. Reed, PhD thesis, page 106

The most general function here is freeVarMap, which returns returns occurrence information about every free variable. It is also a legacy function in the sense that it used to be the only function for computing free occurrences, and it came with various functions for taking views of the resulting VarMap. You can find these under the "Legacy API" here.

There are also also a bunch of specialized and optimized implementations of traversals here.

If you want to write a new function for computing free occurrence information, your task is essentially to write a ComputeFree instance and invoke freeVars. These together implement a generic traversal which has a good chance of being converted to decent code by GHC. You can look at examples here and also look at the comments at ComputeFree.

Documentation

type FlexRig = FlexRig' MetaSet

data FlexRig' a

Depending on the surrounding context of a variable, it's occurrence can be classified as flexible or rigid, with finer distinctions.

The constructors are listed in increasing order (wrt. information content).

Constructors

Flexible !a
In arguments of metas. The set of metas is used by 'NonLinMatch' to generate the right blocking information. The semantics is that the status of a variable occurrence may change if one of the metas in the set gets solved. We may say the occurrence is tainted by the meta variables in the set.

WeaklyRigid
In arguments to variables and definitions.

Unguarded
In top position, or only under inductive record constructors (unit).

StronglyRigid
Under at least one and only inductive constructors.

class Free t where

Methods

freeVars :: ComputeFree r => t -> Reader r (Collect r)

type FreeEnv c = FreeEnv' MetaSet IgnoreSorts c

data FreeEnv' a b c

The current context.

Constructors

FreeEnv

Fields

feExtra :: !b

Additional context, e.g., whether to ignore free variables in sorts.

feFlexRig :: !(FlexRig' a)

Are we flexible or rigid?

feModality :: !Modality

What is the current relevance and quantity?

feSingleton :: !(b -> Maybe Variable -> c)

Method to return a single variable.

data IgnoreSorts

Where should we skip sorts in free variable analysis?

Constructors

IgnoreNot
Do not skip.

IgnoreInAnnotations
Skip when annotation to a type.

IgnoreAll
Skip unconditionally.

class (Singleton MetaId a, Monoid a, Monoid c) => IsVarSet a c | c -> a where

Any representation c of a set of variables need to be able to be modified by a variable occurrence. This is to ensure that free variable analysis is compositional. For instance, it should be possible to compute `fv (v [u/x])` from `fv v` and `fv u`.

In algebraic terminology, a variable set a needs to be (almost) a left semimodule to the semiring VarOcc.

Methods

withVarOcc :: VarOcc' a -> c -> c

Laws * Respects monoid operations: ``` withVarOcc o mempty == mempty withVarOcc o (x <> y) == withVarOcc o x <> withVarOcc o y ``` * Respects VarOcc composition: ``` withVarOcc oneVarOcc = id withVarOcc (composeVarOcc o1 o2) = withVarOcc o1 . withVarOcc o2 ``` * Respects VarOcc aggregation: ``` withVarOcc (o1 <> o2) x = withVarOcc o1 x <> withVarOcc o2 x ``` Since the corresponding unit law may fail, ``` withVarOcc mempty x = mempty ``` it is not quite a semimodule.

class LensFlexRig o a | o -> a where

Methods

lensFlexRig :: Lens' o (FlexRig' a)

newtype VarCounts

Constructors

VarCounts

Fields

varCounts :: IntMap Int

type VarMap = VarMap' MetaSet

newtype VarMap' a

Constructors

VarMap

Fields

theVarMap :: TheVarMap' a

type VarOcc = VarOcc' MetaSet

data VarOcc' a

Occurrence of free variables is classified by several dimensions. Currently, we have FlexRig and Modality.

Constructors

VarOcc

Fields

varFlexRig :: !(FlexRig' a)

varModality :: !Modality

composeFlexRig :: Semigroup a => FlexRig' a -> FlexRig' a -> FlexRig' a

FlexRig composition (multiplicative operation of the semiring). For accumulating the context of a variable.

Flexible is dominant. Once we are under a meta, we are flexible regardless what else comes. We taint all variable occurrences under a meta by this meta.

WeaklyRigid is next in strength. Destroys strong rigidity.

StronglyRigid is still dominant over Unguarded.

Unguarded is the unit. It is the top (identity) context.

lookupVarMap :: Variable -> VarMap' a -> Maybe (VarOcc' a)

mapVarMap :: (TheVarMap' a -> TheVarMap' b) -> VarMap' a -> VarMap' b

Legacy API

allVars :: VarMap -> VarSet

filterVarMap :: (VarOcc -> Bool) -> VarMap -> VarSet

filterVarMapToList :: (VarOcc -> Bool) -> VarMap -> [Int]

flexibleVars :: VarMap -> IntMap MetaSet

Variables occuring in arguments of metas. These are only potentially free, depending how the meta variable is instantiated. The set contains the id's of the meta variables that this variable is an argument to.

freeVarMap :: Free t => t -> VarMap

Return information about every free variable.

isFlexible :: LensFlexRig o a => o -> Bool

isStronglyRigid :: LensFlexRig o a => o -> Bool

isUnguarded :: LensFlexRig o a => o -> Bool

isWeaklyRigid :: LensFlexRig o a => o -> Bool

rigidVars :: VarMap -> VarSet

Rigid variables: either strongly rigid, unguarded, or weakly rigid.

stronglyRigidVars :: VarMap -> VarSet

Variables under only and at least one inductive constructor(s).

unguardedVars :: VarMap -> VarSet

Variables at top or only under inductive record constructors λs and Πs. The purpose of recording these separately is that they can still become strongly rigid if put under a constructor whereas weakly rigid ones stay weakly rigid.

MetaSet

newtype MetaSet

A set of meta variables. Forms a monoid under union.

Constructors

MetaSet

Fields

theMetaSet :: HashSet MetaId

foldrMetaSet :: (MetaId -> a -> a) -> a -> MetaSet -> a

insertMetaSet :: MetaId -> MetaSet -> MetaSet

metaSetToBlocker :: MetaSet -> Blocker

Specialized traversals

allFreeVar :: Free t => (Int -> Bool) -> t -> Bool

Compute the conjunction of a predicate on free variables.

allFreeVarIgnoreAll :: Free t => (Int -> Bool) -> t -> Bool

Same as allFreeVar except occurrences in sorts of types are ignored.

anyFreeVar :: Free t => (Int -> Bool) -> t -> Bool

Compute the disjunction of a predicate on free variables.

anyFreeVarIgnoreAll :: (Int -> Bool) -> Term -> Bool

Same as anyFreeVar except occurrences in sorts of types are ignored.

closed :: Free t => t -> Bool

Is a term closed?

flexRigOccurrenceIn :: Free a => Nat -> a -> Maybe FlexRig

Compute FlexRig for a single variable.

freeIn :: Free t => Nat -> t -> Bool

Check if single variable occurs freely.

freeInIgnoringSorts :: Free a => Nat -> a -> Bool

Check if a single variable occurs freely outside of sorts of types.

freeVarCounts :: Free t => t -> VarCounts

Return how many times each free variable occurs.

freeVarList :: Free t => t -> [Int]

Compute a (possibly non-unique) list of free variables, in preorder traversal order.

freeVarMapIgnoreAnn :: Free t => t -> VarMap

Return information about every free variable, but ignore occurrences in sorts of types.

freeVarSet :: Free t => t -> VarSet

Compute the set of free variables.

isBinderUsed :: Free a => Abs a -> Bool

Is the variable bound by the abstraction actually used?

relevantInIgnoringSortAnn :: Free t => Nat -> t -> Bool

Does a variable occur relevantly and outside of sorts of types?

setRelInIgnoring :: Free t => VarSet -> t -> VarSet

Test for a set of variables whether each variable occurs relevantly and outside of sort annotations. Return the set of variables that don't occur.