Agda.TypeChecking.Free.Base

Set of meta variables.

newtype MetaSet

A set of meta variables. Forms a monoid under union.

Constructors

MetaSet

Fields

theMetaSet :: HashSet MetaId

insertMetaSet :: MetaId -> MetaSet -> MetaSet

foldrMetaSet :: (MetaId -> a -> a) -> a -> MetaSet -> a

metaSetToBlocker :: MetaSet -> Blocker

Flexible and rigid occurrences (semigroup)

data FlexRig' a

Depending on the surrounding context of a variable, it's occurrence can be classified as flexible or rigid, with finer distinctions.

The constructors are listed in increasing order (wrt. information content).

Constructors

Flexible !a
In arguments of metas. The set of metas is used by 'NonLinMatch' to generate the right blocking information. The semantics is that the status of a variable occurrence may change if one of the metas in the set gets solved. We may say the occurrence is tainted by the meta variables in the set.

WeaklyRigid
In arguments to variables and definitions.

Unguarded
In top position, or only under inductive record constructors (unit).

StronglyRigid
Under at least one and only inductive constructors.

type FlexRig = FlexRig' MetaSet

class LensFlexRig o a | o -> a where

Methods

lensFlexRig :: Lens' o (FlexRig' a)

isFlexible :: LensFlexRig o a => o -> Bool

isUnguarded :: LensFlexRig o a => o -> Bool

isWeaklyRigid :: LensFlexRig o a => o -> Bool

isStronglyRigid :: LensFlexRig o a => o -> Bool

addFlexRig :: Semigroup a => FlexRig' a -> FlexRig' a -> FlexRig' a

FlexRig aggregation (additive operation of the semiring). For combining occurrences of the same variable in subterms. This is a refinement of the max operation for FlexRig which would work if Flexible did not have the MetaSet as an argument. Now, to aggregate two Flexible occurrences, we union the involved MetaSets.

zeroFlexRig :: Monoid a => FlexRig' a

Unit for addFlexRig.

omegaFlexRig :: FlexRig' a

Absorptive for addFlexRig.

composeFlexRig :: Semigroup a => FlexRig' a -> FlexRig' a -> FlexRig' a

FlexRig composition (multiplicative operation of the semiring). For accumulating the context of a variable.

Flexible is dominant. Once we are under a meta, we are flexible regardless what else comes. We taint all variable occurrences under a meta by this meta.

WeaklyRigid is next in strength. Destroys strong rigidity.

StronglyRigid is still dominant over Unguarded.

Unguarded is the unit. It is the top (identity) context.

oneFlexRig :: FlexRig' a

Unit for composeFlexRig.

flexRigToBlocker :: FlexRig -> Blocker

Extract a blocker from an occurrence

Multi-dimensional feature vector for variable occurrence (semigroup)

data VarOcc' a

Occurrence of free variables is classified by several dimensions. Currently, we have FlexRig and Modality.

Constructors

VarOcc

Fields

varFlexRig :: !(FlexRig' a)

varModality :: !Modality

type VarOcc = VarOcc' MetaSet

topVarOcc :: VarOcc' a

The absorptive element of variable occurrence under aggregation: strongly rigid, relevant.

composeVarOcc :: Semigroup a => VarOcc' a -> VarOcc' a -> VarOcc' a

First argument is the outer occurrence (context) and second is the inner. This multiplicative operation is to modify an occurrence under a context.

oneVarOcc :: VarOcc' a

constructorFlexRig :: ConHead -> Elims -> FlexRig' a

What's the rigidity of a constructor?

Storing variable occurrences (semimodule).

class (Singleton MetaId a, Monoid a, Monoid c) => IsVarSet a c | c -> a where

Any representation c of a set of variables need to be able to be modified by a variable occurrence. This is to ensure that free variable analysis is compositional. For instance, it should be possible to compute `fv (v [u/x])` from `fv v` and `fv u`.

In algebraic terminology, a variable set a needs to be (almost) a left semimodule to the semiring VarOcc.

Methods

withVarOcc :: VarOcc' a -> c -> c

Laws * Respects monoid operations: ``` withVarOcc o mempty == mempty withVarOcc o (x <> y) == withVarOcc o x <> withVarOcc o y ``` * Respects VarOcc composition: ``` withVarOcc oneVarOcc = id withVarOcc (composeVarOcc o1 o2) = withVarOcc o1 . withVarOcc o2 ``` * Respects VarOcc aggregation: ``` withVarOcc (o1 <> o2) x = withVarOcc o1 x <> withVarOcc o2 x ``` Since the corresponding unit law may fail, ``` withVarOcc mempty x = mempty ``` it is not quite a semimodule.

type TheVarMap' a = IntMap (VarOcc' a)

Representation of a variable set as map from de Bruijn indices to VarOcc.

newtype VarMap' a

Constructors

VarMap

Fields

theVarMap :: TheVarMap' a

type TheVarMap = TheVarMap' MetaSet

type VarMap = VarMap' MetaSet

mapVarMap :: (TheVarMap' a -> TheVarMap' b) -> VarMap' a -> VarMap' b

lookupVarMap :: Variable -> VarMap' a -> Maybe (VarOcc' a)

Simple flexible/rigid variable collection.

type TheFlexRigMap = IntMap (FlexRig' ())

Keep track of FlexRig for every variable, but forget the involved meta vars.

newtype FlexRigMap

Constructors

FlexRigMap

Fields

theFlexRigMap :: TheFlexRigMap

mapFlexRigMap :: (TheFlexRigMap -> TheFlexRigMap) -> FlexRigMap -> FlexRigMap

Plain variable occurrence counting.

newtype VarCounts

Constructors

VarCounts

Fields

varCounts :: IntMap Int

Environment for collecting free variables.

data IgnoreSorts

Where should we skip sorts in free variable analysis?

Constructors

IgnoreNot
Do not skip.

IgnoreInAnnotations
Skip when annotation to a type.

IgnoreAll
Skip unconditionally.

data FreeEnv' a b c

The current context.

Constructors

FreeEnv

Fields

feExtra :: !b

Additional context, e.g., whether to ignore free variables in sorts.

feFlexRig :: !(FlexRig' a)

Are we flexible or rigid?

feModality :: !Modality

What is the current relevance and quantity?

feSingleton :: !(b -> Maybe Variable -> c)

Method to return a single variable.

type Variable = Int

type SingleVar c = Variable -> c

type FreeEnv c = FreeEnv' MetaSet IgnoreSorts c

feIgnoreSorts :: FreeEnv' a IgnoreSorts c -> IgnoreSorts

Ignore free variables in sorts.

initFreeEnv :: Monoid c => b -> SingleVar c -> FreeEnv' a b c

The initial context.

Orphan instances

Singleton MetaId ()

Instance details

Methods

singleton :: MetaId -> ()