Agda.TypeChecking.Free.Lazy

Description

Computing the free variables of a term lazily.

We implement a reduce (traversal into monoid) over internal syntax for a generic collection (monoid with singletons). This should allow a more efficient test for the presence of a particular variable.

Worst-case complexity does not change (i.e. the case when a variable does not occur), but best case-complexity does matter. For instance, see mkAbs: each time we construct a dependent function type, we check whether it is actually dependent.

The distinction between rigid and strongly rigid occurrences comes from: Jason C. Reed, PhD thesis, 2009, page 96 (see also his LFMTP 2009 paper)

The main idea is that x = t(x) is unsolvable if x occurs strongly rigidly in t. It might have a solution if the occurrence is not strongly rigid, e.g.

x = f -> suc (f (x ( y -> k))) has x = f -> suc (f (suc k))

Jason C. Reed, PhD thesis, page 106

Under coinductive constructors, occurrences are never strongly rigid. Also, function types and lambdas do not establish strong rigidity. Only inductive constructors do so. (See issue 1271).

For further reading on semirings and semimodules for variable occurrence, see e.g. Conor McBrides "I got plenty of nuttin'" (Wadlerfest 2016). There, he treats the "quantity" dimension of variable occurrences.

The semiring has an additive operation for combining occurrences of subterms, and a multiplicative operation of representing function composition. E.g. if variable x appears o in term u, but u appears in context q in term t then occurrence of variable x coming from u is accounted for as q o in t.

Consider example (λ{ x → (x,x)}) y:

Variable x occurs once unguarded in x.

It occurs twice unguarded in the aggregation x x

Inductive constructor , turns this into two strictly rigid occurrences.

If , is a record constructor, then we stay unguarded.

The function ({λ x → (x,x)}) provides a context for variable y. This context can be described as weakly rigid with quantity two.

The final occurrence of y is obtained as composing the context with the occurrence of y in itself (which is the unit for composition). Thus, y occurs weakly rigid with quantity two.

It is not a given that the context can be described in the same way as the variable occurrence. However, for quantity it is the case and we obtain a semiring of occurrences with 0, 1, and even ω, which is an absorptive element for addition.

Set of meta variables.

newtype MetaSet

A set of meta variables. Forms a monoid under union.

Constructors

MetaSet

Fields

theMetaSet :: HashSet MetaId

insertMetaSet :: MetaId -> MetaSet -> MetaSet

foldrMetaSet :: (MetaId -> a -> a) -> a -> MetaSet -> a

metaSetToBlocker :: MetaSet -> Blocker

Flexible and rigid occurrences (semigroup)

data FlexRig' a

Depending on the surrounding context of a variable, it's occurrence can be classified as flexible or rigid, with finer distinctions.

The constructors are listed in increasing order (wrt. information content).

Constructors

Flexible a
In arguments of metas. The set of metas is used by 'NonLinMatch' to generate the right blocking information. The semantics is that the status of a variable occurrence may change if one of the metas in the set gets solved. We may say the occurrence is tainted by the meta variables in the set.

WeaklyRigid
In arguments to variables and definitions.

Unguarded
In top position, or only under inductive record constructors (unit).

StronglyRigid
Under at least one and only inductive constructors.

type FlexRig = FlexRig' MetaSet

class LensFlexRig o a | o -> a where

Methods

lensFlexRig :: Lens' o (FlexRig' a)

isFlexible :: LensFlexRig o a => o -> Bool

isUnguarded :: LensFlexRig o a => o -> Bool

isWeaklyRigid :: LensFlexRig o a => o -> Bool

isStronglyRigid :: LensFlexRig o a => o -> Bool

addFlexRig :: Semigroup a => FlexRig' a -> FlexRig' a -> FlexRig' a

FlexRig aggregation (additive operation of the semiring). For combining occurrences of the same variable in subterms. This is a refinement of the max operation for FlexRig which would work if Flexible did not have the MetaSet as an argument. Now, to aggregate two Flexible occurrences, we union the involved MetaSets.

zeroFlexRig :: Monoid a => FlexRig' a

Unit for addFlexRig.

omegaFlexRig :: FlexRig' a

Absorptive for addFlexRig.

composeFlexRig :: Semigroup a => FlexRig' a -> FlexRig' a -> FlexRig' a

FlexRig composition (multiplicative operation of the semiring). For accumulating the context of a variable.

Flexible is dominant. Once we are under a meta, we are flexible regardless what else comes. We taint all variable occurrences under a meta by this meta.

WeaklyRigid is next in strength. Destroys strong rigidity.

StronglyRigid is still dominant over Unguarded.

Unguarded is the unit. It is the top (identity) context.

oneFlexRig :: FlexRig' a

Unit for composeFlexRig.

Multi-dimensional feature vector for variable occurrence (semigroup)

data VarOcc' a

Occurrence of free variables is classified by several dimensions. Currently, we have FlexRig and Modality.

Constructors

VarOcc

Fields

varFlexRig :: FlexRig' a

varModality :: Modality

type VarOcc = VarOcc' MetaSet

topVarOcc :: VarOcc' a

The absorptive element of variable occurrence under aggregation: strongly rigid, relevant.

composeVarOcc :: Semigroup a => VarOcc' a -> VarOcc' a -> VarOcc' a

First argument is the outer occurrence (context) and second is the inner. This multiplicative operation is to modify an occurrence under a context.

oneVarOcc :: VarOcc' a

Storing variable occurrences (semimodule).

class (Singleton MetaId a, Semigroup a, Monoid a, Semigroup c, Monoid c) => IsVarSet a c | c -> a where

Any representation c of a set of variables need to be able to be modified by a variable occurrence. This is to ensure that free variable analysis is compositional. For instance, it should be possible to compute `fv (v [u/x])` from `fv v` and `fv u`.

In algebraic terminology, a variable set a needs to be (almost) a left semimodule to the semiring VarOcc.

Methods

withVarOcc :: VarOcc' a -> c -> c

Laws * Respects monoid operations: ``` withVarOcc o mempty == mempty withVarOcc o (x <> y) == withVarOcc o x <> withVarOcc o y ``` * Respects VarOcc composition: ``` withVarOcc oneVarOcc = id withVarOcc (composeVarOcc o1 o2) = withVarOcc o1 . withVarOcc o2 ``` * Respects VarOcc aggregation: ``` withVarOcc (o1 <> o2) x = withVarOcc o1 x <> withVarOcc o2 x ``` Since the corresponding unit law may fail, ``` withVarOcc mempty x = mempty ``` it is not quite a semimodule.

type TheVarMap' a = IntMap (VarOcc' a)

Representation of a variable set as map from de Bruijn indices to VarOcc.

newtype VarMap' a

Constructors

VarMap

Fields

theVarMap :: TheVarMap' a

type TheVarMap = TheVarMap' MetaSet

type VarMap = VarMap' MetaSet

mapVarMap :: (TheVarMap' a -> TheVarMap' b) -> VarMap' a -> VarMap' b

lookupVarMap :: Variable -> VarMap' a -> Maybe (VarOcc' a)

Simple flexible/rigid variable collection.

type TheFlexRigMap = IntMap (FlexRig' ())

Keep track of FlexRig for every variable, but forget the involved meta vars.

newtype FlexRigMap

Constructors

FlexRigMap

Fields

theFlexRigMap :: TheFlexRigMap

mapFlexRigMap :: (TheFlexRigMap -> TheFlexRigMap) -> FlexRigMap -> FlexRigMap

Environment for collecting free variables.

data IgnoreSorts

Where should we skip sorts in free variable analysis?

Constructors

IgnoreNot
Do not skip.

IgnoreInAnnotations
Skip when annotation to a type.

IgnoreAll
Skip unconditionally.

data FreeEnv' a b c

The current context.

Constructors

FreeEnv

Fields

feExtra :: !b

Additional context, e.g., whether to ignore free variables in sorts.

feFlexRig :: !(FlexRig' a)

Are we flexible or rigid?

feModality :: !Modality

What is the current relevance and quantity?

feSingleton :: Maybe Variable -> c

Method to return a single variable.

type Variable = Int

type SingleVar c = Variable -> c

type FreeEnv c = FreeEnv' MetaSet IgnoreSorts c

feIgnoreSorts :: FreeEnv' a IgnoreSorts c -> IgnoreSorts

Ignore free variables in sorts.

initFreeEnv :: Monoid c => b -> SingleVar c -> FreeEnv' a b c

The initial context.

type FreeT a b (m :: Type -> Type) c = ReaderT (FreeEnv' a b c) m c

type FreeM a c = Reader (FreeEnv' a IgnoreSorts c) c

runFreeM :: IsVarSet a c => SingleVar c -> IgnoreSorts -> FreeM a c -> c

Run function for FreeM.

variable :: forall (m :: Type -> Type) a c b. (Monad m, IsVarSet a c) => Int -> FreeT a b m c

Base case: a variable.

subVar :: Int -> Maybe Variable -> Maybe Variable

Subtract, but return Nothing if result is negative.

underBinder :: MonadReader (FreeEnv' a b c) m => m z -> m z

Going under a binder.

underBinder' :: MonadReader (FreeEnv' a b c) m => Nat -> m z -> m z

Going under n binders.

underModality :: (MonadReader r m, LensModality r, LensModality o) => o -> m z -> m z

Changing the Modality.

underRelevance :: (MonadReader r m, LensRelevance r, LensRelevance o) => o -> m z -> m z

Changing the Relevance.

underQuantity :: (MonadReader r m, LensQuantity r, LensQuantity o) => o -> m a -> m a

In the given computation the Quantity is locally scaled using the Quantity of the first argument.

underFlexRig :: (MonadReader r m, LensFlexRig r a, Semigroup a, LensFlexRig o a) => o -> m z -> m z

Changing the FlexRig context.

underConstructor :: (MonadReader r m, LensFlexRig r a, Semigroup a) => ConHead -> Elims -> m z -> m z

What happens to the variables occurring under a constructor?

Recursively collecting free variables.

class Free t where

Gather free variables in a collection.

Minimal complete definition

Nothing

Methods

freeVars' :: IsVarSet a c => t -> FreeM a c

default freeVars' :: forall (f :: Type -> Type) b a c. (t ~ f b, Foldable f, Free b, IsVarSet a c) => t -> FreeM a c

Orphan instances

Singleton MetaId ()

Instance details

Methods

singleton :: MetaId -> ()