Agda.TypeChecking.Generalize

Documentation

generalizeType :: Set QName -> TCM Type -> TCM ([Maybe QName], Type)

Generalize a type over a set of (used) generalizable variables.

generalizeType' :: Set QName -> TCM (Type, a) -> TCM ([Maybe QName], Type, a)

Allow returning additional information from the type checking action.

generalizeTelescope :: Map QName Name -> (forall a. (Telescope -> TCM a) -> TCM a) -> ([Maybe Name] -> Telescope -> TCM a) -> TCM a

Generalize a telescope over a set of generalizable variables.