Agda.TypeChecking.Implicit

Description

Functions for inserting implicit arguments at the right places.

Documentation

splitImplicitBinderT :: HasRange a => NamedArg a -> Type -> TCM (Telescope, Type)

Split a given Pi Type until you reach the given named argument, returning the number of arguments skipped to reach the right plicity and name.

implicitArgs

Arguments

:: (PureTCM m, MonadMetaSolver m, MonadTCM m)

=> Int
n, the maximum number of implicts to be inserted.

-> (Hiding -> Bool)
expand, the predicate to test whether we should keep inserting.

-> Type
The (function) type t we are eliminating.

-> m (Args, Type)
The eliminating arguments and the remaining type.

implicitArgs n expand t generates up to n implicit argument metas (unbounded if n<0), as long as t is a function type and expand holds on the hiding info of its domain.

implicitNamedArgs

Arguments

:: (PureTCM m, MonadMetaSolver m, MonadTCM m)

=> Int
n, the maximum number of implicts to be inserted.

-> (Hiding -> ArgName -> Bool)
expand, the predicate to test whether we should keep inserting.

-> Type
The (function) type t we are eliminating.

-> m (NamedArgs, Type)
The eliminating arguments and the remaining type.

implicitNamedArgs n expand t generates up to n named implicit arguments metas (unbounded if n<0), as long as t is a function type and expand holds on the hiding and name info of its domain.

noHeadConstraints :: [CheckedArg] -> TCM [Elim]

Make sure there are no head constraints attached to the eliminations and just return the eliminations.

noHeadConstraint :: CheckedArg -> TCM Elim

implicitCheckedArgs

Arguments

:: (PureTCM m, MonadMetaSolver m, MonadTCM m)

=> Int
n, the maximum number of implicts to be inserted.

-> (Hiding -> ArgName -> Bool)
expand, the predicate to test whether we should keep inserting.

-> Type
The (function) type t we are eliminating.

-> m ([Named_ CheckedArg], Type)
The eliminating arguments and the remaining type.

implicitCheckedArgs n expand t generates up to n named implicit arguments metas (unbounded if n<0), as long as t is a function type and expand holds on the hiding and name info of its domain.

makeLockConstraint :: Type -> Term -> TCM (Maybe (Abs Constraint))

makeLockConstraint u funType(El _ (Pi (Dom{ domInfo=info, unDom=a }) _)) creates a CheckLockedVars constaint for lock u : a if getLock info == IsLock _@.

Precondition: Type is reduced.

newMetaArg

Arguments

:: (PureTCM m, MonadMetaSolver m)

=> MetaKind
Kind of meta.

-> ArgInfo
Rrelevance of meta.

-> ArgName
Name suggestion for meta.

-> Comparison
Check (CmpLeq) or infer (CmpEq) the type.

-> Type
Type of meta.

-> m (MetaId, Term)
The created meta as id and as term.

Create a metavariable of MetaKind.

newInteractionMetaArg

Arguments

:: ArgInfo
Relevance of meta.

-> ArgName
Name suggestion for meta.

-> Comparison
Check (CmpLeq) or infer (CmpEq) the type.

-> Type
Type of meta.

-> TCM (MetaId, Term)
The created meta as id and as term.

Create a questionmark (always UnificationMeta).

data ImplicitInsertion

Possible results of insertImplicit.

Constructors

ImpInsert [Dom ()]
Success: this many implicits have to be inserted (list can be empty).

BadImplicits
Error: hidden argument where there should have been a non-hidden argument.

NoSuchName ArgName
Error: bad named argument.

pattern NoInsertNeeded :: ImplicitInsertion

insertImplicit

Arguments

:: NamedArg e
Next given argument a.

-> [Dom a]
Expected arguments ts.

-> ImplicitInsertion

If the next given argument is a and the expected arguments are ts insertImplicit' a ts returns the prefix of ts that precedes a.

If a is named but this name does not appear in ts, the NoSuchName exception is thrown.

insertImplicit'

Arguments

:: NamedArg e
Next given argument a.

-> [Dom ArgName]
Expected arguments ts.

-> ImplicitInsertion

If the next given argument is a and the expected arguments are ts insertImplicit' a ts returns the prefix of ts that precedes a.

If a is named but this name does not appear in ts, the NoSuchName exception is thrown.