Agda.TypeChecking.Level.Solve

Documentation

defaultOpenLevelsToZero :: (PureTCM m, MonadMetaSolver m) => m a -> m a

Run the given action. If --cumulativity is off, that's it.

Otherwise, at the end, take all new metavariables of type level for which the only constraints are upper bounds on the level, and instantiate them to the lowest level.

defaultLevelsToZero :: (PureTCM m, MonadMetaSolver m) => LocalMetaStore -> m ()