Agda.TypeChecking.Modalities

Documentation

checkModality' :: MonadConversion m => QName -> Definition -> m (Maybe TypeError)

The second argument is the definition of the first.

checkModality :: MonadConversion m => QName -> Definition -> m ()

The second argument is the definition of the first.

checkModalityArgs :: MonadConversion m => Definition -> Args -> m ()

Checks that the given implicitely inserted arguments, are used in a modally correct way.