Agda.TypeChecking.Monad.Base.Types

Description

Data structures for the type checker.

Part of Agda.TypeChecking.Monad.Base, extracted to avoid import cycles.

Context

type Context = [ContextEntry]

The Context is a stack of ContextEntrys.

type ContextEntry = Dom (Name, Type)

Conversion

data Comparison

Constructors

CmpEq

CmpLeq

data Polarity

Polarity for equality and subtype checking.

Constructors

Covariant
monotone

Contravariant
antitone

Invariant
no information (mixed variance)

Nonvariant
constant

Cubical

data IPFace' t

Datatype representing a single boundary condition: x_0 = u_0, ... ,x_n = u_n ⊢ t = ?n es

Constructors

IPFace'

Fields

faceEqns :: [(t, t)]

faceRHS :: t

Highlighting

data HighlightingLevel

How much highlighting should be sent to the user interface?

Constructors

None

NonInteractive

Interactive
This includes both non-interactive highlighting and interactive highlighting of the expression that is currently being type-checked.

data HighlightingMethod

How should highlighting be sent to the user interface?

Constructors

Direct
Via stdout.

Indirect
Both via files and via stdout.

Managing file names

type ModuleToSource = Map TopLevelModuleName AbsolutePath

Maps top-level module names to the corresponding source file names.

Meta variables

data NamedMeta

For printing, we couple a meta with its name suggestion.

Constructors

NamedMeta

Fields

nmSuggestion :: MetaNameSuggestion

nmid :: MetaId