Agda.TypeChecking.Monad.Base.Warning

Description

Types related to warnings raised by Agda.

Documentation

data RecordFieldWarning

Constructors

DuplicateFields (List1 (Name, Range))
Each redundant field comes with a range of associated dead code.

TooManyFields QName [Name] (List1 (Name, Range))
Record type, fields not supplied by user, non-fields but supplied. The redundant fields come with a range of associated dead code.

data UselessPublicReason

Constructors

UselessPublicPreamble

UselessPublicLet

UselessPublicNoOpen

UselessPublicAnonymousModule