Agda.TypeChecking.Monad.Context

Modifying the context

unsafeModifyContext :: MonadTCEnv tcm => (Context -> Context) -> tcm a -> tcm a

Modify a Context in a computation. Warning: does not update the checkpoints. Use updateContext instead.

modifyContextInfo :: MonadTCEnv tcm => (forall e. Dom e -> Dom e) -> tcm a -> tcm a

Modify the Dom part of context entries.

inTopContext :: (MonadTCEnv tcm, ReadTCState tcm) => tcm a -> tcm a

Change to top (=empty) context. Resets the checkpoints.

unsafeInTopContext :: (MonadTCEnv m, ReadTCState m) => m a -> m a

Change to top (=empty) context, but don't update the checkpoints. Totally not safe!

unsafeEscapeContext :: MonadTCM tcm => Int -> tcm a -> tcm a

Delete the last n bindings from the context.

Doesn't update checkpoints! Use escapeContext or `updateContext rho (drop n)` instead, for an appropriate substitution rho.

escapeContext :: MonadAddContext m => Impossible -> Int -> m a -> m a

Delete the last n bindings from the context. Any occurrences of these variables are replaced with the given err.

Manipulating checkpoints --

checkpoint :: (MonadDebug tcm, MonadTCM tcm, MonadFresh CheckpointId tcm, ReadTCState tcm) => Substitution -> tcm a -> tcm a

Add a new checkpoint. Do not use directly!

checkpointSubstitution :: MonadTCEnv tcm => CheckpointId -> tcm Substitution

Get the substitution from the context at a given checkpoint to the current context.

checkpointSubstitution' :: MonadTCEnv tcm => CheckpointId -> tcm (Maybe Substitution)

Get the substitution from the context at a given checkpoint to the current context.

getModuleParameterSub :: (MonadTCEnv m, ReadTCState m) => ModuleName -> m (Maybe Substitution)

Get substitution Γ ⊢ ρ : Γm where Γ is the current context and Γm is the module parameter telescope of module m.

Returns Nothing in case the we don't have a checkpoint for m.

Adding to the context

class MonadTCEnv m => MonadAddContext (m :: Type -> Type) where

Minimal complete definition

Nothing

Methods

addCtx :: Name -> Dom Type -> m a -> m a

addCtx x arg cont add a variable to the context.

Chooses an unused Name.

Warning: Does not update module parameter substitution!

default addCtx :: forall (n :: Type -> Type) (t :: (Type -> Type) -> Type -> Type) a. (MonadAddContext n, MonadTransControl t, t n ~ m) => Name -> Dom Type -> m a -> m a

addLetBinding' :: Origin -> Name -> Term -> Dom Type -> m a -> m a

Add a let bound variable to the context

default addLetBinding' :: forall (n :: Type -> Type) (t :: (Type -> Type) -> Type -> Type) a. (MonadAddContext n, MonadTransControl t, t n ~ m) => Origin -> Name -> Term -> Dom Type -> m a -> m a

updateContext :: Substitution -> (Context -> Context) -> m a -> m a

Update the context. Requires a substitution that transports things living in the old context to the new.

default updateContext :: forall (n :: Type -> Type) (t :: (Type -> Type) -> Type -> Type) a. (MonadAddContext n, MonadTransControl t, t n ~ m) => Substitution -> (Context -> Context) -> m a -> m a

withFreshName :: Range -> ArgName -> (Name -> m a) -> m a

default withFreshName :: forall (n :: Type -> Type) (t :: (Type -> Type) -> Type -> Type) a. (MonadAddContext n, MonadTransControl t, t n ~ m) => Range -> ArgName -> (Name -> m a) -> m a

defaultAddCtx :: MonadAddContext m => Name -> Dom Type -> m a -> m a

Default implementation of addCtx in terms of updateContext

withFreshName_ :: MonadAddContext m => ArgName -> (Name -> m a) -> m a

withShadowingNameTCM :: Name -> TCM b -> TCM b

Run the given TCM action, and register the given variable as being shadowed by all the names with the same root that are added to the context during this TCM action.

addRecordNameContext :: (MonadAddContext m, MonadFresh NameId m) => Dom Type -> m b -> m b

class AddContext b where

Various specializations of addCtx.

Methods

addContext :: MonadAddContext m => b -> m a -> m a

contextSize :: b -> Nat

newtype KeepNames a

Wrapper to tell addContext not to mark names as NotInScope. Used when adding a user-provided, but already type checked, telescope to the context.

Constructors

KeepNames a

underAbstraction :: (Subst a, MonadAddContext m) => Dom Type -> Abs a -> (a -> m b) -> m b

Go under an abstraction. Do not extend context in case of NoAbs.

underAbstraction' :: (Subst a, MonadAddContext m, AddContext (name, Dom Type)) => (String -> name) -> Dom Type -> Abs a -> (a -> m b) -> m b

underAbstractionAbs :: (Subst a, MonadAddContext m) => Dom Type -> Abs a -> (a -> m b) -> m b

Go under an abstraction, treating NoAbs as Abs.

underAbstractionAbs' :: (Subst a, MonadAddContext m, AddContext (name, Dom Type)) => (String -> name) -> Dom Type -> Abs a -> (a -> m b) -> m b

underAbstraction_ :: (Subst a, MonadAddContext m) => Abs a -> (a -> m b) -> m b

Go under an abstract without worrying about the type to add to the context.

mapAbstraction :: (Subst a, Subst b, MonadAddContext m) => Dom Type -> (a -> m b) -> Abs a -> m (Abs b)

Map a monadic function on the thing under the abstraction, adding the abstracted variable to the context.

mapAbstraction_ :: (Subst a, Subst b, MonadAddContext m) => (a -> m b) -> Abs a -> m (Abs b)

getLetBindings :: MonadTCEnv tcm => tcm [(Name, LetBinding)]

defaultAddLetBinding' :: (ReadTCState m, MonadTCEnv m) => Origin -> Name -> Term -> Dom Type -> m a -> m a

Add a let bound variable

addLetBinding :: MonadAddContext m => ArgInfo -> Origin -> Name -> Term -> Type -> m a -> m a

Add a let bound variable

removeLetBinding :: MonadTCEnv m => Name -> m a -> m a

Remove a let bound variable.

removeLetBindingsFrom :: MonadTCEnv m => Name -> m a -> m a

Remove a let bound variable and all let bindings introduced after it. For instance before printing its body to avoid folding the binding itself, or using bindings defined later. Relies on the invariant that names introduced later are sorted after earlier names.

Querying the context

getContext :: MonadTCEnv m => m Context

Get the current context.

getContextSize :: (Applicative m, MonadTCEnv m) => m Nat

Get the size of the current context.

getContextArgs :: (Applicative m, MonadTCEnv m) => m Args

Generate [var (n - 1), ..., var 0] for all declarations in the context.

getContextTerms :: (Applicative m, MonadTCEnv m) => m [Term]

Generate [var (n - 1), ..., var 0] for all declarations in the context.

getContextTelescope :: (Applicative m, MonadTCEnv m) => m Telescope

Get the current context as a Telescope.

getContextNames :: (Applicative m, MonadTCEnv m) => m [Name]

Get the names of all declarations in the context.

lookupBV_ :: Nat -> Context -> Maybe ContextEntry

get type of bound variable (i.e. deBruijn index)

lookupBV' :: MonadTCEnv m => Nat -> m (Maybe ContextEntry)

lookupBV :: (MonadFail m, MonadTCEnv m) => Nat -> m (Dom (Name, Type))

domOfBV :: (Applicative m, MonadFail m, MonadTCEnv m) => Nat -> m (Dom Type)

typeOfBV :: (Applicative m, MonadFail m, MonadTCEnv m) => Nat -> m Type

nameOfBV' :: (Applicative m, MonadFail m, MonadTCEnv m) => Nat -> m (Maybe Name)

nameOfBV :: (Applicative m, MonadFail m, MonadTCEnv m) => Nat -> m Name

getVarInfo :: (MonadFail m, MonadTCEnv m) => Name -> m (Term, Dom Type)

Get the term corresponding to a named variable. If it is a lambda bound variable the deBruijn index is returned and if it is a let bound variable its definition is returned.