Agda.TypeChecking.Monad.Mutual

Documentation

noMutualBlock :: TCM a -> TCM a

inMutualBlock :: (MutualId -> TCM a) -> TCM a

Pass the current mutual block id or create a new mutual block if we are not already inside on.

setMutualBlockInfo :: MutualId -> MutualInfo -> TCM ()

Set the mutual block info for a block, possibly overwriting the existing one.

insertMutualBlockInfo :: MutualId -> MutualInfo -> TCM ()

Set the mutual block info for a block if non-existing.

setMutualBlock :: MutualId -> QName -> TCM ()

Set the mutual block for a definition.

currentOrFreshMutualBlock :: TCM MutualId

Get the current mutual block, if any, otherwise a fresh mutual block is returned.

lookupMutualBlock :: ReadTCState tcm => MutualId -> tcm MutualBlock

mutualBlockOf :: QName -> TCM MutualId

Reverse lookup of a mutual block id for a name.