Agda.TypeChecking.Monad.Open

Documentation

makeOpen :: (ReadTCState m, MonadTCEnv m) => a -> m (Open a)

Create an open term in the current context.

getOpen :: (TermSubst a, MonadTCEnv m) => Open a -> m a

Extract the value from an open term. The checkpoint at which it was created must be in scope.

tryGetOpen :: (TermSubst a, ReadTCState m, MonadTCEnv m) => (Substitution -> a -> Maybe a) -> Open a -> m (Maybe a)

Extract the value from an open term. If the checkpoint is no longer in scope use the provided function to pull the object to the most recent common checkpoint. The function is given the substitution from the common ancestor to the checkpoint of the thing.

isClosed :: Open a -> Bool

An Open is closed if it has checkpoint 0.