Agda.TypeChecking.Monad.Trace

Trace

interestingCall :: Call -> Bool

class (MonadTCEnv m, ReadTCState m) => MonadTrace (m :: Type -> Type) where

Minimal complete definition

traceClosureCall

Methods

traceCall :: Call -> m a -> m a

Record a function call in the trace.

traceCallM :: m Call -> m a -> m a

traceCallCPS :: Call -> ((a -> m b) -> m b) -> (a -> m b) -> m b

Like traceCall, but resets envCall and the current ranges to the previous values in the continuation.

traceClosureCall :: Closure Call -> m a -> m a

printHighlightingInfo :: RemoveTokenBasedHighlighting -> HighlightingInfo -> m ()

Lispify and print the given highlighting information.

default printHighlightingInfo :: forall (t :: (Type -> Type) -> Type -> Type) (n :: Type -> Type). (MonadTrans t, MonadTrace n, t n ~ m) => RemoveTokenBasedHighlighting -> HighlightingInfo -> m ()

getCurrentRange :: MonadTCEnv m => m Range

setCurrentRange :: (MonadTrace m, HasRange x) => x -> m a -> m a

Sets the current range (for error messages etc.) to the range of the given object, if it has a range (i.e., its range is not noRange).

highlightAsTypeChecked

Arguments

:: MonadTrace m

=> Range
rPre

-> Range
r

-> m a

-> m a

highlightAsTypeChecked rPre r m runs m and returns its result. Additionally, some code may be highlighted:

If r is non-empty and not a sub-range of rPre (after continuousPerLine has been applied to both): r is highlighted as being type-checked while m is running (this highlighting is removed if m completes successfully).

Otherwise: Highlighting is removed for rPre - r before m runs, and if m completes successfully, then rPre - r is highlighted as being type-checked.