Agda.TypeChecking.Names

Description

EDSL to construct terms without touching De Bruijn indices.

e.g. if given t, u :: Term, Γ ⊢ t, u : A, we can build "λ f. f t u" like this:

runNames [] $ do -- open binds t and u to computations that know how to weaken themselves in -- an extended context

t,u

<- mapM open [t,u]

- lam gives the illusion of HOAS by providing f as a computation

- It also extends the internal context with the name "f", so that

- t and u will get weakened in the body.

- Then we finish the job using the (@) combinator from Agda.TypeChecking.Primitive

lam "f" $ f -> f @ t @ u

Documentation

newtype NamesT m a

Constructors

NamesT

Fields

unName :: ReaderT Names m a

type Names = [String]

runNamesT :: Names -> NamesT m a -> m a

runNames :: Names -> NamesT Fail a -> a

currentCxt :: Monad m => NamesT m Names

cxtSubst :: MonadFail m => Names -> NamesT m (Substitution' a)

inCxt :: (MonadFail m, Subst a) => Names -> a -> NamesT m a

cl' :: Applicative m => a -> NamesT m a

cl :: Monad m => m a -> NamesT m a

open :: (MonadFail m, Subst a) => a -> NamesT m (NamesT m a)

bind' :: (MonadFail m, Subst b, DeBruijn b, Subst a, Free a) => ArgName -> (NamesT m b -> NamesT m a) -> NamesT m a

bind :: (MonadFail m, Subst b, DeBruijn b, Subst a, Free a) => ArgName -> (NamesT m b -> NamesT m a) -> NamesT m (Abs a)

glam :: MonadFail m => ArgInfo -> ArgName -> (NamesT m Term -> NamesT m Term) -> NamesT m Term

glamN :: (Functor m, MonadFail m) => [Arg ArgName] -> (NamesT m Args -> NamesT m Term) -> NamesT m Term

lam :: MonadFail m => ArgName -> (NamesT m Term -> NamesT m Term) -> NamesT m Term

ilam :: MonadFail m => ArgName -> (NamesT m Term -> NamesT m Term) -> NamesT m Term