Agda.TypeChecking.Opacity

Documentation

saturateOpaqueBlocks :: (MonadTCState m, ReadTCState m, MonadFresh OpaqueId m, MonadDebug m, MonadTrace m, MonadWarning m, MonadIO m) => m ()

Ensure that opaque blocks defined in the current module have saturated unfolding sets.

isAccessibleDef :: TCEnv -> TCState -> Definition -> Bool

Decide whether or not a definition is reducible. Returns True if the definition can step.

hasAccessibleDef :: (ReadTCState m, MonadTCEnv m, HasConstInfo m) => QName -> m Bool

Will the given QName have a proper definition, or will it be wrapped in an AbstractDefn?