Agda.TypeChecking.Polarity

Description

Computing the polarity (variance) of function arguments, for the sake of subtyping.

Polarity computation

computePolarity :: (HasOptions m, HasConstInfo m, HasBuiltins m, MonadTCEnv m, MonadTCState m, MonadReduce m, MonadAddContext m, MonadTCError m, MonadDebug m, MonadPretty m) => [QName] -> m ()

Main function of this module.

Auxiliary functions

composePol :: Polarity -> Polarity -> Polarity

What is the polarity of a function composition?

nextPolarity :: [Polarity] -> (Polarity, [Polarity])

Get the next polarity from a list, Invariant if empty.

purgeNonvariant :: [Polarity] -> [Polarity]

Replace Nonvariant by Covariant. (Arbitrary bias, but better than Invariant, see issue 1596).

polFromOcc :: Occurrence -> Polarity