Agda.TypeChecking.Positivity

Description

Check that a datatype is strictly positive.

Documentation

checkStrictlyPositive :: MutualInfo -> Set QName -> TCM ()

Check that the datatypes in the mutual block containing the given declarations are strictly positive.

Find polarity of datatypes parameters and indices.

Also add information about positivity and recursivity of records to the signature.

getDefArity :: Definition -> TCM Int

hasDefinition :: Defn -> Bool

isDatatype :: Definition -> Maybe (PositivityCheck, DataOrRecord)

preprocessBlock :: [QName] -> TCM (Maybe [(QName, Int, Maybe (PositivityCheck, DataOrRecord), Bool)])

Orphan instances

PrettyTCM Node

Instance details

Methods

prettyTCM :: MonadPretty m => Node -> m Doc

PrettyTCM (Seq OccursWhere)

Instance details

Methods

prettyTCM :: MonadPretty m => Seq OccursWhere -> m Doc

Pretty a => PrettyTCMWithNode (Edge a)

Instance details

Methods

prettyTCMWithNode :: (PrettyTCM n, MonadPretty m) => WithNode n (Edge a) -> m Doc