Agda.TypeChecking.Positivity.Occurrence

Description

Occurrences.

Documentation

type PragmaPolarities = List1 (Ranged Occurrence)

List of polarities stemming from POLARITY pragma.

data Occurrence

Subterm occurrences for positivity checking. The constructors are listed in increasing information they provide: Mixed <= JustPos <= StrictPos <= GuardPos <= Unused Mixed <= JustNeg <= Unused.

Constructors

Mixed
Arbitrary occurrence (positive and negative).

JustNeg
Negative occurrence.

JustPos
Positive occurrence, but not strictly positive.

StrictPos
Strictly positive occurrence.

GuardPos
Guarded strictly positive occurrence (i.e., under ∞). For checking recursive records.

Unused
No occurrence.

modalPolarityToOccurrence :: ModalPolarity -> Occurrence

data OccursPath

Constructors

Root

LeftOfArrow !OccursPath

DefArg !OccursPath !QName !Nat
in the nth argument of a defined constant

MutDefArg !OccursPath !QName !Nat
in the nth argument of a def in the current mutual block. (def given by Int index in the block)

UnderInf !OccursPath
in the principal argument of built-in ∞

VarArg !OccursPath !Nat !Occurrence
as an argument to a bound variable with given polarity. The polarity is only used for warning printing.

MetaArg !OccursPath
as an argument of a metavariable

ConArgType !OccursPath !QName
in the type of a constructor

IndArgType !OccursPath !QName
in a datatype index of a constructor

ConEndpoint !OccursPath !QName
in an endpoint of a higher constructor

InClause !OccursPath !Nat
in the nth clause of a defined function

Matched !OccursPath
matched against in a clause of a defined function

InIndex !OccursPath
is an index of an inductive family

InDefOf !OccursPath !QName
in the definition of a constant

data OccursWhere

Constructors

OccursWhere !Range !OccursPath

data Node

Constructors

DefNode QName
Definition site of a mutual definition

ArgNode QName Int
Occurrence of the nth definition argument inside the definition.

data Edge a

Constructors

Edge !Occurrence !a

mergeEdges :: Edge a -> Edge a -> Edge a