Agda.TypeChecking.Pretty

Documentation

pretty :: (Applicative m, Pretty a) => a -> m Doc

text :: Applicative m => String -> m Doc

(<?>) :: Applicative m => m Doc -> m Doc -> m Doc infixl 6

comma :: Applicative m => m Doc

colon :: Applicative m => m Doc

parens :: Functor m => m Doc -> m Doc

(<+>) :: Applicative m => m Doc -> m Doc -> m Doc infixl 6

hcat :: (Applicative m, Foldable t) => t (m Doc) -> m Doc

hsep :: (Applicative m, Foldable t) => t (m Doc) -> m Doc

vcat :: (Applicative m, Foldable t) => t (m Doc) -> m Doc

brackets :: Functor m => m Doc -> m Doc

braces :: Functor m => m Doc -> m Doc

type Doc = Doc

equals :: Applicative m => m Doc

quotes :: Functor m => m Doc -> m Doc

doubleQuotes :: Functor m => m Doc -> m Doc

nest :: Functor m => Int -> m Doc -> m Doc

hang :: Applicative m => m Doc -> Int -> m Doc -> m Doc

punctuate :: (Applicative m, Semigroup (m Doc), Foldable t) => m Doc -> t (m Doc) -> [m Doc]

($$) :: Applicative m => m Doc -> m Doc -> m Doc infixl 5

($+$) :: Applicative m => m Doc -> m Doc -> m Doc infixl 5

sep :: (Applicative m, Foldable t) => t (m Doc) -> m Doc

fsep :: (Applicative m, Foldable t) => t (m Doc) -> m Doc

class PrettyTCM a where

Methods

prettyTCM :: MonadPretty m => a -> m Doc

dbraces :: Functor m => m Doc -> m Doc

prettyA :: (ToConcrete a, Pretty (ConOfAbs a), MonadAbsToCon m) => a -> m Doc

prettyList_ :: (Applicative m, Semigroup (m Doc), Foldable t) => t (m Doc) -> m Doc

prettyList without the brackets.

type MonadPretty (m :: Type -> Type) = MonadAbsToCon m

pshow :: (Applicative m, Show a) => a -> m Doc

prettyList :: (Applicative m, Semigroup (m Doc), Foldable t) => t (m Doc) -> m Doc

Comma-separated list in brackets.

pwords :: Applicative m => String -> [m Doc]

fwords :: Applicative m => String -> m Doc

parensNonEmpty :: Functor m => m Doc -> m Doc

multiLineText :: Applicative m => String -> m Doc

newtype PrettyContext

Constructors

PrettyContext Context

prettyAs :: (ToConcrete a, ConOfAbs a ~ [ce], Pretty ce, MonadAbsToCon m) => a -> m Doc

superscript :: Applicative m => Int -> m Doc

prettyTCMCtx :: (PrettyTCM a, MonadPretty m) => Precedence -> a -> m Doc

Pretty print with a given context precedence

prettyR :: (ToAbstract r, PrettyTCM (AbsOfRef r), MonadPretty m, MonadError TCErr m) => r -> m Doc

prettyTCMPatterns :: MonadPretty m => [NamedArg DeBruijnPattern] -> m [Doc]

Proper pretty printing of patterns:

prettyTCMPatternList :: MonadPretty m => [NamedArg DeBruijnPattern] -> m Doc

data WithNode n a

Pairing something with a node (for printing only).

Constructors

WithNode n a

class PrettyTCMWithNode a where

Pretty-print something paired with a (printable) node. | This intermediate typeclass exists to avoid UndecidableInstances.

Methods

prettyTCMWithNode :: (PrettyTCM n, MonadPretty m) => WithNode n a -> m Doc

class Semigroup a where

The class of semigroups (types with an associative binary operation).

Instances should satisfy the following:

Associativity

x <> (y <> z) = (x <> y) <> z

You can alternatively define sconcat instead of (<>), in which case the laws are:

Unit

sconcat (pure x) = x

Multiplication

sconcat (join xss) = sconcat (fmap sconcat xss)

Since: base-4.9.0.0

Minimal complete definition

(<>) | sconcat

Methods

(<>) :: a -> a -> a infixr 6

An associative operation.

Examples

Expand

>>> [1,2,3] <> [4,5,6]
[1,2,3,4,5,6]

>>> Just [1, 2, 3] <> Just [4, 5, 6]
Just [1,2,3,4,5,6]

>>> putStr "Hello, " <> putStrLn "World!"
Hello, World!

type MonadAbsToCon (m :: Type -> Type) = (MonadFresh NameId m, MonadInteractionPoints m, MonadStConcreteNames m, HasOptions m, PureTCM m, IsString (m Doc), Null (m Doc), Semigroup (m Doc))

The function runAbsToCon can target any monad that satisfies the constraints of MonadAbsToCon.

Orphan instances

Semigroup (TCM Doc)
This instance is more specific than a generic instance Semigroup a => Semigroup (TCM a).

Instance details

Methods

(<>) :: TCM Doc -> TCM Doc -> TCM Doc

sconcat :: NonEmpty (TCM Doc) -> TCM Doc

stimes :: Integral b => b -> TCM Doc -> TCM Doc