Agda.TypeChecking.Pretty.Warning

Documentation

applyFlagsToTCWarnings :: HasOptions m => Set TCWarning -> m (Set TCWarning)

applyFlagsToTCWarningsPreserving :: HasOptions m => Set WarningName -> Set TCWarning -> m (Set TCWarning)

Depending which flags are set, one may happily ignore some warnings.

filterTCWarnings :: Set TCWarning -> TCM [TCWarning]

If there are several warnings, remove the unsolved-constraints warning in case there are no interesting constraints to list.

getAllUnsolvedWarnings :: (ReadTCState m, MonadWarning m, MonadTCM m) => m [TCWarning]

getAllWarnings :: (ReadTCState m, MonadWarning m, MonadTCM m) => WhichWarnings -> m (Set TCWarning)

Collect all warnings that have accumulated in the state.

getAllWarningsOfTCErr :: TCErr -> TCM (Set TCWarning)

getAllWarningsPreserving :: (ReadTCState m, MonadWarning m, MonadTCM m) => Set WarningName -> WhichWarnings -> m (Set TCWarning)

prettyDuplicateFields :: MonadPretty m => List1 Name -> m Doc

prettyTCWarnings :: Set TCWarning -> TCM String

prettyTCWarnings' :: Set TCWarning -> TCM [Doc]

prettyTooManyFields :: MonadPretty m => QName -> [Name] -> List1 Name -> m Doc

prettyWarning :: MonadPretty m => Warning -> m Doc

tcWarningsToError :: [TCWarning] -> TCM ()

Turns warnings, if any, into errors.

Orphan instances

PrettyTCM DataOrRecord_

Instance details

Methods

prettyTCM :: MonadPretty m => DataOrRecord_ -> m Doc

PrettyTCM TCWarning

Instance details

Methods

prettyTCM :: MonadPretty m => TCWarning -> m Doc

PrettyTCM Warning

Instance details

Methods

prettyTCM :: MonadPretty m => Warning -> m Doc