Agda.TypeChecking.Primitive

Description

Primitive functions, such as addition on builtin integers.

Documentation

module Agda.TypeChecking.Primitive.Base

module Agda.TypeChecking.Primitive.Cubical

buildList :: TCM ([Term] -> Term)

buildList A ts builds a list of type List A. Assumes that the terms ts all have type A.

fromLiteral :: (Literal -> Maybe a) -> TCM (FromTermFunction a)

fromReducedTerm :: (Term -> Maybe a) -> TCM (FromTermFunction a)

genPrimForce :: TCM Type -> (Term -> Arg Term -> Term) -> TCM PrimitiveImpl

Used for both primForce and primForceLemma.

getRefl :: TCM (Arg Term -> Term)

getReflArgInfo :: ConHead -> TCM (Maybe ArgInfo)

Get the ArgInfo of the principal argument of BUILTIN REFL.

Returns Nothing for e.g. data Eq {a} {A : Set a} (x : A) : A → Set a where refl : Eq x x

Returns Just ... for e.g. data Eq {a} {A : Set a} : (x y : A) → Set a where refl : ∀ x → Eq x x

metaToNat :: MetaId -> Nat

Converts MetaIds to natural numbers.

mkPrimFun1 :: (PrimType a, FromTerm a, PrimType b, ToTerm b) => (a -> b) -> TCM PrimitiveImpl

mkPrimFun1TCM :: (FromTerm a, ToTerm b) => TCM Type -> (a -> ReduceM b) -> TCM PrimitiveImpl

mkPrimFun2 :: (PrimType a, FromTerm a, ToTerm a, PrimType b, FromTerm b, PrimType c, ToTerm c) => (a -> b -> c) -> TCM PrimitiveImpl

mkPrimFun3 :: (PrimType a, FromTerm a, ToTerm a, PrimType b, FromTerm b, ToTerm b, PrimType c, FromTerm c, PrimType d, ToTerm d) => (a -> b -> c -> d) -> TCM PrimitiveImpl

mkPrimFun4 :: (PrimType a, FromTerm a, ToTerm a, PrimType b, FromTerm b, ToTerm b, PrimType c, FromTerm c, ToTerm c, PrimType d, FromTerm d, PrimType e, ToTerm e) => (a -> b -> c -> d -> e) -> TCM PrimitiveImpl

mkPrimInjective :: Type -> Type -> QName -> TCM PrimitiveImpl

mkPrimInjective takes two Set0 a and b and a function f of type a -> b and outputs a primitive internalizing the fact that f is injective.

mkPrimLevelMax :: TCM PrimitiveImpl

mkPrimLevelSuc :: TCM PrimitiveImpl

mkPrimLevelZero :: TCM PrimitiveImpl

primCharToNatInjective :: TCM PrimitiveImpl

primEraseEquality :: TCM PrimitiveImpl

primEraseEquality : {a : Level} {A : Set a} {x y : A} -> x ≡ y -> x ≡ y

primFloatToWord64Injective :: TCM PrimitiveImpl

primForce :: TCM PrimitiveImpl

primForceLemma :: TCM PrimitiveImpl

primLockUniv' :: TCM PrimitiveImpl

primMetaToNatInjective :: TCM PrimitiveImpl

primQNameToWord64sInjective :: TCM PrimitiveImpl

primStringFromListInjective :: TCM PrimitiveImpl

primStringToListInjective :: TCM PrimitiveImpl

primWord64ToNatInjective :: TCM PrimitiveImpl

primitiveFunctions :: Map PrimitiveId (TCM PrimitiveImpl)

toTermTCM :: ToTerm a => TCM (a -> TCM Term)

class FromTerm a where

Methods

fromTerm :: TCM (FromTermFunction a)

type FromTermFunction a = Arg Term -> ReduceM (Reduced (MaybeReduced (Arg Term)) a)

type Fun a = a -> a

newtype Lvl

Constructors

Lvl

Fields

unLvl :: Integer

newtype Nat

Constructors

Nat

Fields

unNat :: Integer

type Op a = a -> a -> a

type Pred a = a -> Bool

class PrimType a => PrimTerm a where

Methods

primTerm :: a -> TCM Term

class PrimType a where

Minimal complete definition

Nothing

Methods

primType :: a -> TCM Type

default primType :: PrimTerm a => a -> TCM Type

type Rel a = a -> a -> Bool

class ToTerm a where

Methods

toTerm :: TCM (a -> ReduceM Term)