Agda.TypeChecking.Primitive.Base

Documentation

(-->) :: Applicative m => m Type -> m Type -> m Type infixr 4

(.-->) :: Applicative m => m Type -> m Type -> m Type infixr 4

(..-->) :: Applicative m => m Type -> m Type -> m Type infixr 4

garr :: Applicative m => (Relevance -> Relevance) -> m Type -> m Type -> m Type

gpi :: (MonadAddContext m, MonadDebug m) => ArgInfo -> String -> m Type -> m Type -> m Type

hPi :: (MonadAddContext m, MonadDebug m) => String -> m Type -> m Type -> m Type

nPi :: (MonadAddContext m, MonadDebug m) => String -> m Type -> m Type -> m Type

hPi' :: forall (m :: Type -> Type). (MonadAddContext m, MonadDebug m) => String -> NamesT m Type -> (NamesT m Term -> NamesT m Type) -> NamesT m Type

nPi' :: forall (m :: Type -> Type). (MonadAddContext m, MonadDebug m) => String -> NamesT m Type -> (NamesT m Term -> NamesT m Type) -> NamesT m Type

pPi' :: forall (m :: Type -> Type). (MonadAddContext m, HasBuiltins m, MonadDebug m) => String -> NamesT m Term -> (NamesT m Term -> NamesT m Type) -> NamesT m Type

toFinitePi :: Type -> Type

Turn a Pi type into one whose domain is annotated finite, i.e., one that represents a Partial element rather than an actual function.

el' :: Applicative m => m Term -> m Term -> m Type

els :: Applicative m => m Sort -> m Term -> m Type

el's :: Applicative m => m Term -> m Term -> m Type

elInf :: Functor m => m Term -> m Type

elSSet :: Functor m => m Term -> m Type

nolam :: Term -> Term

varM :: Applicative m => Int -> m Term

gApply :: Applicative m => Hiding -> m Term -> m Term -> m Term

gApply' :: Applicative m => ArgInfo -> m Term -> m Term -> m Term

(<@>) :: Applicative m => m Term -> m Term -> m Term infixl 9

(<#>) :: Applicative m => m Term -> m Term -> m Term infixl 9

(<..>) :: Applicative m => m Term -> m Term -> m Term

(<@@>) :: Applicative m => m Term -> (m Term, m Term, m Term) -> m Term

list :: TCM Term -> TCM Term

tMaybe :: TCM Term -> TCM Term

io :: TCM Term -> TCM Term

path :: TCM Term -> TCM Term

el :: Functor m => m Term -> m Type

tset :: Applicative m => m Type

The universe Set0 as a type.

sSizeUniv :: Sort

SizeUniv as a sort.

tSizeUniv :: Applicative m => m Type

SizeUniv as a type.

tLevelUniv :: Applicative m => m Type

argN :: e -> Arg e

Abbreviation: argN = Arg defaultArgInfo.

domN :: e -> Dom e

argH :: e -> Arg e

Abbreviation: argH = hide Arg defaultArgInfo.

domH :: e -> Dom e

Accessing the primitive functions

lookupPrimitiveFunction :: PrimitiveId -> TCM PrimitiveImpl

lookupPrimitiveFunctionQ :: QName -> TCM (PrimitiveId, PrimitiveImpl)

getBuiltinName :: (HasBuiltins m, MonadReduce m) => BuiltinId -> m (Maybe QName)

getQNameFromTerm :: forall (m :: Type -> Type). MonadReduce m => Term -> MaybeT m QName

Convert a name in Term form back to QName.

isBuiltin :: (HasBuiltins m, MonadReduce m) => QName -> BuiltinId -> m Bool

Builtin Sigma

data SigmaKit

Constructors

SigmaKit

Fields

sigmaName :: QName

sigmaCon :: ConHead

sigmaFst :: QName

sigmaSnd :: QName

getSigmaKit :: (HasBuiltins m, HasConstInfo m) => m (Maybe SigmaKit)