Agda.TypeChecking.Primitive.Cubical

Documentation

requireCubical

Arguments

:: Cubical
Which variant of Cubical Agda is required?

-> String

-> TCM ()

Checks that the correct variant of Cubical Agda is activated. Note that --erased-cubical "counts as" --cubical in erased contexts.

primIntervalType :: (HasBuiltins m, MonadError TCErr m, MonadTCEnv m, ReadTCState m) => m Type

primINeg' :: TCM PrimitiveImpl

primDepIMin' :: TCM PrimitiveImpl

primIBin :: IntervalView -> IntervalView -> TCM PrimitiveImpl

primIMin' :: TCM PrimitiveImpl

primIMax' :: TCM PrimitiveImpl

imax :: HasBuiltins m => m Term -> m Term -> m Term

imin :: HasBuiltins m => m Term -> m Term -> m Term

primIdJ :: TCM PrimitiveImpl

primIdElim' :: TCM PrimitiveImpl

primPOr :: TCM PrimitiveImpl

primPartial' :: TCM PrimitiveImpl

primPartialP' :: TCM PrimitiveImpl

primSubOut' :: TCM PrimitiveImpl

primIdFace' :: TCM PrimitiveImpl

primIdPath' :: TCM PrimitiveImpl

primTrans' :: TCM PrimitiveImpl

primHComp' :: TCM PrimitiveImpl

data TranspOrHComp

Constructors

DoTransp

DoHComp

cmdToName :: TranspOrHComp -> String

data FamilyOrNot a

Constructors

IsFam

Fields

famThing :: a

IsNot

Fields

famThing :: a

mkGComp :: HasBuiltins m => String -> NamesT m (NamesT m Term -> NamesT m Term -> NamesT m Term -> NamesT m Term -> NamesT m Term -> NamesT m Term)

Define a "ghcomp" version of gcomp. Normal comp looks like:

comp^i A [ phi -> u ] u0 = hcomp^i A(1/i) [ phi -> forward A i u ] (forward A 0 u0)

So for "gcomp" we compute:

gcomp^i A [ phi -> u ] u0 = hcomp^i A(1/i) [ phi -> forward A i u, ~ phi -> forward A 0 u0 ] (forward A 0 u0)

The point of this is that gcomp does not produce any empty systems (if phi = 0 it will reduce to "forward A 0 u".

unglueTranspGlue :: PureTCM m => Arg Term -> Arg Term -> FamilyOrNot (Arg Term, Arg Term, Arg Term, Arg Term, Arg Term, Arg Term) -> m Term

data TermPosition

Constructors

Head

Eliminated

headStop :: PureTCM m => TermPosition -> m Term -> m Bool

compGlue :: PureTCM m => TranspOrHComp -> Arg Term -> Maybe (Arg Term) -> Arg Term -> FamilyOrNot (Arg Term, Arg Term, Arg Term, Arg Term, Arg Term, Arg Term) -> TermPosition -> m (Maybe Term)

compHCompU :: PureTCM m => TranspOrHComp -> Arg Term -> Maybe (Arg Term) -> Arg Term -> FamilyOrNot (Arg Term, Arg Term, Arg Term, Arg Term) -> TermPosition -> m (Maybe Term)

primTransHComp :: TranspOrHComp -> [Arg Term] -> Int -> ReduceM (Reduced MaybeReducedArgs Term)

primComp :: TCM PrimitiveImpl

prim_glueU' :: TCM PrimitiveImpl

prim_unglueU' :: TCM PrimitiveImpl

primGlue' :: TCM PrimitiveImpl

prim_glue' :: TCM PrimitiveImpl

prim_unglue' :: TCM PrimitiveImpl

primFaceForall' :: TCM PrimitiveImpl

decomposeInterval :: HasBuiltins m => Term -> m [(Map Int Bool, [Term])]

decomposeInterval' :: HasBuiltins m => Term -> m [(Map Int (Set Bool), [Term])]

transpTel :: Abs Telescope -> Term -> Args -> ExceptT (Closure (Abs Type)) TCM Args

Tries to primTransp a whole telescope of arguments, following the rule for Σ types. If a type in the telescope does not support transp, transpTel throws it as an exception.

trFillTel :: Abs Telescope -> Term -> Args -> Term -> ExceptT (Closure (Abs Type)) TCM Args

Like transpTel but performing a transpFill.