Agda.TypeChecking.RecordPatterns

Description

Code which replaces pattern matching on record constructors with uses of projection functions.

Documentation

translateCompiledClauses :: (HasConstInfo m, MonadChange m) => QName -> CompiledClauses -> m CompiledClauses

translateSplitTree :: SplitTree -> TCM SplitTree

Bottom-up procedure to record-pattern-translate split tree.

recordPatternToProjections :: DeBruijnPattern -> TCM [Term -> Term]

Take a record pattern p and yield a list of projections corresponding to the pattern variables, from left to right.

E.g. for (x , (y , z)) we return [ fst, fst . snd, snd . snd ].

If it is not a record pattern, error ShouldBeRecordPattern is raised.

recordRHSToCopatterns :: (MonadChange m, PureTCM m) => Clause -> m [Clause]