Agda.TypeChecking.Reduce

Documentation

class Instantiate t

Instantiate something. Results in an open meta variable or a non meta. Doesn't do any reduction, and preserves blocking tags (when blocking meta is uninstantiated).

instantiate' :: Instantiate t => t -> ReduceM t

instantiate :: (Instantiate a, MonadReduce m) => a -> m a

instantiateWhen :: (InstantiateFull a, MonadReduce m) => (MetaId -> ReduceM Bool) -> a -> m a

A variant of instantiateFull that only instantiates those meta-variables that satisfy the predicate.

class InstantiateFull t

instantiateFull' instantiates metas everywhere (and recursively) but does not reduce.

instantiateFull' :: InstantiateFull t => t -> ReduceM t

instantiateFull :: (InstantiateFull a, MonadReduce m) => a -> m a

class IsMeta a

Is something (an elimination of) a meta variable? Does not perform any reductions.

Minimal complete definition

isMeta

isMeta :: IsMeta a => a -> Maybe MetaId

class Reduce t

Minimal complete definition

reduce' | reduceB'

reduce' :: Reduce t => t -> ReduceM t

reduceB' :: Reduce t => t -> ReduceM (Blocked t)

reduce :: (Reduce a, MonadReduce m) => a -> m a

reduceB :: (Reduce a, MonadReduce m) => a -> m (Blocked a)

reduceWithBlocker :: (Reduce a, IsMeta a, MonadReduce m) => a -> m (Blocker, a)

reduceIApply' :: (Term -> ReduceM (Blocked Term)) -> ReduceM (Blocked Term) -> [Elim] -> ReduceM (Blocked Term)

reduceDefCopy :: PureTCM m => QName -> Elims -> m (Reduced () Term)

Reduce a non-primitive definition if it is a copy linking to another def.

reduceDefCopyTCM :: QName -> Elims -> TCM (Reduced () Term)

Specialized version to put in boot file.

reduceHead :: PureTCM m => Term -> m (Blocked Term)

Reduce simple (single clause) definitions.

slowReduceTerm :: Term -> ReduceM (Blocked Term)

unfoldCorecursion :: Term -> ReduceM (Blocked Term)

unfoldCorecursionE :: Elim -> ReduceM (Blocked Elim)

unfoldDefinitionE :: (Term -> ReduceM (Blocked Term)) -> Term -> QName -> Elims -> ReduceM (Blocked Term)

unfoldDefinitionStep :: Term -> QName -> Elims -> ReduceM (Reduced (Blocked Term) Term)

unfoldInlined :: PureTCM m => Term -> m Term

Unfold as many copies as possible, and then potentially a single inline function.

appDefE_ :: QName -> Term -> [Clause] -> Maybe CompiledClauses -> RewriteRules -> MaybeReducedElims -> ReduceM (Reduced (Blocked Term) Term)

appDef' :: QName -> Term -> [Clause] -> RewriteRules -> MaybeReducedArgs -> ReduceM (Reduced (Blocked Term) Term)

Apply a defined function to it's arguments, using the original clauses.

appDefE' :: QName -> Term -> [Clause] -> RewriteRules -> MaybeReducedElims -> ReduceM (Reduced (Blocked Term) Term)

abortIfBlocked :: (MonadReduce m, MonadBlock m, IsMeta t, Reduce t) => t -> m t

Throw pattern violation if blocked or a meta.

ifBlocked :: (Reduce t, IsMeta t, MonadReduce m) => t -> (Blocker -> t -> m a) -> (NotBlocked -> t -> m a) -> m a

Case on whether a term is blocked on a meta (or is a meta). That means it can change its shape when the meta is instantiated.

isBlocked :: (Reduce t, IsMeta t, MonadReduce m) => t -> m (Maybe Blocker)

fromBlocked :: MonadBlock m => Blocked a -> m a

Throw a pattern violation if the argument is Blocked, otherwise return the value embedded in the NotBlocked.

blockOnError :: MonadError TCErr m => Blocker -> m a -> m a

Run the given computation but turn any errors into blocked computations with the given blocker

class Simplify t

Only unfold definitions if this leads to simplification which means that a constructor/literal pattern is matched. We include reduction of IApply patterns, as `p i0` is akin to matcing on the i0 constructor of interval.

simplify :: (Simplify a, MonadReduce m) => a -> m a

simplifyBlocked' :: Simplify t => Blocked t -> ReduceM t

class Normalise t

normalise' :: Normalise t => t -> ReduceM t

normalise :: (Normalise a, MonadReduce m) => a -> m a

slowNormaliseArgs :: Term -> ReduceM Term