Agda.TypeChecking.Rewriting.Clause

Converting clauses to rewrite rules

getClausesAsRewriteRules :: (HasConstInfo m, ReadTCState m, MonadFresh NameId m) => QName -> m [RewriteRule]

Get all the clauses of a definition and convert them to rewrite rules.

clauseQName :: (HasConstInfo m, MonadFresh NameId m) => QName -> Int -> m QName

Generate a sensible name for the given clause

clauseToRewriteRule :: (MonadTCEnv m, ReadTCState m) => QName -> QName -> Clause -> m (Maybe RewriteRule)

clauseToRewriteRule f q cl converts the clause cl of the function f to a rewrite rule with name q. Returns Nothing if clauseBody cl is Nothing. Precondition: clauseType cl is not Nothing.

class ToNLPat a b where

Minimal complete definition

Nothing

Methods

toNLPat :: a -> b

default toNLPat :: forall a' b' (f :: Type -> Type). (ToNLPat a' b', Functor f, a ~ f a', b ~ f b') => a -> b