Agda.TypeChecking.Rewriting.NonLinMatch

Description

Non-linear matching of the lhs of a rewrite rule against a neutral term.

Given a lhs

Δ ⊢ lhs : B

and a candidate term

Γ ⊢ t : A

we seek a substitution Γ ⊢ σ : Δ such that

Γ ⊢ B[σ] = A and Γ ⊢ lhs[σ] = t : A

Documentation

newtype NLM a

Monad for non-linear matching.

Constructors

NLM

Fields

unNLM :: ExceptT Blocked_ (StateT NLMState ReduceM) a

data NLMState

Constructors

NLMState

Fields

_nlmSub :: Sub

_nlmEqs :: PostponedEquations

nlmSub :: Lens' NLMState Sub

nlmEqs :: Lens' NLMState PostponedEquations

runNLM :: MonadReduce m => NLM () -> m (Either Blocked_ NLMState)

matchingBlocked :: Blocked_ -> NLM ()

tellSub :: Relevance -> Int -> Type -> Term -> NLM ()

Add substitution i |-> v : a to result of matching.

tellEq :: Telescope -> Context -> Type -> Term -> Term -> NLM ()

type Sub = IntMap (Relevance, Term)

data PostponedEquation

Matching against a term produces a constraint which we have to verify after applying the substitution computed by matching.

Constructors

PostponedEquation

Fields

eqFreeVars :: Context

Context of free variables in the equation

eqType :: Type

Type of the equation, living in same context as the rhs.

eqLhs :: Term

Term from pattern, living in pattern context.

eqRhs :: Term

Term from scrutinee, living in context where matching was invoked.

type PostponedEquations = [PostponedEquation]

class Match a b where

Match a non-linear pattern against a neutral term, returning a substitution.

Methods

match

Arguments

:: Relevance
Are we currently matching in an irrelevant context?

-> Telescope
The telescope of pattern variables

-> Context
The context of lambda-bound variables

-> TypeOf b
The type of the pattern

-> a
The pattern to match

-> b
The term to be matched against the pattern

-> NLM ()

extendContext :: MonadAddContext m => Context -> ArgName -> Dom Type -> m Context

makeSubstitution :: Telescope -> Sub -> Maybe Substitution

checkPostponedEquations :: PureTCM m => Substitution -> PostponedEquations -> m (Maybe Blocked_)

nonLinMatch :: (PureTCM m, Match a b) => Telescope -> TypeOf b -> a -> b -> m (Either Blocked_ Substitution)

equal :: PureTCM m => Type -> Term -> Term -> m (Maybe Blocked_)

Typed βη-equality, also handles empty record types. Returns Nothing if the terms are equal, or `Just b` if the terms are not (where b contains information about possible metas blocking the comparison)

getTypedHead :: PureTCM m => Term -> m (Maybe (QName, Type))

Utility function for getting the name and type of a head term (i.e. a Def or Con with no arguments)