Agda.TypeChecking.Rules.Decl

Documentation

checkDeclCached :: Declaration -> TCM ()

Cached checkDecl

checkDecls :: [Declaration] -> TCM ()

Type check a sequence of declarations.

checkDecl :: Declaration -> TCM ()

Type check a single declaration.

mutualChecks :: MutualInfo -> Declaration -> MutualId -> Set QName -> TCM ()

revisitRecordPatternTranslation :: [QName] -> TCM ()

Check if there is a inferred eta record type in the mutual block. If yes, repeat the record pattern translation for all function definitions in the block. This is necessary since the original record pattern translation will have skipped record patterns of the new record types (as eta was off for them). See issue #2308 (and #2197).

type FinalChecks = Maybe (TCM ())

checkUnquoteDecl :: MutualInfo -> [DefInfo] -> [QName] -> Expr -> TCM FinalChecks

checkUnquoteDef :: [DefInfo] -> [QName] -> Expr -> TCM ()

unquoteTop :: [QName] -> Expr -> TCM [QName]

Run a reflected TCM computatation expected to define a given list of names.

instantiateDefinitionType :: QName -> TCM ()

Instantiate all metas in Definition associated to QName. Makes sense after freezing metas. Some checks, like free variable analysis, are not in TCM, so they will be more precise (see issue 1099) after meta instantiation. Precondition: name has been added to signature already.

data HighlightModuleContents

Constructors

DontHightlightModuleContents

DoHighlightModuleContents

highlight_ :: HighlightModuleContents -> Declaration -> TCM ()

Highlight a declaration. Called after checking a mutual block (to ensure we have the right definitions for all names). For modules inside mutual blocks we haven't highlighted their contents, but for modules not in a mutual block we have. Hence the flag.

checkTermination_ :: Declaration -> TCM ()

Termination check a declaration.

checkPositivity_ :: MutualInfo -> Set QName -> TCM ()

Check a set of mutual names for positivity.

checkInjectivity_ :: Set QName -> TCM ()

Check a set of mutual names for constructor-headedness.

checkProjectionLikeness_ :: Set QName -> TCM ()

Check a set of mutual names for projection likeness.

Only a single, non-abstract function can be projection-like. Making an abstract function projection-like would break the invariant that the type of the principle argument of a projection-like function is always inferable.

whenAbstractFreezeMetasAfter :: DefInfo -> TCM a -> TCM a

Freeze metas created by given computation if in abstract mode.

checkGeneralize :: Set QName -> DefInfo -> ArgInfo -> QName -> Expr -> TCM ()

checkAxiom :: KindOfName -> DefInfo -> ArgInfo -> Maybe PragmaPolarities -> QName -> Expr -> TCM ()

Type check an axiom.

checkAxiom' :: Maybe GeneralizeTelescope -> KindOfName -> DefInfo -> ArgInfo -> Maybe PragmaPolarities -> QName -> Expr -> TCM ()

Data and record type signatures need to remember the generalized parameters for when checking the corresponding definition, so for these we pass in the parameter telescope separately.

checkPrimitive :: DefInfo -> QName -> Arg Expr -> TCM ()

Type check a primitive function declaration.

checkPragma :: Range -> Pragma -> TCM ()

Check a pragma.

checkMutual :: MutualInfo -> List1 Declaration -> TCM (MutualId, Set QName)

Type check a bunch of mutual inductive recursive definitions.

All definitions which have so far been assigned to the given mutual block are returned.

checkSig :: KindOfName -> DefInfo -> Erased -> QName -> GeneralizeTelescope -> Expr -> TCM ()

checkTypeSignature :: TypeSignature -> TCM ()

Type check the type signature of an inductive or recursive definition.

checkTypeSignature' :: Maybe GeneralizeTelescope -> TypeSignature -> TCM ()

checkSection :: Erased -> ModuleName -> GeneralizeTelescope -> [Declaration] -> TCM ()

Type check a module.

checkModuleArity

Arguments

:: ModuleName
Name of applied module.

-> Telescope
The module parameters.

-> [NamedArg Expr]
The arguments this module is applied to.

-> TCM Telescope
The remaining module parameters (has free de Bruijn indices!).

Helper for checkSectionApplication.

Matches the arguments of the module application with the module parameters.

Returns the remaining module parameters as an open telescope. Warning: the returned telescope is not the final result, an actual instantiation of the parameters does not occur.

checkSectionApplication

Arguments

:: ModuleInfo

-> Erased
Should "everything" be treated as erased?

-> ModuleName
Name m1 of module defined by the module macro.

-> ModuleApplication
The module macro λ tel → m2 args.

-> ScopeCopyInfo
Imported names and modules

-> ImportDirective

-> TCM ()

Check an application of a section.

checkSectionApplication'

Arguments

:: ModuleInfo

-> Erased

-> ModuleName
Name m1 of module defined by the module macro.

-> ModuleApplication
The module macro λ tel → m2 args.

-> ScopeCopyInfo
Imported names and modules

-> TCM ()

Check an application of a section. (Do not invoke this procedure directly, use checkSectionApplication.)

checkImportDirective :: ImportDirective -> TCM ()

Checks that open public is not used in hard compile-time mode.

Debugging

class ShowHead a where

Methods

showHead :: a -> String

debugPrintDecl :: Declaration -> TCM ()