Agda.TypeChecking.Rules.LHS.Problem

Documentation

type FlexibleVars = [FlexibleVar Nat]

data FlexibleVarKind

When we encounter a flexible variable in the unifier, where did it come from? The alternatives are ordered such that we will assign the higher one first, i.e., first we try to assign a DotFlex, then...

Constructors

RecordFlex [FlexibleVarKind]
From a record pattern (ConP). Saves the FlexibleVarKind of its subpatterns.

ImplicitFlex
From a hidden formal argument or underscore (WildP).

DotFlex
From a dot pattern (DotP).

OtherFlex
From a non-record constructor or literal (ConP or LitP).

data FlexibleVar a

Flexible variables are equipped with information where they come from, in order to make a choice which one to assign when two flexibles are unified.

Constructors

FlexibleVar

Fields

flexArgInfo :: ArgInfo

flexForced :: IsForced

flexKind :: FlexibleVarKind

flexPos :: Maybe Int

flexVar :: a

allFlexVars :: [IsForced] -> Telescope -> FlexibleVars

data FlexChoice

Constructors

ChooseLeft

ChooseRight

ChooseEither

ExpandBoth

class ChooseFlex a where

Methods

chooseFlex :: a -> a -> FlexChoice

data ProblemEq

A user pattern together with an internal term that it should be equal to after splitting is complete. Special cases: * User pattern is a variable but internal term isn't: this will be turned into an as pattern. * User pattern is a dot pattern: this pattern won't trigger any splitting but will be checked for equality after all splitting is complete and as patterns have been bound. * User pattern is an absurd pattern: emptiness of the type will be checked after splitting is complete. * User pattern is an annotated wildcard: type annotation will be checked after splitting is complete.

Constructors

ProblemEq

Fields

problemInPat :: Pattern

problemInst :: Term

problemType :: Dom Type

data Problem a

The user patterns we still have to split on.

Constructors

Problem

Fields

_problemEqs :: [ProblemEq]

User patterns which are typed (including the ones generated from implicit arguments).

_problemRestPats :: [NamedArg Pattern]

List of user patterns which could not yet be typed. Example: f : (b : Bool) -> if b then Nat else Nat -> Nat f true = zero f false zero = zero f false (suc n) = n In this sitation, for clause 2, we construct an initial problem problemEqs = [false = b] problemRestPats = [zero] As we instantiate b to false, the targetType reduces to Nat -> Nat and we can move pattern zero over to problemEqs.

_problemCont :: LHSState a -> TCM a

The code that checks the RHS.

problemEqs :: forall a f. Functor f => ([ProblemEq] -> f [ProblemEq]) -> Problem a -> f (Problem a)

problemRestPats :: forall a f. Functor f => ([NamedArg Pattern] -> f [NamedArg Pattern]) -> Problem a -> f (Problem a)

problemCont :: forall a f. Functor f => ((LHSState a -> TCM a) -> f (LHSState a -> TCM a)) -> Problem a -> f (Problem a)

problemInPats :: Problem a -> [Pattern]

data AsBinding

Constructors

AsB Name Term (Dom Type)

data DotPattern

Constructors

Dot Expr Term (Dom Type)

data AbsurdPattern

Constructors

Absurd Range Type

data AnnotationPattern

Constructors

Ann Expr Type

data LHSState a

State worked on during the main loop of checking a lhs. [Ulf Norell's PhD, page. 35]

Constructors

LHSState

Fields

_lhsTel :: Telescope

The types of the pattern variables.

_lhsOutPat :: [NamedArg DeBruijnPattern]

Patterns after splitting. The de Bruijn indices refer to positions in the list of abstract syntax patterns in the problem, counted from the back (right-to-left).

_lhsProblem :: Problem a

User patterns of supposed type delta.

_lhsTarget :: Arg Type

Type eliminated by problemRestPats in the problem. Can be Irrelevant to indicate that we came by an irrelevant projection and, hence, the rhs must be type-checked in irrelevant mode.

_lhsPartialSplit :: ![Maybe Int]

have we splitted with a PartialFocus?

_lhsIndexedSplit :: !Bool

Have we split on any indexed inductive types?

lhsTel :: forall a f. Functor f => (Telescope -> f Telescope) -> LHSState a -> f (LHSState a)

lhsOutPat :: forall a f. Functor f => ([NamedArg DeBruijnPattern] -> f [NamedArg DeBruijnPattern]) -> LHSState a -> f (LHSState a)

lhsProblem :: forall a f. Functor f => (Problem a -> f (Problem a)) -> LHSState a -> f (LHSState a)

lhsTarget :: forall a f. Functor f => (Arg Type -> f (Arg Type)) -> LHSState a -> f (LHSState a)

data LeftoverPatterns

Constructors

LeftoverPatterns

Fields

patternVariables :: IntMap [(Arg Name, PatVarPosition)]

asPatterns :: [AsBinding]

dotPatterns :: [DotPattern]

absurdPatterns :: [AbsurdPattern]

typeAnnotations :: [AnnotationPattern]

otherPatterns :: [Pattern]

getLeftoverPatterns :: (PureTCM m, MonadFresh NameId m) => [ProblemEq] -> m LeftoverPatterns

Classify remaining patterns after splitting is complete into pattern variables, as patterns, dot patterns, and absurd patterns. Precondition: there are no more constructor patterns.

getUserVariableNames

Arguments

:: Telescope
The telescope of pattern variables

-> IntMap [(Arg Name, PatVarPosition)]
The list of user names for each pattern variable

-> ([Maybe Name], [AsBinding])

Build a renaming for the internal patterns using variable names from the user patterns. If there are multiple user names for the same internal variable, the unused ones are returned as as-bindings. Names that are not also module parameters are preferred over those that are.

Orphan instances

PrettyTCM ProblemEq

Instance details

Methods

prettyTCM :: MonadPretty m => ProblemEq -> m Doc