Agda.TypeChecking.Serialise

Description

Structure-sharing serialisation of Agda interface files.

Documentation

type InterfacePrefix = (Hash, Hash, Word64)

type Encoded = (Word32, Array Node, Array String, Array Text, Array Text, Array Integer, Array VarSet, Array Double)

Hash-consed encoding of a value.

encode :: EmbPrj a => a -> TCM Encoded

Encodes something as a hash-consed tuple of arrays. To ensure relocatability, file paths in positions are replaced with module names.

encodeFile :: FilePath -> Interface -> TCM Interface

Encodes an interface. To ensure relocatability file paths in positions are replaced with module names. A hash-consed and compacted interface is returned.

decode :: EmbPrj a => Encoded -> MaybeT TCM a

Decode a hash-consed value. The result depends on the include path.

decodeFile :: FilePath -> TCM (Maybe Interface)

decodeInterface :: ByteString -> MaybeT TCM Interface

Decode an interface from a bytestring. We check the stored interface version against the current interface version, but we don't do anything with the hash prefix.

deserializeInterface :: ByteString -> MaybeT TCM Encoded

deserializeHashes :: ByteString -> IO (Maybe (Hash, Hash))