Agda.TypeChecking.SizedTypes.Syntax

Description

Syntax of size expressions and constraints.

Syntax

newtype Offset

Constant finite sizes n >= 0.

Constructors

O Int

newtype Rigid

Fixed size variables i.

Constructors

RigidId

Fields

rigidId :: String

newtype Flex

Size meta variables X to solve for.

Constructors

FlexId

Fields

flexId :: String

data SizeExpr' rigid flex

Size expressions appearing in constraints.

Constructors

Const
Constant number n.

Fields

offset :: Offset

Rigid
Variable plus offset i + n.

Fields

rigid :: rigid

offset :: Offset

Infty
Infinity ∞.

Flex
Meta variable X + n.

Fields

flex :: flex

offset :: Offset

type SizeExpr = SizeExpr' Rigid Flex

data Cmp

Comparison operator, e.g. for size expression.

Constructors

Lt
<.

Le
≤.

data Constraint' rigid flex

Constraint: an inequation between size expressions, e.g. X < ∞ or i + 3 ≤ j.

Constructors

Constraint

Fields

leftExpr :: SizeExpr' rigid flex

cmp :: Cmp

rightExpr :: SizeExpr' rigid flex

type Constraint = Constraint' Rigid Flex

Polarities to specify solutions.

data Polarity

What type of solution are we looking for?

Constructors

Least

Greatest

data PolarityAssignment flex

Assigning a polarity to a flexible variable.

Constructors

PolarityAssignment Polarity flex

type Polarities flex = Map flex Polarity

Type of solution wanted for each flexible.

emptyPolarities :: Polarities flex

polaritiesFromAssignments :: Ord flex => [PolarityAssignment flex] -> Polarities flex

getPolarity :: Ord flex => Polarities flex -> flex -> Polarity

Default polarity is Least.

Solutions.

newtype Solution rigid flex

Partial substitution from flexible variables to size expression.

Constructors

Solution

Fields

theSolution :: Map flex (SizeExpr' rigid flex)

emptySolution :: Solution r f

class Substitute r f a where

Executing a substitution.

Methods

subst :: Solution r f -> a -> a

Constraint simplification

type CTrans r f = Constraint' r f -> Maybe [Constraint' r f]

simplify1 :: (Pretty f, Pretty r, Eq r) => CTrans r f -> CTrans r f

Returns Nothing if the constraint is contradictory.

ifLe :: Cmp -> a -> a -> a

Le acts as True, Lt as False.

compareOffset :: Offset -> Cmp -> Offset -> Bool

Interpret Cmp as relation on Offset.

Printing

Wellformedness

class ValidOffset a where

Offsets + n must be non-negative

Methods

validOffset :: a -> Bool

class TruncateOffset a where

Make offsets non-negative by rounding up.

Methods

truncateOffset :: a -> a

Computing variable sets

class Ord (RigidOf a) => Rigids a where

The rigid variables contained in a pice of syntax.

Associated Types

type RigidOf a

Methods

rigids :: a -> Set (RigidOf a)

class Ord (FlexOf a) => Flexs a where

The flexibe variables contained in a pice of syntax.

Associated Types

type FlexOf a

Methods

flexs :: a -> Set (FlexOf a)

data NamedRigid

Identifiers for rigid variables.

Constructors

NamedRigid

Fields

rigidName :: String

Name for printing in debug messages.

rigidIndex :: Int

De Bruijn index.

data SizeMeta

Size metas in size expressions.

Constructors

SizeMeta

Fields

sizeMetaId :: MetaId

sizeMetaArgs :: [Int]

De Bruijn indices.

type DBSizeExpr = SizeExpr' NamedRigid SizeMeta

Size expression with de Bruijn indices.

type SizeConstraint = Constraint' NamedRigid SizeMeta

data HypSizeConstraint

Size constraint with de Bruijn indices.

Constructors

HypSizeConstraint

Fields

sizeContext :: Context

sizeHypIds :: [Nat]

DeBruijn indices

sizeHypotheses :: [SizeConstraint]

Living in Context.

sizeConstraint :: SizeConstraint

Living in Context.