Agda.TypeChecking.Substitute

Description

This module contains the definition of hereditary substitution and application operating on internal syntax which is in β-normal form (β including projection reductions).

Further, it contains auxiliary functions which rely on substitution but not on reduction.

Documentation

class TeleNoAbs a where

Performs void (noAbs) abstraction over telescope.

Methods

teleNoAbs :: a -> Term -> Term

data TelV a

Constructors

TelV

Fields

theTel :: Tele (Dom a)

theCore :: a

type TelView = TelV Type

applyTermE :: forall t. (Coercible Term t, Apply t, EndoSubst t) => (Empty -> Term -> Elims -> Term) -> t -> Elims -> t

Apply Elims while using the given function to report ill-typed redexes. Recursive calls for applyE and applySubst happen at type t to propagate the same strategy to subtrees.

canProject :: QName -> Term -> Maybe (Arg Term)

If $v$ is a record value, canProject f v returns its field f.

conApp :: forall t. (Coercible t Term, Apply t) => (Empty -> Term -> Elims -> Term) -> ConHead -> ConInfo -> Elims -> Elims -> Term

Eliminate a constructed term.

defApp :: QName -> Elims -> Elims -> Term

defApp f us vs applies Def f us to further arguments vs, eliminating top projection redexes. If us is not empty, we cannot have a projection redex, since the record argument is the first one.

argToDontCare :: Arg Term -> Term

relToDontCare :: LensRelevance a => a -> Term -> Term

piApply :: Type -> Args -> Type

(x:A)->B(x) piApply [u] = B(u)

Precondition: The type must contain the right number of pis without having to perform any reduction.

piApply is potentially unsafe, the monadic piApplyM is preferable.

telVars :: Int -> Telescope -> [Arg DeBruijnPattern]

namedTelVars :: Int -> Telescope -> [NamedArg DeBruijnPattern]

abstractArgs :: Abstract a => Args -> a -> a

renaming :: forall a. DeBruijn a => Impossible -> Permutation -> Substitution' a

If permute π : [a]Γ -> [a]Δ, then applySubst (renaming _ π) : Term Γ -> Term Δ

renamingR :: DeBruijn a => Permutation -> Substitution' a

If permute π : [a]Γ -> [a]Δ, then applySubst (renamingR π) : Term Δ -> Term Γ

renameP :: Subst a => Impossible -> Permutation -> a -> a

The permutation should permute the corresponding context. (right-to-left list)

applySubstTerm :: forall t. (Coercible t Term, EndoSubst t, Apply t) => Substitution' t -> t -> t

applyNLPatSubst :: TermSubst a => Substitution' NLPat -> a -> a

applyNLSubstToDom :: SubstWith NLPat a => Substitution' NLPat -> Dom a -> Dom a

fromPatternSubstitution :: PatternSubstitution -> Substitution

applyPatSubst :: TermSubst a => PatternSubstitution -> a -> a

usePatOrigin :: PatOrigin -> Pattern' a -> Pattern' a

usePatternInfo :: PatternInfo -> Pattern' a -> Pattern' a

projDropParsApply :: Projection -> ProjOrigin -> Relevance -> Args -> Term

projDropParsApply proj o args = projDropPars proj o `apply` args

This function is an optimization, saving us from construction lambdas we immediately remove through application.

telView' :: Type -> TelView

Takes off all exposed function domains from the given type. This means that it does not reduce to expose Pi-types.

telView'UpTo :: Int -> Type -> TelView

telView'UpTo n t takes off the first n exposed function types of t. Takes off all (exposed ones) if n < 0.

bindsToTel' :: (Name -> a) -> [Name] -> Dom Type -> ListTel' a

Turn a typed binding (x1 .. xn : A) into a telescope.

bindsToTel :: [Name] -> Dom Type -> ListTel

bindsToTel'1 :: (Name -> a) -> List1 Name -> Dom Type -> ListTel' a

bindsToTel1 :: List1 Name -> Dom Type -> ListTel

namedBindsToTel :: [NamedArg Name] -> Type -> Telescope

Turn a typed binding (x1 .. xn : A) into a telescope.

namedBindsToTel1 :: List1 (NamedArg Name) -> Type -> Telescope

domFromNamedArgName :: NamedArg Name -> Dom ()

mkPiSort :: Dom Type -> Abs Type -> Sort

mkPi :: Dom (ArgName, Type) -> Type -> Type

mkPi dom t = telePi (telFromList [dom]) t

mkLam :: Arg ArgName -> Term -> Term

lamView :: Term -> ([Arg ArgName], Term)

unlamView :: [Arg ArgName] -> Term -> Term

telePi' :: (Abs Type -> Abs Type) -> Telescope -> Type -> Type

telePi :: Telescope -> Type -> Type

Uses free variable analysis to introduce NoAbs bindings.

telePi_ :: Telescope -> Type -> Type

Everything will be an Abs.

telePiVisible :: Telescope -> Type -> Type

Only abstract the visible components of the telescope, and all that bind variables. Everything will be an Abs! Caution: quadratic time!

teleLam :: Telescope -> Term -> Term

Abstract over a telescope in a term, producing lambdas. Dumb abstraction: Always produces Abs, never NoAbs.

The implementation is sound because Telescope does not use NoAbs.

typeArgsWithTel :: Telescope -> [Term] -> Maybe [Dom Type]

Given arguments vs : tel (vector typing), extract their individual types. Returns Nothing is tel is not long enough.

compiledClauseBody :: Clause -> Maybe Term

In compiled clauses, the variables in the clause body are relative to the pattern variables (including dot patterns) instead of the clause telescope.

univSort' :: Sort -> Maybe Sort

univSort' univInf s gets the next higher sort of s, if it is known (i.e. it is not just UnivSort s).

Precondition: s is reduced

univSort :: Sort -> Sort

sort :: Sort -> Type

ssort :: Level -> Type

isSmallSort :: Sort -> Maybe (Bool, IsFibrant)

Returns Nothing for unknown (meta) sorts, and otherwise returns Just (b,f) where b indicates smallness and f fibrancy. I.e., b is True for (relatively) small sorts like Set l and Prop l, and instead b is False for large sorts such as Setω.

fibrantLub :: IsFibrant -> IsFibrant -> IsFibrant

funSort' :: Sort -> Sort -> Maybe Sort

Compute the sort of a function type from the sorts of its domain and codomain.

funSort :: Sort -> Sort -> Sort

piSort' :: Dom Term -> Sort -> Abs Sort -> Maybe Sort

Compute the sort of a pi type from the sorts of its domain and codomain.

piSort :: Dom Term -> Sort -> Abs Sort -> Sort

levelMax :: Integer -> [PlusLevel] -> Level

levelLub :: Level -> Level -> Level

Given two levels a and b, compute a ⊔ b and return its canonical form.

levelTm :: Level -> Term

module Agda.TypeChecking.Substitute.Class

module Agda.TypeChecking.Substitute.DeBruijn

data Substitution' a

Substitutions.

Constructors

IdS
Identity substitution. Γ ⊢ IdS : Γ

EmptyS Impossible
Empty substitution, lifts from the empty context. First argument is IMPOSSIBLE. Apply this to closed terms you want to use in a non-empty context. Γ ⊢ EmptyS : ()

a :# (Substitution' a) infixr 4
Substitution extension, `cons'. Γ ⊢ u : Aρ Γ ⊢ ρ : Δ ---------------------- Γ ⊢ u :# ρ : Δ, A

Strengthen Impossible (Substitution' a)
Strengthening substitution. First argument is IMPOSSIBLE. Apply this to a term which does not contain variable 0 to lower all de Bruijn indices by one. Γ ⊢ ρ : Δ --------------------------- Γ ⊢ Strengthen ρ : Δ, A

Wk !Int (Substitution' a)
Weakening substitution, lifts to an extended context. Γ ⊢ ρ : Δ ------------------- Γ, Ψ ⊢ Wk |Ψ| ρ : Δ

Lift !Int (Substitution' a)
Lifting substitution. Use this to go under a binder. Lift 1 ρ == var 0 :# Wk 1 ρ. Γ ⊢ ρ : Δ ------------------------- Γ, Ψρ ⊢ Lift |Ψ| ρ : Δ, Ψ

type Substitution = Substitution' Term

Orphan instances

Eq Substitution

Instance details

Methods

(==) :: Substitution -> Substitution -> Bool

(/=) :: Substitution -> Substitution -> Bool

Eq NotBlocked

Instance details

Methods

(==) :: NotBlocked -> NotBlocked -> Bool

(/=) :: NotBlocked -> NotBlocked -> Bool

Eq PlusLevel

Instance details

Methods

(==) :: PlusLevel -> PlusLevel -> Bool

(/=) :: PlusLevel -> PlusLevel -> Bool

Eq Level

Instance details

Methods

(==) :: Level -> Level -> Bool

(/=) :: Level -> Level -> Bool

Eq Sort

Instance details

Methods

(==) :: Sort -> Sort -> Bool

(/=) :: Sort -> Sort -> Bool

Eq Term
Syntactic Term equality, ignores stuff below DontCare and sharing.

Instance details

Methods

(==) :: Term -> Term -> Bool

(/=) :: Term -> Term -> Bool

Eq Candidate

Instance details

Methods

(==) :: Candidate -> Candidate -> Bool

(/=) :: Candidate -> Candidate -> Bool

Eq CandidateKind

Instance details

Methods

(==) :: CandidateKind -> CandidateKind -> Bool

(/=) :: CandidateKind -> CandidateKind -> Bool

Eq Section

Instance details

Methods

(==) :: Section -> Section -> Bool

(/=) :: Section -> Section -> Bool

Ord Substitution

Instance details

Methods

compare :: Substitution -> Substitution -> Ordering

(<) :: Substitution -> Substitution -> Bool

(<=) :: Substitution -> Substitution -> Bool

(>) :: Substitution -> Substitution -> Bool

(>=) :: Substitution -> Substitution -> Bool

max :: Substitution -> Substitution -> Substitution

min :: Substitution -> Substitution -> Substitution

Ord PlusLevel

Instance details

Methods

compare :: PlusLevel -> PlusLevel -> Ordering

(<) :: PlusLevel -> PlusLevel -> Bool

(<=) :: PlusLevel -> PlusLevel -> Bool

(>) :: PlusLevel -> PlusLevel -> Bool

(>=) :: PlusLevel -> PlusLevel -> Bool

max :: PlusLevel -> PlusLevel -> PlusLevel

min :: PlusLevel -> PlusLevel -> PlusLevel

Ord Level

Instance details

Methods

compare :: Level -> Level -> Ordering

(<) :: Level -> Level -> Bool

(<=) :: Level -> Level -> Bool

(>) :: Level -> Level -> Bool

(>=) :: Level -> Level -> Bool

max :: Level -> Level -> Level

min :: Level -> Level -> Level

Ord Sort

Instance details

Methods

compare :: Sort -> Sort -> Ordering

(<) :: Sort -> Sort -> Bool

(<=) :: Sort -> Sort -> Bool

(>) :: Sort -> Sort -> Bool

(>=) :: Sort -> Sort -> Bool

max :: Sort -> Sort -> Sort

min :: Sort -> Sort -> Sort

Ord Term

Instance details

Methods

compare :: Term -> Term -> Ordering

(<) :: Term -> Term -> Bool

(<=) :: Term -> Term -> Bool

(>) :: Term -> Term -> Bool

(>=) :: Term -> Term -> Bool

max :: Term -> Term -> Term

min :: Term -> Term -> Term

DeBruijn BraveTerm

Instance details

Methods

deBruijnVar :: Int -> BraveTerm

debruijnNamedVar :: String -> Int -> BraveTerm

deBruijnView :: BraveTerm -> Maybe Int

DeBruijn NLPat

Instance details

Methods

deBruijnVar :: Int -> NLPat

debruijnNamedVar :: String -> Int -> NLPat

deBruijnView :: NLPat -> Maybe Int

Subst ()

Instance details

Associated Types

type SubstArg ()

Methods

applySubst :: Substitution' (SubstArg ()) -> () -> ()

Subst Range

Instance details

Associated Types

type SubstArg Range

Methods

applySubst :: Substitution' (SubstArg Range) -> Range -> Range

Subst Name

Instance details

Associated Types

type SubstArg Name

Methods

applySubst :: Substitution' (SubstArg Name) -> Name -> Name

Subst EqualityView

Instance details

Associated Types

type SubstArg EqualityView

Methods

applySubst :: Substitution' (SubstArg EqualityView) -> EqualityView -> EqualityView

Subst ConPatternInfo

Instance details

Associated Types

type SubstArg ConPatternInfo

Methods

applySubst :: Substitution' (SubstArg ConPatternInfo) -> ConPatternInfo -> ConPatternInfo

Subst DeBruijnPattern

Instance details

Associated Types

type SubstArg DeBruijnPattern

Methods

applySubst :: Substitution' (SubstArg DeBruijnPattern) -> DeBruijnPattern -> DeBruijnPattern

Subst Pattern

Instance details

Associated Types

type SubstArg Pattern

Methods

applySubst :: Substitution' (SubstArg Pattern) -> Pattern -> Pattern

Subst BraveTerm

Instance details

Associated Types

type SubstArg BraveTerm

Methods

applySubst :: Substitution' (SubstArg BraveTerm) -> BraveTerm -> BraveTerm

Subst Term

Instance details

Associated Types

type SubstArg Term

Methods

applySubst :: Substitution' (SubstArg Term) -> Term -> Term

Subst ProblemEq

Instance details

Associated Types

type SubstArg ProblemEq

Methods

applySubst :: Substitution' (SubstArg ProblemEq) -> ProblemEq -> ProblemEq

Subst Candidate

Instance details

Associated Types

type SubstArg Candidate

Methods

applySubst :: Substitution' (SubstArg Candidate) -> Candidate -> Candidate

Subst RewriteRule

Instance details

Associated Types

type SubstArg RewriteRule

Methods

applySubst :: Substitution' (SubstArg RewriteRule) -> RewriteRule -> RewriteRule

Subst NLPSort

Instance details

Associated Types

type SubstArg NLPSort

Methods

applySubst :: Substitution' (SubstArg NLPSort) -> NLPSort -> NLPSort

Subst NLPType

Instance details

Associated Types

type SubstArg NLPType

Methods

applySubst :: Substitution' (SubstArg NLPType) -> NLPType -> NLPType

Subst NLPat

Instance details

Associated Types

type SubstArg NLPat

Methods

applySubst :: Substitution' (SubstArg NLPat) -> NLPat -> NLPat

Subst DisplayTerm

Instance details

Associated Types

type SubstArg DisplayTerm

Methods

applySubst :: Substitution' (SubstArg DisplayTerm) -> DisplayTerm -> DisplayTerm

Subst DisplayForm

Instance details

Associated Types

type SubstArg DisplayForm

Methods

applySubst :: Substitution' (SubstArg DisplayForm) -> DisplayForm -> DisplayForm

Subst CompareAs

Instance details

Associated Types

type SubstArg CompareAs

Methods

applySubst :: Substitution' (SubstArg CompareAs) -> CompareAs -> CompareAs

Subst Constraint

Instance details

Associated Types

type SubstArg Constraint

Methods

applySubst :: Substitution' (SubstArg Constraint) -> Constraint -> Constraint

Abstract Permutation

Instance details

Methods

abstract :: Telescope -> Permutation -> Permutation

Abstract Clause

Instance details

Methods

abstract :: Telescope -> Clause -> Clause

Abstract Sort

Instance details

Methods

abstract :: Telescope -> Sort -> Sort

Abstract Telescope

Instance details

Methods

abstract :: Telescope -> Telescope -> Telescope

Abstract Type

Instance details

Methods

abstract :: Telescope -> Type -> Type

Abstract Term

Instance details

Methods

abstract :: Telescope -> Term -> Term

Abstract CompiledClauses

Instance details

Methods

abstract :: Telescope -> CompiledClauses -> CompiledClauses

Abstract FunctionInverse

Instance details

Methods

abstract :: Telescope -> FunctionInverse -> FunctionInverse

Abstract PrimFun

Instance details

Methods

abstract :: Telescope -> PrimFun -> PrimFun

Abstract Defn

Instance details

Methods

abstract :: Telescope -> Defn -> Defn

Abstract ProjLams

Instance details

Methods

abstract :: Telescope -> ProjLams -> ProjLams

Abstract Projection

Instance details

Methods

abstract :: Telescope -> Projection -> Projection

Abstract System

Instance details

Methods

abstract :: Telescope -> System -> System

Abstract NumGeneralizableArgs

Instance details

Methods

abstract :: Telescope -> NumGeneralizableArgs -> NumGeneralizableArgs

Abstract Definition

Instance details

Methods

abstract :: Telescope -> Definition -> Definition

Abstract RewriteRule
tel ⊢ (Γ ⊢ lhs ↦ rhs : t) becomes tel, Γ ⊢ lhs ↦ rhs : t) we do not need to change lhs, rhs, and t since they live in Γ. See 'Abstract Clause'.

Instance details

Methods

abstract :: Telescope -> RewriteRule -> RewriteRule

Apply Permutation

Instance details

Methods

apply :: Permutation -> Args -> Permutation

applyE :: Permutation -> Elims -> Permutation

Apply Clause

Instance details

Methods

apply :: Clause -> Args -> Clause

applyE :: Clause -> Elims -> Clause

Apply BraveTerm

Instance details

Methods

apply :: BraveTerm -> Args -> BraveTerm

applyE :: BraveTerm -> Elims -> BraveTerm

Apply Sort

Instance details

Methods

apply :: Sort -> Args -> Sort

applyE :: Sort -> Elims -> Sort

Apply Term

Instance details

Methods

apply :: Term -> Args -> Term

applyE :: Term -> Elims -> Term

Apply CompiledClauses

Instance details

Methods

apply :: CompiledClauses -> Args -> CompiledClauses

applyE :: CompiledClauses -> Elims -> CompiledClauses

Apply FunctionInverse

Instance details

Methods

apply :: FunctionInverse -> Args -> FunctionInverse

applyE :: FunctionInverse -> Elims -> FunctionInverse

Apply PrimFun

Instance details

Methods

apply :: PrimFun -> Args -> PrimFun

applyE :: PrimFun -> Elims -> PrimFun

Apply Defn

Instance details

Methods

apply :: Defn -> Args -> Defn

applyE :: Defn -> Elims -> Defn

Apply ProjLams

Instance details

Methods

apply :: ProjLams -> Args -> ProjLams

applyE :: ProjLams -> Elims -> ProjLams

Apply Projection

Instance details

Methods

apply :: Projection -> Args -> Projection

applyE :: Projection -> Elims -> Projection

Apply ExtLamInfo

Instance details

Methods

apply :: ExtLamInfo -> Args -> ExtLamInfo

applyE :: ExtLamInfo -> Elims -> ExtLamInfo

Apply System

Instance details

Methods

apply :: System -> Args -> System

applyE :: System -> Elims -> System

Apply NumGeneralizableArgs

Instance details

Methods

apply :: NumGeneralizableArgs -> Args -> NumGeneralizableArgs

applyE :: NumGeneralizableArgs -> Elims -> NumGeneralizableArgs

Apply Definition

Instance details

Methods

apply :: Definition -> Args -> Definition

applyE :: Definition -> Elims -> Definition

Apply RewriteRule

Instance details

Methods

apply :: RewriteRule -> Args -> RewriteRule

applyE :: RewriteRule -> Elims -> RewriteRule

Apply DisplayTerm

Instance details

Methods

apply :: DisplayTerm -> Args -> DisplayTerm

applyE :: DisplayTerm -> Elims -> DisplayTerm

(Subst a, Eq a) => Eq (Elim' a)

Instance details

Methods

(==) :: Elim' a -> Elim' a -> Bool

(/=) :: Elim' a -> Elim' a -> Bool

Eq a => Eq (Pattern' a)

Instance details

Methods

(==) :: Pattern' a -> Pattern' a -> Bool

(/=) :: Pattern' a -> Pattern' a -> Bool

Eq t => Eq (Blocked t)

Instance details

Methods

(==) :: Blocked t -> Blocked t -> Bool

(/=) :: Blocked t -> Blocked t -> Bool

(Subst a, Eq a) => Eq (Tele a)

Instance details

Methods

(==) :: Tele a -> Tele a -> Bool

(/=) :: Tele a -> Tele a -> Bool

Eq a => Eq (Type' a)
Syntactic Type equality, ignores sort annotations.

Instance details

Methods

(==) :: Type' a -> Type' a -> Bool

(/=) :: Type' a -> Type' a -> Bool

(Subst a, Eq a) => Eq (Abs a)
Equality of binders relies on weakening which is a special case of renaming which is a special case of substitution.

Instance details

Methods

(==) :: Abs a -> Abs a -> Bool

(/=) :: Abs a -> Abs a -> Bool

(Subst a, Ord a) => Ord (Elim' a)

Instance details

Methods

compare :: Elim' a -> Elim' a -> Ordering

(<) :: Elim' a -> Elim' a -> Bool

(<=) :: Elim' a -> Elim' a -> Bool

(>) :: Elim' a -> Elim' a -> Bool

(>=) :: Elim' a -> Elim' a -> Bool

max :: Elim' a -> Elim' a -> Elim' a

min :: Elim' a -> Elim' a -> Elim' a

(Subst a, Ord a) => Ord (Tele a)

Instance details

Methods

compare :: Tele a -> Tele a -> Ordering

(<) :: Tele a -> Tele a -> Bool

(<=) :: Tele a -> Tele a -> Bool

(>) :: Tele a -> Tele a -> Bool

(>=) :: Tele a -> Tele a -> Bool

max :: Tele a -> Tele a -> Tele a

min :: Tele a -> Tele a -> Tele a

Ord a => Ord (Type' a)

Instance details

Methods

compare :: Type' a -> Type' a -> Ordering

(<) :: Type' a -> Type' a -> Bool

(<=) :: Type' a -> Type' a -> Bool

(>) :: Type' a -> Type' a -> Bool

(>=) :: Type' a -> Type' a -> Bool

max :: Type' a -> Type' a -> Type' a

min :: Type' a -> Type' a -> Type' a

(Subst a, Ord a) => Ord (Abs a)

Instance details

Methods

compare :: Abs a -> Abs a -> Ordering

(<) :: Abs a -> Abs a -> Bool

(<=) :: Abs a -> Abs a -> Bool

(>) :: Abs a -> Abs a -> Bool

(>=) :: Abs a -> Abs a -> Bool

max :: Abs a -> Abs a -> Abs a

min :: Abs a -> Abs a -> Abs a

Ord a => Ord (Dom a)

Instance details

Methods

compare :: Dom a -> Dom a -> Ordering

(<) :: Dom a -> Dom a -> Bool

(<=) :: Dom a -> Dom a -> Bool

(>) :: Dom a -> Dom a -> Bool

(>=) :: Dom a -> Dom a -> Bool

max :: Dom a -> Dom a -> Dom a

min :: Dom a -> Dom a -> Dom a

DeBruijn a => DeBruijn (Pattern' a)

Instance details

Methods

deBruijnVar :: Int -> Pattern' a

debruijnNamedVar :: String -> Int -> Pattern' a

deBruijnView :: Pattern' a -> Maybe Int

Subst a => Subst [a]

Instance details

Associated Types

type SubstArg [a]

Methods

applySubst :: Substitution' (SubstArg [a]) -> [a] -> [a]

Subst a => Subst (Maybe a)

Instance details

Associated Types

type SubstArg (Maybe a)

Methods

applySubst :: Substitution' (SubstArg (Maybe a)) -> Maybe a -> Maybe a

Subst a => Subst (Arg a)

Instance details

Associated Types

type SubstArg (Arg a)

Methods

applySubst :: Substitution' (SubstArg (Arg a)) -> Arg a -> Arg a

Subst a => Subst (WithHiding a)

Instance details

Associated Types

type SubstArg (WithHiding a)

Methods

applySubst :: Substitution' (SubstArg (WithHiding a)) -> WithHiding a -> WithHiding a

Subst a => Subst (Elim' a)

Instance details

Associated Types

type SubstArg (Elim' a)

Methods

applySubst :: Substitution' (SubstArg (Elim' a)) -> Elim' a -> Elim' a

EndoSubst a => Subst (Substitution' a)

Instance details

Associated Types

type SubstArg (Substitution' a)

Methods

applySubst :: Substitution' (SubstArg (Substitution' a)) -> Substitution' a -> Substitution' a

Subst a => Subst (Blocked a)

Instance details

Associated Types

type SubstArg (Blocked a)

Methods

applySubst :: Substitution' (SubstArg (Blocked a)) -> Blocked a -> Blocked a

Subst a => Subst (PlusLevel' a)

Instance details

Associated Types

type SubstArg (PlusLevel' a)

Methods

applySubst :: Substitution' (SubstArg (PlusLevel' a)) -> PlusLevel' a -> PlusLevel' a

Subst a => Subst (Level' a)

Instance details

Associated Types

type SubstArg (Level' a)

Methods

applySubst :: Substitution' (SubstArg (Level' a)) -> Level' a -> Level' a

(Coercible a Term, Subst a) => Subst (Sort' a)

Instance details

Associated Types

type SubstArg (Sort' a)

Methods

applySubst :: Substitution' (SubstArg (Sort' a)) -> Sort' a -> Sort' a

Subst a => Subst (Tele a)

Instance details

Associated Types

type SubstArg (Tele a)

Methods

applySubst :: Substitution' (SubstArg (Tele a)) -> Tele a -> Tele a

Subst a => Subst (Abs a)

Instance details

Associated Types

type SubstArg (Abs a)

Methods

applySubst :: Substitution' (SubstArg (Abs a)) -> Abs a -> Abs a

Abstract t => Abstract [t]

Instance details

Methods

abstract :: Telescope -> [t] -> [t]

Abstract [Polarity]

Instance details

Methods

abstract :: Telescope -> [Polarity] -> [Polarity]

Abstract [Occurrence]

Instance details

Methods

abstract :: Telescope -> [Occurrence] -> [Occurrence]

Abstract t => Abstract (Maybe t)

Instance details

Methods

abstract :: Telescope -> Maybe t -> Maybe t

DoDrop a => Abstract (Drop a)

Instance details

Methods

abstract :: Telescope -> Drop a -> Drop a

Abstract a => Abstract (Case a)

Instance details

Methods

abstract :: Telescope -> Case a -> Case a

Abstract a => Abstract (WithArity a)

Instance details

Methods

abstract :: Telescope -> WithArity a -> WithArity a

Apply t => Apply [t]

Instance details

Methods

apply :: [t] -> Args -> [t]

applyE :: [t] -> Elims -> [t]

Apply [NamedArg (Pattern' a)]
Make sure we only drop variable patterns.

Instance details

Methods

apply :: [NamedArg (Pattern' a)] -> Args -> [NamedArg (Pattern' a)]

applyE :: [NamedArg (Pattern' a)] -> Elims -> [NamedArg (Pattern' a)]

Apply [Polarity]

Instance details

Methods

apply :: [Polarity] -> Args -> [Polarity]

applyE :: [Polarity] -> Elims -> [Polarity]

Apply [Occurrence]

Instance details

Methods

apply :: [Occurrence] -> Args -> [Occurrence]

applyE :: [Occurrence] -> Elims -> [Occurrence]

Apply t => Apply (Maybe t)

Instance details

Methods

apply :: Maybe t -> Args -> Maybe t

applyE :: Maybe t -> Elims -> Maybe t

Apply t => Apply (Maybe t)

Instance details

Methods

apply :: Maybe t -> Args -> Maybe t

applyE :: Maybe t -> Elims -> Maybe t

DoDrop a => Apply (Drop a)

Instance details

Methods

apply :: Drop a -> Args -> Drop a

applyE :: Drop a -> Elims -> Drop a

Apply t => Apply (Blocked t)

Instance details

Methods

apply :: Blocked t -> Args -> Blocked t

applyE :: Blocked t -> Elims -> Blocked t

TermSubst a => Apply (Tele a)

Instance details

Methods

apply :: Tele a -> Args -> Tele a

applyE :: Tele a -> Elims -> Tele a

Apply a => Apply (Case a)

Instance details

Methods

apply :: Case a -> Args -> Case a

applyE :: Case a -> Elims -> Case a

Apply a => Apply (WithArity a)

Instance details

Methods

apply :: WithArity a -> Args -> WithArity a

applyE :: WithArity a -> Elims -> WithArity a

(Subst a, Subst b, SubstArg a ~ SubstArg b) => Subst (a, b)

Instance details

Associated Types

type SubstArg (a, b)

Methods

applySubst :: Substitution' (SubstArg (a, b)) -> (a, b) -> (a, b)

(Ord k, Subst a) => Subst (Map k a)

Instance details

Associated Types

type SubstArg (Map k a)

Methods

applySubst :: Substitution' (SubstArg (Map k a)) -> Map k a -> Map k a

Subst a => Subst (Named name a)

Instance details

Associated Types

type SubstArg (Named name a)

Methods

applySubst :: Substitution' (SubstArg (Named name a)) -> Named name a -> Named name a

(Coercible a Term, Subst a, Subst b, SubstArg a ~ SubstArg b) => Subst (Type'' a b)

Instance details

Associated Types

type SubstArg (Type'' a b)

Methods

applySubst :: Substitution' (SubstArg (Type'' a b)) -> Type'' a b -> Type'' a b

(Subst a, Subst b, SubstArg a ~ SubstArg b) => Subst (Dom' a b)

Instance details

Associated Types

type SubstArg (Dom' a b)

Methods

applySubst :: Substitution' (SubstArg (Dom' a b)) -> Dom' a b -> Dom' a b

Abstract v => Abstract (Map k v)

Instance details

Methods

abstract :: Telescope -> Map k v -> Map k v

Abstract v => Abstract (HashMap k v)

Instance details

Methods

abstract :: Telescope -> HashMap k v -> HashMap k v

(Apply a, Apply b) => Apply (a, b)

Instance details

Methods

apply :: (a, b) -> Args -> (a, b)

applyE :: (a, b) -> Elims -> (a, b)

Apply v => Apply (Map k v)

Instance details

Methods

apply :: Map k v -> Args -> Map k v

applyE :: Map k v -> Elims -> Map k v

Apply v => Apply (HashMap k v)

Instance details

Methods

apply :: HashMap k v -> Args -> HashMap k v

applyE :: HashMap k v -> Elims -> HashMap k v

(Subst a, Subst b, Subst c, SubstArg a ~ SubstArg b, SubstArg b ~ SubstArg c) => Subst (a, b, c)

Instance details

Associated Types

type SubstArg (a, b, c)

Methods

applySubst :: Substitution' (SubstArg (a, b, c)) -> (a, b, c) -> (a, b, c)

(Apply a, Apply b, Apply c) => Apply (a, b, c)

Instance details

Methods

apply :: (a, b, c) -> Args -> (a, b, c)

applyE :: (a, b, c) -> Elims -> (a, b, c)

(Subst a, Subst b, Subst c, Subst d, SubstArg a ~ SubstArg b, SubstArg b ~ SubstArg c, SubstArg c ~ SubstArg d) => Subst (a, b, c, d)

Instance details

Associated Types

type SubstArg (a, b, c, d)

Methods

applySubst :: Substitution' (SubstArg (a, b, c, d)) -> (a, b, c, d) -> (a, b, c, d)