Agda.TypeChecking.Substitute.Class

Application

class Apply t where

Apply something to a bunch of arguments. Preserves blocking tags (application can never resolve blocking).

Minimal complete definition

applyE

Methods

apply :: t -> Args -> t

applyE :: t -> Elims -> t

applys :: Apply t => t -> [Term] -> t

Apply to some default arguments.

apply1 :: Apply t => t -> Term -> t

Apply to a single default argument.

apply2 :: Apply t => t -> Term -> Term -> t

Apply to two default arguments.

Abstraction

class Abstract t where

(abstract args v) apply args --> v[args].

Methods

abstract :: Telescope -> t -> t

Substitution and shifting/weakening/strengthening

class DeBruijn (SubstArg a) => Subst a where

Apply a substitution.

Minimal complete definition

Nothing

Associated Types

type SubstArg a

Methods

applySubst :: Substitution' (SubstArg a) -> a -> a

default applySubst :: forall (f :: Type -> Type) b. (a ~ f b, Functor f, Subst b, SubstArg a ~ SubstArg b) => Substitution' (SubstArg a) -> a -> a

type SubstWith t a = (Subst a, SubstArg a ~ t)

Simple constraint alias for a Subst instance a with arg type t.

type EndoSubst a = SubstWith a a

Subst instance whose agument type is itself

type TermSubst a = SubstWith Term a

Subst instance whose argument type is Term

raise :: Subst a => Nat -> a -> a

Raise de Bruijn index, i.e. weakening

raiseFrom :: Subst a => Nat -> Nat -> a -> a

subst :: Subst a => Int -> SubstArg a -> a -> a

Replace de Bruijn index i by a Term in something.

strengthen :: Subst a => Impossible -> a -> a

substUnder :: Subst a => Nat -> SubstArg a -> a -> a

Replace what is now de Bruijn index 0, but go under n binders. substUnder n u == subst n (raise n u).

isNoAbs :: (Free a, Subst a) => Abs a -> Maybe a

Checks whether the variable bound by the abstraction is actually used, and, if not, returns the term within, strengthened to live in the context outside the abstraction. See also isBinderUsed.

Identity instances

Explicit substitutions

idS :: Substitution' a

wkS :: Int -> Substitution' a -> Substitution' a

raiseS :: Int -> Substitution' a

consS :: DeBruijn a => a -> Substitution' a -> Substitution' a

singletonS :: DeBruijn a => Int -> a -> Substitution' a

To replace index n by term u, do applySubst (singletonS n u). Γ, Δ ⊢ u : A --------------------------------- Γ, Δ ⊢ singletonS |Δ| u : Γ, A, Δ

inplaceS :: EndoSubst a => Int -> a -> Substitution' a

Single substitution without disturbing any deBruijn indices. Γ, A, Δ ⊢ u : A --------------------------------- Γ, A, Δ ⊢ inplace |Δ| u : Γ, A, Δ

liftS :: Int -> Substitution' a -> Substitution' a

Lift a substitution under k binders.

dropS :: Int -> Substitution' a -> Substitution' a

        Γ ⊢ ρ : Δ, Ψ
     -------------------
     Γ ⊢ dropS |Ψ| ρ : Δ
  

composeS :: EndoSubst a => Substitution' a -> Substitution' a -> Substitution' a

applySubst (ρ composeS σ) v == applySubst ρ (applySubst σ v)

splitS :: Int -> Substitution' a -> (Substitution' a, Substitution' a)

(++#) :: DeBruijn a => [a] -> Substitution' a -> Substitution' a infixr 4

prependS :: DeBruijn a => Impossible -> [Maybe a] -> Substitution' a -> Substitution' a

     Γ ⊢ ρ : Δ  Γ ⊢ reverse vs : Θ
     ----------------------------- (treating Nothing as having any type)
       Γ ⊢ prependS vs ρ : Δ, Θ
  

parallelS :: DeBruijn a => [a] -> Substitution' a

       Γ ⊢ reverse vs : Δ
     -----------------------------
       Γ ⊢ parallelS vs ρ : Γ, Δ
  

Note the Γ in Γ, Δ.

strengthenS :: Impossible -> Int -> Substitution' a

Γ ⊢ (strengthenS ⊥ |Δ|) : Γ,Δ

strengthenS' :: Impossible -> Int -> Substitution' a -> Substitution' a

A "smart" variant of Strengthen. If strengthenS is applied to a substitution with an outermost Strengthen constructor, then the "error message" of that constructor is discarded in favour of the Impossible argument of this function.

lookupS :: EndoSubst a => Substitution' a -> Nat -> a

listS :: EndoSubst a => [(Int, a)] -> Substitution' a

lookupS (listS [(x0,t0)..(xn,tn)]) xi = ti, assuming x0 < .. < xn.

raiseFromS :: Nat -> Nat -> Substitution' a

Γ, Ξ, Δ ⊢ raiseFromS |Δ| |Ξ| : Γ, Δ

Functions on abstractions

absApp :: Subst a => Abs a -> SubstArg a -> a

Instantiate an abstraction. Strict in the term.

lazyAbsApp :: Subst a => Abs a -> SubstArg a -> a

Instantiate an abstraction. Lazy in the term, which allow it to be IMPOSSIBLE in the case where the variable shouldn't be used but we cannot use noabsApp. Used in Apply.

noabsApp :: Subst a => Impossible -> Abs a -> a

Instantiate an abstraction that doesn't use its argument.

absBody :: Subst a => Abs a -> a

mkAbs :: (Subst a, Free a) => ArgName -> a -> Abs a

reAbs :: (Subst a, Free a) => Abs a -> Abs a