Agda.TypeChecking.Telescope

Documentation

flattenTel :: TermSubst a => Tele (Dom a) -> [Dom a]

Flatten telescope: (Γ : Tel) -> [Type Γ].

flattenTel is lazy in both the spine and values of the resulting list.

flattenRevTel :: TermSubst a => Tele (Dom a) -> [Dom a]

Flatten telescope: (Γ : Tel) -> [Type Γ] into a reversed list.

flattenRevTel is lazy in both the spine and values of the resulting list.

flattenContext :: Context -> [ContextEntry]

Turn a context into a flat telescope: all entries live in the whole context. (Γ : Context) -> [Type Γ]

reorderTel :: [Dom Type] -> Maybe Permutation

Order a flattened telescope in the correct dependeny order: Γ -> Permutation (Γ -> Γ~)

Since reorderTel tel uses free variable analysis of type in tel, the telescope should be normalised.

reorderTel_ :: [Dom Type] -> Permutation

unflattenTel :: [ArgName] -> [Dom Type] -> Telescope

Unflatten: turns a flattened telescope into a proper telescope. Must be properly ordered.

unflattenTel' :: Int -> [ArgName] -> [Dom Type] -> Telescope

A variant of unflattenTel which takes the size of the last argument as an argument.

renameTel :: [Maybe ArgName] -> Telescope -> Telescope

Rename the variables in the telescope to the given names Precondition: size xs == size tel.

teleNames :: Telescope -> [ArgName]

Get the suggested names from a telescope

teleArgNames :: Telescope -> [Arg ArgName]

teleArgs :: DeBruijn a => Tele (Dom t) -> [Arg a]

Convert a telescope to a list of Arg in descending order.

teleDoms :: DeBruijn a => Tele (Dom t) -> [Dom a]

Convert a telescope to a list of Dom in descending order.

teleNamedArgs :: DeBruijn a => Tele (Dom t) -> [NamedArg a]

tele2NamedArgs :: DeBruijn a => Telescope -> Telescope -> [NamedArg a]

A variant of teleNamedArgs which takes the argument names (and the argument info) from the first telescope and the variable names from the second telescope.

Precondition: the two telescopes have the same length.

splitTelescopeAt :: Int -> Telescope -> (Telescope, Telescope)

Split the telescope at the specified position.

permuteTel :: Permutation -> Telescope -> Telescope

Permute telescope: permutes or drops the types in the telescope according to the given permutation. Assumes that the permutation preserves the dependencies in the telescope.

For example (Andreas, 2016-12-18, issue #2344): tel = (A : Set) (X : _18 A) (i : Fin (_m_23 A X)) tel (de Bruijn) = 2:Set, 1:_18 0, 0:Fin(_m_23 1 0) flattenTel tel = 2:Set, 1:_18 0, 0:Fin(_m_23 1 0) |- [ Set, _18 2, Fin (_m_23 2 1) ] perm = 0,1,2 -> 0,1 (picks the first two) renaming _ perm = [var 0, var 1, error] -- THE WRONG RENAMING! renaming _ (flipP perm) = [error, var 1, var 0] -- The correct renaming! apply to flattened tel = ... |- [ Set, _18 1, Fin (_m_23 1 0) ] permute perm it = ... |- [ Set, _18 1 ] unflatten (de Bruijn) = 1:Set, 0: _18 0 unflatten = (A : Set) (X : _18 A)

permuteContext :: Permutation -> Context -> Telescope

Like permuteTel, but start with a context.

varDependencies :: Telescope -> VarSet -> VarSet

Recursively computes dependencies of a set of variables in a given telescope. Any dependencies outside of the telescope are ignored.

Note that varDependencies considers a variable to depend on itself.

varDependents :: Telescope -> VarSet -> VarSet

Computes the set of variables in a telescope whose type depend on one of the variables in the given set (including recursive dependencies). Any dependencies outside of the telescope are ignored.

Unlike varDependencies, a variable is *not* considered to depend on itself.

data SplitTel

A telescope split in two.

Constructors

SplitTel

Fields

firstPart :: Telescope

secondPart :: Telescope

splitPerm :: Permutation

The permutation takes us from the original telescope to firstPart ++ secondPart.

splitTelescope

Arguments

:: VarSet
A set of de Bruijn indices.

-> Telescope
Original telescope.

-> SplitTel
firstPart mentions the given variables, secondPart not.

Split a telescope into the part that defines the given variables and the part that doesn't.

See prop_splitTelescope.

splitTelescopeExact

Arguments

:: [Int]
A list of de Bruijn indices

-> Telescope
The telescope to split

-> Maybe SplitTel
firstPart mentions the given variables in the given order, secondPart contains all other variables

As splitTelescope, but fails if any additional variables or reordering would be needed to make the first part well-typed.

instantiateTelescope

Arguments

:: Telescope
⊢ Γ

-> Int
Γ ⊢ var k : A de Bruijn _level_

-> DeBruijnPattern
Γ ⊢ u : A

-> Maybe (Telescope, PatternSubstitution, Permutation)

Try to instantiate one variable in the telescope (given by its de Bruijn level) with the given value, returning the new telescope and a substitution to the old one. Returns Nothing if the given value depends (directly or indirectly) on the variable.

expandTelescopeVar :: Telescope -> Int -> Telescope -> ConHead -> (Telescope, PatternSubstitution)

Try to eta-expand one variable in the telescope (given by its de Bruijn level)

telView :: (MonadReduce m, MonadAddContext m) => Type -> m TelView

Gather leading Πs of a type in a telescope.

telViewUpTo :: (MonadReduce m, MonadAddContext m) => Int -> Type -> m TelView

telViewUpTo n t takes off the first n function types of t. Takes off all if n < 0.

telViewUpTo' :: (MonadReduce m, MonadAddContext m) => Int -> (Dom Type -> Bool) -> Type -> m TelView

telViewUpTo' n p t takes off $t$ the first n (or arbitrary many if n < 0) function domains as long as they satify p.

telViewPath :: PureTCM m => Type -> m TelView

telViewUpToPath :: PureTCM m => Int -> Type -> m TelView

telViewUpToPath n t takes off t the first n (or arbitrary many if n < 0) function domains or Path types.

telViewUpToPath n t = fst $ telViewUpToPathBoundary' n t

newtype Boundary' x a

Boundary conditions [[ (i,(x,y)) ]] = [(i=0) -> x, (i=1) -> y]. For instance, if p : Path A a b, then p i has boundary condition (i,(a,b)). We call i the dimension and (x,y) its boundary.

Constructors

Boundary

Fields

theBoundary :: [(x, (a, a))]

type Boundary = Boundary' Int Term

Usually, the dimensions of a boundary condition are interval variables, represented by a de Bruijn index Int.

type TmBoundary = Boundary' Term Term

Substitution formally creates dimensions that are interval expressions, represented by a Term. However, in practice these terms should be of the form var i.

tmBoundary :: Boundary' Int a -> Boundary' Term a

Turn dimension variables i into dimension expressions var i.

varBoundary :: Boundary' Term a -> Boundary' Int a

Turn dimension expressions into dimension variables. Formally this is a partial operation, but should only be called when the precondition is met.

Precondition: the dimension terms in the boundary are all of the form var i.

telViewUpToPathBoundary' :: PureTCM m => Int -> Type -> m (TelView, Boundary)

Like telViewUpToPath but also returns the Boundary expected by the Path types encountered. The boundary terms live in the telescope given by the TelView. Each point of the boundary has the type of the codomain of the Path type it got taken from, see fullBoundary.

 (TelV Γ b, [(i,t_i,u_i)]) <- telViewUpToPathBoundary' n a
 Input:  Δ ⊢ a
 Output: Δ.Γ ⊢ b
         Δ.Γ ⊢ T is the codomain of the PathP at variable i
         Δ.Γ ⊢ i : I
         Δ.Γ ⊢ [ (i=0) -> t_i; (i=1) -> u_i ] : T

Useful to reconstruct IApplyP patterns after teleNamedArgs Γ.

fullBoundary :: Telescope -> Boundary -> Boundary

telViewUpToPathBoundary :: PureTCM m => Int -> Type -> m (TelView, Boundary)

(TelV Γ b, [(i,t_i,u_i)]) <- telViewUpToPathBoundary n a Input: Δ ⊢ a Output: ΔΓ ⊢ b ΔΓ ⊢ i : I ΔΓ ⊢ [ (i=0) -> t_i; (i=1) -> u_i ] : b

telViewPathBoundary :: PureTCM m => Type -> m (TelView, Boundary)

teleElims :: DeBruijn a => Telescope -> Boundary' Int a -> [Elim' a]

teleElimsB args bs = es Input: Δ.Γ ⊢ args : Γ Δ.Γ ⊢ T is the codomain of the PathP at variable i Δ.Γ ⊢ i : I Δ.Γ ⊢ bs = [ (i=0) -> t_i; (i=1) -> u_i ] : T Output: Δ.Γ | PiPath Γ bs A ⊢ es : A

pathViewAsPi :: PureTCM m => Type -> m (Either (Dom Type, Abs Type) Type)

Reduces Type.

pathViewAsPi' :: PureTCM m => Type -> m (Either ((Dom Type, Abs Type), (Term, Term)) Type)

Reduces Type.

pathViewAsPi'whnf :: HasBuiltins m => m (Type -> Either ((Dom Type, Abs Type), (Term, Term)) Type)

piOrPath :: HasBuiltins m => Type -> m (Either (Dom Type, Abs Type) Type)

Returns Left (a,b) in case the type is Pi a b or PathP b _ _. Assumes the Type is in whnf.

telView'UpToPath :: Int -> Type -> TCM TelView

Assumes Type is in whnf.

telView'Path :: Type -> TCM TelView

isPath :: PureTCM m => Type -> m (Maybe (Dom Type, Abs Type))

ifPath :: PureTCM m => Type -> (Dom Type -> Abs Type -> m a) -> (Type -> m a) -> m a

ifPathB :: PureTCM m => Type -> (Dom Type -> Abs Type -> m a) -> (Blocked Type -> m a) -> m a

ifNotPathB :: PureTCM m => Type -> (Blocked Type -> m a) -> (Dom Type -> Abs Type -> m a) -> m a

ifPiOrPathB :: PureTCM m => Type -> (Dom Type -> Abs Type -> m a) -> (Blocked Type -> m a) -> m a

ifNotPiOrPathB :: PureTCM m => Type -> (Blocked Type -> m a) -> (Dom Type -> Abs Type -> m a) -> m a

telePatterns :: DeBruijn a => Telescope -> Boundary -> [NamedArg (Pattern' a)]

telePatterns' :: DeBruijn a => (forall a1. DeBruijn a1 => Telescope -> [NamedArg a1]) -> Telescope -> Boundary -> [NamedArg (Pattern' a)]

mustBePi :: MonadReduce m => Type -> m (Dom Type, Abs Type)

Decomposing a function type.

ifPi :: MonadReduce m => Term -> (Dom Type -> Abs Type -> m a) -> (Term -> m a) -> m a

If the given type is a Pi, pass its parts to the first continuation. If not (or blocked), pass the reduced type to the second continuation.

ifPiB :: MonadReduce m => Term -> (Dom Type -> Abs Type -> m a) -> (Blocked Term -> m a) -> m a

ifPiTypeB :: MonadReduce m => Type -> (Dom Type -> Abs Type -> m a) -> (Blocked Type -> m a) -> m a

ifPiType :: MonadReduce m => Type -> (Dom Type -> Abs Type -> m a) -> (Type -> m a) -> m a

If the given type is a Pi, pass its parts to the first continuation. If not (or blocked), pass the reduced type to the second continuation.

ifNotPi :: MonadReduce m => Term -> (Term -> m a) -> (Dom Type -> Abs Type -> m a) -> m a

If the given type is blocked or not a Pi, pass it reduced to the first continuation. If it is a Pi, pass its parts to the second continuation.

ifNotPiType :: MonadReduce m => Type -> (Type -> m a) -> (Dom Type -> Abs Type -> m a) -> m a

If the given type is blocked or not a Pi, pass it reduced to the first continuation. If it is a Pi, pass its parts to the second continuation.

ifNotPiOrPathType :: (MonadReduce tcm, HasBuiltins tcm) => Type -> (Type -> tcm a) -> (Dom Type -> Abs Type -> tcm a) -> tcm a

shouldBePath :: (PureTCM m, MonadBlock m, MonadTCError m) => Type -> m (Dom Type, Abs Type)

shouldBePi :: (PureTCM m, MonadBlock m, MonadTCError m) => Type -> m (Dom Type, Abs Type)

shouldBePiOrPath :: (PureTCM m, MonadBlock m, MonadTCError m) => Type -> m (Dom Type, Abs Type)

class PiApplyM a where

A safe variant of piApply.

Minimal complete definition

piApplyM'

Methods

piApplyM' :: (MonadReduce m, HasBuiltins m) => m Empty -> Type -> a -> m Type

piApplyM :: (MonadReduce m, HasBuiltins m) => Type -> a -> m Type

typeArity :: Type -> TCM Nat

Compute type arity

foldrTelescopeM :: MonadAddContext m => (Dom (ArgName, Type) -> m b -> m b) -> m b -> Telescope -> m b

Fold a telescope into a monadic computation, adding variables to the context at each step.