Agda.TypeChecking.Telescope.Path

Documentation

telePiPath :: (Abs Type -> Abs Type) -> ([Arg ArgName] -> Term -> Term) -> Telescope -> Type -> Boundary -> TCM Type

In an ambient context Γ, telePiPath f lams Δ t bs builds a type that can be telViewPathBoundaryed into (TelV Δ t, bs'). Γ.Δ ⊢ t bs = [(i,u_i)] Δ = Δ0,(i : I),Δ1 ∀ b ∈ {0,1}. Γ.Δ0 | lams Δ1 (u_i .b) : (telePiPath f Δ1 t bs)(i = b) -- kinda: see lams Γ ⊢ telePiPath f Δ t bs

telePiPath_ :: Telescope -> Type -> Boundary -> TCM Type

telePiPath_ Δ t [(i,u)] Δ ⊢ t i ∈ Δ Δ ⊢ u_b : t for b ∈ {0,1}

arityPiPath :: Type -> TCM Int

arity of the type, including both Pi and Path. Does not reduce the type.

class IApplyVars p where

Collect the interval copattern variables as list of de Bruijn indices.

Methods

iApplyVars :: p -> [Int]

isInterval :: (MonadTCM m, MonadReduce m) => Type -> m Bool

Check whether a type is the built-in interval type.