Agda.TypeChecking.Unquote

Documentation

agdaTermType :: TCM Type

agdaTypeType :: TCM Type

qNameType :: TCM Type

data Dirty

Constructors

Dirty

Clean

type UnquoteState = (Dirty, TCState)

type UnquoteM = ReaderT Context (StateT UnquoteState (WriterT [QName] (ExceptT UnquoteError TCM)))

type UnquoteRes a = Either UnquoteError ((a, UnquoteState), [QName])

unpackUnquoteM :: UnquoteM a -> Context -> UnquoteState -> TCM (UnquoteRes a)

packUnquoteM :: (Context -> UnquoteState -> TCM (UnquoteRes a)) -> UnquoteM a

runUnquoteM :: UnquoteM a -> TCM (Either UnquoteError (a, [QName]))

liftU1 :: (TCM (UnquoteRes a) -> TCM (UnquoteRes b)) -> UnquoteM a -> UnquoteM b

liftU2 :: (TCM (UnquoteRes a) -> TCM (UnquoteRes b) -> TCM (UnquoteRes c)) -> UnquoteM a -> UnquoteM b -> UnquoteM c

inOriginalContext :: UnquoteM a -> UnquoteM a

isCon :: ConHead -> TCM (Maybe Term) -> UnquoteM Bool

isDef :: QName -> TCM (Maybe Term) -> UnquoteM Bool

reduceQuotedTerm :: Term -> UnquoteM Term

class Unquote a where

Methods

unquote :: Term -> UnquoteM a

unquoteN :: (HasCallStack, Unquote a) => Arg Term -> UnquoteM a

Unquote an Arg Term assuming the ArgInfo does not contain information. (This means, it should be visible, relevant, etc., like defaultArg).

choice :: Monad m => [(m Bool, m a)] -> m a -> m a

ensureDef :: QName -> UnquoteM QName

ensureCon :: QName -> UnquoteM QName

pickName :: Type -> String

unquoteString :: Term -> UnquoteM String

unquoteNString :: Arg Term -> UnquoteM Text

data ErrorPart

Constructors

StrPart String

TermPart Expr

PattPart Pattern

NamePart QName

renderErrorParts :: [ErrorPart] -> TCM Doc

We do a little bit of work here to make it possible to generate nice layout for multi-line error messages. Specifically we split the parts into lines (indicated by n in a string part) and vcat all the lines.

unquoteTCM :: Term -> Term -> UnquoteM Term

Argument should be a term of type Term → TCM A for some A. Returns the resulting term of type A. The second argument is the term for the hole, which will typically be a metavariable. This is passed to the computation (quoted).

evalTCM :: Term -> UnquoteM Term

Trusted executables

type ExeArg = Text

type StdIn = Text

type StdOut = Text

type StdErr = Text

requireAllowExec :: TCM ()

Raise an error if the --allow-exec option was not specified.

exitCodeToNat :: ExitCode -> Nat

Convert an ExitCode to an Agda natural number.

tcExec :: ExeName -> [ExeArg] -> StdIn -> TCM Term

Call a trusted executable with the given arguments and input.

Returns the exit code, stdout, and stderr.