Agda.TypeChecking.Warnings

Documentation

class (MonadPretty m, MonadError TCErr m) => MonadWarning (m :: Type -> Type) where

Minimal complete definition

Nothing

Methods

addWarning :: TCWarning -> m ()

Store a warning and generate highlighting from it.

default addWarning :: forall (n :: Type -> Type) (t :: (Type -> Type) -> Type -> Type). (MonadWarning n, MonadTrans t, t n ~ m) => TCWarning -> m ()

warning'_ :: MonadWarning m => CallStack -> Warning -> m TCWarning

warning_ :: (HasCallStack, MonadWarning m) => Warning -> m TCWarning

warning' :: MonadWarning m => CallStack -> Warning -> m ()

warning :: (HasCallStack, MonadWarning m) => Warning -> m ()

warnings :: (HasCallStack, MonadWarning m) => [Warning] -> m ()

raiseWarningsOnUsage :: (MonadWarning m, ReadTCState m) => QName -> m ()

Raise every WARNING_ON_USAGE connected to a name.

isUnsolvedWarning :: Warning -> Bool

isMetaWarning :: Warning -> Bool

isMetaTCWarning :: TCWarning -> Bool

onlyShowIfUnsolved :: Warning -> Bool

data WhichWarnings

Classifying warnings: some are benign, others are (non-fatal) errors

Constructors

ErrorWarnings
warnings that will be turned into errors

AllWarnings
all warnings, including errors and benign ones Note: order of constructors is important for the derived Ord instance

classifyWarning :: Warning -> WhichWarnings

data WarningsAndNonFatalErrors

Assorted warnings and errors to be displayed to the user

tcWarnings :: WarningsAndNonFatalErrors -> [TCWarning]

nonFatalErrors :: WarningsAndNonFatalErrors -> [TCWarning]

emptyWarningsAndNonFatalErrors :: WarningsAndNonFatalErrors

The only way to construct a empty WarningsAndNonFatalErrors

classifyWarnings :: [TCWarning] -> WarningsAndNonFatalErrors

runPM :: PM a -> TCM a

running the Parse monad