Agda.Utils.Applicative

Documentation

(?*>) :: Alternative f => Bool -> f a -> f a

Guard: return the action f only if the boolean is True

(?$>) :: Alternative f => Bool -> a -> f a

Guard: return the value a only if the boolean is True

foldA :: (Alternative f, Foldable t) => t a -> f a

Branch over a Foldable collection of values.

foldMapA :: (Alternative f, Foldable t) => (a -> f b) -> t a -> f b

Branch over a Foldable collection of values using the supplied action.

forA :: (Traversable t, Applicative f) => t a -> (a -> f b) -> f (t b)

Better name for for.