Agda.Utils.Bag

Description

A simple overlay over Data.Map to manage unordered sets with duplicates.

Documentation

newtype Bag a

A set with duplicates. Faithfully stores elements which are equal with regard to (==).

Constructors

Bag

Fields

bag :: Map a (NonEmpty a)

The list contains all occurrences of a (not just the duplicates!). Hence, the invariant: the list is never empty.

Query

null :: Bag a -> Bool

Is the bag empty?

size :: Bag a -> Int

Number of elements in the bag. Duplicates count. O(n).

(!) :: Ord a => Bag a -> a -> [a]

(bag ! a) finds all elements equal to a. O(log n). Total function, returns [] if none are.

member :: Ord a => a -> Bag a -> Bool

O(log n).

notMember :: Ord a => a -> Bag a -> Bool

O(log n).

count :: Ord a => a -> Bag a -> Int

Return the multiplicity of the given element. O(log n + count _ _).

Construction

empty :: Bag a

O(1)

singleton :: a -> Bag a

O(1)

union :: Ord a => Bag a -> Bag a -> Bag a

unions :: Ord a => [Bag a] -> Bag a

insert :: Ord a => a -> Bag a -> Bag a

insert a b = union b (singleton a)

fromList :: Ord a => [a] -> Bag a

fromList = unions . map singleton

Destruction

groups :: Bag a -> [NonEmpty a]

Returns the elements of the bag, grouped by equality (==).

toList :: Bag a -> [a]

Returns the bag, with duplicates.

keys :: Bag a -> [a]

Returns the bag without duplicates.

elems :: Bag a -> [a]

Returns the bag, with duplicates.

toAscList :: Bag a -> [a]

Traversal

map :: Ord b => (a -> b) -> Bag a -> Bag b

O(n).

traverse' :: forall a b m. (Applicative m, Ord b) => (a -> m b) -> Bag a -> m (Bag b)