Agda.Utils.Boolean

Description

Boolean algebras and types isomorphic to Bool.

There are already solutions for Boolean algebras in the Haskell ecosystem, but they do not offer easy instantiations for types isomorphic to Bool. In particular, if type a is isomorphic to Bool, so it satisfies `IsBool a`, we would like to instantiate 'Boolean a' by just giving true and false. To facilitate this within the limits of the Haskell class system, we define the class Boolean mutually with class IsBool, so that operations not, (&&), and (||) can get default implementations.

Usage: import Prelude hiding ( not, (&&), (||) ) import Agda.Utils.Boolean

Documentation

class Boolean a where

Boolean algebras.

Minimal complete definition

fromBool

Methods

fromBool :: Bool -> a

true :: a

false :: a

not :: a -> a

default not :: IsBool a => a -> a

(&&) :: a -> a -> a infixr 3

default (&&) :: IsBool a => a -> a -> a

(||) :: a -> a -> a infixr 2

default (||) :: IsBool a => a -> a -> a

implies :: a -> a -> a

butNot :: a -> a -> a

Set difference, dual to implies.

class (Boolean a, Eq a) => IsBool a where

Types isomorphic to Bool.

Minimal complete definition

toBool

Methods

toBool :: a -> Bool

ifThenElse :: a -> b -> b -> b

fromBool1 :: (Bool -> Bool) -> a -> a

fromBool2 :: (Bool -> Bool -> Bool) -> a -> a -> a