Agda.Utils.CallStack

Documentation

data CallStack

data SrcLoc

Constructors

SrcLoc

Fields

srcLocPackage :: [Char]

srcLocModule :: [Char]

srcLocFile :: [Char]

srcLocStartLine :: Int

srcLocStartCol :: Int

srcLocEndLine :: Int

srcLocEndCol :: Int

getCallStack :: CallStack -> [([Char], SrcLoc)]

type HasCallStack = ?callStack :: CallStack

fromCallSiteList :: [([Char], SrcLoc)] -> CallStack

prettySrcLoc :: SrcLoc -> String

callStack :: HasCallStack => CallStack

type CallSiteFilter = CallSite -> Bool

Type of a filter for CallSite

type CallSite = (SrcFun, SrcLoc)

Type of an entry in a CallStack

type SrcLocCol = Int

Type of a column of a SrcLoc

type SrcLocLine = Int

Type of a line number of a SrcLoc

type SrcLocFile = String

Type of a filename of a SrcLoc | e.g. `srcfullAgdaUtilsFoo.hs`

type SrcFun = String

Type of the name of a function in a CallSite | e.g. proveEverything

type SrcLocModule = String

Type of the module name of a SrcLoc | e.g. Foo

type SrcLocPackage = String

Type of the package name of a SrcLoc | e.g. `Agda-2.…`

prettyCallSite :: CallSite -> String

The same as the un-exported internal function in GHC.Exceptions (prettyCallStackLines) Prints like: doFoo, called at foo.hs:190:24 in main:Main

prettyCallStack :: CallStack -> String

Pretty-print a CallStack. This has a few differences from GHC.Stack.prettyCallStackLines. We omit the "CallStack (from GetCallStack)" header line for brevity. If there is only one entry (which is common, due to the manual nature of the HasCallStack constraint), shows the entry on one line. If there are multiple, then the following lines are indented.

headCallSite :: CallStack -> Maybe CallSite

Get the most recent CallSite in a CallStack, if there is one.

truncatedCallStack :: CallStack -> CallStack

CallStack comprising only the most recent CallSite

overCallSites :: ([CallSite] -> [CallSite]) -> CallStack -> CallStack

Transform a CallStack by transforming its list of CallSite

filterCallStack :: CallSiteFilter -> CallStack -> CallStack

Transform a CallStack by filtering each CallSite

popnCallStack :: Word -> CallStack -> CallStack

Pops n entries off a CallStack using popCallStack. Note that frozen callstacks are unaffected.

withNBackCallStack :: HasCallStack => Word -> (CallStack -> b) -> b

withCurrentCallStack :: HasCallStack => (CallStack -> b) -> b

withCallerCallStack :: HasCallStack => (CallStack -> b) -> b