Agda.Utils.Cluster

Description

Create clusters of non-overlapping things.

Documentation

cluster :: Ord c => (a -> NonEmpty c) -> [a] -> [NonEmpty a]

Given a function f :: a -> NonEmpty c which returns a non-empty list of characteristics of a, partition a list of as into groups such that each element in a group shares at least one characteristic with at least one other element of the group.

cluster' :: Ord c => [(a, NonEmpty c)] -> [NonEmpty a]

Partition a list of as paired with a non-empty list of characteristics into groups such that each element in a group shares at least one characteristic with at least one other element of the group.

cluster1 :: Ord c => (a -> NonEmpty c) -> NonEmpty a -> NonEmpty (NonEmpty a)

Given a function f :: a -> NonEmpty c which returns a non-empty list of characteristics of a, partition a non-empty list of as into groups such that each element in a group shares at least one characteristic with at least one other element of the group.

cluster1' :: Ord c => NonEmpty (a, NonEmpty c) -> NonEmpty (NonEmpty a)

Partition a non-empty list of as paired with a non-empty list of characteristics into groups such that each element in a group shares at least one characteristic with at least one other element of the group.