Agda.Utils.CompressedTrie

Description

Strict compressed tries (based on Data.Map.Strict and Agda.Utils.Maybe.Strict).

In contrast to Agda.Utils.Trie, branches from a node can be labeled with non-empty lists of keys that do not overlap, meaning they start with different first elements. This way, non-branching unlabelled pathes can be stored more efficiently.

As a side effect, the notion of height changes which we utilize in the truncate function.

An Trie is usually constructed from an Trie via the compress function.

Documentation

data Trie k v

Finite map from [k] to v.

With the strict Maybe type, Trie is also strict in v.

Invariant: only the root of the tree may be of the form Trie Nothing (Map.singleton ks t). A node inside of the trie of that form must be fused into its branch.

Constructors

Trie

Fields

trieValue :: !(Maybe v)

trieBranches :: !(Map (List1 k) (Trie k v))

Construction

root :: v -> Trie k v

Singleton tree with value at the root.

singleton :: [k] -> v -> Trie k v

Singleton trie.

singleton1 :: List1 k -> v -> Trie k v

Singleton tree with empty root.

singleBranch :: List1 k -> Trie k v -> (List1 k, Trie k v)

A single branch with the given subtree. Precondition: the given subtree has a value at the root.

isSingleBranch :: Trie k v -> Maybe (List1 k, Trie k v)

Is the trie starting with a single branch?

compress :: Ord k => Trie k v -> Trie k v

Turn a trie into a compressed trie.

Deconstruction

uncompress :: Ord k => Trie k v -> Trie k v

Turn a compressed trie into an uncompressed trie.

toList :: Ord k => Trie k v -> [([k], v)]

Convert to ascending list.

toAscList :: Ord k => Trie k v -> [([k], v)]

Convert to ascending list.

size :: Trie k v -> Int

The number of labeled nodes.

sizeWithLeaves :: Trie k v -> Int

The number of nodes that are labeled or leaves..

Modification

substLeaves :: (Maybe v -> Trie k v) -> Trie k v -> Trie k v

Modify the leaves (input and output may be non-canonical).

truncate :: Int -> Trie k v -> Trie k v

Truncate tree at given depth. truncate 0 t returns the root.

truncateSize :: (Trie k v -> Int) -> Int -> Trie k v -> Trie k v

Truncate the tree so that it has at most the given size according to the given sizing function. truncateSize sizing 0 t returns the trie truncate 0 t.

truncateSizeWithLeaves :: Int -> Trie k v -> Trie k v

truncateSize sizeWithLeaves