Agda.Utils.Empty

Description

An empty type with some useful instances.

Documentation

data Empty

absurd :: Empty -> a

toImpossible :: Empty -> IO Impossible

toImpossible e extracts the Impossible value raised via IMPOSSIBLE to create the element e of type Empty. It proceeds by evaluating e to weak head normal form and catching the exception. We are forced to wrap things in a Maybe because of catchImpossible's type.