Agda.Utils.FileId

Description

Translating between file paths and ids.

This module allows you to build a dictionary from file paths to some unique identifier and to look up both file paths and identifiers in that dictionary.

Documentation

type File = AbsolutePath

newtype FileId

Unique identifier of a file.

Constructors

FileId

Fields

theFileId :: Word32

Mapping between files and their unique identifiers.

type FileToId = Map File FileId

type IdToFile = EnumMap FileId File

data FileDict

Constructors

FileDict

Fields

fileToId :: FileToId

idToFile :: IdToFile

class GetFileId a where

Translate a file to an ID; mapping must exist.

Methods

getFileId :: a -> File -> FileId

class GetIdFile a where

Translate a ID to a file; mapping must exist.

Methods

getIdFile :: a -> FileId -> File

Constructing a mapping between files and their unique identifiers.

data FileDictBuilder

Constructors

FileDictBuilder

Fields

nextFileId :: FileId

fileDict :: FileDict

registerFileId :: File -> FileDictBuilder -> (FileId, FileDictBuilder)

Register a new file identifier or retrieve an existing one.

registerFileId' :: File -> FileDictBuilder -> ((FileId, Bool), FileDictBuilder)

Register a new file identifier (True) or retrieve an existing one (False).

registerFileId'' :: File -> FileDictBuilder -> (FileId, Maybe FileDictBuilder)

Register a new file identifier or retrieve an existing one.

If Nothing is returned, the file was already registered.

Monadic interface

class Monad m => MonadFileId (m :: Type -> Type) where

Minimal complete definition

Nothing

Methods

fileFromId :: FileId -> m File

default fileFromId :: forall (t :: (Type -> Type) -> Type -> Type) (n :: Type -> Type). (MonadTrans t, MonadFileId n, m ~ t n) => FileId -> m File

idFromFile :: File -> m FileId

default idFromFile :: forall (t :: (Type -> Type) -> Type -> Type) (n :: Type -> Type). (MonadTrans t, MonadFileId n, m ~ t n) => File -> m FileId