Agda.Utils.Functor

Description

Utilities for functors.

Documentation

(<.>) :: Functor m => (b -> c) -> (a -> m b) -> a -> m c infixr 9

Composition: pure function after functorial (monadic) function.

for :: Functor m => m a -> (a -> b) -> m b

The true pure for loop. for is a misnomer, it should be forA.

class Functor t => Decoration t where

A decoration is a functor that is traversable into any functor.

The Functor superclass is given because of the limitations of the Haskell class system. traverseF actually implies functoriality.

Minimal complete definition: traverseF or distributeF.

Minimal complete definition

Nothing

Methods

traverseF :: Functor m => (a -> m b) -> t a -> m (t b)

traverseF is the defining property.

distributeF :: Functor m => t (m a) -> m (t a)

Decorations commute into any functor.

dmap :: Decoration t => (a -> b) -> t a -> t b

Any decoration is traversable with traverse = traverseF. Just like any Traversable is a functor, so is any decoration, given by just traverseF, a functor.

dget :: Decoration t => t a -> a

Any decoration is a lens. set is a special case of dmap.

(<$>) :: Functor f => (a -> b) -> f a -> f b

($>) :: Functor f => f a -> b -> f b

(<&>) :: Functor m => m a -> (a -> b) -> m b infixl 1

Infix version of for.