Agda.Utils.IO.UTF8

Description

Text IO using the UTF8 character encoding.

Documentation

data ReadException

A kind of exception that can be thrown by readTextFile and readFile.

readTextFile :: FilePath -> IO Text

Reads a UTF8-encoded text file and converts many character sequences which may be interpreted as line or paragraph separators into 'n'.

If the file cannot be decoded, then a ReadException is raised.

readFile :: FilePath -> IO String

Reads a UTF8-encoded text file and converts many character sequences which may be interpreted as line or paragraph separators into 'n'.

If the file cannot be decoded, then a ReadException is raised.

writeFile :: FilePath -> String -> IO ()

Writes a UTF8-encoded text file. The native convention for line endings is used.

writeTextToFile :: FilePath -> Text -> IO ()

Writes a UTF8-encoded text file. The native convention for line endings is used.