Agda.Utils.Impossible

Description

An interface for reporting "impossible" errors

Documentation

data Impossible

"Impossible" errors, annotated with a file name and a line number corresponding to the source code location of the error.

Constructors

Impossible CallStack
We reached a program point which should be unreachable.

Unreachable CallStack
Impossible with a different error message. Used when we reach a program point which can in principle be reached, but not for a certain run.

ImpMissingDefinitions [String] String
We reached a program point without all the required primitives or BUILTIN to proceed forward. ImpMissingDefinitions neededDefs forThis

throwImpossible :: Impossible -> a

Abort by throwing an "impossible" error. You should not use this function directly. Instead use IMPOSSIBLE

class CatchImpossible (m :: Type -> Type) where

Monads in which we can catch an "impossible" error, if possible.

Minimal complete definition

catchImpossibleJust | handleImpossibleJust

Methods

catchImpossible :: m a -> (Impossible -> m a) -> m a

Catch any Impossible exception.

catchImpossibleJust :: (Impossible -> Maybe b) -> m a -> (b -> m a) -> m a

Catch only Impossible exceptions selected by the filter.

handleImpossible :: (Impossible -> m a) -> m a -> m a

Version of catchImpossible with argument order suiting short handlers.

handleImpossibleJust :: (Impossible -> Maybe b) -> (b -> m a) -> m a -> m a

Version of catchImpossibleJust with argument order suiting short handlers.

impossible :: HasCallStack => Impossible

Construct a value of Impossible reporting the location where you call this function.

__IMPOSSIBLE__ :: HasCallStack => a

Throw an Impossible error reporting the location where you place __IMPOSSIBLE__.

__IMPOSSIBLE__1 :: HasCallStack => a

Throw an Impossible error reporting the *caller's* call site.

__UNREACHABLE__ :: HasCallStack => a

Throw an Unreachable error reporting the *caller's* call site. Note that this call to "withFileAndLine" will be filtered out due its filter on the srcLocModule.