Agda.Utils.IndexMap

Description

Maps from de Bruijn indices, implemented using IntMaps from de Bruijn levels.

Documentation

data IndexMap a

Maps from de Bruijn indices.

invariant :: IndexMap a -> Bool

The IndexMap invariant.

nil :: IndexMap a

An empty map.

snoc :: IndexMap a -> a -> IndexMap a

Inserts a new binding, which will have de Bruijn index 0, and shifts other bindings one step.

append :: IndexMap a -> [a] -> IndexMap a

Inserts a number of new bindings. The last element in the list will have de Bruijn index 0.

fromList :: [a] -> IndexMap a

Converts a list to an IndexMap. The last element in the list will have de Bruijn index 0.

update :: Int -> a -> IndexMap a -> IndexMap a

index :: Int -> IndexMap a -> a