Agda.Utils.List1

Description

Non-empty lists.

Better name List1 for non-empty lists, plus missing functionality.

Import: @

{-# LANGUAGE PatternSynonyms #-}

import Agda.Utils.List1 (List1, pattern (:|)) import qualified Agda.Utils.List1 as List1

@

Documentation

type List1 = NonEmpty

type String1 = List1 Char

catMaybes :: List1 (Maybe a) -> [a]

Like catMaybes.

mapMaybe :: (a -> Maybe b) -> List1 a -> [b]

Like filter.

zipWithM :: Applicative m => (a -> b -> m c) -> List1 a -> List1 b -> m (List1 c)

Like zipWithM.

partitionEithers :: List1 (Either a b) -> ([a], [b])

Like partitionEithers.

zipWithM_ :: Applicative m => (a -> b -> m c) -> List1 a -> List1 b -> m ()

Like zipWithM.

ifNull :: [a] -> b -> (List1 a -> b) -> b

union :: Eq a => List1 a -> List1 a -> List1 a

Like union. Duplicates in the first list are not removed. O(nm).

initLast :: List1 a -> ([a], a)

Return the last element and the rest.

lefts :: List1 (Either a b) -> [a]

Like lefts.

snoc :: [a] -> a -> List1 a

Build a list with one element.

More precise type for snoc.

concat :: [List1 a] -> [a]

Concatenate one or more non-empty lists.

foldr :: (a -> b -> b) -> (a -> b) -> List1 a -> b

List foldr but with a base case for the singleton list.

rights :: List1 (Either a b) -> [b]

Like rights.

ifNotNull :: [a] -> (List1 a -> b) -> b -> b

unlessNull :: Null m => [a] -> (List1 a -> m) -> m

breakAfter :: (a -> Bool) -> List1 a -> (List1 a, [a])

Breaks a list just after an element satisfying the predicate is found.

>>> breakAfter even [1,3,5,2,4,7,8]
([1,3,5,2],[4,7,8])

groupBy' :: forall a. (a -> a -> Bool) -> [a] -> [List1 a]

More precise type for groupBy'.

A variant of groupBy which applies the predicate to consecutive pairs. O(n).

allEqual :: Eq a => List1 a -> Bool

Checks if all the elements in the list are equal. Assumes that the Eq instance stands for an equivalence relation. O(n).

nubM :: Monad m => (a -> a -> m Bool) -> List1 a -> m (List1 a)

Non-efficient, monadic nub. O(n²).

wordsBy :: (a -> Bool) -> [a] -> [List1 a]

Split a list into sublists. Generalisation of the prelude function words. Same as wordsBy and wordsBy, but with the non-emptyness guarantee on the chunks. O(n).

words xs == wordsBy isSpace xs

fromListSafe

Arguments

:: List1 a
Default value if convertee is empty.

-> [a]
List to convert, supposedly non-empty.

-> List1 a
Converted list.

Safe version of fromList.

data NonEmpty a

Constructors

a :| [a]

group :: (Foldable f, Eq a) => f a -> [NonEmpty a]

(!!) :: HasCallStack => NonEmpty a -> Int -> a

sortBy :: (a -> a -> Ordering) -> NonEmpty a -> NonEmpty a

isPrefixOf :: Eq a => [a] -> NonEmpty a -> Bool

partition :: (a -> Bool) -> NonEmpty a -> ([a], [a])

singleton :: a -> NonEmpty a

(<|) :: a -> NonEmpty a -> NonEmpty a

map :: (a -> b) -> NonEmpty a -> NonEmpty b

insert :: (Foldable f, Ord a) => a -> f a -> NonEmpty a

length :: NonEmpty a -> Int

sort :: Ord a => NonEmpty a -> NonEmpty a

nonEmpty :: [a] -> Maybe (NonEmpty a)

transpose :: NonEmpty (NonEmpty a) -> NonEmpty (NonEmpty a)

tails :: Foldable f => f a -> NonEmpty [a]

nub :: Eq a => NonEmpty a -> NonEmpty a

unfoldr :: (a -> (b, Maybe a)) -> a -> NonEmpty b

unzip :: Functor f => f (a, b) -> (f a, f b)

filter :: (a -> Bool) -> NonEmpty a -> [a]

drop :: Int -> NonEmpty a -> [a]

splitAt :: Int -> NonEmpty a -> ([a], [a])

break :: (a -> Bool) -> NonEmpty a -> ([a], [a])

cycle :: NonEmpty a -> NonEmpty a

dropWhile :: (a -> Bool) -> NonEmpty a -> [a]

head :: NonEmpty a -> a

init :: NonEmpty a -> [a]

iterate :: (a -> a) -> a -> NonEmpty a

last :: NonEmpty a -> a

repeat :: a -> NonEmpty a

reverse :: NonEmpty a -> NonEmpty a

scanl :: Foldable f => (b -> a -> b) -> b -> f a -> NonEmpty b

scanl1 :: (a -> a -> a) -> NonEmpty a -> NonEmpty a

scanr :: Foldable f => (a -> b -> b) -> b -> f a -> NonEmpty b

scanr1 :: (a -> a -> a) -> NonEmpty a -> NonEmpty a

span :: (a -> Bool) -> NonEmpty a -> ([a], [a])

tail :: NonEmpty a -> [a]

take :: Int -> NonEmpty a -> [a]

takeWhile :: (a -> Bool) -> NonEmpty a -> [a]

zipWith :: (a -> b -> c) -> NonEmpty a -> NonEmpty b -> NonEmpty c

zip :: NonEmpty a -> NonEmpty b -> NonEmpty (a, b)

cons :: a -> NonEmpty a -> NonEmpty a

uncons :: NonEmpty a -> (a, Maybe (NonEmpty a))

groupBy :: Foldable f => (a -> a -> Bool) -> f a -> [NonEmpty a]

inits :: Foldable f => f a -> NonEmpty [a]

intersperse :: a -> NonEmpty a -> NonEmpty a

nubBy :: (a -> a -> Bool) -> NonEmpty a -> NonEmpty a

xor :: NonEmpty Bool -> Bool

append :: NonEmpty a -> NonEmpty a -> NonEmpty a

unfold :: (a -> (b, Maybe a)) -> a -> NonEmpty b

appendList :: NonEmpty a -> [a] -> NonEmpty a

group1 :: Eq a => NonEmpty a -> NonEmpty (NonEmpty a)

groupAllWith :: Ord b => (a -> b) -> [a] -> [NonEmpty a]

groupAllWith1 :: Ord b => (a -> b) -> NonEmpty a -> NonEmpty (NonEmpty a)

groupBy1 :: (a -> a -> Bool) -> NonEmpty a -> NonEmpty (NonEmpty a)

groupWith :: (Foldable f, Eq b) => (a -> b) -> f a -> [NonEmpty a]

groupWith1 :: Eq b => (a -> b) -> NonEmpty a -> NonEmpty (NonEmpty a)

prependList :: [a] -> NonEmpty a -> NonEmpty a

some1 :: Alternative f => f a -> f (NonEmpty a)

sortWith :: Ord o => (a -> o) -> NonEmpty a -> NonEmpty a

class IsList l where

Minimal complete definition

fromList, toList

Associated Types

type Item l

Methods

fromList :: [Item l] -> l

fromListN :: Int -> [Item l] -> l

toList :: l -> [Item l]

type family Item l