Agda.Utils.ListT

Description

ListT done right, see https://www.haskell.org/haskellwiki/ListT_done_right_alternative

There is also the list-t package on hackage (Nikita Volkov) but it again depends on other packages we do not use yet, so we rather implement the few bits we need afresh.

Documentation

newtype ListT (m :: Type -> Type) a

Lazy monadic computation of a list of results.

Constructors

ListT

Fields

runListT :: m (Maybe (a, ListT m a))

mapListT :: (m (Maybe (a, ListT m a)) -> n (Maybe (b, ListT n b))) -> ListT m a -> ListT n b

Boilerplate function to lift MonadReader through the ListT transformer.

unmapListT :: (ListT m a -> ListT n b) -> m (Maybe (a, ListT m a)) -> n (Maybe (b, ListT n b))

Inverse to mapListT.

List operations

nilListT :: forall (m :: Type -> Type) a. Monad m => ListT m a

The empty lazy list.

consListT :: forall (m :: Type -> Type) a. Monad m => a -> ListT m a -> ListT m a

Consing a value to a lazy list.

sgListT :: forall (m :: Type -> Type) a. Monad m => a -> ListT m a

Singleton lazy list.

caseListT :: Monad m => ListT m a -> m b -> (a -> ListT m a -> m b) -> m b

Case distinction over lazy list.

foldListT :: Monad m => (a -> m b -> m b) -> m b -> ListT m a -> m b

Folding a lazy list, effects left-to-right.

anyListT :: Monad m => ListT m a -> (a -> m Bool) -> m Bool

Lazy monadic disjunction of lazy monadic list, effects left-to-right

allListT :: Monad m => ListT m a -> (a -> m Bool) -> m Bool

Lazy monadic conjunction of lazy monadic list, effects left-to-right

sequenceListT :: Monad m => ListT m a -> m [a]

Force all values in the lazy list, effects left-to-right

concatListT :: forall (m :: Type -> Type) a. Monad m => ListT m (ListT m a) -> ListT m a

The join operation of the ListT m monad.

Monadic list operations.

runMListT :: Monad m => m (ListT m a) -> ListT m a

We can `run' a computation of a ListT as it is monadic itself.

consMListT :: Monad m => m a -> ListT m a -> ListT m a

Monadic cons.

sgMListT :: Monad m => m a -> ListT m a

Monadic singleton.

mapMListT :: Monad m => (a -> m b) -> ListT m a -> ListT m b

Extending a monadic function to ListT.

mapMListT_alt :: Monad m => (a -> m b) -> ListT m a -> ListT m b

Alternative implementation using foldListT.

liftListT :: (Monad m, Monad m') => (forall a1. m a1 -> m' a1) -> ListT m a -> ListT m' a

Change from one monad to another