Agda.Utils.Map1

Description

Non-empty maps.

Provides type Map1 of non-empty maps.

Import: @

import Agda.Utils.Map1 (Map1) import qualified Agda.Utils.Map1 as Map1

@

Documentation

ifNull :: Map k a -> b -> (Map1 k a -> b) -> b

unlessNull :: Applicative m => Map k a -> (Map1 k a -> m ()) -> m ()

A more general type would be Null m => Map k a -> (Map1 k a -> m) -> m but this type is problematic as we do not have a general instance Applicative m => Null (m ()).

unlessNullM :: Monad m => m (Map k a) -> (Map1 k a -> m ()) -> m ()

type Map1 = NEMap

module Data.Map.NonEmpty