Agda.Utils.Maybe

Description

Extend Maybe by common operations for the Maybe type.

Note: since this module is usually imported unqualified, we do not use short names, but all names contain Maybe, Just, or 'Nothing.

Documentation

allJustM :: Monad m => [m (Maybe a)] -> m (Maybe [a])

Lazy version of allJust . sequence. (allJust = mapM for the Maybe monad.) Only executes monadic effect while isJust.

boolToMaybe :: Bool -> a -> Maybe a

Retain object when tag is True.

caseMaybe :: Maybe a -> b -> (a -> b) -> b

Version of maybe with different argument ordering. Often, we want to case on a Maybe, do something interesting in the Just case, but only a default action in the Nothing case. Then, the argument ordering of caseMaybe is preferable.

caseMaybe m d f = flip (maybe d) m f

caseMaybeM :: Monad m => m (Maybe a) -> m b -> (a -> m b) -> m b

Monadic version of caseMaybe. That is, maybeM with a different argument ordering.

filterMaybe :: (a -> Bool) -> a -> Maybe a

Filtering a singleton list.

filterMaybe p a = listToMaybe (filter p [a])

forMaybe :: [a] -> (a -> Maybe b) -> [b]

Version of mapMaybe with different argument ordering.

fromMaybeM :: Monad m => m a -> m (Maybe a) -> m a

Monadic version of fromMaybe.

fromMaybeT :: Monad m => a -> MaybeT m a -> m a

Run a MaybeT with a default value for Nothing.

ifJust :: Maybe a -> (a -> b) -> b -> b

caseMaybe with flipped branches.

ifJustM :: Monad m => m (Maybe a) -> (a -> m b) -> m b -> m b

caseMaybeM with flipped branches.

liftMaybe :: Alternative f => Maybe a -> f a

Lift a maybe to an Alternative.

maybeM :: Monad m => m b -> (a -> m b) -> m (Maybe a) -> m b

Monadic version of maybe.

orEmpty :: Null a => Maybe a -> a

Unwrap a Maybe value, returning empty for Nothing.

predicateToMaybe :: (a -> Bool) -> a -> Maybe a

Retain object when it passes the given test.

spanMaybe :: (a -> Maybe b) -> [a] -> ([b], [a])

Like span, takes the prefix of a list satisfying a predicate. Returns the run of Justs until the first Nothing, and the tail of the list.

unionMaybeWith :: (a -> a -> a) -> Maybe a -> Maybe a -> Maybe a

unionWith for collections of size <= 1.

unionsMaybeWith :: (a -> a -> a) -> [Maybe a] -> Maybe a

unionsWith for collections of size <= 1.

unzipMaybe :: Maybe (a, b) -> (Maybe a, Maybe b)

Unzipping a list of length <= 1.

whenJust :: Monad m => Maybe a -> (a -> m ()) -> m ()

A more telling name for forM_ for the Maybe collection type. Or: caseMaybe without the Nothing case.

whenJustM :: Monad m => m (Maybe a) -> (a -> m ()) -> m ()

caseMaybeM without the Nothing case.

whenNothing :: Monad m => Maybe a -> m () -> m ()

Pendent to whenJust.

whenNothingM :: Monad m => m (Maybe a) -> m () -> m ()

caseMaybeM without the Just case.

module Data.Maybe