Agda.Utils.Maybe.Strict

Description

A strict version of the Maybe type.

Import qualified, as in import qualified Agda.Utils.Maybe.Strict as Strict

Copyright : (c) 2006-2007 Roman Leshchinskiy (c) 2013 Simon Meier License : BSD-style (see the file LICENSE)

Copyright : (c) 2014 Andreas Abel

Documentation

fromJust :: Maybe a -> a

fromMaybe :: a -> Maybe a -> a

isJust :: Maybe a -> Bool

isNothing :: Maybe a -> Bool

maybe :: b -> (a -> b) -> Maybe a -> b

data Maybe a

Constructors

Nothing

Just !a

toStrict :: Maybe a -> Maybe a

toLazy :: Maybe a -> Maybe a

listToMaybe :: [a] -> Maybe a

Analogous to listToMaybe in Data.Maybe.

maybeToList :: Maybe a -> [a]

Analogous to maybeToList in Data.Maybe.

catMaybes :: [Maybe a] -> [a]

Analogous to catMaybes in Data.Maybe.

mapMaybe :: (a -> Maybe b) -> [a] -> [b]

Analogous to mapMaybe in Data.Maybe.

unionMaybeWith :: (a -> a -> a) -> Maybe a -> Maybe a -> Maybe a

unionWith for collections of size <= 1.

unzipMaybe :: Maybe (a, b) -> (Maybe a, Maybe b)

Unzipping a list of length <= 1.

filterMaybe :: (a -> Bool) -> a -> Maybe a

Filtering a singleton list.

filterMaybe p a = listToMaybe (filter p [a])

forMaybe :: [a] -> (a -> Maybe b) -> [b]

Version of mapMaybe with different argument ordering.

caseMaybe :: Maybe a -> b -> (a -> b) -> b

Version of maybe with different argument ordering. Often, we want to case on a Maybe, do something interesting in the Just case, but only a default action in the Nothing case. Then, the argument ordering of caseMaybe is preferable.

caseMaybe m err f = flip (maybe err) m f

maybeM :: Monad m => m b -> (a -> m b) -> m (Maybe a) -> m b

Monadic version of maybe.

fromMaybeM :: Monad m => m a -> m (Maybe a) -> m a

Monadic version of fromMaybe.

caseMaybeM :: Monad m => m (Maybe a) -> m b -> (a -> m b) -> m b

Monadic version of caseMaybe. That is, maybeM with a different argument ordering.

ifJustM :: Monad m => m (Maybe a) -> (a -> m b) -> m b -> m b

caseMaybeM with flipped branches.

whenJust :: Monad m => Maybe a -> (a -> m ()) -> m ()

A more telling name for forM for the Maybe collection type. Or: caseMaybe without the Nothing case.

whenJustM :: Monad m => m (Maybe a) -> (a -> m ()) -> m ()

caseMaybeM without the Nothing case.

Orphan instances

Applicative Maybe
Note that strict Maybe is an Applicative only modulo strictness. The laws only hold in the strict semantics. Eg. pure f * pure _|_ = _|_, but according to the laws for Applicative it should be pure (f _|_). We ignore this issue here, it applies also to Foldable and Traversable.

Instance details

Methods

pure :: a -> Maybe a

(<*>) :: Maybe (a -> b) -> Maybe a -> Maybe b

liftA2 :: (a -> b -> c) -> Maybe a -> Maybe b -> Maybe c

(*>) :: Maybe a -> Maybe b -> Maybe b

(<*) :: Maybe a -> Maybe b -> Maybe a

Null (Maybe a)

Instance details

Methods

empty :: Maybe a

null :: Maybe a -> Bool