Agda.Utils.Null

Description

Overloaded null and empty for collections and sequences.

Documentation

class Null a where

Minimal complete definition

empty

Methods

empty :: a

null :: a -> Bool

Satisfying null empty == True.

default null :: Eq a => a -> Bool

The default implementation of null compares with empty, first trying pointer equality, then falling back to Eq equality.

Testing for null.

ifNull :: Null a => a -> b -> (a -> b) -> b

ifNotNull :: Null a => a -> (a -> b) -> b -> b

ifNullM :: (Monad m, Null a) => m a -> m b -> (a -> m b) -> m b

ifNotNullM :: (Monad m, Null a) => m a -> (a -> m b) -> m b -> m b

whenNull :: (Monad m, Null a) => a -> m () -> m ()

unlessNull :: (Monad m, Null a) => a -> (a -> m ()) -> m ()

whenNullM :: (Monad m, Null a) => m a -> m () -> m ()

unlessNullM :: (Monad m, Null a) => m a -> (a -> m ()) -> m ()

applyUnlessNull :: Null a => a -> (a -> b -> b) -> b -> b

catchNull :: Null a => a -> a -> a

Disjunction (interpreting null _ as False).