Agda.Utils.Parser.MemoisedCPS

Description

Parser combinators with support for left recursion, following Johnson's "Memoization in Top-Down Parsing".

This implementation is based on an implementation due to Atkey (attached to an edlambda-members mailing list message from 2011-02-15 titled 'Slides for "Introduction to Parser Combinators"').

Note that non-memoised left recursion is not guaranteed to work.

The code contains an important deviation from Johnson's paper: the check for subsumed results is not included. This means that one can get the same result multiple times when parsing using ambiguous grammars. As an example, parsing the empty string using S ∷= ε | ε succeeds twice. This change also means that parsing fails to terminate for some cyclic grammars that would otherwise be handled successfully, such as S ∷= S | ε. However, the library is not intended to handle infinitely ambiguous grammars. (It is unclear to the author of this module whether the change leads to more non-termination for grammars that are not cyclic.)

Documentation

class (Functor p, Applicative p, Alternative p, Monad p) => ParserClass (p :: Type -> Type) k r tok | p -> k, p -> r, p -> tok where

Methods

parse :: p a -> [tok] -> [a]

Runs the parser.

grammar :: Show k => p a -> Doc

Tries to print the parser, or returns empty, depending on the implementation. This function might not terminate.

sat' :: (tok -> Maybe a) -> p a

Parses a token satisfying the given predicate. The computed value is returned.

annotate :: (DocP -> DocP) -> p a -> p a

Uses the given function to modify the printed representation (if any) of the given parser.

memoise :: k -> p r -> p r

Memoises the given parser.

Every memoised parser must be annotated with a unique key. (Parametrised parsers must use distinct keys for distinct inputs.)

memoiseIfPrinting :: k -> p r -> p r

Memoises the given parser, but only if printing, not if parsing.

Every memoised parser must be annotated with a unique key. (Parametrised parsers must use distinct keys for distinct inputs.)

sat :: ParserClass p k r tok => (tok -> Bool) -> p tok

Parses a token satisfying the given predicate.

token :: ParserClass p k r tok => p tok

Parses a single token.

tok :: (ParserClass p k r tok, Eq tok, Show tok) => tok -> p tok

Parses a given token.

doc :: ParserClass p k r tok => Doc -> p a -> p a

Uses the given document as the printed representation of the given parser. The document's precedence is taken to be atomP.

type DocP = (Doc, Int)

Documents paired with precedence levels.

bindP :: Int

Precedence of >>=.

choiceP :: Int

Precedence of |.

seqP :: Int

Precedence of *.

starP :: Int

Precedence of ⋆ and +.

atomP :: Int

Precedence of atoms.

data Parser k r tok a

The parser type.

The parameters of the type Parser k r tok a have the following meanings:

k

Type used for memoisation keys.

r

The type of memoised values. (Yes, all memoised values have to have the same type.)

tok

The token type.

a

The result type.

data ParserWithGrammar k r tok a

An extended parser type, with some support for printing parsers.