Agda.Utils.PartialOrd

Documentation

data PartialOrdering

The result of comparing two things (of the same type).

Constructors

POLT
Less than.

POLE
Less or equal than.

POEQ
Equal

POGE
Greater or equal.

POGT
Greater than.

POAny
No information (incomparable).

leqPO :: PartialOrdering -> PartialOrdering -> Bool

Comparing the information content of two elements of PartialOrdering. More precise information is smaller.

Includes equality: x leqPO x == True.

oppPO :: PartialOrdering -> PartialOrdering

Opposites.

related a po b iff related b (oppPO po) a.

orPO :: PartialOrdering -> PartialOrdering -> PartialOrdering

Combining two pieces of information (picking the least information). Used for the dominance ordering on tuples.

orPO is associative, commutative, and idempotent. orPO has dominant element POAny, but no neutral element.

seqPO :: PartialOrdering -> PartialOrdering -> PartialOrdering

Chains (transitivity) x R y S z.

seqPO is associative, commutative, and idempotent. seqPO has dominant element POAny and neutral element (unit) POEQ.

fromOrdering :: Ordering -> PartialOrdering

Embed Ordering.

fromOrderings :: [Ordering] -> PartialOrdering

Represent a non-empty disjunction of Orderings as PartialOrdering.

toOrderings :: PartialOrdering -> [Ordering]

A PartialOrdering information is a disjunction of Ordering informations.

Comparison with partial result

type Comparable a = a -> a -> PartialOrdering

class PartialOrd a where

Decidable partial orderings.

Methods

comparable :: Comparable a

comparableOrd :: Ord a => Comparable a

Any Ord is a PartialOrd.

related :: PartialOrd a => a -> PartialOrdering -> a -> Bool

Are two elements related in a specific way?

related a o b holds iff comparable a b is contained in o.

Totally ordered types.

Generic partially ordered types.

newtype Pointwise a

Pointwise comparison wrapper.

Constructors

Pointwise

Fields

pointwise :: a

newtype Inclusion a

Inclusion comparison wrapper.

Constructors

Inclusion

Fields

inclusion :: a

PartialOrdering is itself partially ordered!