Agda.Utils.Pointer

Documentation

data Ptr a

newPtr :: a -> Ptr a

derefPtr :: Ptr a -> a

setPtr :: a -> Ptr a -> Ptr a

updatePtr :: (a -> a) -> Ptr a -> Ptr a

updatePtrM :: Functor f => (a -> f a) -> Ptr a -> f (Ptr a)

If f a contains many copies of a they will all be the same pointer in the result. If the function is well-behaved (i.e. preserves the implicit equivalence, this shouldn't matter).