Agda.Utils.Size

Description

Collection size.

For TermSize see Agda.Syntax.Internal.

Documentation

class Sized a where

The size of a collection (i.e., its length).

Minimal complete definition

Nothing

Methods

size :: a -> Int

Strict size computation.

Anti-patterns: size xs == n where n is 0, 1 or another number that is likely smaller than size xs. Similar for size xs >= 1 etc. Use natSize instead.

See https://wiki.haskell.org/Haskell_programming_tips#Don.27t_ask_for_the_length_of_a_list_when_you_don.27t_need_it .

default size :: forall (t :: Type -> Type) b. (Foldable t, t b ~ a) => a -> Int

natSize :: a -> Peano

Lazily compute a (possibly infinite) size.

Use when comparing a size against a fixed number.

default natSize :: forall (t :: Type -> Type) b. (Foldable t, t b ~ a) => a -> Peano

data SizedThing a

Thing decorated with its size. The thing should fit into main memory, thus, the size is an Int.

Constructors

SizedThing

Fields

theSize :: !Int

sizedThing :: a

sizeThing :: Sized a => a -> SizedThing a

Cache the size of an object.

data Peano

The natural numbers in (lazy) unary notation.

Constructors

Zero

Succ Peano