Agda.Utils.StrictState

Description

This is a plain strict state monad, where state update, monadic binding and return are all strict. If you only need a single strict state effect, use this module.

Do not use Control.Monad.State.Strict for the same purpose; it's not even strict in state updates and is much less amenable to GHC optimizations than this module.

Documentation

class Monad m => MonadState s (m :: Type -> Type) | m -> s where

Minimal definition is either both of get and put or just state

Minimal complete definition

state | get, put

Methods

get :: m s

Return the state from the internals of the monad.

put :: s -> m ()

Replace the state inside the monad.

state :: (s -> (a, s)) -> m a

Embed a simple state action into the monad.

evalState :: State s a -> s -> a

evalStateT :: Monad m => StateT s m a -> s -> m a

execState :: State s a -> s -> s

execStateT :: Monad m => StateT s m a -> s -> m s

gets :: MonadState s m => (s -> a) -> m a

modify :: MonadState s m => (s -> s) -> m ()

runState :: State s a -> s -> (a, s)

runStateT :: Monad m => StateT s m a -> s -> m (a, s)

newtype State s a

Constructors

State

Fields

runState# :: s -> (# a, s #)

newtype StateT s (m :: Type -> Type) a

Constructors

StateT

Fields

runStateT# :: s -> m (Pair a s)