Agda.Utils.StrictState2

Description

This is a strict state monad, where state update, monadic binding and return are all strict. It is specialized to the case where we have (s, t) for state, i.e. two state components. The reason for specialization is that GHC cannot reliably unbox State (s, t) for any definition of plain State, so we need to directly use unboxed tuples for two state components in the internals.

If you need two strict state components and no other effects, use this module. Do not use Control.Monad.State.Strict for the same purpose; it's not even strict in state updates and is much less amenable to GHC optimizations than this module.

Documentation

newtype State s t a

Constructors

State

Fields

runState# :: s -> t -> (# a, s, t #)

put :: s -> t -> State s t ()

put1 :: s -> State s t ()

put2 :: t -> State s t ()

get :: State s t (s, t)

get1 :: State s t s

get2 :: State s t t

gets :: ((s, t) -> a) -> State s t a

modify :: ((s, t) -> (s, t)) -> State s t ()

modify1 :: (s -> s) -> State s t ()

modify2 :: (t -> t) -> State s t ()

execState :: State s t a -> (s, t) -> (s, t)

runState :: State s t a -> (s, t) -> (a, (s, t))

evalState :: State s t a -> (s, t) -> a