Agda.Utils.StrictWriter

Description

This is a very strict Writer monad, as a wrapper on Agda.Utils.StrictState.

Documentation

class (Monoid w, Monad m) => MonadWriter w (m :: Type -> Type) | m -> w where

Minimal complete definition

(writer | tell), listen, pass

Methods

writer :: (a, w) -> m a

writer (a,w) embeds a simple writer action.

tell :: w -> m ()

tell w is an action that produces the output w.

listen :: m a -> m (a, w)

listen m is an action that executes the action m and adds its output to the value of the computation.

pass :: m (a, w -> w) -> m a

pass m is an action that executes the action m, which returns a value and a function, and returns the value, applying the function to the output.

censor :: MonadWriter w m => (w -> w) -> m a -> m a

censor f m is an action that executes the action m and applies the function f to its output, leaving the return value unchanged.

censor f m = pass (liftM (\x -> (x,f)) m)

execWriter :: Monoid w => Writer w a -> w

execWriterT :: (Monoid w, Monad m) => WriterT w m a -> m w

runWriter :: Monoid w => Writer w a -> (a, w)

runWriterT :: (Monoid w, Monad m) => WriterT w m a -> m (a, w)

newtype Writer w a

Constructors

Writer

Fields

unWriter :: State w a

newtype WriterT w (m :: Type -> Type) a

Constructors

WriterT

Fields

unWriterT :: StateT w m a