Agda.Utils.Trie

Description

Strict tries (based on Data.Map.Strict and Agda.Utils.Maybe.Strict).

Note that if delete or adjust are used, one may end up with non-canonical subtries, i.e., those that do not contain any value (Data.List.null . toList) but may be Eq-different to other empty tries (/= empty).

Documentation

data Trie k v

Finite map from [k] to v.

With the strict Maybe type, Trie is also strict in v.

Constructors

Trie !(Maybe v) !(Map k (Trie k v))

empty :: Null a => a

singleton :: [k] -> v -> Trie k v

Singleton trie.

everyPrefix :: [k] -> v -> Trie k v

everyPrefix k v is a trie where every prefix of k (including k itself) is mapped to v.

insert :: Ord k => [k] -> v -> Trie k v -> Trie k v

Insert. Overwrites existing value if present.

insert = insertWith ( new old -> new)

insertWith :: Ord k => (v -> v -> v) -> [k] -> v -> Trie k v -> Trie k v

Insert with function merging new value with old value.

union :: Ord k => Trie k v -> Trie k v -> Trie k v

Left biased union.

union = unionWith ( new old -> new).

unionWith :: Ord k => (v -> v -> v) -> Trie k v -> Trie k v -> Trie k v

Pointwise union with merge function for values.

adjust :: Ord k => [k] -> (Maybe v -> Maybe v) -> Trie k v -> Trie k v

Adjust value at key, leave subtree intact.

Disclaimer: may return a non-canoncial trie because it does not clean up subtries that become empty.

delete :: Ord k => [k] -> Trie k v -> Trie k v

Delete value at key, but leave subtree intact.

Disclaimer: may return a non-canoncial trie because it does not clean up subtries that become empty.

toList :: Ord k => Trie k v -> [([k], v)]

Convert to ascending list.

toAscList :: Ord k => Trie k v -> [([k], v)]

Convert to ascending list.

toListOrderedBy :: Ord k => (v -> v -> Ordering) -> Trie k v -> [([k], v)]

Convert to list where nodes at the same level are ordered according to the given ordering.

lookup :: Ord k => [k] -> Trie k v -> Maybe v

Returns the value associated with the given key, if any.

member :: Ord k => [k] -> Trie k v -> Bool

Is the given key present in the trie?

lookupPath :: Ord k => [k] -> Trie k v -> [v]

Collect all values along a given path.

lookupTrie :: Ord k => [k] -> Trie k v -> Trie k v

Get the subtrie rooted at the given key.

mapSubTries :: Ord k => (Trie k u -> Maybe v) -> Trie k u -> Trie k v

Create new values based on the entire subtrie. Almost, but not quite comonad extend.

filter :: Ord k => (v -> Bool) -> Trie k v -> Trie k v

Filter a trie.

valueAt :: Ord k => [k] -> Lens' (Trie k v) (Maybe v)

Key lens.