Agda.Utils.TypeLevel

Documentation

type family All (p :: k -> Constraint) (as :: [k]) where ...

All p as ensures that the constraint p is satisfied by all the types in as. (Types is between scare-quotes here because the code is actually kind polymorphic)

Equations

All (p :: k -> Constraint) ('[] :: [k]) = ()

All (p :: k -> Constraint) (a ': as :: [k]) = (p a, All p as)

type family If (b :: Bool) (l :: k) (r :: k) :: k where ...

On Booleans

Equations

If 'True (l :: k) (r :: k) = l

If 'False (l :: k) (r :: k) = r

type family Foldr (c :: k -> l -> l) (n :: l) (as :: [k]) :: l where ...

On Lists

Equations

Foldr (c :: k -> l -> l) (n :: l) ('[] :: [k]) = n

Foldr (c :: k -> l -> l) (n :: l) (a ': as :: [k]) = c a (Foldr c n as)

type family Foldr' (c :: Function k (Function l l -> Type) -> Type) (n :: l) (as :: [k]) :: l where ...

Version of Foldr taking a defunctionalised argument so that we can use partially applied functions.

Equations

Foldr' (c :: Function k (Function l l -> Type) -> Type) (n :: l) ('[] :: [k]) = n

Foldr' (c :: Function k (Function l l -> Type) -> Type) (n :: l) (a ': as :: [k]) = Apply (Apply c a) (Foldr' c n as)

type family Map (f :: Function k l -> Type) (as :: [k]) :: [l] where ...

Equations

Map (f :: Function k l -> Type) (as :: [k]) = Foldr' (ConsMap0 f) ('[] :: [l]) as

data ConsMap0 (a :: Function k l -> Type) (b :: Function k (Function [l] [l] -> Type))

data ConsMap1 (a :: Function k l -> Type) (b :: k) (c :: Function [l] [l])

type family Constant b (as :: [k]) :: [Type] where ...

Equations

Constant b (as :: [k]) = Map (Constant1 b :: Function k Type -> Type) as

type Arrows (as :: [Type]) r = Foldr (->) r as

Arrows [a1,..,an] r corresponds to a1 -> .. -> an -> r | Products [a1,..,an] corresponds to (a1, (..,( an, ())..))

type Products (as :: [Type]) = Foldr (,) () as

data StrictPair a b

Constructors

Pair a b

type StrictProducts (as :: [Type]) = Foldr StrictPair () as

strictCurry :: (StrictPair a b -> c) -> a -> b -> c

strictUncurry :: (a -> b -> c) -> StrictPair a b -> c

type family IsBase t :: Bool where ...

IsBase t is 'True whenever t is *not* a function space.

Equations

IsBase (a -> t3) = 'False

IsBase a = 'True

type family Domains t :: [Type] where ...

Using IsBase we can define notions of Domains and CoDomains which *reduce* under positive information IsBase t ~ 'True even though the shape of t is not formally exposed

Equations

Domains t = If (IsBase t) ('[] :: [Type]) (Domains' t)

type family Domains' t :: [Type] where ...

Equations

Domains' (a -> t) = a ': Domains t

type family CoDomain t where ...

Equations

CoDomain t = If (IsBase t) t (CoDomain' t)

type family CoDomain' t where ...

Equations

CoDomain' (a -> t2) = CoDomain t2

class Currying (as :: [Type]) b where

Currying as b witnesses the isomorphism between Arrows as b and Products as -> b. It is defined as a type class rather than by recursion on a singleton for as so all of that these conversions are inlined at compile time for concrete arguments.

Methods

uncurrys :: Proxy as -> Proxy b -> Arrows as b -> Products as -> b

currys :: Proxy as -> Proxy b -> (Products as -> b) -> Arrows as b

class StrictCurrying (as :: [Type]) b where

Methods

strictUncurrys :: Proxy as -> Proxy b -> Arrows as b -> StrictProducts as -> b

strictCurrys :: Proxy as -> Proxy b -> (StrictProducts as -> b) -> Arrows as b

data Function a b

data Constant0 (c :: Function a (Function b a -> Type))

data Constant1 c (d :: Function b a)

type family Apply (t :: Function k l -> Type) (u :: k) :: l