Agda.Utils.Update

Documentation

data ChangeT m a

The ChangeT monad transformer.

runChangeT :: Functor m => ChangeT m a -> m (a, Bool)

Run a ChangeT computation, returning result plus change flag.

mapChangeT :: (m (a, Any) -> n (b, Any)) -> ChangeT m a -> ChangeT n b

Map a ChangeT computation (monad transformer action).

type UpdaterT m a = a -> ChangeT m a

runUpdaterT :: Functor m => UpdaterT m a -> a -> m (a, Bool)

Blindly run an updater.

type Change a = ChangeT Identity a

class Monad m => MonadChange m where

The class of change monads.

Methods

tellDirty

Arguments

:: m ()
Mark computation as having changed something.

listenDirty :: m a -> m (a, Bool)

runChange :: Change a -> (a, Bool)

Run a Change computation, returning result plus change flag.

type Updater a = UpdaterT Identity a

sharing :: Monad m => UpdaterT m a -> UpdaterT m a

Replace result of updating with original input if nothing has changed.

runUpdater :: Updater a -> a -> (a, Bool)

Blindly run an updater.

dirty :: Monad m => UpdaterT m a

Mark a computation as dirty.

ifDirty :: (Monad m, MonadChange m) => m a -> (a -> m b) -> (a -> m b) -> m b

class Traversable f => Updater1 f where

Like Functor, but preserving sharing.

Minimal complete definition

Nothing

Methods

updater1 :: Updater a -> Updater (f a)

updates1

Arguments

:: Updater a

-> Updater (f a)
= sharing . updater1

update1 :: Updater a -> EndoFun (f a)

class Updater2 f where

Like Bifunctor, but preserving sharing.

Minimal complete definition

updater2

Methods

updater2 :: Updater a -> Updater b -> Updater (f a b)

updates2 :: Updater a -> Updater b -> Updater (f a b)

update2 :: Updater a -> Updater b -> EndoFun (f a b)