Agda.Utils.Zip

Description

Overloaded zip and zipWith.

This module provides the Zip class to overload zipping functions for variants of the list type ('[]', List1, List2, ListInf).

Motivation: In practice, we often zip a finite list with an infinite list, e.g. zipWith f xs (ys ++ repeat z). We can know statically in this case that the resulting considers each xs by rewriting this to zipWith f xs (ListInf.pad ys z) statically exposing the infinite nature of the second list. To avoid cluttering the namespace with 4*4 variants of zipWith for all combination of our list types, we introduce an overloaded version.

Import this module as follows:

import Prelude hiding (zip, zipWith)
import Agda.Utils.Zip

Documentation

class Zip (f :: Type -> Type) (g :: Type -> Type) (h :: Type -> Type) | f g -> h where

Minimal complete definition

zipWith

Methods

zip :: f a -> g b -> h (a, b)

zipWith :: (a -> b -> c) -> f a -> g b -> h c