Agda.Version

Documentation

version :: String

The version of Agda.

package :: String

This package name. This is mainly intended for use in the test suites to filter ephemeral hash-fingerprinted package names like Agda-2.6.2-5ceeWeguf1QFMaHLput4zw.

docsUrl :: String -> String

Returns a URL corresponding to the given section in the documentation for the current version.