Agda.Compiler.MAlonzo.Misc

Documentation

data HsModuleEnv

Constructors

HsModuleEnv

Fields

mazModuleName :: TopLevelModuleName

The name of the Agda module

mazIsMainModule :: Bool

Whether this is the compilation root and therefore should have the main function. This corresponds to the IsMain flag provided to the backend, not necessarily whether the GHC module has a main function defined.

data GHCOptions

The options derived from GHCFlags and other shared options.

Constructors

GHCOptions

Fields

optGhcCallGhc :: Bool

optGhcBin :: FilePath

Use the compiler at PATH instead of "ghc"

optGhcFlags :: [String]

optGhcCompileDir :: FilePath

optGhcStrictData :: Bool

Make inductive constructors strict?

optGhcStrict :: Bool

Make functions strict?

data GHCEnv

A static part of the GHC backend's environment that does not change from module to module.

Constructors

GHCEnv

Fields

ghcEnvOpts :: GHCOptions

ghcEnvBool :: Maybe QName

ghcEnvTrue :: Maybe QName

ghcEnvFalse :: Maybe QName

ghcEnvMaybe :: Maybe QName

ghcEnvNothing :: Maybe QName

ghcEnvJust :: Maybe QName

ghcEnvList :: Maybe QName

ghcEnvNil :: Maybe QName

ghcEnvCons :: Maybe QName

ghcEnvNat :: Maybe QName

ghcEnvInteger :: Maybe QName

ghcEnvWord64 :: Maybe QName

ghcEnvInf :: Maybe QName

ghcEnvSharp :: Maybe QName

ghcEnvFlat :: Maybe QName

ghcEnvInterval :: Maybe QName

ghcEnvIZero :: Maybe QName

ghcEnvIOne :: Maybe QName

ghcEnvIsOne :: Maybe QName

ghcEnvItIsOne :: Maybe QName

ghcEnvIsOne1 :: Maybe QName

ghcEnvIsOne2 :: Maybe QName

ghcEnvIsOneEmpty :: Maybe QName

ghcEnvPathP :: Maybe QName

ghcEnvSub :: Maybe QName

ghcEnvSubIn :: Maybe QName

ghcEnvId :: Maybe QName

ghcEnvConId :: Maybe QName

ghcEnvIsTCBuiltin :: QName -> Bool

ghcEnvListArity :: Maybe Int

ghcEnvMaybeArity :: Maybe Int

data GHCModuleEnv

Module compilation environment, bundling the overall backend session options along with the module's basic readable properties.

Constructors

GHCModuleEnv

Fields

ghcModEnv :: GHCEnv

ghcModHsModuleEnv :: HsModuleEnv

class Monad m => ReadGHCModuleEnv (m :: Type -> Type) where

Monads that can produce a GHCModuleEnv.

Minimal complete definition

Nothing

Methods

askGHCModuleEnv :: m GHCModuleEnv

default askGHCModuleEnv :: forall (t :: (Type -> Type) -> Type -> Type) (n :: Type -> Type). (MonadTrans t, Monad n, m ~ t n, ReadGHCModuleEnv n) => m GHCModuleEnv

askHsModuleEnv :: m HsModuleEnv

askGHCEnv :: m GHCEnv

newtype HsCompileState

Constructors

HsCompileState

Fields

mazAccumlatedImports :: Set TopLevelModuleName

type HsCompileT (m :: Type -> Type) = ReaderT GHCModuleEnv (StateT HsCompileState m)

Transformer adding read-only module info and a writable set of imported modules

type HsCompileM = HsCompileT TCM

The default compilation monad is the entire TCM (☹️) enriched with our state and module info

runHsCompileT' :: HsCompileT m a -> GHCModuleEnv -> HsCompileState -> m (a, HsCompileState)

runHsCompileT :: HsCompileT m a -> GHCModuleEnv -> m (a, HsCompileState)

curIsMainModule :: ReadGHCModuleEnv m => m Bool

Whether the current module is expected to have the main function. This corresponds to the IsMain flag provided to the backend, not necessarily whether the GHC module actually has a main function defined.

curAgdaMod :: ReadGHCModuleEnv m => m TopLevelModuleName

This is the same value as curMName, but does not rely on the TCM's state. (curMName and co. should be removed, but the current Backend interface is not sufficient yet to allow that)

curHsMod :: ReadGHCModuleEnv m => m ModuleName

Get the Haskell module name of the currently-focused Agda module

data FunctionKind

There are two kinds of functions: those definitely without unused arguments, and those that might have unused arguments.

Constructors

NoUnused

PossiblyUnused

data VariableKind

Different kinds of variables: those starting with a, those starting with v, and those starting with x.

Constructors

A

V

X

data NameKind

Different kinds of names.

Constructors

TypeK
Types.

ConK
Constructors.

VarK VariableKind
Variables.

CoverK
Used for coverage checking.

CheckK
Used for constructor type checking.

FunK FunctionKind
Other functions.

encodeString :: NameKind -> String -> String

Turns strings into valid Haskell identifiers.

In order to avoid clashes with names of regular Haskell definitions (those not generated from Agda definitions), make sure that the Haskell names are always used qualified, with the exception of names from the prelude.

ihname :: VariableKind -> Nat -> Name

unqhname :: NameKind -> QName -> Name

tlmodOf :: ReadTCState m => ModuleName -> m ModuleName

xqual :: QName -> Name -> HsCompileM QName

xhqn :: NameKind -> QName -> HsCompileM QName

hsName :: String -> QName

conhqn :: QName -> HsCompileM QName

bltQual :: BuiltinId -> String -> HsCompileM QName

dname :: QName -> Name

duname :: QName -> Name

Name for definition stripped of unused arguments

hsPrimOp :: String -> QOp

hsPrimOpApp :: String -> Exp -> Exp -> Exp

hsInt :: Integer -> Exp

hsTypedInt :: Integral a => a -> Exp

hsTypedDouble :: Real a => a -> Exp

hsLet :: Name -> Exp -> Exp -> Exp

hsVarUQ :: Name -> Exp

hsAppView :: Exp -> [Exp]

hsOpToExp :: QOp -> Exp

hsLambda :: [Pat] -> Exp -> Exp

hsMapAlt :: (Exp -> Exp) -> Alt -> Alt

hsMapRHS :: (Exp -> Exp) -> Rhs -> Rhs

mazstr :: String

mazName :: Name

mazMod' :: String -> ModuleName

mazMod :: TopLevelModuleName -> ModuleName

mazCoerceName :: String

mazErasedName :: String

mazAnyTypeName :: String

mazCoerce :: Exp

mazUnreachableError :: Exp

rtmUnreachableError :: Exp

mazHole :: Exp

rtmHole :: String -> Exp

mazAnyType :: Type

mazRTE :: ModuleName

mazRTEFloat :: ModuleName

rtmQual :: String -> QName

rtmVar :: String -> Exp

rtmError :: Text -> Exp

unsafeCoerceMod :: ModuleName

fakeD :: Name -> String -> Decl

fakeDS :: String -> String -> Decl

fakeDQ :: QName -> String -> Decl

fakeType :: String -> Type

fakeExp :: String -> Exp

fakeDecl :: String -> Decl

emptyBinds :: Maybe Binds

isModChar :: Char -> Bool

Can the character be used in a Haskell module name part (conid)? This function is more restrictive than what the Haskell report allows.