Agda.Interaction.Highlighting.Precise

Description

Types used for precise syntax highlighting.

Highlighting information

data Aspect

Constructors

Comment

Keyword

String

Number

Hole

Symbol
Symbols like forall, =, ->, etc.

PrimitiveType
Things like Set and Prop.

Name (Maybe NameKind) Bool
Is the name an operator part?

Pragma
Text occurring in pragmas that does not have a more specific aspect.

Background
Non-code contents in literate Agda

Markup
Delimiters used to separate the Agda code blocks from the other contents in literate Agda

data NameKind

NameKinds are figured out during scope checking.

Constructors

Bound
Bound variable.

Generalizable
Generalizable variable. (This includes generalizable variables that have been generalized).

Constructor Induction
Inductive or coinductive constructor.

Datatype

Field
Record field.

Function

Module
Module name.

Postulate

Primitive
Primitive.

Record
Record type.

Argument
Named argument, like x in {x = v}

Macro
Macro.

data OtherAspect

Other aspects, generated by type checking. (These can overlap with each other and with Aspects.)

Constructors

Error

ErrorWarning
A warning that is considered fatal in the end.

DottedPattern

UnsolvedMeta

UnsolvedConstraint
Unsolved constraint not connected to meta-variable. This could for instance be an emptyness constraint.

TerminationProblem

PositivityProblem

Deadcode
Used for highlighting unreachable clauses, unreachable RHS (because of an absurd pattern), etc.

ShadowingInTelescope
Used for shadowed repeated variable names in telescopes.

CoverageProblem

IncompletePattern
When this constructor is used it is probably a good idea to include a note explaining why the pattern is incomplete.

TypeChecks
Code which is being type-checked.

MissingDefinition
Function declaration without matching definition.

InstanceProblem
Unusable instance etc.

CosmeticProblem
Nothing serious, just a beauty flaw.

CatchallClause

ConfluenceProblem

data Aspects

Syntactic aspects of the code. (These cannot overlap.)

Meta information which can be associated with a character/character range.

Constructors

Aspects

Fields

aspect :: Maybe Aspect

otherAspects :: Set OtherAspect

note :: String

This note, if not null, can be displayed as a tool-tip or something like that. It should contain useful information about the range (like the module containing a certain identifier, or the fixity of an operator).

definitionSite :: Maybe DefinitionSite

The definition site of the annotated thing, if applicable and known.

tokenBased :: !TokenBased

Is this entry token-based?

data DefinitionSite

Constructors

DefinitionSite

Fields

defSiteModule :: TopLevelModuleName' Range

The defining module.

defSitePos :: Int

The file position in that module. File positions are counted from 1.

defSiteHere :: Bool

Has this DefinitionSite been created at the defining site of the name?

defSiteAnchor :: Maybe String

A pretty name for the HTML linking.

data TokenBased

Is the highlighting "token-based", i.e. based only on information from the lexer?

Constructors

TokenBased

NotOnlyTokenBased

newtype RangePair

A limited kind of syntax highlighting information: a pair consisting of Ranges and Aspects.

Note the invariant which RangePairs should satisfy (rangePairInvariant).

Constructors

RangePair

Fields

rangePair :: (Ranges, Aspects)

rangePairInvariant :: RangePair -> Bool

Invariant for RangePair.

newtype PositionMap

Syntax highlighting information, represented by maps from positions to Aspects.

The first position in the file has number 1.

Constructors

PositionMap

Fields

positionMap :: IntMap Aspects

newtype DelayedMerge hl

Highlighting info with delayed merging.

Merging large sets of highlighting info repeatedly might be costly. The idea of this type is to accumulate small pieces of highlighting information, and then to merge them all at the end.

Note the invariant which values of this type should satisfy (delayedMergeInvariant).

Constructors

DelayedMerge (Endo [hl])

delayedMergeInvariant :: (hl -> Bool) -> DelayedMerge hl -> Bool

Invariant for DelayedMerge hl, parametrised by the invariant for hl.

Additionally the endofunction should be extensionally equal to (fs ++) for some list fs.

type HighlightingInfo = RangeMap Aspects

Highlighting information.

Note the invariant which values of this type should satisfy (highlightingInfoInvariant).

This is a type synonym in order to make it easy to change to another representation.

highlightingInfoInvariant :: HighlightingInfo -> Bool

The invariant for HighlightingInfo.

type HighlightingInfoBuilder = DelayedMerge RangePair

A type that is intended to be used when constructing highlighting information.

Note the invariant which values of this type should satisfy (highlightingInfoBuilderInvariant).

This is a type synonym in order to make it easy to change to another representation.

The type should be an instance of IsBasicRangeMap Aspects, Semigroup and Monoid, and there should be an instance of Convert HighlightingInfoBuilder HighlightingInfo.

highlightingInfoBuilderInvariant :: HighlightingInfoBuilder -> Bool

The invariant for HighlightingInfoBuilder.

Additionally the endofunction should be extensionally equal to (fs ++) for some list fs.

Operations

parserBased :: Aspects

A variant of mempty with tokenBased set to NotOnlyTokenBased.

kindOfNameToNameKind :: KindOfName -> NameKind

Conversion from classification of the scope checker.

class IsBasicRangeMap a m | m -> a where

A class that is intended to make it easy to swap between different range map implementations.

Note that some RangeMap operations are not included in this class.

Minimal complete definition

singleton, toMap, toList

Methods

singleton :: Ranges -> a -> m

The map singleton rs x contains the ranges from rs, and every position in those ranges is associated with x.

toMap :: m -> IntMap a

Converts range maps to IntMaps from positions to values.

toList :: m -> [(Range, a)]

Converts the map to a list. The ranges are non-overlapping and non-empty, and earlier ranges precede later ones in the list.

coveringRange :: m -> Maybe Range

Returns the smallest range covering everything in the map (or Nothing, if the range would be empty).

Note that the default implementation of this operation might be inefficient.

several :: (IsBasicRangeMap a hl, Monoid hl) => [Ranges] -> a -> hl

Like singleton, but with several Ranges instead of only one.

class Convert a b where

Conversion between different types.

Methods

convert :: a -> b

insideAndOutside :: Range -> RangeMap a -> (RangeMap a, RangeMap a)

Returns a RangeMap overlapping the given range, as well as the rest of the map.

restrictTo :: Range -> RangeMap a -> RangeMap a

Restricts the RangeMap to the given range.

Orphan instances

NFData Aspect

Instance details

Methods

rnf :: Aspect -> ()

NFData Aspects

Instance details

Methods

rnf :: Aspects -> ()

NFData DefinitionSite

Instance details

Methods

rnf :: DefinitionSite -> ()

NFData OtherAspect

Instance details

Methods

rnf :: OtherAspect -> ()

Monoid Aspects

Instance details

Methods

mempty :: Aspects

mappend :: Aspects -> Aspects -> Aspects

mconcat :: [Aspects] -> Aspects

Monoid TokenBased

Instance details

Methods

mempty :: TokenBased

mappend :: TokenBased -> TokenBased -> TokenBased

mconcat :: [TokenBased] -> TokenBased

Semigroup Aspects

Instance details

Methods

(<>) :: Aspects -> Aspects -> Aspects

sconcat :: NonEmpty Aspects -> Aspects

stimes :: Integral b => b -> Aspects -> Aspects

Semigroup DefinitionSite

Instance details

Methods

(<>) :: DefinitionSite -> DefinitionSite -> DefinitionSite

sconcat :: NonEmpty DefinitionSite -> DefinitionSite

stimes :: Integral b => b -> DefinitionSite -> DefinitionSite

Semigroup TokenBased

Instance details

Methods

(<>) :: TokenBased -> TokenBased -> TokenBased

sconcat :: NonEmpty TokenBased -> TokenBased

stimes :: Integral b => b -> TokenBased -> TokenBased