Agda.Interaction.Library

Description

Library management.

Sample use:

    -- Get libraries as listed in .agda/libraries file.
    libs <- getInstalledLibraries Nothing

    -- Get the libraries (and immediate paths) relevant for projectRoot.
    -- This involves locating and processing the .agda-lib file for the project.
    (libNames, includePaths) <-  getDefaultLibraries projectRoot True

    -- Get include paths of depended-on libraries.
    resolvedPaths <- libraryIncludePaths Nothing libs libNames

    let allPaths = includePaths ++ resolvedPaths
  

Documentation

findProjectRoot :: FilePath -> LibM (Maybe FilePath)

Get project root

getDefaultLibraries

Arguments

:: FilePath
Project root.

-> Bool
Use defaults if no .agda-lib file exists for this project?

-> LibM ([LibName], [FilePath])
The returned LibNames are all non-empty strings.

Get dependencies and include paths for given project root:

Look for .agda-lib files according to findAgdaLibFiles. If none are found, use default dependencies (according to defaults file) and current directory (project root).

getInstalledLibraries

Arguments

:: Maybe FilePath
Override the default libraries file?

-> LibM [AgdaLibFile]
Content of library files. (Might have empty LibNames.)

Parse the descriptions of the libraries Agda knows about.

Returns none if there is no libraries file.

getTrustedExecutables

Arguments

:: LibM (Map ExeName FilePath)
Content of executables files.

Return the trusted executables Agda knows about.

Returns none if there is no executables file.

libraryIncludePaths

Arguments

:: Maybe FilePath
libraries file (error reporting only).

-> [AgdaLibFile]
Libraries Agda knows about.

-> [LibName]
(Non-empty) library names to be resolved to (lists of) pathes.

-> LibM [FilePath]
Resolved pathes (no duplicates). Contains "." if [LibName] does.

Get all include pathes for a list of libraries to use.

getAgdaLibFile :: FilePath -> LibM [AgdaLibFile]

Get the content of the .agda-lib file in the given project root.

getPrimitiveLibDir :: IO AbsolutePath

Returns the absolute default lib dir. This directory is used to store the Primitive.agda file.

classifyBuiltinModule_ :: AbsolutePath -> AbsolutePath -> Maybe IsBuiltinModule

Determine whether the second absolute path refers to one of Agda's primitive modules. The first argument should be the result of getPrimitiveLibDir.

builtinModules :: Set FilePath

All builtin modules.

builtinModulesWithSafePostulates :: Set FilePath

These builtins may use postulates, and are still considered --safe.

builtinModulesWithUnsafePostulates :: Set FilePath

These builtins may not use postulates under --safe. They are not automatically unsafe, but will be if they use an unsafe feature.

primitiveModules :: Set FilePath

The very magical, auto-imported modules.

data LibName

A symbolic library name.

Library names are structured into the base name and a suffix of version numbers, e.g. mylib-1.2.3. The version suffix is optional.

parseLibName :: String -> LibName

Split a library name into basename and a list of version numbers.

parseLibName "foo-1.2.3"    == LibName "foo" [1, 2, 3]
parseLibName "foo-01.002.3" == LibName "foo" [1, 2, 3]

Note that because of leading zeros, parseLibName is not injective. (prettyShow . parseLibName would produce a normal form.)

data OptionsPragma

The options from an OPTIONS pragma (or a .agda-lib file).

In the future it might be nice to switch to a more structured representation. Note that, currently, there is not a one-to-one correspondence between list elements and options.

Constructors

OptionsPragma

Fields

pragmaStrings :: [String]

The options.

pragmaRange :: Range

The range of the options in the pragma (not including things like an OPTIONS keyword).

data AgdaLibFile

Content of a .agda-lib file.

Constructors

AgdaLibFile

Fields

_libName :: LibName

The symbolic name of the library.

_libFile :: FilePath

Path to this .agda-lib file (not content of the file).

_libAbove :: !Int

How many directories above the Agda file is the .agda-lib file located?

_libIncludes :: [FilePath]

Roots where to look for the modules of the library.

_libDepends :: [LibName]

Dependencies.

_libPragmas :: OptionsPragma

Default pragma options for all files in the library.

type ExeName = Text

A symbolic executable name.

type LibM = ExceptT LibErrors (WriterT [LibWarning] (StateT LibState IO))

Throws LibErrors exceptions, still collects LibWarnings.

mkLibM :: [AgdaLibFile] -> LibErrorIO a -> LibM a

Raise collected LibErrors as exception.

data LibWarning

Constructors

LibWarning (Maybe LibPositionInfo) LibWarning'

data LibPositionInfo

Information about which .agda-lib file we are reading and from where in the libraries file it came from.

Constructors

LibPositionInfo

Fields

libFilePos :: Maybe FilePath

Name of libraries file.

lineNumPos :: LineNumber

Line number in libraries file.

filePos :: FilePath

Library file.

libraryWarningName :: LibWarning -> WarningName

data ProjectConfig

A file can either belong to a project located at a given root containing an .agda-lib file, or be part of the default project.

Constructors

ProjectConfig

Fields

configRoot :: FilePath

Directory which contains the .agda-lib file for the current project.

configAgdaLibFile :: FilePath

.agda-lib file relative to configRoot (filename only, no directory).

configAbove :: !Int

How many directories above the Agda file is the .agda-lib file located?

DefaultProjectConfig

Exported for testing

findLib' :: (a -> LibName) -> LibName -> [a] -> [a]

Generalized version of findLib for testing.

findLib' id "a"   [ "a-1", "a-02", "a-2", "b" ] == [ "a-02", "a-2" ]

findLib' id "a"   [ "a", "a-1", "a-01", "a-2", "b" ] == [ "a" ]
findLib' id "a-1" [ "a", "a-1", "a-01", "a-2", "b" ] == [ "a-1", "a-01" ]
findLib' id "a-2" [ "a", "a-1", "a-01", "a-2", "b" ] == [ "a-2" ]
findLib' id "c"   [ "a", "a-1", "a-01", "a-2", "b" ] == []