Agda.Interaction.Library.Base

Description

Basic data types for library management.

Documentation

data LibName

A symbolic library name.

Library names are structured into the base name and a suffix of version numbers, e.g. mylib-1.2.3. The version suffix is optional.

Constructors

LibName

Fields

libNameBase :: Text

Actual library name.

libNameVersion :: [Integer]

Major version, minor version, subminor version, etc., all non-negative.

parseLibName :: String -> LibName

Split a library name into basename and a list of version numbers.

parseLibName "foo-1.2.3"    == LibName "foo" [1, 2, 3]
parseLibName "foo-01.002.3" == LibName "foo" [1, 2, 3]

Note that because of leading zeros, parseLibName is not injective. (prettyShow . parseLibName would produce a normal form.)

data LibrariesFile

Constructors

LibrariesFile

Fields

lfPath :: FilePath

E.g. ~.agdalibraries.

lfExists :: Bool

The libraries file might not exist, but we may print its assumed location in error messages.

type ExeName = Text

A symbolic executable name.

type ExeMap = Map ExeName FilePath

data ExecutablesFile

Constructors

ExecutablesFile

Fields

efPath :: FilePath

E.g. ~.agdaexecutables.

efExists :: Bool

The executables file might not exist, but we may print its assumed location in error messages.

libNameForCurrentDir :: LibName

The special name "." is used to indicated that the current directory should count as a project root.

data ProjectConfig

A file can either belong to a project located at a given root containing an .agda-lib file, or be part of the default project.

Constructors

ProjectConfig

Fields

configRoot :: FilePath

Directory which contains the .agda-lib file for the current project.

configAgdaLibFile :: FilePath

.agda-lib file relative to configRoot (filename only, no directory).

configAbove :: !Int

How many directories above the Agda file is the .agda-lib file located?

DefaultProjectConfig

data OptionsPragma

The options from an OPTIONS pragma (or a .agda-lib file).

In the future it might be nice to switch to a more structured representation. Note that, currently, there is not a one-to-one correspondence between list elements and options.

Constructors

OptionsPragma

Fields

pragmaStrings :: [String]

The options.

pragmaRange :: Range

The range of the options in the pragma (not including things like an OPTIONS keyword).

data AgdaLibFile

Content of a .agda-lib file.

Constructors

AgdaLibFile

Fields

_libName :: LibName

The symbolic name of the library.

_libFile :: FilePath

Path to this .agda-lib file (not content of the file).

_libAbove :: !Int

How many directories above the Agda file is the .agda-lib file located?

_libIncludes :: [FilePath]

Roots where to look for the modules of the library.

_libDepends :: [LibName]

Dependencies.

_libPragmas :: OptionsPragma

Default pragma options for all files in the library.

emptyLibFile :: AgdaLibFile

Lenses

Lenses for ProjectConfig

lensConfigAbove :: Lens' ProjectConfig Int

Lenses for AgdaLibFile

libName :: Lens' AgdaLibFile LibName

libFile :: Lens' AgdaLibFile FilePath

libAbove :: Lens' AgdaLibFile Int

libIncludes :: Lens' AgdaLibFile [FilePath]

libDepends :: Lens' AgdaLibFile [LibName]

libPragmas :: Lens' AgdaLibFile OptionsPragma

Library warnings and errors

Position information

type LineNumber = Int

data LibPositionInfo

Information about which .agda-lib file we are reading and from where in the libraries file it came from.

Constructors

LibPositionInfo

Fields

libFilePos :: Maybe FilePath

Name of libraries file.

lineNumPos :: LineNumber

Line number in libraries file.

filePos :: FilePath

Library file.

Warnings

data LibWarning

Constructors

LibWarning (Maybe LibPositionInfo) LibWarning'

data LibWarning'

Library Warnings.

Constructors

UnknownField String

libraryWarningName :: LibWarning -> WarningName

Errors

data LibError

Constructors

LibError (Maybe LibPositionInfo) LibError'

data LibError'

Collected errors while processing library files.

Constructors

LibrariesFileNotFound FilePath
The user specified replacement for the default libraries file does not exist.

LibNotFound LibrariesFile LibName
Raised when a library name could not successfully be resolved to an .agda-lib file.

AmbiguousLib LibName (List2 AgdaLibFile)
Raised when a library name is defined in several .agda-lib files.

SeveralAgdaLibFiles FilePath (List2 FilePath)
The given project root contains more than one .agda-lib file.

LibParseError LibParseError
The .agda-lib file could not be parsed.

ReadError
An I/O Error occurred when reading a file.

Fields

IOException

The caught exception

String

Explanation when this error occurred.

DuplicateExecutable
The executables file contains duplicate entries.

Fields

FilePath

Name of the executables file.

Text

Name of the executable that is defined twice.

(List2 (LineNumber, FilePath))

The resolutions of the executable.

data LibParseError

Exceptions thrown by the .agda-lib parser.

Constructors

BadLibraryName String
An invalid library name, e.g., containing spaces.

ReadFailure FilePath IOException
I/O error while reading file.

MissingFields (List1 String)
Missing these mandatory fields.

DuplicateFields (List1 String)
These fields occur each more than once.

MissingFieldName LineNumber
At the given line number, a field name is missing before the :.

BadFieldName LineNumber String
At the given line number, an invalid field name is encountered before the :. (E.g., containing spaces.)

MissingColonForField LineNumber String
At the given line number, the given field is not followed by :.

ContentWithoutField LineNumber
At the given line number, indented text (content) is not preceded by a field.

Raising warnings and errors

type LibErrWarns = [Either LibError LibWarning]

Collection of LibErrors and LibWarnings.

warnings :: MonadWriter LibErrWarns m => List1 LibWarning -> m ()

warnings' :: MonadWriter LibErrWarns m => List1 LibWarning' -> m ()

raiseErrors' :: MonadWriter LibErrWarns m => List1 LibError' -> m ()

raiseErrors :: MonadWriter LibErrWarns m => List1 LibError -> m ()

Library Monad

type LibErrorIO = WriterT LibErrWarns (StateT LibState IO)

Collects LibErrors and LibWarnings.

type LibM = ExceptT LibErrors (WriterT [LibWarning] (StateT LibState IO))

Throws LibErrors exceptions, still collects LibWarnings.

type LibState = LibCache

data LibCache

Cache locations of project configurations and parsed .agda-lib files.

Constructors

LibCache

Fields

projectConfigs :: !(Map FilePath ProjectConfig)

Map from directories to paths of closest enclosing .agda-lib files (or DefaultProjectConfig if there are none).

agdaLibFiles :: !(Map FilePath AgdaLibFile)

Contents of .agda-lib files that have already been parsed.

data LibErrors

Collected errors when processing an .agda-lib file.

Constructors

LibErrors

Fields

libErrorsInstalledLibraries :: [AgdaLibFile]

libErrors :: List1 LibError

runLibM :: LibM a -> LibState -> IO ((Either LibErrors a, [LibWarning]), LibState)

getCachedProjectConfig :: (MonadState LibState m, MonadIO m) => FilePath -> m (Maybe ProjectConfig)

storeCachedProjectConfig :: (MonadState LibState m, MonadIO m) => FilePath -> ProjectConfig -> m ()

getCachedAgdaLibFile :: (MonadState LibState m, MonadIO m) => FilePath -> m (Maybe AgdaLibFile)

storeCachedAgdaLibFile :: (MonadState LibState m, MonadIO m) => FilePath -> AgdaLibFile -> m ()

Prettyprinting errors and warnings

formatLibError :: [AgdaLibFile] -> LibError -> Doc

Pretty-print LibError.

formatLibErrors :: LibErrors -> Doc

Pretty-print LibErrors.

hasLineNumber :: LibParseError -> Maybe LineNumber

Does a parse error contain a line number?

formatLibPositionInfo :: LibPositionInfo -> LibParseError -> Doc

Compute a position position prefix.

Depending on the error to be printed, it will

either give the name of the libraries file and a line inside it,

or give the name of the .agda-lib file.

prettyInstalledLibraries :: [AgdaLibFile] -> Doc

Orphan instances

NFData IOException

Instance details

Methods

rnf :: IOException -> ()