Agda.Interaction.Options

Documentation

checkOpts :: MonadError OptionError m => CommandLineOptions -> m CommandLineOptions

Checks that the given options are consistent. Also makes adjustments (e.g. when one option implies another).

deadStandardOptions :: [OptDescr (Flag CommandLineOptions)]

Command line options of previous versions of Agda. Should not be listed in the usage info, put parsed by GetOpt for good error messaging.

defaultInteractionOptions :: PragmaOptions

defaultOptions :: CommandLineOptions

defaultPragmaOptions :: PragmaOptions

getOptSimple

Arguments

:: [String]
command line argument words

-> [OptDescr (Flag opts)]
options handlers

-> (String -> Flag opts)
handler of non-options (only one is allowed)

-> Flag opts
combined opts data structure transformer

Simple interface for System.Console.GetOpt Could be moved to Agda.Utils.Options (does not exist yet)

impliedPragmaOptions :: [ImpliedPragmaOption]

infectiveCoinfectiveOptions :: [InfectiveCoinfectiveOption]

Infective and coinfective options.

Note that --cubical and --erased-cubical are "jointly infective": if one of them is used in one module, then one or the other must be used in all modules that depend on this module.

inputFlag :: FilePath -> Flag CommandLineOptions

lensOptAllowExec :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptAllowIncompleteMatch :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptAllowUnsolved :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptAutoInline :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptBacktrackingInstances :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptCaching :: Functor f => (WithDefault 'True -> f (WithDefault 'True)) -> PragmaOptions -> f PragmaOptions

lensOptCallByName :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptCohesion :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptCompileMain :: Functor f => (WithDefault 'True -> f (WithDefault 'True)) -> PragmaOptions -> f PragmaOptions

lensOptConfluenceCheck :: Functor f => (Maybe ConfluenceCheck -> f (Maybe ConfluenceCheck)) -> PragmaOptions -> f PragmaOptions

lensOptCopatterns :: Functor f => (WithDefault 'True -> f (WithDefault 'True)) -> PragmaOptions -> f PragmaOptions

lensOptCountClusters :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptCubical :: Functor f => (Maybe Cubical -> f (Maybe Cubical)) -> PragmaOptions -> f PragmaOptions

lensOptCubicalCompatible :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptCumulativity :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptDoubleCheck :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptEraseRecordParameters :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptErasedMatches :: Functor f => (WithDefault 'True -> f (WithDefault 'True)) -> PragmaOptions -> f PragmaOptions

lensOptErasure :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptEta :: Functor f => (WithDefault 'True -> f (WithDefault 'True)) -> PragmaOptions -> f PragmaOptions

lensOptExactSplit :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptExperimentalIrrelevance :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptExperimentalLazyInstances :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptFastReduce :: Functor f => (WithDefault 'True -> f (WithDefault 'True)) -> PragmaOptions -> f PragmaOptions

lensOptFirstOrder :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptFlatSplit :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptForcing :: Functor f => (WithDefault 'True -> f (WithDefault 'True)) -> PragmaOptions -> f PragmaOptions

lensOptGuarded :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptGuardedness :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptHiddenArgumentPuns :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptImportSorts :: Functor f => (WithDefault 'True -> f (WithDefault 'True)) -> PragmaOptions -> f PragmaOptions

lensOptInferAbsurdClauses :: Functor f => (WithDefault 'True -> f (WithDefault 'True)) -> PragmaOptions -> f PragmaOptions

lensOptInjectiveTypeConstructors :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptInstanceSearchDepth :: Functor f => (Int -> f Int) -> PragmaOptions -> f PragmaOptions

lensOptInversionMaxDepth :: Functor f => (Int -> f Int) -> PragmaOptions -> f PragmaOptions

lensOptIrrelevantProjections :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptKeepCoveringClauses :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptKeepPatternVariables :: Functor f => (WithDefault 'True -> f (WithDefault 'True)) -> PragmaOptions -> f PragmaOptions

lensOptLevelUniverse :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptLoadPrimitives :: Functor f => (WithDefault 'True -> f (WithDefault 'True)) -> PragmaOptions -> f PragmaOptions

lensOptNoUniverseCheck :: Functor f => (WithDefault' Bool 'True -> f (WithDefault' Bool 'True)) -> PragmaOptions -> f PragmaOptions

lensOptOmegaInOmega :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptPatternMatching :: Functor f => (WithDefault 'True -> f (WithDefault 'True)) -> PragmaOptions -> f PragmaOptions

lensOptPolarity :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptPositivityCheck :: Functor f => (WithDefault 'True -> f (WithDefault 'True)) -> PragmaOptions -> f PragmaOptions

lensOptPostfixProjections :: Functor f => (WithDefault 'True -> f (WithDefault 'True)) -> PragmaOptions -> f PragmaOptions

lensOptPrintPatternSynonyms :: Functor f => (WithDefault 'True -> f (WithDefault 'True)) -> PragmaOptions -> f PragmaOptions

lensOptProfiling :: Functor f => (ProfileOptions -> f ProfileOptions) -> PragmaOptions -> f PragmaOptions

lensOptProjectionLike :: Functor f => (WithDefault 'True -> f (WithDefault 'True)) -> PragmaOptions -> f PragmaOptions

lensOptProp :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptQualifiedInstances :: Functor f => (WithDefault 'True -> f (WithDefault 'True)) -> PragmaOptions -> f PragmaOptions

lensOptRequireUniqueMetaSolutions :: Functor f => (WithDefault 'True -> f (WithDefault 'True)) -> PragmaOptions -> f PragmaOptions

lensOptRewriting :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptSafe :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptSaveMetas :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptShowIdentitySubstitutions :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptShowImplicit :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptShowIrrelevant :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptSizedTypes :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptSyntacticEquality :: Functor f => (Maybe Int -> f (Maybe Int)) -> PragmaOptions -> f PragmaOptions

lensOptTerminationCheck :: Functor f => (WithDefault 'True -> f (WithDefault 'True)) -> PragmaOptions -> f PragmaOptions

lensOptTerminationDepth :: Functor f => (CutOff -> f CutOff) -> PragmaOptions -> f PragmaOptions

lensOptTwoLevel :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

lensOptUniverseCheck :: Functor f => (WithDefault 'True -> f (WithDefault 'True)) -> PragmaOptions -> f PragmaOptions

lensOptUniversePolymorphism :: Functor f => (WithDefault 'True -> f (WithDefault 'True)) -> PragmaOptions -> f PragmaOptions

lensOptUseUnicode :: Functor f => (WithDefault' UnicodeOrAscii 'True -> f (WithDefault' UnicodeOrAscii 'True)) -> PragmaOptions -> f PragmaOptions

lensOptVerbose :: Functor f => (Verbosity -> f Verbosity) -> PragmaOptions -> f PragmaOptions

lensOptWarningMode :: Functor f => (WarningMode -> f WarningMode) -> PragmaOptions -> f PragmaOptions

lensOptWithoutK :: Functor f => (WithDefault 'False -> f (WithDefault 'False)) -> PragmaOptions -> f PragmaOptions

mapFlag :: (String -> String) -> OptDescr a -> OptDescr a

Map a function over the long options. Also removes the short options. Will be used to add the plugin name to the plugin options.

optAllowExec :: PragmaOptions -> Bool

optAllowIncompleteMatch :: PragmaOptions -> Bool

optAllowUnsolved :: PragmaOptions -> Bool

optAutoInline :: PragmaOptions -> Bool

optBacktrackingInstances :: PragmaOptions -> Bool

optCaching :: PragmaOptions -> Bool

optCallByName :: PragmaOptions -> Bool

optCohesion :: PragmaOptions -> Bool

optCohesion is implied by optFlatSplit.

optCompileNoMain :: PragmaOptions -> Bool

optConfluenceCheck :: PragmaOptions -> Maybe ConfluenceCheck

optCopatterns :: PragmaOptions -> Bool

optCountClusters :: PragmaOptions -> Bool

optCubical :: PragmaOptions -> Maybe Cubical

optCubicalCompatible :: PragmaOptions -> Bool

optCumulativity :: PragmaOptions -> Bool

optDoubleCheck :: PragmaOptions -> Bool

optEraseRecordParameters :: PragmaOptions -> Bool

optErasedMatches :: PragmaOptions -> Bool

optErasure :: PragmaOptions -> Bool

optErasure is implied by optEraseRecordParameters. optErasure is also implied by an explicitly given `--erased-matches`.

optEta :: PragmaOptions -> Bool

optExperimentalIrrelevance :: PragmaOptions -> Bool

optFastReduce :: PragmaOptions -> Bool

optFirstOrder :: PragmaOptions -> Bool

optFlatSplit :: PragmaOptions -> Bool

optForcedArgumentRecursion :: PragmaOptions -> Bool

optForcing :: PragmaOptions -> Bool

optGuarded :: PragmaOptions -> Bool

optGuardedness :: PragmaOptions -> Bool

optHiddenArgumentPuns :: PragmaOptions -> Bool

optImportSorts :: PragmaOptions -> Bool

optImportSorts requires optLoadPrimitives.

optInferAbsurdClauses :: PragmaOptions -> Bool

optInjectiveTypeConstructors :: PragmaOptions -> Bool

optInstanceSearchDepth :: PragmaOptions -> Int

optInversionMaxDepth :: PragmaOptions -> Int

optIrrelevantProjections :: PragmaOptions -> Bool

optKeepCoveringClauses :: PragmaOptions -> Bool

optKeepPatternVariables :: PragmaOptions -> Bool

optLargeIndices :: PragmaOptions -> Bool

optLevelUniverse :: PragmaOptions -> Bool

optLoadPrimitives :: PragmaOptions -> Bool

optOmegaInOmega :: PragmaOptions -> Bool

optPatternMatching :: PragmaOptions -> Bool

optPolarity :: PragmaOptions -> Bool

optPositivityCheck :: PragmaOptions -> Bool

optPostfixProjections :: PragmaOptions -> Bool

optPrintPatternSynonyms :: PragmaOptions -> Bool

optProfiling :: PragmaOptions -> ProfileOptions

optProjectionLike :: PragmaOptions -> Bool

optProp :: PragmaOptions -> Bool

optQualifiedInstances :: PragmaOptions -> Bool

optRequireUniqueMetaSolutions :: PragmaOptions -> Bool

optRewriting :: PragmaOptions -> Bool

optSafe :: PragmaOptions -> Bool

optSaveMetas :: PragmaOptions -> Bool

optShowGeneralized :: PragmaOptions -> Bool

optShowIdentitySubstitutions :: PragmaOptions -> Bool

optShowImplicit :: PragmaOptions -> Bool

optShowIrrelevant :: PragmaOptions -> Bool

optSizedTypes :: PragmaOptions -> Bool

optSyntacticEquality :: PragmaOptions -> Maybe Int

optTerminationCheck :: PragmaOptions -> Bool

optTerminationDepth :: PragmaOptions -> CutOff

optTwoLevel :: PragmaOptions -> Bool

optUniverseCheck :: PragmaOptions -> Bool

optUniversePolymorphism :: PragmaOptions -> Bool

optUseUnicode :: PragmaOptions -> UnicodeOrAscii

optVerbose :: PragmaOptions -> Verbosity

optWarningMode :: PragmaOptions -> WarningMode

optWithoutK :: PragmaOptions -> Bool

optionWarningName :: OptionWarning -> WarningName

parsePluginOptions :: [String] -> [OptDescr (Flag opts)] -> Flag opts

Parse options for a plugin.

parsePragmaOptions

Arguments

:: OptionsPragma
Pragma options.

-> CommandLineOptions
Command-line options which should be updated.

-> OptM PragmaOptions

Parse options from an options pragma.

parseVerboseKey :: VerboseKey -> [VerboseKeyItem]

recheckBecausePragmaOptionsChanged

Arguments

:: PragmaOptions
The options that were used to check the file.

-> PragmaOptions
The options that are currently in effect.

-> Bool

This function returns True if the file should be rechecked.

runOptM :: OptM opts -> (Either OptionError opts, OptionWarnings)

safeFlag :: Flag PragmaOptions

standardOptions :: [OptDescr (Flag CommandLineOptions)]

standardOptions_ :: [OptDescr ()]

Used for printing usage info. Does not include the dead options.

stripRTS :: [String] -> [String]

Removes RTS options from a list of options.

unsafePragmaOptions :: PragmaOptions -> [String]

Check for unsafe pragmas. Gives a list of used unsafe flags.

usage :: [OptDescr ()] -> String -> Help -> String

The usage info message. The argument is the program name (probably agda).

defaultCutOff :: CutOff

The default termination depth.

type Flag opts = opts -> OptM opts

f :: Flag opts is an action on the option record that results from parsing an option. f opts produces either an error message or an updated options record

data ImpliedPragmaOption where

Constructors

ImpliesPragmaOption

Fields

:: forall (a :: Bool) (b :: Bool). String

-> Bool

-> (PragmaOptions -> WithDefault a)

-> String

-> Bool

-> (PragmaOptions -> WithDefault b)

-> ImpliedPragmaOption

The first option having the given value implies the second option having its given value. For instance, `ImpliesPragmaOption "lossy-unification" True _optFirstOrder "require-unique-meta-solutions" False _optRequireUniqueMetaSolutions` encodes the fact that --lossy-unification implies --no-require-unique-meta-solutions.

data InfectiveCoinfectiveOption

Descriptions of infective and coinfective options.

Constructors

ICOption

Fields

icOptionActive :: PragmaOptions -> Bool

Is the option active?

icOptionDescription :: String

A description of the option (typically a flag that activates the option).

icOptionKind :: InfectiveCoinfective

Is the option (roughly speaking) infective or coinfective?

icOptionOK :: PragmaOptions -> PragmaOptions -> Bool

This function returns True exactly when, from the perspective of the option in question, the options in the current module (the first argument) are compatible with the options in a given imported module (the second argument).

icOptionWarning :: TopLevelModuleName -> Doc

A warning message that should be used if this option is not used correctly. The given module name is the name of an imported module for which icOptionOK failed.

data OptM a

The options parse monad OptM collects warnings that are not discarded when a fatal error occurrs

type OptionError = String

data OptionWarning

Warnings when parsing options.

Constructors

OptionRenamed
Name of option changed in a newer version of Agda.

Fields

oldOptionName :: String

newOptionName :: String

WarningProblem WarningModeError
A problem with setting or unsetting a warning.

data CommandLineOptions

Constructors

Options

Fields

optProgramName :: String

optInputFile :: Maybe FilePath

optIncludePaths :: [FilePath]

optAbsoluteIncludePaths :: [AbsolutePath]

The list should not contain duplicates.

optLibraries :: [LibName]

optOverrideLibrariesFile :: Maybe FilePath

Use this (if Just) instead of ~/.agda/libraries.

optDefaultLibs :: Bool

Use ~/.agda/defaults.

optUseLibs :: Bool

look for .agda-lib files.

optTraceImports :: Integer

Configure notifications about imported modules.

optTrustedExecutables :: Map ExeName FilePath

Map names of trusted executables to absolute paths.

optPrintAgdaDataDir :: Bool

optPrintAgdaAppDir :: Bool

optPrintVersion :: Maybe PrintAgdaVersion

optPrintHelp :: Maybe Help

optBuildLibrary :: Bool

--build-library: Build all modules in the current library.

optSetup :: Bool

Force Agda to self-setup at startup.

optEmacsMode :: Set EmacsModeCommand

Emacs mode administration (formerly done by the agda-mode executable). Accepted subcommands: locate, setup, compile. Can be given several times, subcommands are accumulated in the set.

optInteractive :: Bool

Agda REPL (-I).

optGHCiInteraction :: Bool

--interaction mode.

optJSONInteraction :: Bool

--interaction-json mode.

optExitOnError :: !Bool

Exit if an interactive command fails.

optCompileDir :: Maybe FilePath

In the absence of a path the project root is used.

optGenerateVimFile :: Bool

optIgnoreInterfaces :: Bool

optIgnoreAllInterfaces :: Bool

optPragmaOptions :: PragmaOptions

optOnlyScopeChecking :: Bool

Should the top-level module only be scope-checked, and not type-checked?

optTransliterate :: Bool

Should code points that are not supported by the locale be transliterated?

optDiagnosticsColour :: DiagnosticsColours

Configure colour output.

data ConfluenceCheck

Constructors

LocalConfluenceCheck

GlobalConfluenceCheck

data DiagnosticsColours

Constructors

AlwaysColour

NeverColour

AutoColour

data EmacsModeCommand

If several --emacs-mode commands are given, they are executed in the order as given in this datatype.

Constructors

EmacsModeCompile
Compile the .el files to .elc.

EmacsModeSetup
Add the initialization lines to .emacs.

EmacsModeLocate
Print the installation location of agda2.el.

data InfectiveCoinfective

Infective or coinfective?

Constructors

Infective

Coinfective

data PragmaOptions

Options which can be set in a pragma.

Constructors

PragmaOptions

Fields

_optShowImplicit :: WithDefault 'False

_optShowGeneralized :: WithDefault 'True

Show generalized parameters in Pi types

_optShowIrrelevant :: WithDefault 'False

_optUseUnicode :: WithDefault' UnicodeOrAscii 'True

_optVerbose :: !Verbosity

_optProfiling :: ProfileOptions

_optProp :: WithDefault 'False

_optLevelUniverse :: WithDefault 'False

_optTwoLevel :: WithDefault 'False

_optAllowUnsolved :: WithDefault 'False

_optAllowIncompleteMatch :: WithDefault 'False

_optPositivityCheck :: WithDefault 'True

_optTerminationCheck :: WithDefault 'True

_optTerminationDepth :: CutOff

Cut off structural order comparison at some depth in termination checker?

_optUniverseCheck :: WithDefault 'True

_optOmegaInOmega :: WithDefault 'False

_optCumulativity :: WithDefault 'False

_optSizedTypes :: WithDefault 'False

_optGuardedness :: WithDefault 'False

_optInjectiveTypeConstructors :: WithDefault 'False

_optUniversePolymorphism :: WithDefault 'True

_optIrrelevantProjections :: WithDefault 'False

_optExperimentalIrrelevance :: WithDefault 'False

irrelevant levels, irrelevant data matching

_optWithoutK :: WithDefault 'False

_optCubicalCompatible :: WithDefault 'False

_optCopatterns :: WithDefault 'True

Allow definitions by copattern matching?

_optPatternMatching :: WithDefault 'True

Is pattern matching allowed in the current file?

_optExactSplit :: WithDefault 'False

_optHiddenArgumentPuns :: WithDefault 'False

Should patterns of the form {x} or ⦃ x ⦄ be interpreted as puns?

_optEta :: WithDefault 'True

_optForcing :: WithDefault 'True

Perform the forcing analysis on data constructors?

_optProjectionLike :: WithDefault 'True

Perform the projection-likeness analysis on functions?

_optErasure :: WithDefault 'False

_optErasedMatches :: WithDefault 'True

Allow matching in erased positions for single-constructor, non-indexed data/record types. (This kind of matching is always allowed for record types with η-equality.)

_optEraseRecordParameters :: WithDefault 'False

Mark parameters of record modules as erased?

_optRewriting :: WithDefault 'False

Can rewrite rules be added and used?

_optCubical :: Maybe Cubical

_optGuarded :: WithDefault 'False

_optFirstOrder :: WithDefault 'False

Should we speculatively unify function applications as if they were injective? Implies optRequireUniqueMetaSolutions.

_optRequireUniqueMetaSolutions :: WithDefault 'True

Forbid non-unique meta solutions allowed. For instance from INJECTIVE_FOR_INFERENCE pragmas.

_optPostfixProjections :: WithDefault 'True

Should system generated projections ProjSystem be printed postfix (True) or prefix (False).

_optKeepPatternVariables :: WithDefault 'True

Should case splitting replace variables with dot patterns (False) or keep them as variables (True).

_optInferAbsurdClauses :: WithDefault 'True

Should case splitting and coverage checking try to discharge absurd clauses? Default: True, but False might make coverage checking considerably faster in some cases.

_optInstanceSearchDepth :: Int

_optBacktrackingInstances :: WithDefault 'False

_optQualifiedInstances :: WithDefault 'True

Should instance search consider instances with qualified names?

_optInversionMaxDepth :: Int

_optSafe :: WithDefault 'False

_optDoubleCheck :: WithDefault 'False

_optSyntacticEquality :: !(Maybe Int)

Should the conversion checker use the syntactic equality shortcut? Nothing means that it should. Just n, for a non-negative number n, means that syntactic equality checking gets n units of fuel. If the fuel becomes zero, then syntactic equality checking is turned off. The fuel counter is decreased in the failure continuation of checkSyntacticEquality.

_optWarningMode :: WarningMode

_optCompileMain :: WithDefault 'True

Treat the module given at the command line or via interaction as main module in compilation?

_optCaching :: WithDefault 'True

_optCountClusters :: WithDefault 'False

Count extended grapheme clusters rather than code points when generating LaTeX.

_optAutoInline :: WithDefault 'False

Automatic compile-time inlining for simple definitions (unless marked NOINLINE).

_optPrintPatternSynonyms :: WithDefault 'True

_optFastReduce :: WithDefault 'True

Use the Agda abstract machine (fastReduce)?

_optCallByName :: WithDefault 'False

Use call-by-name instead of call-by-need.

_optConfluenceCheck :: Maybe ConfluenceCheck

Check confluence of rewrite rules?

_optCohesion :: WithDefault 'False

Are the cohesion modalities available?

_optFlatSplit :: WithDefault 'False

Can we split on a (@flat x : A) argument?

_optPolarity :: WithDefault 'False

Can we use modal polarities (++, +, etc.)?

_optImportSorts :: WithDefault 'True

Should every top-level module start with an implicit statement open import Agda.Primitive using (Set; Prop)?

_optLoadPrimitives :: WithDefault 'True

Should we load the primitive modules at all? This is a stronger form of optImportSorts.

_optAllowExec :: WithDefault 'False

Allow running external executables from meta programs.

_optSaveMetas :: WithDefault 'False

Save meta-variables to interface files.

_optShowIdentitySubstitutions :: WithDefault 'False

Show identity substitutions when pretty-printing terms (i.e. always show all arguments of a metavariable).

_optKeepCoveringClauses :: WithDefault 'False

Do not discard clauses constructed by the coverage checker (needed for some external backends).

_optLargeIndices :: WithDefault 'False

Allow large indices, and large forced arguments in constructors.

_optForcedArgumentRecursion :: WithDefault 'True

Allow recursion on forced constructor arguments.

_optExperimentalLazyInstances :: WithDefault 'False

Defer instance search based on discrimination tree results

data PrintAgdaVersion

Options --version and --numeric-version (last wins).

Constructors

PrintAgdaVersion
Print Agda version information.

PrintAgdaNumericVersion
Print Agda version number.

type VerboseKey = String

type VerboseLevel = Int

type Verbosity = Maybe (Trie VerboseKeyItem VerboseLevel)

Nothing is used if no verbosity options have been given, thus making it possible to handle the default case relatively quickly. Note that Nothing corresponds to a trie with verbosity level 1 for the empty path.

data WarningMode

A WarningMode has two components: a set of warnings to be displayed and a flag stating whether warnings should be turned into fatal errors.

Constructors

WarningMode

Fields

_warningSet :: Set WarningName

_warn2Error :: Bool

data UnicodeOrAscii

We want to know whether we are allowed to insert unicode characters or not.

Constructors

UnicodeOk
true: Unicode characters are allowed.

AsciiOnly
'false: Stick to ASCII.

data ArgDescr a

Describes whether an option takes an argument or not, and if so how the argument is injected into a value of type a.

Constructors

NoArg a
no argument expected

ReqArg (String -> a) String
option requires argument

OptArg (Maybe String -> a) String
optional argument

data OptDescr a

Each OptDescr describes a single option.

The arguments to Option are:

list of short option characters

list of long option strings (without "--")

argument descriptor

explanation of option for user

Constructors

Option [Char] [String] (ArgDescr a) String

class (Functor m, Applicative m, Monad m) => HasOptions (m :: Type -> Type) where

Minimal complete definition

Nothing

Methods

pragmaOptions :: m PragmaOptions

Returns the pragma options which are currently in effect.

default pragmaOptions :: forall (n :: Type -> Type) (t :: (Type -> Type) -> Type -> Type). (HasOptions n, MonadTrans t, m ~ t n) => m PragmaOptions

commandLineOptions :: m CommandLineOptions

Returns the command line options which are currently in effect.

default commandLineOptions :: forall (n :: Type -> Type) (t :: (Type -> Type) -> Type -> Type). (HasOptions n, MonadTrans t, m ~ t n) => m CommandLineOptions

data CommandLineOptions

Constructors

Options

Fields

optProgramName :: String

optInputFile :: Maybe FilePath

optIncludePaths :: [FilePath]

optAbsoluteIncludePaths :: [AbsolutePath]

The list should not contain duplicates.

optLibraries :: [LibName]

optOverrideLibrariesFile :: Maybe FilePath

Use this (if Just) instead of ~/.agda/libraries.

optDefaultLibs :: Bool

Use ~/.agda/defaults.

optUseLibs :: Bool

look for .agda-lib files.

optTraceImports :: Integer

Configure notifications about imported modules.

optTrustedExecutables :: Map ExeName FilePath

Map names of trusted executables to absolute paths.

optPrintAgdaDataDir :: Bool

optPrintAgdaAppDir :: Bool

optPrintVersion :: Maybe PrintAgdaVersion

optPrintHelp :: Maybe Help

optBuildLibrary :: Bool

--build-library: Build all modules in the current library.

optSetup :: Bool

Force Agda to self-setup at startup.

optEmacsMode :: Set EmacsModeCommand

Emacs mode administration (formerly done by the agda-mode executable). Accepted subcommands: locate, setup, compile. Can be given several times, subcommands are accumulated in the set.

optInteractive :: Bool

Agda REPL (-I).

optGHCiInteraction :: Bool

--interaction mode.

optJSONInteraction :: Bool

--interaction-json mode.

optExitOnError :: !Bool

Exit if an interactive command fails.

optCompileDir :: Maybe FilePath

In the absence of a path the project root is used.

optGenerateVimFile :: Bool

optIgnoreInterfaces :: Bool

optIgnoreAllInterfaces :: Bool

optPragmaOptions :: PragmaOptions

optOnlyScopeChecking :: Bool

Should the top-level module only be scope-checked, and not type-checked?

optTransliterate :: Bool

Should code points that are not supported by the locale be transliterated?

optDiagnosticsColour :: DiagnosticsColours

Configure colour output.

data ConfluenceCheck

Constructors

LocalConfluenceCheck

GlobalConfluenceCheck

data DiagnosticsColours

Constructors

AlwaysColour

NeverColour

AutoColour

data EmacsModeCommand

If several --emacs-mode commands are given, they are executed in the order as given in this datatype.

Constructors

EmacsModeCompile
Compile the .el files to .elc.

EmacsModeSetup
Add the initialization lines to .emacs.

EmacsModeLocate
Print the installation location of agda2.el.

data InfectiveCoinfective

Infective or coinfective?

Constructors

Infective

Coinfective

class LensPragmaOptions a where

Minimal complete definition

lensPragmaOptions

Methods

getPragmaOptions :: a -> PragmaOptions

setPragmaOptions :: PragmaOptions -> a -> a

mapPragmaOptions :: (PragmaOptions -> PragmaOptions) -> a -> a

lensPragmaOptions :: Lens' a PragmaOptions

data PragmaOptions

Options which can be set in a pragma.

Constructors

PragmaOptions

Fields

_optShowImplicit :: WithDefault 'False

_optShowGeneralized :: WithDefault 'True

Show generalized parameters in Pi types

_optShowIrrelevant :: WithDefault 'False

_optUseUnicode :: WithDefault' UnicodeOrAscii 'True

_optVerbose :: !Verbosity

_optProfiling :: ProfileOptions

_optProp :: WithDefault 'False

_optLevelUniverse :: WithDefault 'False

_optTwoLevel :: WithDefault 'False

_optAllowUnsolved :: WithDefault 'False

_optAllowIncompleteMatch :: WithDefault 'False

_optPositivityCheck :: WithDefault 'True

_optTerminationCheck :: WithDefault 'True

_optTerminationDepth :: CutOff

Cut off structural order comparison at some depth in termination checker?

_optUniverseCheck :: WithDefault 'True

_optOmegaInOmega :: WithDefault 'False

_optCumulativity :: WithDefault 'False

_optSizedTypes :: WithDefault 'False

_optGuardedness :: WithDefault 'False

_optInjectiveTypeConstructors :: WithDefault 'False

_optUniversePolymorphism :: WithDefault 'True

_optIrrelevantProjections :: WithDefault 'False

_optExperimentalIrrelevance :: WithDefault 'False

irrelevant levels, irrelevant data matching

_optWithoutK :: WithDefault 'False

_optCubicalCompatible :: WithDefault 'False

_optCopatterns :: WithDefault 'True

Allow definitions by copattern matching?

_optPatternMatching :: WithDefault 'True

Is pattern matching allowed in the current file?

_optExactSplit :: WithDefault 'False

_optHiddenArgumentPuns :: WithDefault 'False

Should patterns of the form {x} or ⦃ x ⦄ be interpreted as puns?

_optEta :: WithDefault 'True

_optForcing :: WithDefault 'True

Perform the forcing analysis on data constructors?

_optProjectionLike :: WithDefault 'True

Perform the projection-likeness analysis on functions?

_optErasure :: WithDefault 'False

_optErasedMatches :: WithDefault 'True

Allow matching in erased positions for single-constructor, non-indexed data/record types. (This kind of matching is always allowed for record types with η-equality.)

_optEraseRecordParameters :: WithDefault 'False

Mark parameters of record modules as erased?

_optRewriting :: WithDefault 'False

Can rewrite rules be added and used?

_optCubical :: Maybe Cubical

_optGuarded :: WithDefault 'False

_optFirstOrder :: WithDefault 'False

Should we speculatively unify function applications as if they were injective? Implies optRequireUniqueMetaSolutions.

_optRequireUniqueMetaSolutions :: WithDefault 'True

Forbid non-unique meta solutions allowed. For instance from INJECTIVE_FOR_INFERENCE pragmas.

_optPostfixProjections :: WithDefault 'True

Should system generated projections ProjSystem be printed postfix (True) or prefix (False).

_optKeepPatternVariables :: WithDefault 'True

Should case splitting replace variables with dot patterns (False) or keep them as variables (True).

_optInferAbsurdClauses :: WithDefault 'True

Should case splitting and coverage checking try to discharge absurd clauses? Default: True, but False might make coverage checking considerably faster in some cases.

_optInstanceSearchDepth :: Int

_optBacktrackingInstances :: WithDefault 'False

_optQualifiedInstances :: WithDefault 'True

Should instance search consider instances with qualified names?

_optInversionMaxDepth :: Int

_optSafe :: WithDefault 'False

_optDoubleCheck :: WithDefault 'False

_optSyntacticEquality :: !(Maybe Int)

Should the conversion checker use the syntactic equality shortcut? Nothing means that it should. Just n, for a non-negative number n, means that syntactic equality checking gets n units of fuel. If the fuel becomes zero, then syntactic equality checking is turned off. The fuel counter is decreased in the failure continuation of checkSyntacticEquality.

_optWarningMode :: WarningMode

_optCompileMain :: WithDefault 'True

Treat the module given at the command line or via interaction as main module in compilation?

_optCaching :: WithDefault 'True

_optCountClusters :: WithDefault 'False

Count extended grapheme clusters rather than code points when generating LaTeX.

_optAutoInline :: WithDefault 'False

Automatic compile-time inlining for simple definitions (unless marked NOINLINE).

_optPrintPatternSynonyms :: WithDefault 'True

_optFastReduce :: WithDefault 'True

Use the Agda abstract machine (fastReduce)?

_optCallByName :: WithDefault 'False

Use call-by-name instead of call-by-need.

_optConfluenceCheck :: Maybe ConfluenceCheck

Check confluence of rewrite rules?

_optCohesion :: WithDefault 'False

Are the cohesion modalities available?

_optFlatSplit :: WithDefault 'False

Can we split on a (@flat x : A) argument?

_optPolarity :: WithDefault 'False

Can we use modal polarities (++, +, etc.)?

_optImportSorts :: WithDefault 'True

Should every top-level module start with an implicit statement open import Agda.Primitive using (Set; Prop)?

_optLoadPrimitives :: WithDefault 'True

Should we load the primitive modules at all? This is a stronger form of optImportSorts.

_optAllowExec :: WithDefault 'False

Allow running external executables from meta programs.

_optSaveMetas :: WithDefault 'False

Save meta-variables to interface files.

_optShowIdentitySubstitutions :: WithDefault 'False

Show identity substitutions when pretty-printing terms (i.e. always show all arguments of a metavariable).

_optKeepCoveringClauses :: WithDefault 'False

Do not discard clauses constructed by the coverage checker (needed for some external backends).

_optLargeIndices :: WithDefault 'False

Allow large indices, and large forced arguments in constructors.

_optForcedArgumentRecursion :: WithDefault 'True

Allow recursion on forced constructor arguments.

_optExperimentalLazyInstances :: WithDefault 'False

Defer instance search based on discrimination tree results

data PrintAgdaVersion

Options --version and --numeric-version (last wins).

Constructors

PrintAgdaVersion
Print Agda version information.

PrintAgdaNumericVersion
Print Agda version number.

type VerboseKey = String

type VerboseKeyItem = String1

type VerboseLevel = Int

type Verbosity = Maybe (Trie VerboseKeyItem VerboseLevel)

Nothing is used if no verbosity options have been given, thus making it possible to handle the default case relatively quickly. Note that Nothing corresponds to a trie with verbosity level 1 for the empty path.