Agda.Interaction.Options

Documentation

data CommandLineOptions

Constructors

Options

Fields

optProgramName :: String

optInputFile :: Maybe FilePath

optIncludePaths :: [FilePath]

optAbsoluteIncludePaths :: [AbsolutePath]

optLibraries :: [LibName]

optOverrideLibrariesFile :: Maybe FilePath

Use this (if Just) instead of .agda/libraries

optDefaultLibs :: Bool

Use ~.agdadefaults

optUseLibs :: Bool

look for .agda-lib files

optTrustedExecutables :: Map ExeName FilePath

Map names of trusted executables to absolute paths

optPrintAgdaDir :: Bool

optPrintVersion :: Bool

optPrintHelp :: Maybe Help

optInteractive :: Bool

Agda REPL (-I).

optGHCiInteraction :: Bool

optJSONInteraction :: Bool

optOptimSmashing :: Bool

optCompileDir :: Maybe FilePath

In the absence of a path the project root is used.

optGenerateVimFile :: Bool

optIgnoreInterfaces :: Bool

optIgnoreAllInterfaces :: Bool

optLocalInterfaces :: Bool

optPragmaOptions :: PragmaOptions

optOnlyScopeChecking :: Bool

Should the top-level module only be scope-checked, and not type-checked?

data PragmaOptions

Options which can be set in a pragma.

Constructors

PragmaOptions

Fields

optShowImplicit :: Bool

optShowIrrelevant :: Bool

optUseUnicode :: UnicodeOrAscii

optVerbose :: Verbosity

optProp :: Bool

optTwoLevel :: WithDefault 'False

optAllowUnsolved :: Bool

optAllowIncompleteMatch :: Bool

optDisablePositivity :: Bool

optTerminationCheck :: Bool

optTerminationDepth :: CutOff

Cut off structural order comparison at some depth in termination checker?

optCompletenessCheck :: Bool

optUniverseCheck :: Bool

optOmegaInOmega :: Bool

optCumulativity :: Bool

optSizedTypes :: WithDefault 'False

optGuardedness :: WithDefault 'False

optInjectiveTypeConstructors :: Bool

optUniversePolymorphism :: Bool

optIrrelevantProjections :: Bool

optExperimentalIrrelevance :: Bool

irrelevant levels, irrelevant data matching

optWithoutK :: WithDefault 'False

optCopatterns :: Bool

Allow definitions by copattern matching?

optPatternMatching :: Bool

Is pattern matching allowed in the current file?

optExactSplit :: Bool

optEta :: Bool

optForcing :: Bool

Perform the forcing analysis on data constructors?

optProjectionLike :: Bool

Perform the projection-likeness analysis on functions?

optRewriting :: Bool

Can rewrite rules be added and used?

optCubical :: Maybe Cubical

optGuarded :: Bool

optFirstOrder :: Bool

Should we speculatively unify function applications as if they were injective?

optPostfixProjections :: Bool

Should system generated projections ProjSystem be printed postfix (True) or prefix (False).

optKeepPatternVariables :: Bool

Should case splitting replace variables with dot patterns (False) or keep them as variables (True).

optInstanceSearchDepth :: Int

optOverlappingInstances :: Bool

optQualifiedInstances :: Bool

Should instance search consider instances with qualified names?

optInversionMaxDepth :: Int

optSafe :: Bool

optDoubleCheck :: Bool

optSyntacticEquality :: Bool

Should conversion checker use syntactic equality shortcut?

optWarningMode :: WarningMode

optCompileNoMain :: Bool

optCaching :: Bool

optCountClusters :: Bool

Count extended grapheme clusters rather than code points when generating LaTeX.

optAutoInline :: Bool

Automatic compile-time inlining for simple definitions (unless marked NOINLINE).

optPrintPatternSynonyms :: Bool

optFastReduce :: Bool

Use the Agda abstract machine (fastReduce)?

optCallByName :: Bool

Use call-by-name instead of call-by-need

optConfluenceCheck :: Maybe ConfluenceCheck

Check confluence of rewrite rules?

optFlatSplit :: Bool

Can we split on a (@flat x : A) argument?

optImportSorts :: Bool

Should every top-level module start with an implicit statement open import Agda.Primitive using (Set; Prop)?

optAllowExec :: Bool

optShowIdentitySubstitutions :: Bool

Show identity substitutions when pretty-printing terms (i.e. always show all arguments of a metavariable)

type OptionsPragma = [String]

The options from an OPTIONS pragma.

In the future it might be nice to switch to a more structured representation. Note that, currently, there is not a one-to-one correspondence between list elements and options.

type Flag opts = opts -> OptM opts

f :: Flag opts is an action on the option record that results from parsing an option. f opts produces either an error message or an updated options record

type OptM = Except String

runOptM :: Monad m => OptM opts -> m (Either String opts)

data OptDescr a

Constructors

Option [Char] [String] (ArgDescr a) String

data ArgDescr a

Constructors

NoArg a

ReqArg (String -> a) String

OptArg (Maybe String -> a) String

type Verbosity = Trie VerboseKey VerboseLevel

type VerboseKey = String

type VerboseLevel = Int

data WarningMode

A WarningMode has two components: a set of warnings to be displayed and a flag stating whether warnings should be turned into fatal errors.

Constructors

WarningMode

Fields

_warningSet :: Set WarningName

_warn2Error :: Bool

data ConfluenceCheck

Constructors

LocalConfluenceCheck

GlobalConfluenceCheck

data UnicodeOrAscii

We want to know whether we are allowed to insert unicode characters or not.

Constructors

UnicodeOk

AsciiOnly

checkOpts :: Flag CommandLineOptions

Checks that the given options are consistent.

parsePragmaOptions

Arguments

:: [String]
Pragma options.

-> CommandLineOptions
Command-line options which should be updated.

-> OptM PragmaOptions

Parse options from an options pragma.

parsePluginOptions :: [String] -> [OptDescr (Flag opts)] -> Flag opts

Parse options for a plugin.

stripRTS :: [String] -> [String]

Removes RTS options from a list of options.

defaultOptions :: CommandLineOptions

defaultInteractionOptions :: PragmaOptions

defaultVerbosity :: Verbosity

defaultCutOff :: CutOff

The default termination depth.

defaultPragmaOptions :: PragmaOptions

standardOptions_ :: [OptDescr ()]

Used for printing usage info. Does not include the dead options.

unsafePragmaOptions :: CommandLineOptions -> PragmaOptions -> [String]

Check for unsafe pragmas. Gives a list of used unsafe flags.

restartOptions :: [(PragmaOptions -> RestartCodomain, String)]

If any these options have changed, then the file will be rechecked. Boolean options are negated to mention non-default options, where possible.

infectiveOptions :: [(PragmaOptions -> Bool, String)]

An infective option is an option that if used in one module, must be used in all modules that depend on this module.

Note that --cubical and --erased-cubical are "jointly infective": if one of them is used in one module, then one or the other must be used in all modules that depend on this module.

coinfectiveOptions :: [(PragmaOptions -> Bool, String)]

A coinfective option is an option that if used in one module, must be used in all modules that this module depends on.

safeFlag :: Flag PragmaOptions

mapFlag :: (String -> String) -> OptDescr a -> OptDescr a

Map a function over the long options. Also removes the short options. Will be used to add the plugin name to the plugin options.

usage :: [OptDescr ()] -> String -> Help -> String

The usage info message. The argument is the program name (probably agda).

inputFlag :: FilePath -> Flag CommandLineOptions

standardOptions :: [OptDescr (Flag CommandLineOptions)]

deadStandardOptions :: [OptDescr (Flag CommandLineOptions)]

Command line options of previous versions of Agda. Should not be listed in the usage info, put parsed by GetOpt for good error messaging.

getOptSimple

Arguments

:: [String]
command line argument words

-> [OptDescr (Flag opts)]
options handlers

-> (String -> Flag opts)
handler of non-options (only one is allowed)

-> Flag opts
combined opts data structure transformer

Simple interface for System.Console.GetOpt Could be moved to Agda.Utils.Options (does not exist yet)

class (Functor m, Applicative m, Monad m) => HasOptions m where

Minimal complete definition

Nothing

Methods

pragmaOptions :: m PragmaOptions

Returns the pragma options which are currently in effect.

default pragmaOptions :: (HasOptions n, MonadTrans t, m ~ t n) => m PragmaOptions

commandLineOptions :: m CommandLineOptions

Returns the command line options which are currently in effect.

default commandLineOptions :: (HasOptions n, MonadTrans t, m ~ t n) => m CommandLineOptions