Agda.Syntax.Abstract

Description

The abstract syntax. This is what you get after desugaring and scope analysis of the concrete syntax. The type checker works on abstract syntax, producing internal syntax (Agda.Syntax.Internal).

Documentation

pattern Def :: QName -> Expr

Pattern synonym for regular Def.

app :: Expr -> [NamedArg Expr] -> Expr

axiomName :: Declaration -> QName

The name defined by the given axiom.

Precondition: The declaration has to be a (scoped) Axiom.

clauseSpine :: Clause -> ClauseSpine

The clause spine corresponding to a clause.

declarationSpine :: Declaration -> DeclarationSpine

The declaration spine corresponding to a declaration.

extractPattern :: Binder' a -> Maybe (Pattern, a)

generalized :: Set QName -> Type -> Type

Smart constructor for Generalized.

insertImplicitPatSynArgs

Arguments

:: HasRange a

=> (Hiding -> Range -> a)
Thing to insert (wildcard).

-> Range
Range of the whole pattern synonym expression/pattern.

-> [WithHiding Name]
The parameters of the pattern synonym (from its definition).

-> [NamedArg a]
The arguments it is used with.

-> Maybe ([(Name, a)], [WithHiding Name])
Substitution and left-over parameters.

insertedBinder :: a -> Binder' a

insertedBinder_ :: Name -> Binder

lambdaLiftExpr :: [WithHiding Name] -> Expr -> Expr

mkBindName :: Name -> BindName

mkBinder :: a -> Binder' a

mkBinder_ :: Name -> Binder

mkDomainFree :: NamedArg Binder -> LamBinding

mkLet :: ExprInfo -> [LetBinding] -> Expr -> Expr

mkPi :: ExprInfo -> Telescope -> Type -> Type

mkTBind :: Range -> List1 (NamedArg Binder) -> Type -> TypedBinding

mkTLet :: Range -> [LetBinding] -> Maybe TypedBinding

noDataDefParams :: DataDefParams

noWhereDecls :: WhereDeclarations

renamingSize :: Ren a -> Int

Size of the range of the renaming.

rhsSpine :: RHS -> RHSSpine

The right-hand side spine corresponding to a right-hand side.

whereDeclarationsSpine :: WhereDeclarations -> WhereDeclarationsSpine

The spine corresponding to a WhereDeclarations value.

class AnyAbstract a where

Are we in an abstract block?

In that case some definition is abstract.

Minimal complete definition

Nothing

Methods

anyAbstract :: a -> Bool

default anyAbstract :: forall (t :: Type -> Type) b. (Foldable t, AnyAbstract b, t b ~ a) => a -> Bool

type Args = [NamedArg Expr]

type Assign = FieldAssignment' Expr

Record field assignment f = e.

type Assigns = [Assign]

newtype BindName

A name in a binding position: we also compare the nameConcrete when comparing the binders for equality.

With --caching on we compare abstract syntax to determine if we can reuse previous typechecking results: during that comparison two names can have the same nameId but be semantically different, e.g. in {_ : A} -> .. vs. {r : A} -> ...

Constructors

BindName

Fields

unBind :: Name

type Binder = Binder' BindName

data Binder' a

Constructors

Binder

Fields

binderPattern :: Maybe Pattern

binderNameOrigin :: BinderNameOrigin

binderName :: a

type Clause = Clause' LHS

data Clause' lhs

We could throw away where clauses at this point and translate them to let. It's not obvious how to remember that the let was really a where clause though, so for the time being we keep it here.

Constructors

Clause

Fields

clauseLHS :: lhs

clauseStrippedPats :: [ProblemEq]

Only in with-clauses where we inherit some already checked patterns from the parent. These live in the context of the parent clause left-hand side.

clauseRHS :: RHS

clauseWhereDecls :: WhereDeclarations

clauseCatchall :: Catchall

data ClauseSpine

Clause spines.

Constructors

ClauseS RHSSpine WhereDeclarationsSpine

type Constructor = TypeSignature

data DataDefParams

Constructors

DataDefParams

Fields

dataDefGeneralizedParams :: Set Name

We don't yet know the position of generalized parameters from the data sig, so we keep these in a set on the side.

dataDefParams :: [LamBinding]

data Declaration

Constructors

Axiom KindOfName DefInfo ArgInfo (Maybe PragmaPolarities) QName Type
Type signature (can be irrelevant, but not hidden).

The fourth argument contains an optional assignment of polarities to arguments.

Generalize (Set QName) DefInfo ArgInfo QName Type
The first argument is the (possibly empty) set of generalizable variables used in the type.

Field DefInfo QName (Arg Type)
record field

Primitive DefInfo QName (Arg Type)
primitive function

Mutual MutualInfo (List1 Declaration)
a bunch of mutually recursive definitions

Section Range Erased ModuleName GeneralizeTelescope [Declaration]

Apply ModuleInfo Erased ModuleName ModuleApplication ScopeCopyInfo ImportDirective
The ImportDirective is for highlighting purposes.

Import ModuleInfo ModuleName ImportDirective
The ImportDirective is for highlighting purposes.

Pragma Range Pragma

Open ModuleInfo ModuleName ImportDirective

FunDef DefInfo QName (List1 Clause)
sequence of function clauses

DataSig DefInfo Erased QName GeneralizeTelescope Type
lone data signature

DataDef DefInfo QName UniverseCheck DataDefParams [Constructor]

RecSig DefInfo Erased QName GeneralizeTelescope Type
lone record signature

RecDef DefInfo QName UniverseCheck RecordDirectives DataDefParams Type [Declaration]
The Type gives the constructor type telescope, (x1 : A1)..(xn : An) -> Dummy, and the optional name is the constructor's name. The optional Range is for the pattern attribute.

PatternSynDef QName [WithHiding BindName] (Pattern' Void)
Only for highlighting purposes

UnquoteDecl MutualInfo [DefInfo] [QName] Expr

UnquoteDef [DefInfo] [QName] Expr

UnquoteData [DefInfo] QName UniverseCheck [DefInfo] [QName] Expr

ScopedDecl ScopeInfo (List1 Declaration)
scope annotation

UnfoldingDecl Range [QName]
Only for highlighting the unfolded names

data DeclarationSpine

Declaration spines. Used in debugging to make it easy to see where constructors such as ScopedDecl and Mutual are placed.

Constructors

AxiomS

GeneralizeS

FieldS

PrimitiveS

MutualS (List1 DeclarationSpine)

SectionS [DeclarationSpine]

ApplyS

ImportS

PragmaS

OpenS

FunDefS (List1 ClauseSpine)

DataSigS

DataDefS

RecSigS

RecDefS [DeclarationSpine]

PatternSynDefS

UnquoteDeclS

UnquoteDefS

UnquoteDataS

ScopedDeclS (List1 DeclarationSpine)

UnfoldingDeclS

type DefInfo = DefInfo' Expr

data Expr

Expressions after scope checking (operators parsed, names resolved).

Constructors

Var Name
Bound variable.

Def' QName Suffix
Constant: axiom, function, data or record type, with a possible suffix.

Proj ProjOrigin AmbiguousQName
Projection (overloaded).

Con AmbiguousQName
Constructor (overloaded).

PatternSyn AmbiguousQName
Pattern synonym.

Macro QName
Macro.

Lit ExprInfo Literal
Literal.

QuestionMark MetaInfo InteractionId
Meta variable for interaction. The InteractionId is usually identical with the metaNumber of MetaInfo. However, if you want to print an interaction meta as just ? instead of ?n, you should set the metaNumber to Nothing while keeping the InteractionId.

Underscore MetaInfo
Meta variable for hidden argument (must be inferred locally).

Dot ExprInfo Expr
.e, for postfix projection.

App AppInfo Expr (NamedArg Expr)
Ordinary (binary) application.

WithApp ExprInfo Expr (List1 Expr)
With application.

Lam ExprInfo LamBinding Expr
λ bs → e.

AbsurdLam ExprInfo Hiding
λ() or λ{}.

ExtendedLam ExprInfo DefInfo Erased QName (List1 Clause)

Pi ExprInfo Telescope1 Type
Dependent function space Γ → A.

Generalized (Set1 QName) Type
Like a Pi, but the ordering is not known

Fun ExprInfo (Arg Type) Type
Non-dependent function space.

Let ExprInfo (List1 LetBinding) Expr
let bs in e.

Rec KwRange ExprInfo RecordAssigns
Record construction. The KwRange is for the record kewyword.

RecUpdate KwRange ExprInfo Expr Assigns
Record update. The KwRange is for the record kewyword.

RecWhere KwRange ExprInfo [LetBinding] Assigns
record where expression, the set of names is the set of names that should become record fields. The KwRange is for the record keyword.

RecUpdateWhere KwRange ExprInfo Expr [LetBinding] Assigns
record where expression, the set of names is the set of names that should become record fields. The KwRange is for the record keyword.

Quote ExprInfo
Quote an identifier QName.

QuoteTerm ExprInfo
Quote a term.

Unquote ExprInfo
The splicing construct: unquote ...

ScopedExpr ScopeInfo Expr
Scope annotation.

Qualified ModuleName Expr
For remembering the module name of a qualified do block.

Highlighted Aspects Expr
For printing annotated terms produced by the type checker.

DontCare Expr
For printing DontCare from Agda.Syntax.Internal.

type Field = TypeSignature

data GeneralizeTelescope

Constructors

GeneralizeTel

Fields

generalizeTelVars :: Map QName Name

Maps generalize variables to the corresponding bound variable (to be introduced by the generalisation).

generalizeTel :: Telescope

type HoleContent = HoleContent' () BindName Pattern Expr

type ImportDirective = ImportDirective' QName ModuleName

type ImportedName = ImportedName' QName ModuleName

data LHS

The lhs of a clause in focused (projection-application) view (outside-in). Projection patters are represented as LHSProjs.

Constructors

LHS

Fields

lhsInfo :: LHSInfo

Range.

lhsCore :: LHSCore

Copatterns.

type LHSCore = LHSCore' Expr

data LHSCore' e

The lhs in projection-application and with-pattern view. Parameterised over the type e of dot patterns.

Constructors

LHSHead
The head applied to ordinary patterns.

Fields

lhsDefName :: QName

Head f.

lhsPats :: [NamedArg (Pattern' e)]

Applied to patterns ps.

LHSProj
Projection.

Fields

lhsDestructor :: AmbiguousQName

Record projection identifier.

lhsFocus :: NamedArg (LHSCore' e)

Main argument of projection.

lhsPats :: [NamedArg (Pattern' e)]

Applied to patterns ps.

LHSWith
With patterns.

Fields

lhsHead :: LHSCore' e

E.g. the LHSHead.

lhsWithPatterns :: List1 (Arg (Pattern' e))

Applied to with patterns | p1 | ... | pn. These patterns are not prefixed with WithP!

lhsPats :: [NamedArg (Pattern' e)]

Applied to patterns ps.

data LamBinding

A lambda binding is either domain free or typed.

Constructors

DomainFree TacticAttribute (NamedArg Binder)
. x or {x} or .x or {x = y} or x@p or (p)

DomainFull TypedBinding
. (xs:e) or {xs:e} or (let Ds)

data LetBinding

Bindings that are valid in a let.

Constructors

LetBind LetInfo ArgInfo BindName Type Expr
LetBind info rel name type defn

LetAxiom LetInfo ArgInfo BindName Type
Function declarations in a let with no matching body.

LetPatBind LetInfo ArgInfo Pattern Expr
Irrefutable pattern binding.

LetApply ModuleInfo Erased ModuleName ModuleApplication ScopeCopyInfo ImportDirective
LetApply mi newM (oldM args) renamings dir. The ImportDirective is for highlighting purposes.

LetOpen ModuleInfo ModuleName ImportDirective
only for highlighting and abstractToConcrete

data ModuleApplication

Constructors

SectionApp Telescope ModuleName [NamedArg Expr]
tel. M args: applies M to args and abstracts tel.

RecordModuleInstance ModuleName
M {{...}}

type NAPs e = [NamedArg (Pattern' e)]

type NAPs1 e = List1 (NamedArg (Pattern' e))

class NameToExpr a where

Turn a name into an expression.

Methods

nameToExpr :: a -> Expr

type Pattern = Pattern' Expr

data Pattern' e

Parameterised over the type of dot patterns.

Constructors

VarP BindName

ConP ConPatInfo AmbiguousQName (NAPs e)

ProjP PatInfo ProjOrigin AmbiguousQName
Destructor pattern d.

DefP PatInfo AmbiguousQName (NAPs e)
Defined pattern: function definition f ps. It is also abused to convert destructor patterns into concrete syntax thus, we put AmbiguousQName here as well.

WildP PatInfo
Underscore pattern entered by user. Or generated at type checking for implicit arguments.

AsP PatInfo BindName (Pattern' e)

DotP PatInfo e
Dot pattern .e

AbsurdP PatInfo

LitP PatInfo Literal

PatternSynP PatInfo AmbiguousQName (NAPs e)

RecP KwRange ConPatInfo [FieldAssignment' (Pattern' e)]

EqualP PatInfo (List1 (e, e))

WithP PatInfo (Pattern' e)
| p, for with-patterns.

type PatternSynDefn = ([WithHiding Name], Pattern' Void)

type PatternSynDefns = Map QName PatternSynDefn

type Patterns = [NamedArg Pattern]

data Pragma

Constructors

OptionsPragma [String]

BuiltinPragma RString ResolvedName
ResolvedName is not UnknownName. Name can be ambiguous e.g. for built-in constructors.

BuiltinNoDefPragma RString KindOfName QName
Builtins that do not come with a definition, but declare a name for an Agda concept.

RewritePragma Range [QName]
Range is range of REWRITE keyword.

CompilePragma (Ranged BackendName) QName String

StaticPragma QName

EtaPragma QName
For coinductive records, use pragma instead of regular eta-equality definition (as it is might make Agda loop).

InjectivePragma QName

InjectiveForInferencePragma QName

InlinePragma Bool QName

NotProjectionLikePragma QName
Mark the definition as not being projection-like

OverlapPragma QName OverlapMode
If the definition is an instance, set its overlap mode.

DisplayPragma QName [NamedArg Pattern] Expr

data ProblemEq

A user pattern together with an internal term that it should be equal to after splitting is complete. Special cases: * User pattern is a variable but internal term isn't: this will be turned into an as pattern. * User pattern is a dot pattern: this pattern won't trigger any splitting but will be checked for equality after all splitting is complete and as patterns have been bound. * User pattern is an absurd pattern: emptiness of the type will be checked after splitting is complete. * User pattern is an annotated wildcard: type annotation will be checked after splitting is complete.

Constructors

ProblemEq

Fields

problemInPat :: Pattern

problemInst :: Term

problemType :: Dom Type

data RHS

Constructors

RHS

Fields

rhsExpr :: Expr

rhsConcrete :: Maybe Expr

We store the original concrete expression in case we have to reproduce it during interactive case splitting. Nothing for internally generated rhss.

AbsurdRHS

WithRHS QName (List1 WithExpr) (List1 Clause)
The QName is the name of the with function.

RewriteRHS

Fields

rewriteExprs :: [RewriteEqn]

The QNames are the names of the generated with functions, one for each Expr.

rewriteStrippedPats :: [ProblemEq]

The patterns stripped by with-desugaring. These are only present if this rewrite follows a with.

rewriteRHS :: RHS

The RHS should not be another RewriteRHS.

rewriteWhereDecls :: WhereDeclarations

The where clauses are attached to the RewriteRHS by

data RHSSpine

Right-hand side spines.

Constructors

RHSS

AbsurdRHSS

WithRHSS (List1 ClauseSpine)

RewriteRHSS RHSSpine WhereDeclarationsSpine

type RecordAssign = Either Assign ModuleName

type RecordAssigns = [RecordAssign]

data RecordConName

How did we get our hands on the QName for the constructor of this record?

Constructors

NamedRecCon
The user wrote it.

Fields

recordConName :: !QName

FreshRecCon
We made it up.

Fields

recordConName :: !QName

type RecordDirectives = RecordDirectives' RecordConName

type Ren a = Map a (List1 a)

Renaming (generic).

type Renaming = Renaming' QName ModuleName

type RewriteEqn = RewriteEqn' QName BindName Pattern Expr

data ScopeCopyInfo

Information created by the scope checker necessary for type-checking a module copy.

Constructors

ScopeCopyInfo

Fields

renModules :: Ren ModuleName

Associates to each (original) module name the list of copies that should be created.

renNames :: Ren QName

Same as for renModules, but for definitions.

type SpineClause = Clause' SpineLHS

data SpineLHS

The lhs of a clause in spine view (inside-out). Projection patterns are contained in spLhsPats, represented as ProjP d.

Constructors

SpineLHS

Fields

spLhsInfo :: LHSInfo

Range.

spLhsDefName :: QName

Name of function we are defining.

spLhsPats :: [NamedArg Pattern]

Elimination by pattern, projections, with-patterns.

class SubstExpr a where

Minimal complete definition

Nothing

Methods

substExpr :: [(Name, Expr)] -> a -> a

default substExpr :: forall (t :: Type -> Type) b. (Functor t, SubstExpr b, t b ~ a) => [(Name, Expr)] -> a -> a

type TacticAttribute = TacticAttribute' Expr

type Telescope = [TypedBinding]

type Telescope1 = List1 TypedBinding

type Type = Expr

Types are just expressions. Use this type synonym for hinting that an expression should be a type.

type TypeSignature = Declaration

Only Axioms.

data TypedBinding

A typed binding. Appears in dependent function spaces, typed lambdas, and telescopes. It might be tempting to simplify this to only bind a single name at a time, and translate, say, (x y : A) to (x : A)(y : A) before type-checking. However, this would be slightly problematic:

We would have to typecheck the type A several times.

If A contains a meta variable or hole, it would be duplicated by such a translation.

While 1. is only slightly inefficient, 2. would be an outright bug. Duplicating A could not be done naively, we would have to make sure that the metas of the copy are aliases of the metas of the original.

Constructors

TBind Range TypedBindingInfo (List1 (NamedArg Binder)) Type
As in telescope (x y z : A) or type (x y z : A) -> B.

TLet Range (List1 LetBinding)
E.g. (let x = e) or (let open M).

data TypedBindingInfo

Extra information that is attached to a typed binding, that plays a role during type checking but strictly speaking is not part of the name : type" relation which a makes up a binding.

Constructors

TypedBindingInfo

Fields

tbTacticAttr :: TacticAttribute

Does this binding have a tactic annotation?

tbFinite :: Bool

Does this binding correspond to a Partial binder, rather than to a Pi binder? Must be present here to be reflected into abstract syntax later (and to be printed to the user later).

data WhereDeclarations

Constructors

WhereDecls

Fields

whereModule :: Maybe ModuleName

whereAnywhere :: Bool

is it an ordinary unnamed where?

whereDecls :: Maybe Declaration

The declaration is a Section.

data WhereDeclarationsSpine

Spines corresponding to WhereDeclarations values.

Constructors

WhereDeclsS (Maybe DeclarationSpine)

type WithExpr = WithExpr' Expr

type WithExpr' e = Named BindName (Arg e)

module Agda.Syntax.Abstract.Name

Orphan instances

KillRange Suffix

Instance details

Methods

killRange :: KillRangeT Suffix