Agda.Syntax.Abstract.Name

Description

Abstract names carry unique identifiers and stuff.

Documentation

commonParentModule :: ModuleName -> ModuleName -> ModuleName

getUnambiguous :: AmbiguousQName -> Maybe QName

Get the name if unambiguous.

headAmbQ :: AmbiguousQName -> QName

Get the first of the ambiguous names.

isAmbiguous :: AmbiguousQName -> Bool

Is a name ambiguous.

isAnonymousModuleName :: ModuleName -> Bool

A module is anonymous if the qualification path ends in an underscore.

isInModule :: QName -> ModuleName -> Bool

isLeChildModuleOf :: ModuleName -> ModuleName -> Bool

Is the first module a weak child of the second?

isLeParentModuleOf :: ModuleName -> ModuleName -> Bool

Is the first module a weak parent of the second?

isLtChildModuleOf :: ModuleName -> ModuleName -> Bool

Is the first module a proper child of the second?

isLtParentModuleOf :: ModuleName -> ModuleName -> Bool

Is the first module a proper parent of the second?

isOperator :: QName -> Bool

Is the name an operator?

lensQNameName :: Lens' QName Name

makeName :: NameId -> Name -> Range -> Fixity' -> Bool -> Name

mnameFromList :: [Name] -> ModuleName

mnameFromList1 :: List1 Name -> ModuleName

mnameToConcrete :: ModuleName -> QName

mnameToList1 :: ModuleName -> List1 Name

mnameToQName :: ModuleName -> QName

nameToArgName :: Name -> ArgName

namedArgName :: NamedArg Name -> ArgName

nextName :: FreshNameMode -> Name -> Name

Get the next version of the concrete name. For instance, nextName "x" = "x₁". The name must not be a NoName.

noModuleName :: ModuleName

qnameFromList :: List1 Name -> QName

qnameToConcrete :: QName -> QName

Turn a qualified name into a concrete name. This should only be used as a fallback when looking up the right concrete name in the scope fails.

qnameToList :: QName -> List1 Name

qnameToList0 :: QName -> [Name]

qnameToMName :: QName -> ModuleName

qualify :: ModuleName -> Name -> QName

qualifyM :: ModuleName -> ModuleName -> ModuleName

qualifyQ :: ModuleName -> QName -> QName

qualify_ :: Name -> QName

Convert a Name to a QName (add no module name).

sameRoot :: Name -> Name -> Bool

showQNameId :: QName -> String

uglyShowName :: Name -> String

Useful for debugging scoping problems

unambiguous :: QName -> AmbiguousQName

A singleton "ambiguous" name.

withRangesOf :: ModuleName -> List1 Name -> ModuleName

Sets the ranges of the individual names in the module name to match those of the corresponding concrete names. If the concrete names are fewer than the number of module name name parts, then the initial name parts get the range noRange.

C.D.E `withRangesOf` [A, B] returns C.D.E but with ranges set as follows:

C: noRange.

D: the range of A.

E: the range of B.

Precondition: The number of module name name parts has to be at least as large as the length of the list.

withRangesOfQ :: ModuleName -> QName -> ModuleName

Like withRangesOf, but uses the name parts (qualifier + name) of the qualified name as the list of concrete names.

newtype AmbiguousQName

Ambiguous qualified names. Used for overloaded constructors.

Invariant: All the names in the list must have the same concrete, unqualified name. (This implies that they all have the same Range).

Constructors

AmbQ

Fields

unAmbQ :: List1 QName

class IsProjP a where

Check whether we are a projection pattern.

Methods

isProjP :: a -> Maybe (ProjOrigin, AmbiguousQName)

class MkName a where

Make a Name from some kind of string.

Minimal complete definition

mkName

Methods

mkName :: Range -> NameId -> a -> Name

The Range sets the definition site of the name, not the use site.

mkName_ :: NameId -> a -> Name

newtype ModuleName

A module name is just a qualified name.

The SetRange instance for module names sets all individual ranges to the given one.

Constructors

MName

Fields

mnameToList :: [Name]

data Name

A name is a unique identifier and a suggestion for a concrete name. The concrete name contains the source location (if any) of the name. The source location of the binding site is also recorded.

Constructors

Name

Fields

nameId :: !NameId

nameConcrete :: Name

The concrete name used for this instance

nameCanonical :: Name

The concrete name in the original definition (needed by primShowQName, see #4735)

nameBindingSite :: Range

nameFixity :: Fixity'

nameIsRecordName :: Bool

Is this the name of the invisible record variable self? Should not be printed or displayed in the context, see issue #3584.

data QName

Qualified names are non-empty lists of names. Equality on qualified names are just equality on the last name, i.e. the module part is just for show.

The SetRange instance for qualified names sets all individual ranges (including those of the module prefix) to the given one.

Constructors

QName

Fields

qnameModule :: ModuleName

qnameName :: Name

data QNamed a

Something preceeded by a qualified name.

Constructors

QNamed

Fields

qname :: QName

qnamed :: a

data Suffix

A name suffix

Constructors

NoSuffix

Suffix !Integer

class IsNoName a where

Check whether a name is the empty name "_".

Minimal complete definition

Nothing

Methods

isNoName :: a -> Bool

default isNoName :: forall (t :: Type -> Type) b. (Foldable t, IsNoName b, t b ~ a) => a -> Bool

data FreshNameMode

Method by which to generate fresh unshadowed names.

Constructors

UnicodeSubscript
Append an integer Unicode subscript: x, x₁, x₂, …

AsciiCounter
Append an integer ASCII counter: x, x1, x2, …