Agda.Syntax.Builtin

Description

This module defines the names of all builtin and primitives used in Agda.

See Agda.TypeChecking.Monad.Builtin

Documentation

data SomeBuiltin

Either a BuiltinId or PrimitiveId, used for some lookups.

Constructors

BuiltinName !BuiltinId

PrimitiveName !PrimitiveId

class IsBuiltin a where

The class of types which can be converted to SomeBuiltin.

Methods

someBuiltin :: a -> SomeBuiltin

Convert this value to a builtin.

getBuiltinId :: a -> String

Get the identifier for this builtin, generally used for error messages.

Builtins

data BuiltinId

A builtin name, defined by the BUILTIN pragma.

Constructors

BuiltinNat

BuiltinSuc

BuiltinZero

BuiltinNatPlus

BuiltinNatMinus

BuiltinNatTimes

BuiltinNatDivSucAux

BuiltinNatModSucAux

BuiltinNatEquals

BuiltinNatLess

BuiltinWord64

BuiltinInteger

BuiltinIntegerPos

BuiltinIntegerNegSuc

BuiltinFloat

BuiltinChar

BuiltinString

BuiltinUnit

BuiltinUnitUnit

BuiltinSigma

BuiltinSigmaCon

BuiltinBool

BuiltinTrue

BuiltinFalse

BuiltinList

BuiltinNil

BuiltinCons

BuiltinMaybe

BuiltinNothing

BuiltinJust

BuiltinIO

BuiltinPath

BuiltinPathP

BuiltinIntervalUniv

BuiltinInterval

BuiltinIZero

BuiltinIOne

BuiltinPartial

BuiltinPartialP

BuiltinIsOne

BuiltinItIsOne

BuiltinEquiv

BuiltinEquivFun

BuiltinEquivProof

BuiltinTranspProof

BuiltinIsOne1

BuiltinIsOne2

BuiltinIsOneEmpty

BuiltinSub

BuiltinSubIn

BuiltinSizeUniv

BuiltinSize

BuiltinSizeLt

BuiltinSizeSuc

BuiltinSizeInf

BuiltinSizeMax

BuiltinInf

BuiltinSharp

BuiltinFlat

BuiltinEquality

BuiltinRefl

BuiltinRewrite

BuiltinLevelMax

BuiltinLevel

BuiltinLevelZero

BuiltinLevelSuc

BuiltinProp

BuiltinSet

BuiltinStrictSet

BuiltinPropOmega

BuiltinSetOmega

BuiltinSSetOmega

BuiltinLevelUniv

BuiltinFromNat

BuiltinFromNeg

BuiltinFromString

BuiltinQName

BuiltinAgdaSort

BuiltinAgdaSortSet

BuiltinAgdaSortLit

BuiltinAgdaSortProp

BuiltinAgdaSortPropLit

BuiltinAgdaSortInf

BuiltinAgdaSortUnsupported

BuiltinHiding

BuiltinHidden

BuiltinInstance

BuiltinVisible

BuiltinRelevance

BuiltinRelevant

BuiltinIrrelevant

BuiltinQuantity

BuiltinQuantity0

BuiltinQuantityω

BuiltinModality

BuiltinModalityConstructor

BuiltinAssoc

BuiltinAssocLeft

BuiltinAssocRight

BuiltinAssocNon

BuiltinPrecedence

BuiltinPrecRelated

BuiltinPrecUnrelated

BuiltinFixity

BuiltinFixityFixity

BuiltinArg

BuiltinArgInfo

BuiltinArgArgInfo

BuiltinArgArg

BuiltinAbs

BuiltinAbsAbs

BuiltinAgdaTerm

BuiltinAgdaTermVar

BuiltinAgdaTermLam

BuiltinAgdaTermExtLam

BuiltinAgdaTermDef

BuiltinAgdaTermCon

BuiltinAgdaTermPi

BuiltinAgdaTermSort

BuiltinAgdaTermLit

BuiltinAgdaTermUnsupported

BuiltinAgdaTermMeta

BuiltinAgdaErrorPart

BuiltinAgdaErrorPartString

BuiltinAgdaErrorPartTerm

BuiltinAgdaErrorPartPatt

BuiltinAgdaErrorPartName

BuiltinAgdaLiteral

BuiltinAgdaLitNat

BuiltinAgdaLitWord64

BuiltinAgdaLitFloat

BuiltinAgdaLitChar

BuiltinAgdaLitString

BuiltinAgdaLitQName

BuiltinAgdaLitMeta

BuiltinAgdaClause

BuiltinAgdaClauseClause

BuiltinAgdaClauseAbsurd

BuiltinAgdaPattern

BuiltinAgdaPatVar

BuiltinAgdaPatCon

BuiltinAgdaPatDot

BuiltinAgdaPatLit

BuiltinAgdaPatProj

BuiltinAgdaPatAbsurd

BuiltinAgdaDefinitionFunDef

BuiltinAgdaDefinitionDataDef

BuiltinAgdaDefinitionRecordDef

BuiltinAgdaDefinitionDataConstructor

BuiltinAgdaDefinitionPostulate

BuiltinAgdaDefinitionPrimitive

BuiltinAgdaDefinition

BuiltinAgdaMeta

BuiltinAgdaTCM

BuiltinAgdaTCMReturn

BuiltinAgdaTCMBind

BuiltinAgdaTCMUnify

BuiltinAgdaTCMTypeError

BuiltinAgdaTCMInferType

BuiltinAgdaTCMCheckType

BuiltinAgdaTCMNormalise

BuiltinAgdaTCMReduce

BuiltinAgdaTCMCatchError

BuiltinAgdaTCMGetContext

BuiltinAgdaTCMExtendContext

BuiltinAgdaTCMInContext

BuiltinAgdaTCMFreshName

BuiltinAgdaTCMDeclareDef

BuiltinAgdaTCMDeclarePostulate

BuiltinAgdaTCMDeclareData

BuiltinAgdaTCMDefineData

BuiltinAgdaTCMDefineFun

BuiltinAgdaTCMGetType

BuiltinAgdaTCMGetDefinition

BuiltinAgdaTCMBlock

BuiltinAgdaTCMCommit

BuiltinAgdaTCMQuoteTerm

BuiltinAgdaTCMUnquoteTerm

BuiltinAgdaTCMQuoteOmegaTerm

BuiltinAgdaTCMIsMacro

BuiltinAgdaTCMWithNormalisation

BuiltinAgdaTCMWithReconstructed

BuiltinAgdaTCMWithExpandLast

BuiltinAgdaTCMWithReduceDefs

BuiltinAgdaTCMAskNormalisation

BuiltinAgdaTCMAskReconstructed

BuiltinAgdaTCMAskExpandLast

BuiltinAgdaTCMAskReduceDefs

BuiltinAgdaTCMFormatErrorParts

BuiltinAgdaTCMDebugPrint

BuiltinAgdaTCMNoConstraints

BuiltinAgdaTCMWorkOnTypes

BuiltinAgdaTCMRunSpeculative

BuiltinAgdaTCMExec

BuiltinAgdaTCMCheckFromString

BuiltinAgdaTCMGetInstances

BuiltinAgdaTCMSolveInstances

BuiltinAgdaTCMPragmaForeign

BuiltinAgdaTCMPragmaCompile

BuiltinAgdaBlocker

BuiltinAgdaBlockerAny

BuiltinAgdaBlockerAll

BuiltinAgdaBlockerMeta

isBuiltinNoDef :: BuiltinId -> Bool

Builtins that come without a definition in Agda syntax. These are giving names to Agda internal concepts which cannot be assigned an Agda type.

An example would be a user-defined name for Set.

{-# BUILTIN TYPE Type #-}

The type of Type would be Type : Level → Setω which is not valid Agda.

builtinsNoDef :: [BuiltinId]

sizeBuiltins :: [BuiltinId]

builtinNat :: BuiltinId

builtinSuc :: BuiltinId

builtinZero :: BuiltinId

builtinNatPlus :: BuiltinId

builtinNatMinus :: BuiltinId

builtinNatTimes :: BuiltinId

builtinNatDivSucAux :: BuiltinId

builtinNatModSucAux :: BuiltinId

builtinNatEquals :: BuiltinId

builtinNatLess :: BuiltinId

builtinWord64 :: BuiltinId

builtinInteger :: BuiltinId

builtinIntegerPos :: BuiltinId

builtinIntegerNegSuc :: BuiltinId

builtinFloat :: BuiltinId

builtinChar :: BuiltinId

builtinString :: BuiltinId

builtinUnit :: BuiltinId

builtinUnitUnit :: BuiltinId

builtinSigma :: BuiltinId

builtinBool :: BuiltinId

builtinTrue :: BuiltinId

builtinFalse :: BuiltinId

builtinList :: BuiltinId

builtinNil :: BuiltinId

builtinCons :: BuiltinId

builtinMaybe :: BuiltinId

builtinNothing :: BuiltinId

builtinJust :: BuiltinId

builtinIO :: BuiltinId

builtinPath :: BuiltinId

builtinPathP :: BuiltinId

builtinIntervalUniv :: BuiltinId

builtinInterval :: BuiltinId

builtinIZero :: BuiltinId

builtinIOne :: BuiltinId

builtinPartial :: BuiltinId

builtinPartialP :: BuiltinId

builtinIsOne :: BuiltinId

builtinItIsOne :: BuiltinId

builtinEquiv :: BuiltinId

builtinEquivFun :: BuiltinId

builtinEquivProof :: BuiltinId

builtinTranspProof :: BuiltinId

builtinIsOne1 :: BuiltinId

builtinIsOne2 :: BuiltinId

builtinIsOneEmpty :: BuiltinId

builtinSub :: BuiltinId

builtinSubIn :: BuiltinId

builtinSizeUniv :: BuiltinId

builtinSize :: BuiltinId

builtinSizeLt :: BuiltinId

builtinSizeSuc :: BuiltinId

builtinSizeInf :: BuiltinId

builtinSizeMax :: BuiltinId

builtinInf :: BuiltinId

builtinSharp :: BuiltinId

builtinFlat :: BuiltinId

builtinEquality :: BuiltinId

builtinRefl :: BuiltinId

builtinRewrite :: BuiltinId

builtinLevelMax :: BuiltinId

builtinLevel :: BuiltinId

builtinLevelZero :: BuiltinId

builtinLevelSuc :: BuiltinId

builtinProp :: BuiltinId

builtinSet :: BuiltinId

builtinStrictSet :: BuiltinId

builtinPropOmega :: BuiltinId

builtinSetOmega :: BuiltinId

builtinSSetOmega :: BuiltinId

builtinLevelUniv :: BuiltinId

builtinFromNat :: BuiltinId

builtinFromNeg :: BuiltinId

builtinFromString :: BuiltinId

builtinQName :: BuiltinId

builtinAgdaSort :: BuiltinId

builtinAgdaSortSet :: BuiltinId

builtinAgdaSortLit :: BuiltinId

builtinAgdaSortProp :: BuiltinId

builtinAgdaSortPropLit :: BuiltinId

builtinAgdaSortInf :: BuiltinId

builtinAgdaSortUnsupported :: BuiltinId

builtinHiding :: BuiltinId

builtinHidden :: BuiltinId

builtinInstance :: BuiltinId

builtinVisible :: BuiltinId

builtinRelevance :: BuiltinId

builtinRelevant :: BuiltinId

builtinIrrelevant :: BuiltinId

builtinQuantity :: BuiltinId

builtinQuantity0 :: BuiltinId

builtinQuantityω :: BuiltinId

builtinModality :: BuiltinId

builtinModalityConstructor :: BuiltinId

builtinAssoc :: BuiltinId

builtinAssocLeft :: BuiltinId

builtinAssocRight :: BuiltinId

builtinAssocNon :: BuiltinId

builtinPrecedence :: BuiltinId

builtinPrecRelated :: BuiltinId

builtinPrecUnrelated :: BuiltinId

builtinFixity :: BuiltinId

builtinFixityFixity :: BuiltinId

builtinArg :: BuiltinId

builtinArgInfo :: BuiltinId

builtinArgArgInfo :: BuiltinId

builtinArgArg :: BuiltinId

builtinAbs :: BuiltinId

builtinAbsAbs :: BuiltinId

builtinAgdaTerm :: BuiltinId

builtinAgdaTermVar :: BuiltinId

builtinAgdaTermLam :: BuiltinId

builtinAgdaTermExtLam :: BuiltinId

builtinAgdaTermDef :: BuiltinId

builtinAgdaTermCon :: BuiltinId

builtinAgdaTermPi :: BuiltinId

builtinAgdaTermSort :: BuiltinId

builtinAgdaTermLit :: BuiltinId

builtinAgdaTermUnsupported :: BuiltinId

builtinAgdaTermMeta :: BuiltinId

builtinAgdaErrorPart :: BuiltinId

builtinAgdaErrorPartString :: BuiltinId

builtinAgdaErrorPartTerm :: BuiltinId

builtinAgdaErrorPartPatt :: BuiltinId

builtinAgdaErrorPartName :: BuiltinId

builtinAgdaLiteral :: BuiltinId

builtinAgdaLitNat :: BuiltinId

builtinAgdaLitWord64 :: BuiltinId

builtinAgdaLitFloat :: BuiltinId

builtinAgdaLitChar :: BuiltinId

builtinAgdaLitString :: BuiltinId

builtinAgdaLitQName :: BuiltinId

builtinAgdaLitMeta :: BuiltinId

builtinAgdaClause :: BuiltinId

builtinAgdaClauseClause :: BuiltinId

builtinAgdaClauseAbsurd :: BuiltinId

builtinAgdaPattern :: BuiltinId

builtinAgdaPatVar :: BuiltinId

builtinAgdaPatCon :: BuiltinId

builtinAgdaPatDot :: BuiltinId

builtinAgdaPatLit :: BuiltinId

builtinAgdaPatProj :: BuiltinId

builtinAgdaPatAbsurd :: BuiltinId

builtinAgdaDefinitionFunDef :: BuiltinId

builtinAgdaDefinitionDataDef :: BuiltinId

builtinAgdaDefinitionRecordDef :: BuiltinId

builtinAgdaDefinitionDataConstructor :: BuiltinId

builtinAgdaDefinitionPostulate :: BuiltinId

builtinAgdaDefinitionPrimitive :: BuiltinId

builtinAgdaDefinition :: BuiltinId

builtinAgdaMeta :: BuiltinId

builtinAgdaTCM :: BuiltinId

builtinAgdaTCMReturn :: BuiltinId

builtinAgdaTCMBind :: BuiltinId

builtinAgdaTCMUnify :: BuiltinId

builtinAgdaTCMTypeError :: BuiltinId

builtinAgdaTCMInferType :: BuiltinId

builtinAgdaTCMCheckType :: BuiltinId

builtinAgdaTCMNormalise :: BuiltinId

builtinAgdaTCMReduce :: BuiltinId

builtinAgdaTCMCatchError :: BuiltinId

builtinAgdaTCMGetContext :: BuiltinId

builtinAgdaTCMExtendContext :: BuiltinId

builtinAgdaTCMInContext :: BuiltinId

builtinAgdaTCMFreshName :: BuiltinId

builtinAgdaTCMDeclareDef :: BuiltinId

builtinAgdaTCMDeclarePostulate :: BuiltinId

builtinAgdaTCMDeclareData :: BuiltinId

builtinAgdaTCMDefineData :: BuiltinId

builtinAgdaTCMDefineFun :: BuiltinId

builtinAgdaTCMGetType :: BuiltinId

builtinAgdaTCMGetDefinition :: BuiltinId

builtinAgdaTCMBlock :: BuiltinId

builtinAgdaTCMCommit :: BuiltinId

builtinAgdaTCMQuoteTerm :: BuiltinId

builtinAgdaTCMUnquoteTerm :: BuiltinId

builtinAgdaTCMQuoteOmegaTerm :: BuiltinId

builtinAgdaTCMIsMacro :: BuiltinId

builtinAgdaTCMWithNormalisation :: BuiltinId

builtinAgdaTCMWithReconstructed :: BuiltinId

builtinAgdaTCMWithExpandLast :: BuiltinId

builtinAgdaTCMWithReduceDefs :: BuiltinId

builtinAgdaTCMAskNormalisation :: BuiltinId

builtinAgdaTCMAskReconstructed :: BuiltinId

builtinAgdaTCMAskExpandLast :: BuiltinId

builtinAgdaTCMAskReduceDefs :: BuiltinId

builtinAgdaTCMFormatErrorParts :: BuiltinId

builtinAgdaTCMDebugPrint :: BuiltinId

builtinAgdaTCMNoConstraints :: BuiltinId

builtinAgdaTCMWorkOnTypes :: BuiltinId

builtinAgdaTCMRunSpeculative :: BuiltinId

builtinAgdaTCMExec :: BuiltinId

builtinAgdaTCMCheckFromString :: BuiltinId

builtinAgdaTCMGetInstances :: BuiltinId

builtinAgdaTCMSolveInstances :: BuiltinId

builtinAgdaTCMPragmaForeign :: BuiltinId

builtinAgdaTCMPragmaCompile :: BuiltinId

builtinAgdaBlocker :: BuiltinId

builtinAgdaBlockerAny :: BuiltinId

builtinAgdaBlockerAll :: BuiltinId

builtinAgdaBlockerMeta :: BuiltinId

builtinById :: String -> Maybe BuiltinId

Lookup a builtin by the string used in the BUILTIN pragma.

Primitives

data PrimitiveId

A primitive name, defined by the primitive block.

Constructors

PrimIMin

PrimIMax

PrimINeg

PrimPartial

PrimPartialP

PrimSubOut

PrimGlue

Prim_glue

Prim_unglue

Prim_glueU

Prim_unglueU

PrimFaceForall

PrimComp

PrimPOr

PrimTrans

PrimHComp

PrimShowInteger

PrimNatPlus

PrimNatMinus

PrimNatTimes

PrimNatDivSucAux

PrimNatModSucAux

PrimNatEquality

PrimNatLess

PrimShowNat

PrimWord64FromNat

PrimWord64ToNat

PrimWord64ToNatInjective

PrimLevelZero

PrimLevelSuc

PrimLevelMax

PrimFloatEquality

PrimFloatInequality

PrimFloatLess

PrimFloatIsInfinite

PrimFloatIsNaN

PrimFloatIsNegativeZero

PrimFloatIsSafeInteger

PrimFloatToWord64

PrimFloatToWord64Injective

PrimNatToFloat

PrimIntToFloat

PrimFloatRound

PrimFloatFloor

PrimFloatCeiling

PrimFloatToRatio

PrimRatioToFloat

PrimFloatDecode

PrimFloatEncode

PrimShowFloat

PrimFloatPlus

PrimFloatMinus

PrimFloatTimes

PrimFloatNegate

PrimFloatDiv

PrimFloatPow

PrimFloatSqrt

PrimFloatExp

PrimFloatLog

PrimFloatSin

PrimFloatCos

PrimFloatTan

PrimFloatASin

PrimFloatACos

PrimFloatATan

PrimFloatATan2

PrimFloatSinh

PrimFloatCosh

PrimFloatTanh

PrimFloatASinh

PrimFloatACosh

PrimFloatATanh

PrimCharEquality

PrimIsLower

PrimIsDigit

PrimIsAlpha

PrimIsSpace

PrimIsAscii

PrimIsLatin1

PrimIsPrint

PrimIsHexDigit

PrimToUpper

PrimToLower

PrimCharToNat

PrimCharToNatInjective

PrimNatToChar

PrimShowChar

PrimStringToList

PrimStringToListInjective

PrimStringFromList

PrimStringFromListInjective

PrimStringAppend

PrimStringEquality

PrimShowString

PrimStringUncons

PrimErase

PrimEraseEquality

PrimForce

PrimForceLemma

PrimQNameEquality

PrimQNameLess

PrimShowQName

PrimQNameFixity

PrimQNameToWord64s

PrimQNameToWord64sInjective

PrimMetaEquality

PrimMetaLess

PrimShowMeta

PrimMetaToNat

PrimMetaToNatInjective

PrimLockUniv

builtinIMin :: PrimitiveId

builtinIMax :: PrimitiveId

builtinINeg :: PrimitiveId

builtinSubOut :: PrimitiveId

builtinGlue :: PrimitiveId

builtin_glue :: PrimitiveId

builtin_unglue :: PrimitiveId

builtin_glueU :: PrimitiveId

builtin_unglueU :: PrimitiveId

builtinFaceForall :: PrimitiveId

builtinComp :: PrimitiveId

builtinPOr :: PrimitiveId

builtinTrans :: PrimitiveId

builtinHComp :: PrimitiveId

builtinLockUniv :: PrimitiveId

primitiveById :: String -> Maybe PrimitiveId

Lookup a primitive by its identifier.