Agda.Syntax.Concrete.Definitions

Description

Preprocess Declarations, producing NiceDeclarations.

Attach fixity and syntax declarations to the definition they refer to.

Distribute the following attributes to the individual definitions: abstract, instance, postulate, primitive, private, termination pragmas.

Gather the function clauses belonging to one function definition.

Expand ellipsis ... in function clauses following with.

Infer mutual blocks. A block starts when a lone signature is encountered, and ends when all lone signatures have seen their definition.

Handle interleaved mutual blocks. In an `interleaved mutual' block we:

leave the data and fun sigs in place

classify signatures in constructor block based on their return type and group them all as a data def at the position in the block where the first constructor for the data sig in question occured

classify fun clauses based on the declared function used and group them all as a fundef at the position in the block where the first such fun clause appeared

Report basic well-formedness error, when one of the above transformation fails. When possible, errors should be deferred to the scope checking phase (ConcreteToAbstract), where we are in the TCM and can produce more informative error messages.

Documentation

data NiceDeclaration

The nice declarations. No fixity declarations and function definitions are contained in a single constructor instead of spread out between type signatures and clauses. The private, postulate, abstract and instance modifiers have been distributed to the individual declarations.

Observe the order of components:

Range Fixity' Access IsAbstract IsInstance TerminationCheck PositivityCheck

further attributes

(Q)Name

content (Expr, Declaration ...)

Constructors

Axiom Range Access IsAbstract IsInstance ArgInfo Name Expr
IsAbstract argument: We record whether a declaration was made in an abstract block.

ArgInfo argument: Axioms and functions can be declared irrelevant. (Hiding should be NotHidden.)

NiceField Range Access IsAbstract IsInstance TacticAttribute Name (Arg Expr)

PrimitiveFunction Range Access IsAbstract Name (Arg Expr)

NiceMutual KwRange TerminationCheck CoverageCheck PositivityCheck (List1 NiceDeclaration)

NiceModule Range Access IsAbstract Erased QName Telescope [Declaration]

NiceModuleMacro Range Access Erased Name ModuleApplication OpenShortHand ImportDirective

NiceOpen KwRange QName ImportDirective

NiceImport OpenShortHand KwRange QName (Either AsName RawOpenArgs) ImportDirective

NicePragma Range Pragma

NiceRecSig Range Erased Access IsAbstract PositivityCheck UniverseCheck Name Parameters Expr

NiceDataSig Range Erased Access IsAbstract PositivityCheck UniverseCheck Name Parameters Expr

NiceFunClause Range Access IsAbstract TerminationCheck CoverageCheck Catchall Declaration
An uncategorized function clause, could be a function clause without type signature or a pattern lhs (e.g. for irrefutable let). The Declaration is the actual FunClause.

FunSig Range Access IsAbstract IsInstance IsMacro ArgInfo TerminationCheck CoverageCheck Name Expr

FunDef Range (List1 Declaration) IsAbstract IsInstance TerminationCheck CoverageCheck Name (List1 Clause)
Block of function clauses (we have seen the type signature before). The Declarations are the original declarations that were processed into this FunDef and are only used in notSoNiceDeclaration. Andreas, 2017-01-01: Because of issue #2372, we add IsInstance here. An alias should know that it is an instance.

NiceDataDef Range Origin IsAbstract PositivityCheck UniverseCheck Name DefParameters [NiceConstructor]

NiceLoneConstructor KwRange [NiceConstructor]

NiceRecDef Range Origin IsAbstract PositivityCheck UniverseCheck Name [RecordDirective] DefParameters [Declaration]
(Maybe Range) gives range of the 'pattern' declaration.

NicePatternSyn Range Access Name [WithHiding Name] Pattern

NiceGeneralize Range Access ArgInfo TacticAttribute Name Expr

NiceUnquoteDecl Range Access IsAbstract IsInstance TerminationCheck CoverageCheck [Name] Expr

NiceUnquoteDef Range Access IsAbstract TerminationCheck CoverageCheck [Name] Expr

NiceUnquoteData Range Access IsAbstract PositivityCheck UniverseCheck Name [Name] Expr

NiceOpaque KwRange [QName] [NiceDeclaration]

type NiceConstructor = NiceTypeSignature

Only Axioms.

type NiceTypeSignature = NiceDeclaration

Only Axioms.

data Clause

One clause in a function definition. There is no guarantee that the LHS actually declares the Name. We will have to check that later.

Constructors

Clause Name Catchall ArgInfo LHS RHS WhereClause [Clause]

data DeclarationException

Exception with internal source code callstack

Constructors

DeclarationException

Fields

deLocation :: CallStack

deException :: DeclarationException'

data DeclarationWarning

Constructors

DeclarationWarning

Fields

dwLocation :: CallStack

dwWarning :: DeclarationWarning'

data DeclarationWarning'

Non-fatal errors encountered in the Nicifier.

Constructors

DivergentModalityInClause ArgInfo ArgInfo
The modality of the clause (second ArgInfo, range of the warning) is different from the modality of the type signature (first ArgInfo).

EmptyAbstract KwRange
Empty abstract block.

EmptyConstructor KwRange
Empty data _ where block.

EmptyField KwRange
Empty field block.

EmptyGeneralize KwRange
Empty variable block.

EmptyInstance KwRange
Empty instance block

EmptyMacro KwRange
Empty macro block.

EmptyMutual KwRange
Empty mutual block.

EmptyPostulate KwRange
Empty postulate block.

EmptyPrivate KwRange
Empty private block.

EmptyPrimitive KwRange
Empty primitive block.

EmptyPolarityPragma Range
POLARITY pragma without any polarities.

HiddenGeneralize Range
A Hidden identifier in a variable declaration. Hiding has no effect there as generalized variables are always hidden (or instance variables).

InvalidCatchallPragma Range
A {-# CATCHALL #-} pragma that does not precede a function clause.

InvalidConstructorBlock Range
Invalid constructor block (not inside an interleaved mutual block)

InvalidCoverageCheckPragma Range
A {-# NON_COVERING #-} pragma that does not apply to any function.

InvalidDataOrRecDefParameter Range DataOrRecord_ Text Text
An entry in a data or record definition parameter that is not allowed here. Such parameters can only be names with hiding. The first Text says what is wrong and thus ignored. The second Text may contain extra details or clarification.

InvalidNoPositivityCheckPragma Range
A {-# NO_POSITIVITY_CHECK #-} pragma that does not apply to any data or record type.

InvalidNoUniverseCheckPragma Range
A {-# NO_UNIVERSE_CHECK #-} pragma that does not apply to a data or record type.

InvalidTerminationCheckPragma Range
A {-# TERMINATING #-} and {-# NON_TERMINATING #-} pragma that does not apply to any function.

InvalidTacticAttribute Range
A misplaced tactic attribute.

MissingDataDeclaration Name
A data definition without a data signature.

MissingDefinitions (List1 (Name, Range))
Declarations (e.g. type signatures) without a definition.

NotAllowedInMutual Range String

OpenImportPrivate Range KwRange KwRange OpenOrImport
private has no effect on open (import) public. (But the user might think so.) Range is the range of the open public or open import public declaration. The first KwRange is the range of the public keyword. The second KwRange is the range of the private keyword.

OpenImportAbstract Range KwRange OpenOrImport
abstract has no effect on open or import. (But the user might think so.) Range is the range of the open or import declaration. KwRange is the range of the abstract keyword.

PolarityPragmasButNotPostulates (Set1 Name)

PragmaNoTerminationCheck Range
Pragma {-# NO_TERMINATION_CHECK #-} has been replaced by {-# TERMINATING #-} and {-# NON_TERMINATING #-}.

PragmaCompiled Range
COMPILE pragmas are not allowed in safe mode.

SafeFlagEta Range
ETA pragma is unsafe.

SafeFlagInjective Range
INJECTIVE pragma is unsafe.

SafeFlagNoCoverageCheck Range
NON_COVERING pragma is unsafe.

SafeFlagNoPositivityCheck Range
NO_POSITIVITY_CHECK pragma is unsafe.

SafeFlagNoUniverseCheck Range
NO_UNIVERSE_CHECK pragma is unsafe.

SafeFlagNonTerminating Range
NON_TERMINATING pragma is unsafe.

SafeFlagPolarity Range
POLARITY pragma is unsafe.

SafeFlagTerminating Range
TERMINATING pragma is unsafe.

ShadowingInTelescope (List1 (Name, List2 Range))

UnknownFixityInMixfixDecl (Set1 Name)
Public mixfix identifiers without a fixity declaration.

UnknownNamesInFixityDecl (Set1 Name)

UnknownNamesInPolarityPragmas (Set1 Name)

UselessAbstract KwRange
abstract block with nothing that can (newly) be made abstract.

UselessInstance KwRange
instance block with nothing that can (newly) become an instance.

UselessMacro KwRange
macro block with nothing that can (newly) be made macro.

UselessPrivate KwRange
private block with nothing that can (newly) be made private.

UselessImport Range
An import that is applied to parameters by neither bound not opened.

unsafeDeclarationWarning :: DeclarationWarning -> Bool

Nicifier warnings turned into errors in --safe mode.

data Nice a

Nicifier monad. Preserve the state when throwing an exception.

data NiceEnv

Nicifier parameters.

Constructors

NiceEnv

Fields

safeButNotBuiltin :: Bool

We are in a module declared --safe which is not a builtin module.

checkingWhere :: WhereClause_

Are we checking a where module?

runNice :: NiceEnv -> Nice a -> (Either DeclarationException a, NiceWarnings)

Run a Nicifier computation, return result and warnings (in chronological order).

niceDeclarations :: Fixities -> [Declaration] -> Nice [NiceDeclaration]

Main. Fixities (or more precisely syntax declarations) are needed when grouping function clauses.

notSoNiceDeclarations :: NiceDeclaration -> List1 Declaration

(Approximately) convert a NiceDeclaration back to a list of Declarations.

niceHasAbstract :: NiceDeclaration -> Maybe IsAbstract

Has the NiceDeclaration a field of type IsAbstract?

type Measure = Name

Termination measure is, for now, a variable name.

declarationWarningName :: DeclarationWarning -> WarningName