Agda.Syntax.Concrete.Definitions.Errors

Documentation

data DeclarationException

Exception with internal source code callstack

Constructors

DeclarationException

Fields

deLocation :: CallStack

deException :: DeclarationException'

data DeclarationException'

The exception type.

Constructors

MultipleEllipses Pattern

DuplicateDefinition Name Range
The Range is the range of the previous definition.

DuplicateAnonDeclaration Range Range
The first Range is the range of the current declaration of _. The second Range is the range of the previous declaration of _.

MissingWithClauses Name LHS

WrongDefinition Name DataRecOrFun DataRecOrFun

WrongContentBlock KindOfBlock Range

AmbiguousFunClauses LHS (List1 Name)
In a mutual block, a clause could belong to any of the ≥2 type signatures (Name).

AmbiguousConstructor Range Name [Name]
In an interleaved mutual block, a constructor could belong to any of the data signatures (Name)

InvalidMeasureMutual Range
In a mutual block, all or none need a MEASURE pragma. Range is the one of the offending pragma or the mutual block.

UnquoteDefRequiresSignature (List1 Name)

BadMacroDef NiceDeclaration

UnfoldingOutsideOpaque KwRange
An unfolding declaration was not the first declaration contained in an opaque block.

OpaqueInMutual KwRange
opaque block nested in a mutual block. This can never happen, even with reordering. The KwRange is the one of the opaque keyword.

DisallowedInterleavedMutual KwRange String (List1 Name)
A declaration that breaks an implicit mutual block (named by the String argument) was present while the given lone type signatures were still without their definitions.

declarationExceptionString :: DeclarationException' -> String

The name of the error.

data DeclarationWarning

Constructors

DeclarationWarning

Fields

dwLocation :: CallStack

dwWarning :: DeclarationWarning'

data DeclarationWarning'

Non-fatal errors encountered in the Nicifier.

Constructors

DivergentModalityInClause ArgInfo ArgInfo
The modality of the clause (second ArgInfo, range of the warning) is different from the modality of the type signature (first ArgInfo).

EmptyAbstract KwRange
Empty abstract block.

EmptyConstructor KwRange
Empty data _ where block.

EmptyField KwRange
Empty field block.

EmptyGeneralize KwRange
Empty variable block.

EmptyInstance KwRange
Empty instance block

EmptyMacro KwRange
Empty macro block.

EmptyMutual KwRange
Empty mutual block.

EmptyPostulate KwRange
Empty postulate block.

EmptyPrivate KwRange
Empty private block.

EmptyPrimitive KwRange
Empty primitive block.

EmptyPolarityPragma Range
POLARITY pragma without any polarities.

HiddenGeneralize Range
A Hidden identifier in a variable declaration. Hiding has no effect there as generalized variables are always hidden (or instance variables).

InvalidCatchallPragma Range
A {-# CATCHALL #-} pragma that does not precede a function clause.

InvalidConstructorBlock Range
Invalid constructor block (not inside an interleaved mutual block)

InvalidCoverageCheckPragma Range
A {-# NON_COVERING #-} pragma that does not apply to any function.

InvalidDataOrRecDefParameter Range DataOrRecord_ Text Text
An entry in a data or record definition parameter that is not allowed here. Such parameters can only be names with hiding. The first Text says what is wrong and thus ignored. The second Text may contain extra details or clarification.

InvalidNoPositivityCheckPragma Range
A {-# NO_POSITIVITY_CHECK #-} pragma that does not apply to any data or record type.

InvalidNoUniverseCheckPragma Range
A {-# NO_UNIVERSE_CHECK #-} pragma that does not apply to a data or record type.

InvalidTerminationCheckPragma Range
A {-# TERMINATING #-} and {-# NON_TERMINATING #-} pragma that does not apply to any function.

InvalidTacticAttribute Range
A misplaced tactic attribute.

MissingDataDeclaration Name
A data definition without a data signature.

MissingDefinitions (List1 (Name, Range))
Declarations (e.g. type signatures) without a definition.

NotAllowedInMutual Range String

OpenImportPrivate Range KwRange KwRange OpenOrImport
private has no effect on open (import) public. (But the user might think so.) Range is the range of the open public or open import public declaration. The first KwRange is the range of the public keyword. The second KwRange is the range of the private keyword.

OpenImportAbstract Range KwRange OpenOrImport
abstract has no effect on open or import. (But the user might think so.) Range is the range of the open or import declaration. KwRange is the range of the abstract keyword.

PolarityPragmasButNotPostulates (Set1 Name)

PragmaNoTerminationCheck Range
Pragma {-# NO_TERMINATION_CHECK #-} has been replaced by {-# TERMINATING #-} and {-# NON_TERMINATING #-}.

PragmaCompiled Range
COMPILE pragmas are not allowed in safe mode.

SafeFlagEta Range
ETA pragma is unsafe.

SafeFlagInjective Range
INJECTIVE pragma is unsafe.

SafeFlagNoCoverageCheck Range
NON_COVERING pragma is unsafe.

SafeFlagNoPositivityCheck Range
NO_POSITIVITY_CHECK pragma is unsafe.

SafeFlagNoUniverseCheck Range
NO_UNIVERSE_CHECK pragma is unsafe.

SafeFlagNonTerminating Range
NON_TERMINATING pragma is unsafe.

SafeFlagPolarity Range
POLARITY pragma is unsafe.

SafeFlagTerminating Range
TERMINATING pragma is unsafe.

ShadowingInTelescope (List1 (Name, List2 Range))

UnknownFixityInMixfixDecl (Set1 Name)
Public mixfix identifiers without a fixity declaration.

UnknownNamesInFixityDecl (Set1 Name)

UnknownNamesInPolarityPragmas (Set1 Name)

UselessAbstract KwRange
abstract block with nothing that can (newly) be made abstract.

UselessInstance KwRange
instance block with nothing that can (newly) become an instance.

UselessMacro KwRange
macro block with nothing that can (newly) be made macro.

UselessPrivate KwRange
private block with nothing that can (newly) be made private.

UselessImport Range
An import that is applied to parameters by neither bound not opened.

data OpenOrImport

open or import

Constructors

OpenNotImport
open.

ImportMayOpen
import or open import.

declarationWarningName :: DeclarationWarning -> WarningName

declarationWarningName' :: DeclarationWarning' -> WarningName

unsafeDeclarationWarning :: DeclarationWarning -> Bool

Nicifier warnings turned into errors in --safe mode.

unsafeDeclarationWarning' :: DeclarationWarning' -> Bool

unsafePragma :: CMaybe DeclarationWarning' m => Pragma -> m

Pragmas not allowed in --safe mode produce an unsafeDeclarationWarning.