Agda.Syntax.Concrete.Definitions.Errors

Documentation

data DeclarationException

Exception with internal source code callstack

Constructors

DeclarationException

Fields

deLocation :: CallStack

deException :: DeclarationException'

data DeclarationException'

The exception type.

Constructors

MultipleEllipses Pattern

InvalidName Name

DuplicateDefinition Name

DuplicateAnonDeclaration Range

MissingWithClauses Name LHS

WrongDefinition Name DataRecOrFun DataRecOrFun

DeclarationPanic String

WrongContentBlock KindOfBlock Range

AmbiguousFunClauses LHS (List1 Name)
In a mutual block, a clause could belong to any of the ≥2 type signatures (Name).

AmbiguousConstructor Range Name [Name]
In an interleaved mutual block, a constructor could belong to any of the data signatures (Name)

InvalidMeasureMutual Range
In a mutual block, all or none need a MEASURE pragma. Range is of mutual block.

UnquoteDefRequiresSignature (List1 Name)

BadMacroDef NiceDeclaration

data DeclarationWarning

Constructors

DeclarationWarning

Fields

dwLocation :: CallStack

dwWarning :: DeclarationWarning'

data DeclarationWarning'

Non-fatal errors encountered in the Nicifier.

Constructors

EmptyAbstract Range
Empty abstract block.

EmptyConstructor Range
Empty constructor block.

EmptyField Range
Empty field block.

EmptyGeneralize Range
Empty variable block.

EmptyInstance Range
Empty instance block

EmptyMacro Range
Empty macro block.

EmptyMutual Range
Empty mutual block.

EmptyPostulate Range
Empty postulate block.

EmptyPrivate Range
Empty private block.

EmptyPrimitive Range
Empty primitive block.

InvalidCatchallPragma Range
A {-# CATCHALL #-} pragma that does not precede a function clause.

InvalidConstructor Range
Invalid definition in a constructor block

InvalidConstructorBlock Range
Invalid constructor block (not inside an interleaved mutual block)

InvalidCoverageCheckPragma Range
A {-# NON_COVERING #-} pragma that does not apply to any function.

InvalidNoPositivityCheckPragma Range
A {-# NO_POSITIVITY_CHECK #-} pragma that does not apply to any data or record type.

InvalidNoUniverseCheckPragma Range
A {-# NO_UNIVERSE_CHECK #-} pragma that does not apply to a data or record type.

InvalidRecordDirective Range
A record directive outside of a record / below existing fields.

InvalidTerminationCheckPragma Range
A {-# TERMINATING #-} and {-# NON_TERMINATING #-} pragma that does not apply to any function.

MissingDeclarations [(Name, Range)]
Definitions (e.g. constructors or functions) without a declaration.

MissingDefinitions [(Name, Range)]
Declarations (e.g. type signatures) without a definition.

NotAllowedInMutual Range String

OpenPublicPrivate Range
private has no effect on open public. (But the user might think so.)

OpenPublicAbstract Range
abstract has no effect on open public. (But the user might think so.)

PolarityPragmasButNotPostulates [Name]

PragmaNoTerminationCheck Range
Pragma {-# NO_TERMINATION_CHECK #-} has been replaced by {-# TERMINATING #-} and {-# NON_TERMINATING #-}.

PragmaCompiled Range
COMPILE pragmas are not allowed in safe mode

ShadowingInTelescope (List1 (Name, List2 Range))

UnknownFixityInMixfixDecl [Name]

UnknownNamesInFixityDecl [Name]

UnknownNamesInPolarityPragmas [Name]

UselessAbstract Range
abstract block with nothing that can (newly) be made abstract.

UselessInstance Range
instance block with nothing that can (newly) become an instance.

UselessPrivate Range
private block with nothing that can (newly) be made private.

declarationWarningName :: DeclarationWarning -> WarningName

declarationWarningName' :: DeclarationWarning' -> WarningName

unsafeDeclarationWarning :: DeclarationWarning -> Bool

Nicifier warnings turned into errors in --safe mode.

unsafeDeclarationWarning' :: DeclarationWarning' -> Bool