Agda.Syntax.Concrete.Operators.Parser.Monad

Description

The parser monad used by the operator parser

Documentation

data MemoKey

Memoisation keys.

Constructors

NodeK PrecedenceKey

PostLeftsK PrecedenceKey

PreRightsK PrecedenceKey

TopK

AppK

NonfixK

type PrecedenceKey = Either PrecedenceLevel PrecedenceLevel

type Parser tok a = Parser MemoKey tok (MaybePlaceholder tok) a

The parser monad.

parse :: Parser tok a -> [MaybePlaceholder tok] -> [a]

Runs the parser.

sat' :: (MaybePlaceholder tok -> Maybe a) -> Parser tok a

Parses a token satisfying the given predicate. The computed value is returned.

sat :: (MaybePlaceholder tok -> Bool) -> Parser tok (MaybePlaceholder tok)

Parses a token satisfying the given predicate.

doc :: Doc -> Parser tok a -> Parser tok a

Uses the given document as the printed representation of the given parser. The document's precedence is taken to be atomP.

memoise :: MemoKey -> Parser tok tok -> Parser tok tok

Memoises the given parser.

Every memoised parser must be annotated with a unique key. (Parametrised parsers must use distinct keys for distinct inputs.)

memoiseIfPrinting :: MemoKey -> Parser tok tok -> Parser tok tok

Memoises the given parser, but only if printing, not if parsing.

Every memoised parser must be annotated with a unique key. (Parametrised parsers must use distinct keys for distinct inputs.)

grammar :: Parser tok a -> Doc

Tries to print the parser, or returns empty, depending on the implementation. This function might not terminate.