Agda.Syntax.Info

Description

An info object contains additional information about a piece of abstract syntax that isn't part of the actual syntax. For instance, it might contain the source code position of an expression or the concrete syntax that an internal expression originates from.

Documentation

data MetaKind

Kind of a meta: the method how to solve it.

Constructors

InstanceMeta
Meta variable solved by instance search.

UnificationMeta
Meta variable solved by unification (default).

hidingToMetaKind :: Hiding -> MetaKind

Default meta kind from its Hiding context.

type MetaNameSuggestion = String

Name suggestion for meta variable. Empty string means no suggestion.

data MetaInfo

Information associated to a meta variable in the abstract syntax.

Constructors

MetaInfo

Fields

metaRange :: Range

metaScope :: ScopeInfo

metaNumber :: Maybe MetaId

metaNameSuggestion :: MetaNameSuggestion

metaKind :: MetaKind

emptyMetaInfo :: MetaInfo

newtype ExprInfo

Constructors

ExprRange Range

exprNoRange :: ExprInfo

data AppInfo

Information about application

Constructors

AppInfo

Fields

appRange :: Range

appOrigin :: Origin

appParens :: ParenPreference

Do we prefer a lambda argument with or without parens?

defaultAppInfo :: Range -> AppInfo

Default is system inserted and prefer parens.

defaultAppInfo_ :: AppInfo

AppInfo with no range information.

data ModuleInfo

Constructors

ModuleInfo

Fields

minfoRange :: Range

minfoAsTo :: Range

The range of the "as" and "to" keywords, if any. Retained for highlighting purposes.

minfoAsName :: Maybe Name

The "as" module name, if any. Retained for highlighting purposes.

minfoOpenShort :: Maybe OpenShortHand

minfoDirective :: Maybe ImportDirective

Retained for abstractToConcrete of ModuleMacro.

newtype LetInfo

Constructors

LetRange Range

data DefInfo' t

Constructors

DefInfo

Fields

defFixity :: Fixity'

defAccess :: Access

defAbstract :: IsAbstract

defOpaque :: IsOpaque

defInstance :: IsInstance

defMacro :: IsMacro

defInfo :: DeclInfo

defTactic :: TacticAttribute' t

mkDefInfo :: Name -> Fixity' -> Access -> IsAbstract -> Range -> DefInfo' t

mkDefInfoInstance :: Name -> Fixity' -> Access -> IsAbstract -> IsInstance -> IsMacro -> Range -> DefInfo' t

Same as mkDefInfo but where we can also give the IsInstance

data DeclInfo

Constructors

DeclInfo

Fields

declName :: Name

declRange :: Range

data MutualInfo

Constructors

MutualInfo

Fields

mutualTerminationCheck :: TerminationCheck Name

mutualCoverageCheck :: CoverageCheck

mutualPositivityCheck :: PositivityCheck

mutualRange :: Range

data LHSInfo

Constructors

LHSInfo

Fields

lhsRange :: Range

lhsEllipsis :: ExpandedEllipsis

newtype PatInfo

For a general pattern we remember the source code position.

Constructors

PatRange Range

patNoRange :: PatInfo

Empty range for patterns.

data ConPatInfo

Constructor pattern info.

Constructors

ConPatInfo

Fields

conPatOrigin :: ConOrigin

Does this pattern come form the eta-expansion of an implicit pattern?

conPatInfo :: PatInfo

conPatLazy :: ConPatLazy

data ConPatLazy

Has the constructor pattern a dotted (forced) constructor?

Constructors

ConPatLazy
Dotted constructor.

ConPatEager
Ordinary constructor.