Agda.Syntax.Internal

Documentation

pattern ClosedLevel :: Integer -> Level

Constant level n

pattern EqualityType :: Range -> Sort -> QName -> Args -> Arg Term -> Arg Term -> Arg Term -> EqualityView

pattern Prop :: Level' t -> Sort' t

pattern SSet :: Level' t -> Sort' t

pattern Type :: Level' t -> Sort' t

pattern Var :: Int -> Elims -> Term

__DUMMY_DOM__ :: HasCallStack => Dom Type

__DUMMY_LEVEL__ :: HasCallStack => Level

__DUMMY_SORT__ :: HasCallStack => Sort

__DUMMY_TERM_WITH__ :: HasCallStack => String -> Term

__DUMMY_TERM__ :: HasCallStack => Term

__DUMMY_TYPE__ :: HasCallStack => Type

absurdBody :: Abs Term

Absurd lambdas are internally represented as identity with variable name "()".

absurdP :: Int -> DeBruijnPattern

Make an absurd pattern with the given de Bruijn index.

absurdPatternName :: PatVarName

argFromDom :: Dom' t a -> Arg a

arity :: Type -> Nat

Doesn't do any reduction.

atomicLevel :: t -> Level' t

clausePats :: Clause -> [Arg DeBruijnPattern]

couldBeRecursive :: ClauseRecursive -> Bool

decideRecursive :: Bool -> ClauseRecursive

defaultArgDom :: ArgInfo -> a -> Dom a

defaultDom :: a -> Dom a

defaultNamedArgDom :: ArgInfo -> String -> a -> Dom a

defaultPatternInfo :: PatternInfo

domFromArg :: Arg a -> Dom a

domFromNamedArg :: NamedArg a -> Dom a

dontCare :: Term -> Term

Add DontCare is it is not already a DontCare.

dotP :: Term -> Pattern' a

dummyDom :: CallStack -> Dom Type

Context entries without a type have this dummy type. Note: use macro DUMMY_DOM !

dummyLevel :: CallStack -> Level

A dummy level to constitute a level/sort created at location. Note: use macro DUMMY_LEVEL !

dummyLocName :: CallStack -> String

Construct a string representing the call-site that created the dummy thing.

dummySort :: CallStack -> Sort

A dummy sort created at location. Note: use macro DUMMY_SORT !

dummyTerm :: CallStack -> Term

A dummy term created at location. Note: use macro DUMMY_TERM !

dummyTermWith :: String -> CallStack -> Term

Aux: A dummy term to constitute a dummy termlevelsort/type.

dummyType :: CallStack -> Type

A dummy type created at location. Note: use macro DUMMY_TYPE !

eqtLhs :: EqualityView -> Arg Term

eqtName :: EqualityView -> QName

eqtParams :: EqualityView -> Args

eqtRange :: EqualityView -> Range

eqtRhs :: EqualityView -> Arg Term

eqtSort :: EqualityView -> Sort

eqtType :: EqualityView -> Arg Term

fromConPatternInfo :: ConPatternInfo -> ConInfo

Build ConInfo from ConPatternInfo.

hasElims :: Term -> Maybe (Elims -> Term, Elims)

A view distinguishing the neutrals Var, Def, and MetaV which can be projected.

impossibleTerm :: CallStack -> Term

isAbsurdBody :: Abs Term -> Bool

isAbsurdPatternName :: PatVarName -> Bool

isEqualityType :: EqualityView -> Bool

isIOne :: IntervalView -> Bool

isPathType :: PathView -> Bool

isProp :: Sort' t -> Bool

Is this a Prop universe? Answers are yes (True) or maybe (False).

isSort :: Term -> Maybe Sort

isStrictDataSort :: Sort' t -> Bool

Is this a strict universe inhabitable by data types?

levelPlus :: Integer -> Level -> Level

Given a constant m and level l, compute m + l

levelSuc :: Level -> Level

listTel :: Lens' Telescope ListTel

Lens to edit a Telescope as a list.

litP :: Literal -> Pattern' a

mapAbsNames :: (ArgName -> ArgName) -> Tele a -> Tele a

mapAbsNamesM :: Applicative m => (ArgName -> m ArgName) -> Tele a -> m (Tele a)

A traversal for the names in a telescope.

mkProp :: Integer -> Sort

mkSSet :: Integer -> Sort

mkType :: Integer -> Sort

nameToPatVarName :: Name -> PatVarName

namedArgFromDom :: Dom' t a -> NamedArg a

namedDBVarP :: Int -> PatVarName -> Named_ DeBruijnPattern

namedVarP :: PatVarName -> Named_ Pattern

noConPatternInfo :: ConPatternInfo

pDom :: LensHiding a => a -> Doc -> Doc

patVarNameToString :: PatVarName -> String

patternInfo :: Pattern' x -> Maybe PatternInfo

Retrieve the PatternInfo from a pattern

patternOrigin :: Pattern' x -> Maybe PatOrigin

Retrieve the origin of a pattern

prettyPrecLevelSucs :: Int -> Integer -> (Int -> Doc) -> Doc

propToType :: Sort' t -> Sort' t

Turn a known UProp sort into a UType sort, leave others unchanged.

sortUniv :: Sort' t -> Maybe Univ

Get the flavor of the universe. Nothing could also mean "don't know".

stripDontCare :: Term -> Term

Removing a topmost DontCare constructor.

suggests :: [Suggestion] -> String

telFromList :: ListTel -> Telescope

Convert a list telescope to a telescope.

telFromList' :: (a -> ArgName) -> ListTel' a -> Telescope

telToList :: Tele (Dom t) -> [Dom (ArgName, t)]

Convert a telescope to its list form.

tmSSort :: Term -> Sort

tmSort :: Term -> Sort

toConPatternInfo :: ConInfo -> ConPatternInfo

Build partial ConPatternInfo from ConInfo

unSpine :: Term -> Term

Convert top-level postfix projections into prefix projections.

unSpine' :: (ProjOrigin -> QName -> Bool) -> Term -> Term

Convert Proj projection eliminations according to their ProjOrigin into Def projection applications.

var :: Nat -> Term

An unapplied variable.

varP :: a -> Pattern' a

varSort :: Int -> Sort

varTable :: Array Term

varTableSize :: Int

data Abs a

Binder.

Abs: The bound variable might appear in the body. NoAbs is pseudo-binder, it does not introduce a fresh variable, similar to the const of Haskell.

Constructors

Abs
The body has (at least) one free variable. Danger: unAbs doesn't shift variables properly

Fields

absName :: ArgName

unAbs :: a

NoAbs

Fields

absName :: ArgName

unAbs :: a

type Args = [Arg Term]

Type of argument lists.

type Blocked = Blocked' Term

type Blocked_ = Blocked ()

'Blocked a without the a.

newtype BraveTerm

Newtypes for terms that produce a dummy, rather than crash, when applied to incompatible eliminations.

Constructors

BraveTerm

Fields

unBrave :: Term

data Clause

A clause is a list of patterns and the clause body.

The telescope contains the types of the pattern variables and the de Bruijn indices say how to get from the order the variables occur in the patterns to the order they occur in the telescope. The body binds the variables in the order they appear in the telescope.

clauseTel ~ permute clausePerm (patternVars namedClausePats)

Terms in dot patterns are valid in the clause telescope.

For the purpose of the permutation and the body dot patterns count as variables. TODO: Change this!

Constructors

Clause

Fields

clauseLHSRange :: Range

clauseFullRange :: Range

clauseTel :: Telescope

Δ: The types of the pattern variables in dependency order.

namedClausePats :: NAPs

Δ ⊢ ps. The de Bruijn indices refer to Δ.

clauseBody :: Maybe Term

Just v with Δ ⊢ v for a regular clause, or Nothing for an absurd one.

clauseType :: Maybe (Arg Type)

Δ ⊢ t. The type of the rhs under clauseTel. Used, e.g., by TermCheck. Can be Irrelevant if we encountered an irrelevant projection pattern on the lhs.

clauseCatchall :: Catchall

Clause has been labelled as CATCHALL.

clauseRecursive :: ClauseRecursive

clauseBody contains recursive calls? Computed by termination checker.

clauseUnreachable :: Maybe Bool

Clause has been labelled as unreachable by the coverage checker. Nothing means coverage checker has not run yet (clause may be unreachable). Just False means clause is not unreachable. Just True means clause is unreachable.

clauseEllipsis :: ExpandedEllipsis

Was this clause created by expansion of an ellipsis?

clauseWhereModule :: Maybe ModuleName

Keeps track of the module name associate with the clause's where clause.

data ClauseRecursive

Does the clause body contain calls to any of the mutually recursive functions?

Constructors

YesRecursive
Definitely a call to a mutually recursive function.

NotRecursive
Definitely no call to a mutually recursive function.

MaybeRecursive
Don't know because the analysis has not run yet.

data ConHead

Store the names of the record fields in the constructor. This allows reduction of projection redexes outside of TCM. For instance, during substitution and application.

Constructors

ConHead

Fields

conName :: QName

The name of the constructor.

conDataRecord :: DataOrRecord

Data or record constructor?

conInductive :: Induction

Record constructors can be coinductive.

conFields :: [Arg QName]

The name of the record fields. Arg is stored since the info in the constructor args might not be accurate because of subtyping (issue #2170).

type ConInfo = ConOrigin

data ConPatternInfo

The ConPatternInfo states whether the constructor belongs to a record type (True) or data type (False). In the former case, the PatOrigin of the conPInfo says whether the record pattern orginates from the expansion of an implicit pattern. The Type is the type of the whole record pattern. The scope used for the type is given by any outer scope plus the clause's telescope (clauseTel).

Constructors

ConPatternInfo

Fields

conPInfo :: PatternInfo

Information on the origin of the pattern.

conPRecord :: Bool

False if data constructor. True if record constructor.

conPFallThrough :: Bool

Should the match block on non-canonical terms or can it proceed to the catch-all clause?

conPType :: Maybe (Arg Type)

The type of the whole constructor pattern. Should be present (Just) if constructor pattern is is generated ordinarily by type-checking. Could be absent (Nothing) if pattern comes from some plugin (like Agsy). Needed e.g. for with-clause stripping.

conPLazy :: Bool

Lazy patterns are generated by the forcing translation in the unifier (unifyStep) and are dropped by the clause compiler (TODO: not yet) (compileClauses) when the variables they bind are unused. The GHC backend compiles lazy matches to lazy patterns in Haskell (TODO: not yet).

data DBPatVar

Type used when numbering pattern variables.

Constructors

DBPatVar

Fields

dbPatVarName :: PatVarName

dbPatVarIndex :: !Int

type DeBruijnPattern = Pattern' DBPatVar

type Dom = Dom' Term

data Dom' t e

Similar to Arg, but we need to distinguish an irrelevance annotation in a function domain (the domain itself is not irrelevant!) from an irrelevant argument.

Dom is used in Pi of internal syntax, in Context and Telescope. Arg is used for actual arguments (Var, Con, Def etc.) and in Abstract syntax and other situations.

cubical

When annFinite (argInfoAnnotation domInfo) = True for the domain of a Pi type, the elements should be compared by tabulating the domain type. Only supported in case the domain type is primIsOne, to obtain the correct equality for partial elements.

Constructors

Dom

Fields

domInfo :: ArgInfo

domName :: Maybe NamedName

e.g. x in {x = y : A} -> B.

domIsFinite :: Bool

Is this a Π-type (False), or a partial type (True)?

domTactic :: Maybe t

"@tactic e".

unDom :: e

data DummyTermKind

Terms used for internal purposes

Constructors

DummyNamed String
Generic dummy term, the String generally describes where the term was generated (as well as giving a disambiguator).

Should not be used to convey actual information.

DummyBrave Term
Collects applications to a BraveTerm.

type Elim = Elim' Term

type Elims

Arguments

= [Elim]
eliminations ordered left-to-right.

data EqualityTypeData

Constructors

EqualityTypeData

Fields

_eqtRange :: Range

Range of the rewrite expression, if any.

_eqtSort :: Sort

Sort of this type.

_eqtName :: QName

Builtin EQUALITY.

_eqtParams :: Args

Hidden. Empty or Level.

_eqtType :: Arg Term

Hidden.

_eqtLhs :: Arg Term

NotHidden.

_eqtRhs :: Arg Term

NotHidden.

data EqualityView

View type as equality type.

Constructors

EqualityViewType EqualityTypeData
A type of the form u ≡ v decomposed into its parts. Used as type for the rewrite expression.

OtherType Type
A reduced type used as type for a with expression.

IdiomType Type
A reduced type used as type for the with inspect idiom.

data IntervalView

Constructors

IZero

IOne

IMin (Arg Term) (Arg Term)

IMax (Arg Term) (Arg Term)

INeg (Arg Term)

OTerm Term

class LensConName a where

Minimal complete definition

getConName

Methods

getConName :: a -> QName

setConName :: QName -> a -> a

mapConName :: (QName -> QName) -> a -> a

class LensSort a where

Minimal complete definition

lensSort

Methods

lensSort :: Lens' a Sort

getSort :: a -> Sort

type Level = Level' Term

data Level' t

A level is a maximum expression of a closed level and 0..n PlusLevel expressions each of which is an atom plus a number.

Constructors

Max !Integer [PlusLevel' t]

type LevelAtom = Term

type ListTel = ListTel' ArgName

type ListTel' a = [Dom (a, Type)]

Telescope as list.

type NAPs = [NamedArg DeBruijnPattern]

Named pattern arguments.

type NamedArgs = [NamedArg Term]

newtype NoSubst t a

Wrapper for types that do not contain variables (so applying a substitution is the identity). Useful if you have a structure of types that support substitution mixed with types that don't and need to apply a substitution to the full structure.

Constructors

NoSubst

Fields

unNoSubst :: a

type NotBlocked = NotBlocked' Term

data PatOrigin

Origin of the pattern: what did the user write in this position?

Constructors

PatOSystem
Pattern inserted by the system

PatOSplit
Pattern generated by case split

PatOSplitArg ArgName
Argument to pattern generated by case split

PatOVar Name
User wrote a variable pattern

PatODot
User wrote a dot pattern

PatOWild
User wrote a wildcard pattern

PatOCon
User wrote a constructor pattern

PatORec
User wrote a record pattern

PatOLit
User wrote a literal pattern

PatOAbsurd
User wrote an absurd pattern

type PatVarName = ArgName

Pattern variables.

data PathView

View type as path type.

Constructors

PathType

Fields

pathSort :: Sort

Sort of this type.

pathName :: QName

Builtin PATH.

pathLevel :: Arg Term

Hidden

pathType :: Arg Term

Hidden

pathLhs :: Arg Term

NotHidden

pathRhs :: Arg Term

NotHidden

OType Type
reduced

type Pattern

Arguments

= Pattern' PatVarName
The PatVarName is a name suggestion.

data Pattern' x

Patterns are variables, constructors, or wildcards. QName is used in ConP rather than Name since a constructor might come from a particular namespace. This also meshes well with the fact that values (i.e. the arguments we are matching with) use QName.

Constructors

VarP PatternInfo x
x

DotP PatternInfo Term
.t

ConP ConHead ConPatternInfo [NamedArg (Pattern' x)]
c ps The subpatterns do not contain any projection copatterns.

LitP PatternInfo Literal
E.g. 5, "hello".

ProjP ProjOrigin QName
Projection copattern. Can only appear by itself.

IApplyP PatternInfo Term Term x
Path elimination pattern, like VarP but keeps track of endpoints.

DefP PatternInfo QName [NamedArg (Pattern' x)]
Used for HITs, the QName should be the one from primHComp.

data PatternInfo

Constructors

PatternInfo

Fields

patOrigin :: PatOrigin

patAsNames :: [Name]

type PatternSubstitution = Substitution' DeBruijnPattern

type family PatternVarOut a

class PatternVars a where

Extract pattern variables in left-to-right order. A DotP is also treated as variable (see docu for Clause).

Associated Types

type PatternVarOut a

Methods

patternVars :: a -> [Arg (Either (PatternVarOut a) Term)]

type PlusLevel = PlusLevel' Term

data PlusLevel' t

Constructors

Plus !Integer t

class SgTel a where

Constructing a singleton telescope.

Methods

sgTel :: a -> Telescope

type Sort = Sort' Term

data Sort' t

Sorts.

Constructors

Univ Univ (Level' t)
Prop ℓ, Set ℓ, SSet ℓ.

Inf Univ !Integer
Propωᵢ, (S)Setωᵢ.

SizeUniv
SizeUniv, a sort inhabited by type Size.

LockUniv
LockUniv, a sort for locks.

LevelUniv
LevelUniv, a sort inhabited by type Level. When --level-universe isn't on, this universe reduces to Set 0

IntervalUniv
IntervalUniv, a sort inhabited by the cubical interval.

PiSort (Dom' t t) (Sort' t) (Abs (Sort' t))
Sort of the pi type.

FunSort (Sort' t) (Sort' t)
Sort of a (non-dependent) function type.

UnivSort (Sort' t)
Sort of another sort.

MetaS !MetaId [Elim' t]

DefS QName [Elim' t]
A postulated sort.

DummyS String
A (part of a) term or type which is only used for internal purposes. Replaces the abuse of Prop for a dummy sort. The String typically describes the location where we create this dummy, but can contain other information as well.

type Substitution = Substitution' Term

data Substitution' a

Substitutions.

Constructors

IdS
Identity substitution. Γ ⊢ IdS : Γ

EmptyS Impossible
Empty substitution, lifts from the empty context. First argument is IMPOSSIBLE. Apply this to closed terms you want to use in a non-empty context. Γ ⊢ EmptyS : ()

a :# (Substitution' a) infixr 4
Substitution extension, `cons'. Γ ⊢ u : Aρ Γ ⊢ ρ : Δ ---------------------- Γ ⊢ u :# ρ : Δ, A

Strengthen Impossible !Int (Substitution' a)
Strengthening substitution. First argument is IMPOSSIBLE. In 'Strengthen err n ρ the number n must be non-negative. This substitution should only be applied to values t for which none of the variables 0 up to n - 1 are free in t[ρ], and in that case n is subtracted from all free de Bruijn indices in t[ρ]. Γ ⊢ ρ : Δ |Θ| = n --------------------------- Γ ⊢ Strengthen n ρ : Δ, Θ @

Wk !Int (Substitution' a)
Weakening substitution, lifts to an extended context. Γ ⊢ ρ : Δ ------------------- Γ, Ψ ⊢ Wk |Ψ| ρ : Δ

Lift !Int (Substitution' a)
Lifting substitution. Use this to go under a binder. Lift 1 ρ == var 0 :# Wk 1 ρ. Γ ⊢ ρ : Δ ------------------------- Γ, Ψρ ⊢ Lift |Ψ| ρ : Δ, Ψ

class Suggest a where

Suggest a name if available (i.e. name is not "_")

Methods

suggestName :: a -> Maybe String

data Suggestion

Constructors

Suggest a => Suggestion a

class TelToArgs a where

Drop the types from a telescope.

Methods

telToArgs :: a -> [Arg ArgName]

data Tele a

Sequence of types. An argument of the first type is bound in later types and so on.

Constructors

EmptyTel

ExtendTel a (Abs (Tele a))
Abs is never NoAbs.

type Telescope = Tele (Dom Type)

data Term

Raw values.

Def is used for both defined and undefined constants. Assume there is a type declaration and a definition for every constant, even if the definition is an empty list of clauses.

Constructors

Var' !Int Elims
x es neutral

Lam ArgInfo (Abs Term)
Terms are beta normal. Relevance is ignored

Lit Literal

Def QName Elims
f es, possibly a delta/iota-redex

Con ConHead ConInfo Elims
c es or record { fs = es } es allows only Apply and IApply eliminations, and IApply only for data constructors.

Pi (Dom Type) (Abs Type)
dependent or non-dependent function space

Sort Sort

Level Level

MetaV !MetaId Elims

DontCare Term
Irrelevant stuff in relevant position, but created in an irrelevant context. Basically, an internal version of the irrelevance axiom .irrAx : .A -> A.

Dummy DummyTermKind Elims
A (part of a) term or type which is only used for internal purposes.

The DummyTermKind describes why this dummy was created; typically, it is a string describing the source location that created it, but dummy terms are also used when *leaving* the type checker (through reification) to convey out-of-band information.

The second field accumulates eliminations in case we apply a dummy term to more of them. Dummy terms should never be used in places where they can affect type checking, so syntactic checks are free to ignore the eliminators, which are only there to ease debugging when a dummy term incorrectly leaks into a relevant position.

class TermSize a where

The size of a term is roughly the number of nodes in its syntax tree. This number need not be precise for logical correctness of Agda, it is only used for reporting (and maybe decisions regarding performance).

Not counting towards the term size are:

sort and color annotations,

projections.

Minimal complete definition

tsize

Methods

termSize :: a -> Int

tsize :: a -> Sum Int

type Type = Type' Term

type Type' a = Type'' Term a

data Type'' t a

Types are terms with a sort annotation.

Constructors

El

Fields

_getSort :: Sort' t

unEl :: a

type family TypeOf a

data UnivSize

Constructors

USmall
PropSetSSet ℓ.

ULarge
(PropSetSSet)ωᵢ.

module Agda.Syntax.Internal.Blockers

module Agda.Syntax.Internal.Elim

module Agda.Syntax.Internal.Univ

module Agda.Syntax.Abstract.Name

data MetaId

Meta-variable identifiers use the same structure as NameIds.

Constructors

MetaId

Fields

metaId :: !Word64

metaModule :: !ModuleNameHash

newtype ProblemId

A "problem" consists of a set of constraints and the same constraint can be part of multiple problems.

Constructors

ProblemId Nat

Orphan instances

KillRange DataOrRecord

Instance details

Methods

killRange :: KillRangeT DataOrRecord

NFData IsFibrant

Instance details

Methods

rnf :: IsFibrant -> ()

NFData a => NFData (DataOrRecord' a)

Instance details

Methods

rnf :: DataOrRecord' a -> ()