Agda.Syntax.Internal.Names

Description

Extract all names and meta-variables from things.

Documentation

namesIn :: (NamesIn a, Collection QName m) => a -> m

Some or all of the QNames that can be found in the given thing.

namesIn' :: (NamesIn a, Monoid m) => (QName -> m) -> a -> m

Some or all of the QNames that can be found in the given thing.

metasIn :: (NamesIn a, Collection MetaId m) => a -> m

Some or all of the meta-variables that can be found in the given thing.

metasIn' :: (NamesIn a, Monoid m) => (MetaId -> m) -> a -> m

Some or all of the meta-variables that can be found in the given thing.

namesAndMetasIn :: (NamesIn a, Collection QName m1, Collection MetaId m2) => a -> (m1, m2)

Some or all of the names and meta-variables that can be found in the given thing.

class NamesIn a where

Minimal complete definition

Nothing

Methods

namesAndMetasIn' :: Monoid m => (Either QName MetaId -> m) -> a -> m

Some or all of the names and meta-variables that can be found in the given thing.

default namesAndMetasIn' :: forall m (f :: Type -> Type) b. (Monoid m, Foldable f, NamesIn b, f b ~ a) => (Either QName MetaId -> m) -> a -> m

newtype PSyn

Constructors

PSyn PatternSynDefn