Agda.Syntax.TopLevelModuleName

Documentation

hashRawTopLevelModuleName :: RawTopLevelModuleName -> ModuleNameHash

Hashes a raw top-level module name.

lensTopLevelModuleNameParts :: Lens' TopLevelModuleName TopLevelModuleNameParts

A lens focusing on the moduleNameParts.

moduleNameToFileName :: TopLevelModuleName -> String -> FilePath

Turns a top-level module name into a file name with the given suffix.

projectRoot :: AbsolutePath -> TopLevelModuleName -> AbsolutePath

Finds the current project's "root" directory, given a project file and the corresponding top-level module name.

Example: If the module "A.B.C" is located in the file "fooABC.agda", then the root is "foo".

Precondition: The module name must be well-formed.

rawTopLevelModuleName :: TopLevelModuleName -> RawTopLevelModuleName

Converts a top-level module name to a raw top-level module name.

rawTopLevelModuleNameForModule :: Module -> RawTopLevelModuleName

Computes the top-level module name.

Precondition: The Module has to be well-formed. This means that there are only allowed declarations before the first module declaration, typically import declarations. See spanAllowedBeforeModule.

rawTopLevelModuleNameForModuleName :: ModuleName -> RawTopLevelModuleName

Computes the RawTopLevelModuleName corresponding to the given module name, which is assumed to represent a top-level module name.

Precondition: The module name must be well-formed.

rawTopLevelModuleNameForQName :: QName -> RawTopLevelModuleName

Turns a qualified name into a RawTopLevelModuleName. The qualified name is assumed to represent a top-level module name.

rawTopLevelModuleNameToString :: RawTopLevelModuleName -> String

Turns a raw top-level module name into a string.

topLevelModuleNameToQName :: TopLevelModuleName -> QName

A corresponding QName. The range of each Name part is the whole range of the TopLevelModuleName.

unsafeTopLevelModuleName :: RawTopLevelModuleName -> ModuleNameHash -> TopLevelModuleName

Converts a raw top-level module name and a hash to a top-level module name.

This function does not ensure that there are no hash collisions, that is taken care of by topLevelModuleName.

data RawTopLevelModuleName

Raw top-level module names (with linear-time comparisons).

Constructors

RawTopLevelModuleName

Fields

rawModuleNameRange :: Range

rawModuleNameParts :: TopLevelModuleNameParts

rawModuleNameInferred :: !Bool

Was this module name constructed from a file name rather than declared in the file?

type TopLevelModuleName = TopLevelModuleName' Range

Top-level module names (with constant-time comparisons).

module Agda.Syntax.TopLevelModuleName.Boot

Orphan instances

Pretty TopLevelModuleName

Instance details

Methods

pretty :: TopLevelModuleName -> Doc

prettyPrec :: Int -> TopLevelModuleName -> Doc

prettyList :: [TopLevelModuleName] -> Doc

Sized TopLevelModuleName

Instance details

Methods

size :: TopLevelModuleName -> Int

natSize :: TopLevelModuleName -> Peano