Agda.Syntax.Treeless

Description

The treeless syntax is intended to be used as input for the compiler backends. It is more low-level than Internal syntax and is not used for type checking.

Some of the features of treeless syntax are: - case expressions instead of case trees - no instantiated datatypes / constructors

Documentation

module Agda.Syntax.Abstract.Name

type Args = [TTerm]

data TTerm

Treeless Term. All local variables are using de Bruijn indices.

Constructors

TVar Int

TPrim TPrim

TDef QName

TApp TTerm Args

TLam TTerm

TLit Literal

TCon QName

TLet TTerm TTerm
introduces a new (non-recursive) local binding. The bound term MUST only be evaluated if it is used inside the body. Sharing may happen, but is optional. It is also perfectly valid to just inline the bound term in the body.

TCase Int CaseInfo TTerm [TAlt]
Case scrutinee (always variable), case type, default value, alternatives First, all TACon alternatives are tried; then all TAGuard alternatives in top to bottom order. TACon alternatives must not overlap.

TUnit

TSort

TErased

TCoerce TTerm
Used by the GHC backend

TError TError
A runtime error, something bad has happened.

data ArgUsage

Usage status of function arguments in treeless code.

Constructors

ArgUsed

ArgUnused

filterUsed :: [ArgUsage] -> [a] -> [a]

filterUsed used args drops those args which are labelled ArgUnused in list used.

Specification:

  filterUsed used args = [ a | (a, ArgUsed) <- zip args $ used ++ repeat ArgUsed ]

Examples:

  filterUsed []                 == id
  filterUsed (repeat ArgUsed)   == id
  filterUsed (repeat ArgUnused) == const []

data EvaluationStrategy

The treeless compiler can behave differently depending on the target language evaluation strategy. For instance, more aggressive erasure for lazy targets.

Constructors

LazyEvaluation

EagerEvaluation

mkLet :: TTerm -> TTerm -> TTerm

Introduces a new binding

data Compiled

Constructors

Compiled

Fields

cTreeless :: TTerm

cArgUsage :: Maybe [ArgUsage]

Nothing if treeless usage analysis has not run yet.

class Unreachable a where

Methods

isUnreachable :: a -> Bool

Checks if the given expression is unreachable or not.

data TPrim

Compiler-related primitives. This are NOT the same thing as primitives in Agda's surface or internal syntax! Some of the primitives have a suffix indicating which type of arguments they take, using the following naming convention: Char | Type C | Character F | Float I | Integer Q | QName S | String

Constructors

PAdd

PAdd64

PSub

PSub64

PMul

PMul64

PQuot

PQuot64

PRem

PRem64

PGeq

PLt

PLt64

PEqI

PEq64

PEqF

PEqS

PEqC

PEqQ

PIf

PSeq

PITo64

P64ToI

data CaseInfo

Constructors

CaseInfo

Fields

caseLazy :: Bool

caseErased :: Erased

Is this a match on an erased argument?

caseType :: CaseType

data TAlt

Constructors

TACon
Matches on the given constructor. If the match succeeds, the pattern variables are prepended to the current environment (pushes all existing variables aArity steps further away)

Fields

aCon :: QName

aArity :: Int

aBody :: TTerm

TAGuard
Binds no variables

The guard must only use the variable that the case expression matches on.

Fields

aGuard :: TTerm

aBody :: TTerm

TALit

Fields

aLit :: Literal

aBody :: TTerm

data TError

Constructors

TUnreachable
Code which is unreachable. E.g. absurd branches or missing case defaults. Runtime behaviour of unreachable code is undefined, but preferably the program will exit with an error message. The compiler is free to assume that this code is unreachable and to remove it.

TMeta String
Code which could not be obtained because of a hole in the program. This should throw a runtime error. The string gives some information about the meta variable that got compiled.

isPrimEq :: TPrim -> Bool

coerceView :: TTerm -> (Bool, TTerm)

Strip leading coercions and indicate whether there were some.

mkTApp :: TTerm -> Args -> TTerm

tAppView :: TTerm -> (TTerm, [TTerm])

coerceAppView :: TTerm -> ((Bool, TTerm), [TTerm])

Expose the format coerce f args.

We fuse coercions, even if interleaving with applications. We assume that coercion is powerful enough to satisfy coerce (coerce f a) b = coerce f a b

tLetView :: TTerm -> ([TTerm], TTerm)

tLamView :: TTerm -> (Int, TTerm)

mkTLam :: Int -> TTerm -> TTerm

tInt :: Integer -> TTerm

intView :: TTerm -> Maybe Integer

word64View :: TTerm -> Maybe Word64

tPlusK :: Integer -> TTerm -> TTerm

tOp :: TPrim -> TTerm -> TTerm -> TTerm

tNegPlusK :: Integer -> TTerm -> TTerm

plusKView :: TTerm -> Maybe (Integer, TTerm)

negPlusKView :: TTerm -> Maybe (Integer, TTerm)

pattern TPOp :: TPrim -> TTerm -> TTerm -> TTerm

pattern TPFn :: TPrim -> TTerm -> TTerm

tUnreachable :: TTerm

tIfThenElse :: TTerm -> TTerm -> TTerm -> TTerm

data CaseType

Constructors

CTData QName

CTNat

CTInt

CTChar

CTString

CTFloat

CTQName