Agda.TypeChecking.CheckInternal

Description

A bidirectional type checker for internal syntax.

Performs checking on unreduced terms. With the exception that projection-like function applications have to be reduced since they break bidirectionality.

Documentation

type MonadCheckInternal (m :: Type -> Type) = MonadConversion m

checkType :: MonadCheckInternal m => Type -> m ()

Entry point for e.g. checking WithFunctionType.

infer :: MonadCheckInternal m => Term -> m Type

Infer type of a neutral term.

inferSpine :: MonadCheckInternal m => Action m -> Type -> (Elims -> Term) -> Elims -> m (Type, Elims)

inferSpine action t hd es checks that spine es eliminates value hd [] of type t and returns the remaining type (target of elimination) and the transformed eliminations.

class CheckInternal a where

Minimal complete definition

checkInternal'

Methods

checkInternal' :: MonadCheckInternal m => Action m -> a -> Comparison -> TypeOf a -> m a

checkInternal :: MonadCheckInternal m => a -> Comparison -> TypeOf a -> m ()

inferInternal' :: (MonadCheckInternal m, TypeOf a ~ ()) => Action m -> a -> m a

inferInternal :: (MonadCheckInternal m, TypeOf a ~ ()) => a -> m ()

data Action (m :: Type -> Type)

checkInternal traverses the whole Term, and we can use this traversal to modify the term.

Constructors

Action

Fields

preAction :: Type -> Term -> m Term

Called on each subterm before the checker runs.

postAction :: Type -> Term -> m Term

Called on each subterm after the type checking.

modalityAction :: Modality -> Modality -> Modality

Called for each ArgInfo. The first Modality is from the type, the second from the term.

elimViewAction :: Term -> m Term

Called for bringing projection-like funs in post-fix form

defaultAction :: forall (m :: Type -> Type). PureTCM m => Action m

The default action is to not change the Term at all.

eraseUnusedAction :: Action TCM