Agda.TypeChecking.Free

Description

Computing the free variables of a term.

The distinction between rigid and strongly rigid occurrences comes from: Jason C. Reed, PhD thesis, 2009, page 96 (see also his LFMTP 2009 paper)

The main idea is that x = t(x) is unsolvable if x occurs strongly rigidly in t. It might have a solution if the occurrence is not strongly rigid, e.g.

x = f -> suc (f (x ( y -> k))) has x = f -> suc (f (suc k))

Jason C. Reed, PhD thesis, page 106

Under coinductive constructors, occurrences are never strongly rigid. Also, function types and lambdas do not establish strong rigidity. Only inductive constructors do so. (See issue 1271).

If you need the occurrence information for all free variables, you can use freeVars which has amoungst others this instance freeVars :: Term -> VarMap From VarMap, specific information can be extracted, e.g., relevantVars :: VarMap -> VarSet relevantVars = filterVarMap isRelevant

To just check the status of a single free variable, there are more efficient methods, e.g., freeIn :: Nat -> Term -> Bool

Tailored optimized variable checks can be implemented as semimodules to VarOcc, see, for example, VarCounts or SingleFlexRig.

Documentation

newtype VarCounts

Constructors

VarCounts

Fields

varCounts :: IntMap Int

class Free t

Gather free variables in a collection.

class (Singleton MetaId a, Semigroup a, Monoid a, Semigroup c, Monoid c) => IsVarSet a c | c -> a where

Any representation c of a set of variables need to be able to be modified by a variable occurrence. This is to ensure that free variable analysis is compositional. For instance, it should be possible to compute `fv (v [u/x])` from `fv v` and `fv u`.

In algebraic terminology, a variable set a needs to be (almost) a left semimodule to the semiring VarOcc.

Methods

withVarOcc :: VarOcc' a -> c -> c

Laws * Respects monoid operations: ``` withVarOcc o mempty == mempty withVarOcc o (x <> y) == withVarOcc o x <> withVarOcc o y ``` * Respects VarOcc composition: ``` withVarOcc oneVarOcc = id withVarOcc (composeVarOcc o1 o2) = withVarOcc o1 . withVarOcc o2 ``` * Respects VarOcc aggregation: ``` withVarOcc (o1 <> o2) x = withVarOcc o1 x <> withVarOcc o2 x ``` Since the corresponding unit law may fail, ``` withVarOcc mempty x = mempty ``` it is not quite a semimodule.

data IgnoreSorts

Where should we skip sorts in free variable analysis?

Constructors

IgnoreNot
Do not skip.

IgnoreInAnnotations
Skip when annotation to a type.

IgnoreAll
Skip unconditionally.

freeVars :: (IsVarSet a c, Singleton Variable c, Free t) => t -> c

Collect all free variables together with information about their occurrence.

Doesn't go inside solved metas, but collects the variables from a metavariable application X ts as flexibleVars.

freeVars' :: (Free t, IsVarSet a c) => t -> FreeM a c

filterVarMap :: (VarOcc -> Bool) -> VarMap -> VarSet

filterVarMapToList :: (VarOcc -> Bool) -> VarMap -> [Variable]

runFree :: (IsVarSet a c, Free t) => SingleVar c -> IgnoreSorts -> t -> c

Compute free variables.

rigidVars :: VarMap -> VarSet

Rigid variables: either strongly rigid, unguarded, or weakly rigid.

stronglyRigidVars :: VarMap -> VarSet

Variables under only and at least one inductive constructor(s).

unguardedVars :: VarMap -> VarSet

Variables at top or only under inductive record constructors λs and Πs. The purpose of recording these separately is that they can still become strongly rigid if put under a constructor whereas weakly rigid ones stay weakly rigid.

allVars :: VarMap -> VarSet

flexibleVars :: VarMap -> IntMap MetaSet

Variables occuring in arguments of metas. These are only potentially free, depending how the meta variable is instantiated. The set contains the id's of the meta variables that this variable is an argument to.

allFreeVars :: Free t => t -> VarSet

Collect all free variables.

allRelevantVars :: Free t => t -> VarSet

Collect all relevant free variables, excluding the "unused" ones.

allRelevantVarsIgnoring :: Free t => IgnoreSorts -> t -> VarSet

Collect all relevant free variables, possibly ignoring sorts.

freeVarsIgnore :: (IsVarSet a c, Singleton Variable c, Free t) => IgnoreSorts -> t -> c

freeIn :: Free a => Nat -> a -> Bool

freeInIgnoringSorts :: Free a => Nat -> a -> Bool

isBinderUsed :: Free a => Abs a -> Bool

Is the variable bound by the abstraction actually used?

relevantIn :: Free t => Nat -> t -> Bool

relevantInIgnoringSortAnn :: Free t => Nat -> t -> Bool

data FlexRig' a

Depending on the surrounding context of a variable, it's occurrence can be classified as flexible or rigid, with finer distinctions.

The constructors are listed in increasing order (wrt. information content).

Constructors

Flexible a
In arguments of metas. The set of metas is used by 'NonLinMatch' to generate the right blocking information. The semantics is that the status of a variable occurrence may change if one of the metas in the set gets solved. We may say the occurrence is tainted by the meta variables in the set.

WeaklyRigid
In arguments to variables and definitions.

Unguarded
In top position, or only under inductive record constructors (unit).

StronglyRigid
Under at least one and only inductive constructors.

type FlexRig = FlexRig' MetaSet

class LensFlexRig o a | o -> a where

Methods

lensFlexRig :: Lens' o (FlexRig' a)

isFlexible :: LensFlexRig o a => o -> Bool

isUnguarded :: LensFlexRig o a => o -> Bool

isStronglyRigid :: LensFlexRig o a => o -> Bool

isWeaklyRigid :: LensFlexRig o a => o -> Bool

data VarOcc' a

Occurrence of free variables is classified by several dimensions. Currently, we have FlexRig and Modality.

Constructors

VarOcc

Fields

varFlexRig :: FlexRig' a

varModality :: Modality

type VarOcc = VarOcc' MetaSet

varOccurrenceIn :: Free a => Nat -> a -> Maybe VarOcc

Get the full occurrence information of a free variable.

flexRigOccurrenceIn :: Free a => Nat -> a -> Maybe FlexRig

Get the full occurrence information of a free variable.

closed :: Free t => t -> Bool

Is the term entirely closed (no free variables)?

data MetaSet

A set of meta variables. Forms a monoid under union.

insertMetaSet :: MetaId -> MetaSet -> MetaSet

foldrMetaSet :: (MetaId -> a -> a) -> a -> MetaSet -> a

metaSetToBlocker :: MetaSet -> Blocker

Orphan instances

IsVarSet () VarSet

Instance details

Methods

withVarOcc :: VarOcc' () -> VarSet -> VarSet

IsVarSet () All

Instance details

Methods

withVarOcc :: VarOcc' () -> All -> All

IsVarSet () Any

Instance details

Methods

withVarOcc :: VarOcc' () -> Any -> Any

IsVarSet () [Int]

Instance details

Methods

withVarOcc :: VarOcc' () -> [Int] -> [Int]