Agda.TypeChecking.MetaVars.Occurs

Description

The occurs check for unification. Does pruning on the fly.

When hitting a meta variable:

Compute flex/rigid for its arguments.

Compare to allowed variables.

Mark arguments with rigid occurrences of disallowed variables for deletion.

Attempt to delete marked arguments.

We don't need to check for success, we can just continue occurs checking.

MetaOccursCheck: going into definitions to exclude cyclic solutions

modifyOccursCheckDefs :: (Set QName -> Set QName) -> TCM ()

initOccursCheck :: MetaVariable -> TCM ()

Set the names of definitions to be looked at to the defs in the current mutual block.

defNeedsChecking :: QName -> TCM Bool

Is a def in the list of stuff to be checked?

tallyDef :: QName -> TCM ()

Remove a def from the list of defs to be looked at.

OccursM monad and its services

data OccursExtra

Extra environment for the occurs check. (Complements FreeEnv.)

Constructors

OccursExtra

Fields

occUnfold :: UnfoldStrategy

occVars :: VarMap

The allowed variables with their variance.

occMeta :: MetaId

The meta we want to solve.

occCxtSize :: Nat

The size of the typing context upon invocation.

type OccursCtx = FreeEnv' () OccursExtra AllowedVar

type OccursM = ReaderT OccursCtx TCM

Modality handling.

type AllowedVar = Modality -> All

The passed modality is the one of the current context.

variableCheck :: VarMap -> Maybe Variable -> AllowedVar

Check whether a free variable is allowed in the context as specified by the modality.

definitionCheck :: QName -> OccursM ()

Occurs check fails if a defined name is not available since it was declared in irrelevant or erased context.

metaCheck :: MetaId -> OccursM MetaId

allowedVars :: OccursM (Nat -> Bool)

Construct a test whether a de Bruijn index is allowed or needs to be pruned.

Unfolding during occurs check.

data UnfoldStrategy

Unfold definitions during occurs check? This effectively runs the occurs check on the normal form.

Constructors

YesUnfold

NoUnfold

defArgs :: OccursM a -> OccursM a

conArgs :: Elims -> OccursM a -> OccursM a

For a path constructor `c : ... -> Path D a b`, we have that e.g. `c es i0` reduces to a. So we have to consider its arguments as flexible when we do not actually unfold.

unfoldB :: (Instantiate t, Reduce t) => t -> OccursM (Blocked t)

unfold :: (Instantiate t, Reduce t) => t -> OccursM t

Managing rigidiy during occurs check.

weakly :: OccursM a -> OccursM a

Leave the strongly rigid position.

strongly :: OccursM a -> OccursM a

flexibly :: OccursM a -> OccursM a

Error throwing during occurs check.

patternViolation' :: MonadTCM m => Blocker -> Int -> String -> m a

abort :: Blocker -> TypeError -> OccursM a

Implementation of the occurs check.

class Occurs t where

Extended occurs check.

Minimal complete definition

occurs

Methods

occurs :: t -> OccursM t

metaOccurs :: MetaId -> t -> TCM ()

default metaOccurs :: forall (f :: Type -> Type) a. (Foldable f, Occurs a, f a ~ t) => MetaId -> t -> TCM ()

occurs_ :: (Occurs t, TypeOf t ~ ()) => t -> OccursM t

metaOccurs2 :: (Occurs a, Occurs b) => MetaId -> a -> b -> TCM ()

metaOccurs3 :: (Occurs a, Occurs b, Occurs c) => MetaId -> a -> b -> c -> TCM ()

occursCheck :: MetaId -> VarMap -> Term -> TCM Term

When assigning m xs := v, check that m does not occur in v and that the free variables of v are contained in xs.

metaOccursQName :: MetaId -> QName -> TCM ()

Pruning: getting rid of flexible occurrences.

prune

Arguments

:: (PureTCM m, MonadMetaSolver m)

=> MetaId
Meta to prune.

-> Args
Arguments to meta variable.

-> (Nat -> Bool)
Test for allowed variable (de Bruijn index).

-> m PruneResult

prune m' vs xs attempts to remove all arguments from vs whose free variables are not contained in xs. If successful, m' is solved by the new, pruned meta variable and we return True else False.

Issue 1147: If any of the meta args vs is matchable, e.g., is a constructor term, we cannot prune, because the offending variables could be removed by reduction for a suitable instantiation of the meta variable.

hasBadRigid

Arguments

:: forall (m :: Type -> Type). PureTCM m

=> (Nat -> Bool)
Test for allowed variable (de Bruijn index).

-> Term
Argument of meta variable.

-> ExceptT () m Bool
Exception if argument is matchable.

hasBadRigid xs v = Just True iff one of the rigid variables in v is not in xs. Actually we can only prune if a bad variable is in the head. See issue 458. Or in a non-eliminateable position (see succeed/PruningNonMillerPattern).

hasBadRigid xs v = Nothing means that we cannot prune at all as one of the meta args is matchable. (See issue 1147.)

isNeutral :: HasConstInfo m => Blocked t -> QName -> Elims -> m Bool

Check whether a term Def f es is finally stuck. Currently, we give only a crude approximation.

rigidVarsNotContainedIn

Arguments

:: (PureTCM m, AnyRigid a)

=> a

-> (Nat -> Bool)
Test for allowed variable (de Bruijn index).

-> m Bool

Check whether any of the variables (given as de Bruijn indices) occurs *definitely* in the term in a rigid position. Reduces the term successively to remove variables in dead subterms. This fixes issue 1386.

class AnyRigid a where

Collect the *definitely* rigid variables in a monoid. We need to successively reduce the expression to do this.

Methods

anyRigid :: PureTCM tcm => (Nat -> tcm Bool) -> a -> tcm Bool

data PruneResult

Constructors

NothingToPrune
the kill list is empty or only Falses

PrunedNothing
there is no possible kill (because of type dep.)

PrunedSomething
managed to kill some args in the list

PrunedEverything
all prescribed kills where performed

killArgs :: MonadMetaSolver m => [Bool] -> MetaId -> m PruneResult

killArgs [k1,...,kn] X prunes argument i from metavar X if ki==True. Pruning is carried out whenever > 0 arguments can be pruned.

killedType :: MonadReduce m => [(Dom (ArgName, Type), Bool)] -> Type -> m ([Arg Bool], Type)

killedType [((x1,a1),k1)..((xn,an),kn)] b = ([k'1..k'n],t') (ignoring Dom). Let t' = (xs:as) -> b. Invariant: k'i == True iff ki == True and pruning the ith argument from type b is possible without creating unbound variables. t' is type t after pruning all k'i==True.

reallyNotFreeIn :: MonadReduce m => IntSet -> Type -> m (IntSet, Type)

performKill

Arguments

:: MonadMetaSolver m

=> [Arg Bool]
Arguments to old meta var in left to right order with Bool indicating whether they can be pruned.

-> MetaId
The old meta var to receive pruning.

-> Type
The pruned type of the new meta var.

-> m ()

Instantiate a meta variable with a new one that only takes the arguments which are not pruneable.

Orphan instances

IsVarSet () AllowedVar

Instance details

Methods

withVarOcc :: VarOcc' () -> AllowedVar -> AllowedVar