Agda.TypeChecking.Monad.Base.Types

Description

Data structures for the type checker.

Part of Agda.TypeChecking.Monad.Base, extracted to avoid import cycles.

Documentation

lensFileDictBuiltinModuleIds :: Lens' FileDictWithBuiltins BuiltinModuleIds

lensFileDictFileDictBuilder :: Lens' FileDictWithBuiltins FileDictBuilder

lensFileDictPrimitiveLibDir :: Lens' FileDictWithBuiltins PrimitiveLibDir

lensPairModuleToSource :: Lens' (FileDictWithBuiltins, ModuleToSourceId) ModuleToSource

type BuiltinModuleIds = EnumMap FileId IsBuiltinModule

Collection of FileIds of primitive modules.

data Comparison

Constructors

CmpEq

CmpLeq

type Context = [ContextEntry]

The Context is a stack of ContextEntrys.

data ContextEntry

Constructors

CtxVar

Fields

ceName :: Name

ceType :: Dom Type

data FileDictWithBuiltins

Translation between AbsolutePath and FileId that also knows about Agda's builtin modules.

Constructors

FileDictWithBuiltins

Fields

fileDictBuilder :: !FileDictBuilder

(Building a) translation between AbsolutePath and FileId.

builtinModuleIds :: !BuiltinModuleIds

For the known FileIds, remember whether they refer to Agda's builtin modules.

primitiveLibDir :: !PrimitiveLibDir

The absolute path to the directory with the builtin modules. Needs to be set upon initialization.

data HighlightingLevel

How much highlighting should be sent to the user interface?

Constructors

None

NonInteractive

Interactive
This includes both non-interactive highlighting and interactive highlighting of the expression that is currently being type-checked.

data HighlightingMethod

How should highlighting be sent to the user interface?

Constructors

Direct
Via stdout.

Indirect
Both via files and via stdout.

data IPFace' t

Datatype representing a single boundary condition: x_0 = u_0, ... ,x_n = u_n ⊢ t = ?n es

Constructors

IPFace'

Fields

faceEqns :: [(t, t)]

faceRHS :: t

data IsBuiltinModule

Discern Agda's primitive modules from other file modules. @IsPrimitiveModule implies IsBuiltinModuleWithSafePostulate implies isBuiltinModule.

Constructors

IsPrimitiveModule
Very magical module, e.g. Agda.Primitive.

IsBuiltinModuleWithSafePostulates
Safe module, e.g. Agda.Builtin.Equality.

IsBuiltinModule
Any builtin module.

data ModuleToSource

Constructors

ModuleToSource

Fields

fileDict :: !FileDictWithBuiltins

moduleToSourceId :: !ModuleToSourceId

type ModuleToSourceId = Map TopLevelModuleName SourceFile

Maps top-level module names to the corresponding source file ids.

data NamedMeta

For printing, we couple a meta with its name suggestion.

Constructors

NamedMeta

Fields

nmSuggestion :: MetaNameSuggestion

nmid :: MetaId

data Polarity

Polarity for equality and subtype checking.

Constructors

Covariant
monotone

Contravariant
antitone

Invariant
no information (mixed variance)

Nonvariant
constant

type PrimitiveLibDir = AbsolutePath

newtype SourceFile

SourceFiles must exist and be registered in our file dictionary.

Constructors

SourceFile

Fields

srcFileId :: FileId

type TopLevelModuleName = TopLevelModuleName' Range

Top-level module names (with constant-time comparisons).

data FileDictBuilder

data FileId

Unique identifier of a file.

data AbsolutePath

Paths which are known to be absolute.

Note that the Eq and Ord instances do not check if different paths point to the same files or directories.