Agda.TypeChecking.Monad.Benchmark

Description

Measure CPU time for individual phases of the Agda pipeline.

Documentation

module Agda.Benchmarking

class (Ord (BenchPhase m), Functor m, MonadIO m) => MonadBench (m :: Type -> Type)

Monad with access to benchmarking data.

Minimal complete definition

getBenchmark, (putBenchmark | modifyBenchmark), finally

type family BenchPhase (m :: Type -> Type)

getBenchmark :: MonadBench m => m (Benchmark (BenchPhase m))

updateBenchmarkingStatus :: TCM ()

When profile options are set or changed, we need to turn benchmarking on or off.

billTo :: MonadBench m => Account (BenchPhase m) -> m c -> m c

Bill a computation to a specific account. Works even if the computation is aborted by an exception.

billPureTo :: MonadBench m => Account (BenchPhase m) -> c -> m c

Bill a pure computation to a specific account.

billToCPS :: MonadBench m => Account (BenchPhase m) -> ((b -> m c) -> m c) -> (b -> m c) -> m c

Bill a CPS function to an account. Can't handle exceptions.

reset :: MonadBench m => m ()

Resets the account and the timing information.

print :: MonadTCM tcm => tcm ()

Prints the accumulated benchmark results. Does nothing if no benchmark profiling is enabled.