Agda.TypeChecking.Names

Description

EDSL to construct terms without touching De Bruijn indices.

e.g. given t, u :: Term, Γ ⊢ t, u : A, we can build "λ f. f t u" like this:

runNames [] $ do -- open binds t and u to computations that know how to weaken themselves in -- an extended context

t,u

<- mapM open [t,u]

- lam gives the illusion of HOAS by providing f as a computation.

- It also extends the internal context with the name "f", so that

- t and u will get weakened in the body.

- We apply f with the (@) combinator from Agda.TypeChecking.Primitive.

lam "f" $ f -> f @ t @ u

Documentation

newtype NamesT (m :: Type -> Type) a

Constructors

NamesT

Fields

unName :: ReaderT Names m a

type Names = [String]

A list of variable names from a context.

runNamesT :: Names -> NamesT m a -> m a

runNames :: Names -> NamesT Identity a -> a

currentCxt :: forall (m :: Type -> Type). Monad m => NamesT m Names

cxtSubst :: forall (m :: Type -> Type) a. Monad m => Names -> NamesT m (Substitution' a)

cxtSubst Γ returns the substitution needed to go from Γ to the current context.

Precondition Γ is not a subcontext of the current one.

inCxt :: forall (m :: Type -> Type) a. (Monad m, Subst a) => Names -> a -> NamesT m a

inCxt Γ t takes a t in context Γ and produce an action that will return t weakened to the current context.

Fails whenever cxtSubst Γ would.

cl' :: forall (m :: Type -> Type) a. Applicative m => a -> NamesT m a

Closed terms

cl :: Monad m => m a -> NamesT m a

open :: forall (m :: Type -> Type) a. (Monad m, Subst a) => a -> NamesT m (NamesT m a)

Open terms in the current context.

type Var (m :: Type -> Type) = forall b. (Subst b, DeBruijn b) => NamesT m b

Monadic actions standing for variables.

b is quantified over so the same variable can be used e.g. both as a pattern and as an expression.

bind :: forall (m :: Type -> Type) a. Monad m => ArgName -> ((forall b. (Subst b, DeBruijn b) => NamesT m b) -> NamesT m a) -> NamesT m (Abs a)

bind n f provides f with a fresh variable, which can be used in any extended context.

Returns an Abs which binds the extra variable.

bind' :: forall (m :: Type -> Type) a. Monad m => ArgName -> ((forall b. (Subst b, DeBruijn b) => NamesT m b) -> NamesT m a) -> NamesT m a

Like bind but returns a bare term.

Helpers to build lambda abstractions.

glam :: forall (m :: Type -> Type). Monad m => ArgInfo -> ArgName -> (NamesT m Term -> NamesT m Term) -> NamesT m Term

lam :: forall (m :: Type -> Type). Monad m => ArgName -> (NamesT m Term -> NamesT m Term) -> NamesT m Term

ilam :: forall (m :: Type -> Type). Monad m => ArgName -> (NamesT m Term -> NamesT m Term) -> NamesT m Term

Combinators for n-ary binders.

data AbsN a

Constructors

AbsN

Fields

absNName :: [ArgName]

unAbsN :: a

toAbsN :: Abs (AbsN a) -> AbsN a

Will crash on NoAbs

absAppN :: Subst a => AbsN a -> [SubstArg a] -> a

type ArgVars (m :: Type -> Type) = forall b. (Subst b, DeBruijn b) => [NamesT m (Arg b)]

type Vars (m :: Type -> Type) = forall b. (Subst b, DeBruijn b) => [NamesT m b]

bindN :: forall (m :: Type -> Type) a. Monad m => [ArgName] -> (Vars m -> NamesT m a) -> NamesT m (AbsN a)

bindNArg :: forall (m :: Type -> Type) a. Monad m => [Arg ArgName] -> (ArgVars m -> NamesT m a) -> NamesT m (AbsN a)

type Vars1 (m :: Type -> Type) = forall b. (Subst b, DeBruijn b) => List1 (NamesT m b)

bindN1 :: forall (m :: Type -> Type) a. Monad m => List1 ArgName -> (Vars1 m -> NamesT m a) -> NamesT m (AbsN a)

glamN :: forall (m :: Type -> Type). (Functor m, Monad m) => [Arg ArgName] -> (NamesT m Args -> NamesT m Term) -> NamesT m Term

applyN :: forall (m :: Type -> Type) a. (Monad m, Subst a) => NamesT m (AbsN a) -> [NamesT m (SubstArg a)] -> NamesT m a

applyN' :: forall (m :: Type -> Type) a. (Monad m, Subst a) => NamesT m (AbsN a) -> NamesT m [SubstArg a] -> NamesT m a

abstractN :: forall (m :: Type -> Type) a. (Monad m, Abstract a) => NamesT m Telescope -> (Vars m -> NamesT m a) -> NamesT m a

abstractT :: forall (m :: Type -> Type) a. (Monad m, Abstract a) => String -> NamesT m Type -> (Var m -> NamesT m a) -> NamesT m a

lamTel :: forall (m :: Type -> Type). Monad m => NamesT m (Abs [Term]) -> NamesT m [Term]

appTel :: forall (m :: Type -> Type). Monad m => NamesT m [Term] -> NamesT m Term -> NamesT m [Term]