Agda.TypeChecking.Positivity

Description

Check that a datatype is strictly positive.

Documentation

type Graph n e = Graph n e

checkStrictlyPositive :: MutualInfo -> Set QName -> TCM ()

Check that the datatypes in the mutual block containing the given declarations are strictly positive.

Find polarity of datatypes parameters and indices.

Also add information about positivity and recursivity of records to the signature.

getDefArity :: Definition -> TCM Int

data Item

Constructors

AnArg Nat [Occurrence]

ADef QName

type Occurrences = Map Item [OccursWhere]

data OccurrencesBuilder

Used to build Occurrences and occurrence graphs.

Constructors

Concat [OccurrencesBuilder]

OccursAs Where OccurrencesBuilder

OccursHere Item

OnlyVarsUpTo Nat OccurrencesBuilder
OnlyVarsUpTo n occs discards occurrences of de Bruijn index >= n.

data OccurrencesBuilder'

Used to build Occurrences and occurrence graphs.

Constructors

Concat' [OccurrencesBuilder']

OccursAs' Where OccurrencesBuilder'

OccursHere' Item

preprocess :: OccurrencesBuilder -> OccurrencesBuilder'

Removes OnlyVarsUpTo entries.

flatten :: OccurrencesBuilder -> Map Item Integer

An interpreter for OccurrencesBuilder.

WARNING: There can be lots of sharing between the generated OccursWhere entries. Traversing all of these entries could be expensive. (See computeEdges for an example.)

data OccEnv

Context for computing occurrences.

Constructors

OccEnv

Fields

vars :: [Maybe Item]

Items corresponding to the free variables.

Potential invariant: It seems as if the list has the form genericReplicate n Nothing ++ map (Just . AnArg) is, for some n and is, where is is decreasing (non-strictly).

inf :: Maybe QName

Name for ∞ builtin.

type OccM = ReaderT OccEnv ReduceM

Monad for computing occurrences.

withExtendedOccEnv :: Maybe Item -> OccM a -> OccM a

withExtendedOccEnv' :: [Maybe Item] -> OccM a -> OccM a

getOccurrences

Arguments

:: (Show a, PrettyTCM a, ComputeOccurrences a)

=> [Maybe Item]
Extension of the OccEnv, usually a local variable context.

-> a

-> TCM OccurrencesBuilder

Running the monad

class ComputeOccurrences a where

Minimal complete definition

Nothing

Methods

occurrences :: a -> OccM OccurrencesBuilder

default occurrences :: forall (t :: Type -> Type) b. (Foldable t, ComputeOccurrences b, t b ~ a) => a -> OccM OccurrencesBuilder

computeOccurrences :: QName -> TCM (Map Item Integer)

Computes the number of occurrences of different Items in the given definition.

WARNING: There can be lots of sharing between the OccursWhere entries. Traversing all of these entries could be expensive. (See computeEdges for an example.)

computeOccurrences' :: QName -> TCM OccurrencesBuilder

Computes the occurrences in the given definition.

data Node

Constructors

DefNode !QName

ArgNode !QName !Nat

data Edge a

Edge labels for the positivity graph.

Constructors

Edge !Occurrence a

mergeEdges :: Edge a -> Edge a -> Edge a

Merges two edges between the same source and target.

buildOccurrenceGraph :: Set QName -> TCM (Graph Node (Edge OccursWhere))

WARNING: There can be lots of sharing between the OccursWhere entries in the edges. Traversing all of these entries could be expensive. (See computeEdges for an example.)

computeEdges

Arguments

:: Set QName
The names in the current mutual block.

-> QName
The current name.

-> OccurrencesBuilder

-> TCM [Edge Node (Edge OccursWhere)]

Computes all non-ozero occurrence graph edges represented by the given OccurrencesBuilder.

WARNING: There can be lots of sharing between the OccursWhere entries in the edges. Traversing all of these entries could be expensive. For instance, for the function F in benchmarkmiscSlowOccurrences.agda a large number of edges from the argument X to the function F are computed. These edges have polarity StrictPos, JustNeg or JustPos, and contain the following OccursWhere elements:

OccursWhere _ empty (fromList [InDefOf F, InClause 0]),

OccursWhere _ empty (fromList [InDefOf F, InClause 0, LeftOfArrow]),

OccursWhere _ empty (fromList [InDefOf F, InClause 0, LeftOfArrow, LeftOfArrow]),

OccursWhere _ empty (fromList [InDefOf F, InClause 0, LeftOfArrow, LeftOfArrow, LeftOfArrow]),

and so on.

Orphan instances

PrettyTCM (Seq OccursWhere)

Instance details

Methods

prettyTCM :: MonadPretty m => Seq OccursWhere -> m Doc