Agda.TypeChecking.Positivity.Occurrence

Description

Occurrences.

Documentation

type PragmaPolarities = List1 (Ranged Occurrence)

List of polarities stemming from POLARITY pragma.

data Occurrence

Subterm occurrences for positivity checking. The constructors are listed in increasing information they provide: Mixed <= JustPos <= StrictPos <= GuardPos <= Unused Mixed <= JustNeg <= Unused.

Constructors

Mixed
Arbitrary occurrence (positive and negative).

JustNeg
Negative occurrence.

JustPos
Positive occurrence, but not strictly positive.

StrictPos
Strictly positive occurrence.

GuardPos
Guarded strictly positive occurrence (i.e., under ∞). For checking recursive records.

Unused

data OccursWhere

Description of an occurrence.

Constructors

OccursWhere Range (Seq Where) (Seq Where)
The elements of the sequences, read from left to right, explain how to get to the occurrence. The second sequence includes the main information, and if the first sequence is non-empty, then it includes information about the context of the second sequence.

data Where

One part of the description of an occurrence.

Constructors

LeftOfArrow

DefArg QName Nat
in the nth argument of a define constant

UnderInf
in the principal argument of built-in ∞

VarArg Occurrence Nat
as an argument to a bound variable. The polarity, if given, is the polarity of the argument the occurence is in

MetaArg
as an argument of a metavariable

ConArgType QName
in the type of a constructor

IndArgType QName
in a datatype index of a constructor

ConEndpoint QName
in an endpoint of a higher constructor

InClause Nat
in the nth clause of a defined function

Matched
matched against in a clause of a defined function

IsIndex
is an index of an inductive family

InDefOf QName
in the definition of a constant

boundToEverySome :: Map Occurrence [(Occurrence -> Bool, Occurrence -> Bool)]

The map contains bindings of the form bound |-> ess, satisfying the following property: for every non-empty list w, foldr1 otimes w <= bound iff or [ all every w && any some w | (every, some) <- ess ].

productOfEdgesInBoundedWalk :: (SemiRing e, Ord n) => (e -> Occurrence) -> Graph n e -> n -> n -> Occurrence -> Maybe e

productOfEdgesInBoundedWalk occ g u v bound returns a value distinct from Nothing iff there is a walk c (a list of edges) in g, from u to v, for which the product foldr1 otimes (map occ c) <= bound. In this case the returned value is Just (foldr1 otimes c) for one such walk c.

Preconditions: u and v must belong to g, and bound must belong to the domain of boundToEverySome.

modalPolarityToOccurrence :: ModalPolarity -> Occurrence