Agda.TypeChecking.Pretty

Documentation

($$) :: Applicative m => m Doc -> m Doc -> m Doc infixl 5

($+$) :: Applicative m => m Doc -> m Doc -> m Doc infixl 5

(<+>) :: Applicative m => m Doc -> m Doc -> m Doc infixl 6

(<?>) :: Applicative m => m Doc -> m Doc -> m Doc infixl 6

braces :: Functor m => m Doc -> m Doc

brackets :: Functor m => m Doc -> m Doc

colon :: Applicative m => m Doc

comma :: Applicative m => m Doc

dbraces :: Functor m => m Doc -> m Doc

doubleQuotes :: Functor m => m Doc -> m Doc

equals :: Applicative m => m Doc

fsep :: (Applicative m, Foldable t) => t (m Doc) -> m Doc

fwords :: Applicative m => String -> m Doc

githubIssue :: Applicative m => Int -> m Doc

Link to an issue on the Agda bug tracker.

hang :: Applicative m => m Doc -> Int -> m Doc -> m Doc

hcat :: (Applicative m, Foldable t) => t (m Doc) -> m Doc

hsep :: (Applicative m, Foldable t) => t (m Doc) -> m Doc

multiLineText :: Applicative m => String -> m Doc

nest :: Functor m => Int -> m Doc -> m Doc

parens :: Functor m => m Doc -> m Doc

parensNonEmpty :: Functor m => m Doc -> m Doc

pluralS :: (Functor m, Sized a) => a -> m Doc -> m Doc

pretty :: (Applicative m, Pretty a) => a -> m Doc

prettyA :: (ToConcrete a, Pretty (ConOfAbs a), MonadAbsToCon m) => a -> m Doc

prettyAs :: (ToConcrete a, ConOfAbs a ~ [ce], Pretty ce, MonadAbsToCon m) => a -> m Doc

prettyList :: (Applicative m, Foldable t) => t (m Doc) -> m Doc

Comma-separated list in brackets.

prettyList_ :: (Applicative m, Semigroup (m Doc), Foldable t) => t (m Doc) -> m Doc

prettyList without the brackets.

prettyR :: (ToAbstract r, PrettyTCM (AbsOfRef r), MonadPretty m, MonadError TCErr m) => r -> m Doc

For unquote.

prettyTCMCtx :: (PrettyTCM a, MonadPretty m) => Precedence -> a -> m Doc

Pretty print with a given context precedence

prettyTCMPatternList :: MonadPretty m => [NamedArg DeBruijnPattern] -> m Doc

prettyTCMPatterns :: MonadPretty m => [NamedArg DeBruijnPattern] -> m [Doc]

Proper pretty printing of patterns:

pshow :: (Applicative m, Show a) => a -> m Doc

punctuate :: (Applicative m, Semigroup (m Doc), Foldable t) => m Doc -> t (m Doc) -> [m Doc]

pwords :: Applicative m => String -> [m Doc]

quotes :: Functor m => m Doc -> m Doc

sep :: (Applicative m, Foldable t) => t (m Doc) -> m Doc

sequenceAFoldable :: (Applicative m, Foldable t) => t (m a) -> m [a]

sequenceAFoldable == sequenceA . Fold.toList

superscript :: Applicative m => Int -> m Doc

text :: Applicative m => String -> m Doc

url :: Applicative m => String -> m Doc

vcat :: (Applicative m, Foldable t) => t (m Doc) -> m Doc

vsep :: (Applicative m, Foldable t) => t (m Doc) -> m Doc

type Doc = Doc

type MonadPretty (m :: Type -> Type) = MonadAbsToCon m

newtype PrettyContext

Constructors

PrettyContext Context

class PrettyTCM a where

Methods

prettyTCM :: MonadPretty m => a -> m Doc

class PrettyTCMWithNode a where

Pretty-print something paired with a (printable) node. | This intermediate typeclass exists to avoid UndecidableInstances.

Methods

prettyTCMWithNode :: (PrettyTCM n, MonadPretty m) => WithNode n a -> m Doc

data WithNode n a

Pairing something with a node (for printing only).

Constructors

WithNode n a

type MonadAbsToCon (m :: Type -> Type) = (MonadFresh NameId m, MonadInteractionPoints m, MonadStConcreteNames m, HasOptions m, PureTCM m, IsString (m Doc), Null (m Doc), Semigroup (m Doc))

Preconditions to run the AbstractToConcrete translation.

Orphan instances

Semigroup (TCM Doc)
This instance is more specific than a generic instance Semigroup a => Semigroup (TCM a).

Instance details

Methods

(<>) :: TCM Doc -> TCM Doc -> TCM Doc

sconcat :: NonEmpty (TCM Doc) -> TCM Doc

stimes :: Integral b => b -> TCM Doc -> TCM Doc