Agda.TypeChecking.Pretty.Warning

Documentation

prettyWarningName :: MonadPretty m => WarningName -> m Doc

Prefix for a warning text showing name of the warning. E.g. warning: -W[no]warning_name

prettyWarning :: MonadPretty m => Warning -> m Doc

prettyRecordFieldWarning :: MonadPretty m => RecordFieldWarning -> m Doc

prettyDuplicateFields :: MonadPretty m => [Name] -> m Doc

prettyTooManyFields :: MonadPretty m => QName -> [Name] -> [Name] -> m Doc

prettyNotInScopeNames

Arguments

:: (MonadPretty m, Pretty a, HasRange a)

=> Bool
Print range?

-> (a -> m Doc)
Correction suggestion generator.

-> [a]
Names that are not in scope.

-> m Doc

Report a number of names that are not in scope.

didYouMean

Arguments

:: (MonadPretty m, Pretty a, Pretty b)

=> [QName]
Names in scope.

-> (a -> b)
Canonization function for similarity search.

-> a
A name which is not in scope.

-> Maybe (m Doc)
"did you mean" hint.

Suggest some corrections to a misspelled name.

prettyTCWarnings :: [TCWarning] -> TCM String

renderTCWarnings' :: [TCWarning] -> TCM [String]

prettyTCWarnings' :: [TCWarning] -> TCM [Doc]

filterTCWarnings :: [TCWarning] -> [TCWarning]

If there are several warnings, remove the unsolved-constraints warning in case there are no interesting constraints to list.

tcWarningsToError :: [TCWarning] -> TCM ()

Turns warnings, if any, into errors.

applyFlagsToTCWarningsPreserving :: HasOptions m => Set WarningName -> [TCWarning] -> m [TCWarning]

Depending which flags are set, one may happily ignore some warnings.

applyFlagsToTCWarnings :: HasOptions m => [TCWarning] -> m [TCWarning]

isBoundaryConstraint :: (ReadTCState m, MonadTCM m) => ProblemConstraint -> m (Maybe Range)

getAllUnsolvedWarnings :: (MonadFail m, ReadTCState m, MonadWarning m, MonadTCM m) => m [TCWarning]

getAllWarnings :: (MonadFail m, ReadTCState m, MonadWarning m, MonadTCM m) => WhichWarnings -> m [TCWarning]

Collect all warnings that have accumulated in the state.

getAllWarningsPreserving :: (MonadFail m, ReadTCState m, MonadWarning m, MonadTCM m) => Set WarningName -> WhichWarnings -> m [TCWarning]

getAllWarningsOfTCErr :: TCErr -> TCM [TCWarning]

Orphan instances

PrettyTCM TCWarning

Instance details

Methods

prettyTCM :: MonadPretty m => TCWarning -> m Doc