Agda.TypeChecking.Primitive.Cubical

Documentation

requireCubical

Arguments

:: Cubical
Which variant of Cubical Agda is required?

-> String

-> TCM ()

Checks that the correct variant of Cubical Agda is activated. Note that --erased-cubical "counts as" --cubical in erased contexts.

primIntervalType :: (HasBuiltins m, MonadError TCErr m, MonadTCEnv m, ReadTCState m) => m Type

primINeg' :: TCM PrimitiveImpl

primDepIMin' :: TCM PrimitiveImpl

primIBin :: IntervalView -> IntervalView -> TCM PrimitiveImpl

primIMin' :: TCM PrimitiveImpl

primIMax' :: TCM PrimitiveImpl

imax :: HasBuiltins m => m Term -> m Term -> m Term

imin :: HasBuiltins m => m Term -> m Term -> m Term

primIdJ :: TCM PrimitiveImpl

primIdElim' :: TCM PrimitiveImpl

primPOr :: TCM PrimitiveImpl

primPartial' :: TCM PrimitiveImpl

primPartialP' :: TCM PrimitiveImpl

primSubOut' :: TCM PrimitiveImpl

primConId' :: TCM PrimitiveImpl

primIdFace' :: TCM PrimitiveImpl

primIdPath' :: TCM PrimitiveImpl

primTrans' :: TCM PrimitiveImpl

primHComp' :: TCM PrimitiveImpl

data TranspOrHComp

Constructors

DoTransp

DoHComp

cmdToName :: TranspOrHComp -> String

data FamilyOrNot a

Constructors

IsFam

Fields

famThing :: a

IsNot

Fields

famThing :: a

familyOrNot :: IsString p => FamilyOrNot a -> p

mkComp :: HasBuiltins m => String -> NamesT m (NamesT m Term -> NamesT m Term -> NamesT m Term -> NamesT m Term -> NamesT m Term -> NamesT m Term)

mkGComp :: HasBuiltins m => String -> NamesT m (NamesT m Term -> NamesT m Term -> NamesT m Term -> NamesT m Term -> NamesT m Term -> NamesT m Term)

Define a "ghcomp" version of gcomp. Normal comp looks like:

comp^i A [ phi -> u ] u0 = hcomp^i A(1/i) [ phi -> forward A i u ] (forward A 0 u0)

So for "gcomp" we compute:

gcomp^i A [ phi -> u ] u0 = hcomp^i A(1/i) [ phi -> forward A i u, ~ phi -> forward A 0 u0 ] (forward A 0 u0)

The point of this is that gcomp does not produce any empty systems (if phi = 0 it will reduce to "forward A 0 u".

unglueTranspGlue :: PureTCM m => Arg Term -> Arg Term -> FamilyOrNot (Arg Term, Arg Term, Arg Term, Arg Term, Arg Term, Arg Term) -> m Term

data TermPosition

Constructors

Head

Eliminated

headStop :: PureTCM m => TermPosition -> m Term -> m Bool

compGlue :: PureTCM m => TranspOrHComp -> Arg Term -> Maybe (Arg Term) -> Arg Term -> FamilyOrNot (Arg Term, Arg Term, Arg Term, Arg Term, Arg Term, Arg Term) -> TermPosition -> m (Maybe Term)

compHCompU :: PureTCM m => TranspOrHComp -> Arg Term -> Maybe (Arg Term) -> Arg Term -> FamilyOrNot (Arg Term, Arg Term, Arg Term, Arg Term) -> TermPosition -> m (Maybe Term)

primTransHComp :: TranspOrHComp -> [Arg Term] -> Int -> ReduceM (Reduced MaybeReducedArgs Term)

primComp :: TCM PrimitiveImpl

prim_glueU' :: TCM PrimitiveImpl

prim_unglueU' :: TCM PrimitiveImpl

primGlue' :: TCM PrimitiveImpl

prim_glue' :: TCM PrimitiveImpl

prim_unglue' :: TCM PrimitiveImpl

primFaceForall' :: TCM PrimitiveImpl

decomposeInterval :: HasBuiltins m => Term -> m [(IntMap Bool, [Term])]

decomposeInterval' :: HasBuiltins m => Term -> m [(IntMap BoolSet, [Term])]

transpTel :: Abs Telescope -> Term -> Args -> ExceptT (Closure (Abs Type)) TCM Args

Tries to primTransp a whole telescope of arguments, following the rule for Σ types. If a type in the telescope does not support transp, transpTel throws it as an exception.

transpTel' :: (PureTCM m, MonadError TCErr m) => Bool -> Abs Telescope -> Term -> Args -> ExceptT (Closure (Abs Type)) m Args

type LM m a = NamesT (ExceptT (Closure (Abs Type)) m) a

transpSysTel' :: forall m. (PureTCM m, MonadError TCErr m) => Bool -> Abs Telescope -> [(Term, Abs [Term])] -> Term -> Args -> ExceptT (Closure (Abs Type)) m Args

trFillTel :: Abs Telescope -> Term -> Args -> Term -> ExceptT (Closure (Abs Type)) TCM Args

Like transpTel but performing a transpFill.

trFillTel' :: (PureTCM m, MonadError TCErr m) => Bool -> Abs Telescope -> Term -> Args -> Term -> ExceptT (Closure (Abs Type)) m Args

pathTelescope :: forall m. (PureTCM m, MonadError TCErr m) => Telescope -> [Arg Term] -> [Arg Term] -> m Telescope

pathTelescope' :: forall m. (PureTCM m, MonadError (Closure Type) m) => Telescope -> [Arg Term] -> [Arg Term] -> m Telescope

combineSys :: HasBuiltins m => NamesT m Term -> NamesT m Term -> [(NamesT m Term, NamesT m Term)] -> NamesT m Term

combineSys' :: HasBuiltins m => NamesT m Term -> NamesT m Term -> [(NamesT m Term, NamesT m Term)] -> NamesT m (Term, Term)

data TranspError

Constructors

CannotTransp

Fields

errorType :: Closure (Abs Type)

tryTranspError :: TCM a -> TCM (Either (Closure (Abs Type)) a)

transpPathPTel'

Arguments

:: NamesT TCM (Abs (Abs Telescope))
j.i.Δ const on φ

-> [NamesT TCM Term]
x : (i : I) → Δ[0,i] const on φ

-> [NamesT TCM Term]
y : (i : I) → Δ[1,i] const on φ

-> NamesT TCM Term
φ

-> [NamesT TCM Term]
p : PathP (λ j → Δ[j,0]) (x 0) (y 0)

-> NamesT TCM [Arg Term]

transpPathTel'

Arguments

:: NamesT TCM (Abs Telescope)
i.Δ const on φ

-> [NamesT TCM Term]
x : (i : I) → Δ[i] const on φ

-> [NamesT TCM Term]
y : (i : I) → Δ[i] const on φ

-> NamesT TCM Term
φ

-> [NamesT TCM Term]
p : Path (Δ[0]) (x 0) (y 0)

-> NamesT TCM [Arg Term]

trFillPathTel'

Arguments

:: NamesT TCM (Abs Telescope)
i.Δ const on φ

-> [NamesT TCM Term]
x : (i : I) → Δ[i] const on φ

-> [NamesT TCM Term]
y : (i : I) → Δ[i] const on φ

-> NamesT TCM Term
φ

-> [NamesT TCM Term]
p : Path (Δ[0]) (x 0) (y 0)

-> NamesT TCM Term
r

-> NamesT TCM [Arg Term]

trFillPathPTel'

Arguments

:: NamesT TCM (Abs (Abs Telescope))
j.i.Δ const on φ

-> [NamesT TCM Term]
x : (i : I) → Δ[0,i] const on φ

-> [NamesT TCM Term]
y : (i : I) → Δ[1,i] const on φ

-> NamesT TCM Term
φ

-> [NamesT TCM Term]
p : Path ( j -> Δ[j,0]) (x 0) (y 0)

-> NamesT TCM Term
r

-> NamesT TCM [Arg Term]

expTelescope :: Type -> Telescope -> Telescope

expS :: Nat -> Substitution

Γ, Δ^I, i : I |- expS |Δ| : Γ, Δ

Special cases of Type

data LType

A Type with sort Type l Such a type supports both hcomp and transp.

Constructors

LEl Level Term

fromLType :: LType -> Type

lTypeLevel :: LType -> Level

toLType :: MonadReduce m => Type -> m (Maybe LType)

data CType

A Type that either has sort Type l or is a closed definition. Such a type supports some version of transp. In particular we want to allow the Interval as a ClosedType.

Constructors

ClosedType Sort QName

LType LType

fromCType :: CType -> Type

toCType :: MonadReduce m => Type -> m (Maybe CType)

hcomp :: (HasBuiltins m, MonadError TCErr m, MonadReduce m) => NamesT m Type -> [(NamesT m Term, NamesT m Term)] -> NamesT m Term -> NamesT m Term

transpSys :: (HasBuiltins m, MonadError TCErr m, MonadReduce m) => NamesT m (Abs Type) -> [(NamesT m Term, NamesT m Term)] -> NamesT m Term -> NamesT m Term -> NamesT m Term

debugClause :: String -> Clause -> TCM ()