Agda.TypeChecking.Rewriting.NonLinPattern

Description

Various utility functions dealing with the non-linear, higher-order patterns used for rewrite rules.

Documentation

class PatternFrom a b where

Turn a term into a non-linear pattern, treating the free variables as pattern variables. The first argument indicates the relevance we are working under: if this is Irrelevant, then we construct a pattern that never fails to match. The second argument is the number of bound variables (from pattern lambdas). The third argument is the type of the term.

Methods

patternFrom :: Relevance -> Int -> TypeOf a -> a -> TCM b

class NLPatToTerm p a where

Convert from a non-linear pattern to a term.

Minimal complete definition

Nothing

Methods

nlPatToTerm :: PureTCM m => p -> m a

default nlPatToTerm :: forall p' a' (f :: Type -> Type) m. (NLPatToTerm p' a', Traversable f, p ~ f p', a ~ f a', PureTCM m) => p -> m a

class NLPatVars a where

Gather the set of pattern variables of a non-linear pattern

Minimal complete definition

nlPatVarsUnder

Methods

nlPatVarsUnder :: Int -> a -> VarSet

nlPatVars :: a -> VarSet

class GetMatchables a where

Get all symbols that a non-linear pattern matches against

Minimal complete definition

Nothing

Methods

getMatchables :: a -> [QName]

default getMatchables :: forall (f :: Type -> Type) a'. (Foldable f, GetMatchables a', a ~ f a') => a -> [QName]

blockOnMetasIn :: (MonadBlock m, AllMetas t) => t -> m ()

assertPi :: Type -> TCM (Dom Type, Abs Type)

errNotPi :: Type -> TCM a

assertPath :: Type -> TCM (Sort, QName, Arg Term, Arg Term, Arg Term, Arg Term)

errNotPath :: Type -> TCM a

assertProjOf :: QName -> Type -> TCM (Dom Type, Abs Type)

errNotProjOf :: QName -> Type -> TCM a

assertConOf :: ConHead -> Type -> TCM Type

errNotConOf :: ConHead -> Type -> TCM a

Orphan instances

Free NLPSort

Instance details

Methods

freeVars' :: IsVarSet a c => NLPSort -> FreeM a c

Free NLPType

Instance details

Methods

freeVars' :: IsVarSet a c => NLPType -> FreeM a c

Free NLPat
Only computes free variables that are not bound (see nlPatVars), i.e., those in a PTerm.

Instance details

Methods

freeVars' :: IsVarSet a c => NLPat -> FreeM a c