Agda.TypeChecking.Serialise.Base

Documentation

data Node

Constructor tag (maybe omitted) and argument indices.

Constructors

Empty

Cons !Word32 !Node

data MemoEntry

Association lists mapping TypeRep fingerprints to values. In some cases values with different types have the same serialized representation. This structure disambiguates them.

Constructors

MEEmpty

MECons !Fingerprint !(Any :: Type) !MemoEntry

fingerprintNoinline :: TypeRep -> Fingerprint

lookupME :: Proxy a -> Fingerprint -> MemoEntry -> (a -> b) -> b -> b

newtype FreshAndReuse

Structure providing fresh identifiers for hash map and counting hash map hits (i.e. when no fresh identifier required).

Constructors

FreshAndReuse

Fields

farFresh :: Word32

Number of hash map misses.

farEmpty :: FreshAndReuse

lensFresh :: Lens' FreshAndReuse Word32

type QNameId = [NameId]

Two QNames are equal if their QNameId is equal.

qnameId :: QName -> QNameId

Computing a qualified names composed ID.

data Dict

State of the the encoder.

Constructors

Dict

Fields

nodeD :: !(HashTable Node Word32)

Written to interface file.

stringD :: !(HashTable String Word32)

Written to interface file.

lTextD :: !(HashTable Text Word32)

Written to interface file.

sTextD :: !(HashTable Text Word32)

Written to interface file.

integerD :: !(HashTable Integer Word32)

Written to interface file.

doubleD :: !(HashTable Double Word32)

Written to interface file. Dicitionaries which are not serialized, but provide short cuts to speed up serialization: Andreas, Makoto, AIM XXI Memoizing A.Name does not buy us much if we already memoize A.QName.

nameD :: !(HashTable NameId Word32)

Not written to interface file.

qnameD :: !(HashTable QNameId Word32)

Not written to interface file. Fresh UIDs and reuse statistics:

nodeC :: !(IORef FreshAndReuse)

stringC :: !(IORef FreshAndReuse)

lTextC :: !(IORef FreshAndReuse)

sTextC :: !(IORef FreshAndReuse)

integerC :: !(IORef FreshAndReuse)

doubleC :: !(IORef FreshAndReuse)

termC :: !(IORef FreshAndReuse)

nameC :: !(IORef FreshAndReuse)

qnameC :: !(IORef FreshAndReuse)

stats :: !(HashTable String Int)

collectStats :: !Bool

If True collect in stats the quantities of calls to icode for each Typeable a.

emptyDict

Arguments

:: Bool
Collect statistics for icode calls?

-> IO Dict

Creates an empty dictionary.

type Memo = IOArray Word32 MemoEntry

Univeral memo structure, to introduce sharing during decoding

data St

State of the decoder.

Constructors

St

Fields

nodeE :: !(Array Word32 [Word32])

Obtained from interface file.

stringE :: !(Array Word32 String)

Obtained from interface file.

lTextE :: !(Array Word32 Text)

Obtained from interface file.

sTextE :: !(Array Word32 Text)

Obtained from interface file.

integerE :: !(Array Word32 Integer)

Obtained from interface file.

doubleE :: !(Array Word32 Double)

Obtained from interface file.

nodeMemo :: !Memo

Created and modified by decoder. Used to introduce sharing while deserializing objects.

modFile :: !ModuleToSource

Maps module names to file names. Constructed by the decoder.

includes :: !(List1 AbsolutePath)

The include directories.

type S a = ReaderT Dict IO a

Monad used by the encoder.

type R = StateT St IO

Monad used by the decoder.

TCM is not used because the associated overheads would make decoding slower.

malformed :: R a

Throws an error which is suitable when the data stream is malformed.

class Typeable a => EmbPrj a where

Minimal complete definition

Nothing

Methods

icode

Arguments

:: a

-> S Word32
Serialization (wrapper).

icod_

Arguments

:: a

-> S Word32
Serialization (worker).

default icod_ :: (Enum a, Bounded a) => a -> S Word32

value

Arguments

:: Word32

-> R a
Deserialization.

default value :: (Enum a, Bounded a) => Word32 -> R a

goTickIcode :: Typeable a => Proxy a -> S ()

The actual logic of tickICode is cold code, so it's out-of-line, to decrease code size and avoid cache pollution.

tickICode :: Typeable a => a -> S ()

Increase entry for a in stats.

runGetState :: Get a -> ByteString -> ByteOffset -> (a, ByteString, ByteOffset)

Data.Binary.runGetState is deprecated in favour of runGetIncremental. Reimplementing it in terms of the new function. The new Decoder type contains strict byte strings so we need to be careful not to feed the entire lazy byte string to the decoder at once.

icodeX :: (Eq k, Hashable k) => (Dict -> HashTable k Word32) -> (Dict -> IORef FreshAndReuse) -> k -> S Word32

icodeInteger :: Integer -> S Word32

icodeDouble :: Double -> S Word32

icodeString :: String -> S Word32

icodeNode :: Node -> S Word32

icodeMemo

Arguments

:: (Ord a, Hashable a)

=> (Dict -> HashTable a Word32)
Memo structure for thing of key a.

-> (Dict -> IORef FreshAndReuse)
Statistics.

-> a
Key to the thing.

-> S Word32
Fallback computation to encode the thing.

-> S Word32
Encoded thing.

icode only if thing has not seen before.

vcase :: EmbPrj a => ([Word32] -> R a) -> Word32 -> R a

vcase value ix decodes thing represented by ix :: Word32 via the valu function and stores it in nodeMemo. If ix is present in nodeMemo, valu is not used, but the thing is read from nodeMemo instead.

class ICODE t (b :: Bool) where

icodeArgs proxy (a1, ..., an) maps icode over a1, ..., an and returns the corresponding list of Word32.

Methods

icodeArgs :: Proxy t -> StrictProducts (Domains t) -> S Node

icodeN :: (ICODE t (IsBase t), StrictCurrying (Domains t) (S Word32), All EmbPrj (Domains t)) => Word32 -> t -> Arrows (Domains t) (S Word32)

icodeN tag t a1 ... an serialises the arguments a1, ..., an of the constructor t together with a tag tag picked to disambiguate between different constructors. It corresponds to icodeNode . (tag :) =<< mapM icode [a1, ..., an]

icodeN' :: (ICODE t (IsBase t), StrictCurrying (Domains t) (S Word32), All EmbPrj (Domains t)) => t -> Arrows (Domains t) (S Word32)

icodeN' is the same as icodeN except that there is no tag

class VALU t (b :: Bool) where

Methods

valuN' :: t -> StrictProducts (Constant Word32 (Domains t)) -> R (CoDomain t)

valueArgs :: Proxy t -> [Word32] -> Maybe (StrictProducts (Constant Word32 (Domains t)))

valuN :: (VALU t (IsBase t), StrictCurrying (Constant Word32 (Domains t)) (R (CoDomain t)), All EmbPrj (Domains t)) => t -> Arrows (Constant Word32 (Domains t)) (R (CoDomain t))

valueN :: (VALU t (IsBase t), All EmbPrj (CoDomain t ': Domains t)) => t -> Word32 -> R (CoDomain t)