Agda.TypeChecking.Substitute

Description

This module contains the definition of hereditary substitution and application operating on internal syntax which is in β-normal form (β including projection reductions).

Further, it contains auxiliary functions which rely on substitution but not on reduction.

Documentation

pattern LargeSort :: Univ -> Integer -> SizeOfSort

pattern SmallSort :: Univ -> SizeOfSort

absV :: Dom a -> ArgName -> TelV a -> TelV a

Add given binding to the front of the telescope.

abstractArgs :: Abstract a => Args -> a -> a

applyNLPatSubst :: TermSubst a => Substitution' NLPat -> a -> a

applyNLSubstToDom :: SubstWith NLPat a => Substitution' NLPat -> Dom a -> Dom a

applyPatSubst :: TermSubst a => PatternSubstitution -> a -> a

applySubstTerm :: (Coercible t Term, EndoSubst t, Apply t) => Substitution' t -> t -> t

applyTermE :: (Coercible Term t, Apply t, EndoSubst t) => (Empty -> Term -> Elims -> Term) -> t -> Elims -> t

Apply Elims while using the given function to report ill-typed redexes. Recursive calls for applyE and applySubst happen at type t to propagate the same strategy to subtrees.

argToDontCare :: Arg Term -> Term

bindsToTel :: [Name] -> Dom Type -> ListTel

bindsToTel' :: (Name -> a) -> [Name] -> Dom Type -> ListTel' a

Turn a typed binding (x1 .. xn : A) into a telescope.

bindsToTel'1 :: (Name -> a) -> List1 Name -> Dom Type -> ListTel' a

bindsToTel1 :: List1 Name -> Dom Type -> ListTel

canProject :: QName -> Term -> Maybe (Arg Term)

If v is a record or constructed value, canProject f v returns its field f.

compiledClauseBody :: Clause -> Maybe Term

In compiled clauses, the variables in the clause body are relative to the pattern variables (including dot patterns) instead of the clause telescope.

conApp :: (Coercible t Term, Apply t) => (Empty -> Term -> Elims -> Term) -> ConHead -> ConInfo -> Elims -> Elims -> Term

Eliminate a constructed term.

defApp :: QName -> Elims -> Elims -> Term

defApp f us vs applies Def f us to further arguments vs, eliminating top projection redexes. If us is not empty, we cannot have a projection redex, since the record argument is the first one.

domFromNamedArgName :: NamedArg Name -> Dom ()

fromPatternSubstitution :: PatternSubstitution -> Substitution

funSort :: Bool -> Sort -> Sort -> Sort

funSort' :: Bool -> Sort -> Sort -> Either Blocker Sort

Compute the sort of a function type from the sorts of its domain and codomain.

The first argument is the value of isLevelUniverseEnabled

funSortM :: HasOptions m => Sort -> Sort -> m Sort

funSortM' :: HasOptions m => Sort -> Sort -> m (Either Blocker Sort)

isSmallSort :: Sort -> Bool

lamView :: Term -> ([Arg ArgName], Term)

levelLub :: Level -> Level -> Level

Given two levels a and b, compute a ⊔ b and return its canonical form.

levelMax :: Integer -> [PlusLevel] -> Level

levelTm :: Level -> Term

mkLam :: Arg ArgName -> Term -> Term

mkPi :: Dom (ArgName, Type) -> Type -> Type

mkPi dom t = telePi (telFromList [dom]) t

mkPiSort :: Dom Type -> Abs Type -> Sort

namedBindsToTel :: [NamedArg Name] -> Type -> Telescope

Turn a typed binding (x1 .. xn : A) into a telescope.

namedBindsToTel1 :: List1 (NamedArg Name) -> Type -> Telescope

namedTelVars :: Int -> Telescope -> [NamedArg DeBruijnPattern]

piApply :: Type -> Args -> Type

(x:A)->B(x) piApply [u] = B(u)

Precondition: The type must contain the right number of pis without having to perform any reduction.

piApply is potentially unsafe, the monadic piApplyM is preferable.

piSort :: Dom Term -> Sort -> Abs Sort -> Sort

piSort' :: Dom Term -> Sort -> Abs Sort -> Either Blocker Sort

Compute the sort of a pi type from three inputs: 1. The "raw" domain of the pi type (without the sort) 2. The sort of the domain 3. The sort of the codomain (which lives in an extended context)

Note that unlike funSort', we don't care whether --level-universe is enabled here. Instead, we just return a FunSort constructor and assume it will be simplified in the next step. See also: * `instance Reduce Sort` in Agda.TypeChecking.Substitute (this file) * inferPiSort in Agda.TypeChecking.Sort

piSortM :: HasOptions m => Dom Term -> Sort -> Abs Sort -> m Sort

projDropParsApply :: Projection -> ProjOrigin -> Relevance -> Args -> Term

projDropParsApply proj o args = projDropPars proj o `apply' args

This function is an optimization, saving us from construction lambdas we immediately remove through application.

relToDontCare :: LensRelevance a => a -> Term -> Term

renameP :: Subst a => Impossible -> Permutation -> a -> a

The permutation should permute the corresponding context. (right-to-left list)

renaming :: DeBruijn a => Impossible -> Permutation -> Substitution' a

If permute π : [a]Γ -> [a]Δ, then applySubst (renaming _ π) : Term Γ -> Term Δ

renamingR :: DeBruijn a => Permutation -> Substitution' a

If permute π : [a]Γ -> [a]Δ, then applySubst (renamingR π) : Term Δ -> Term Γ

sizeOfSort :: Sort -> Either Blocker SizeOfSort

Returns Left blocker for unknown (blocked) sorts, and otherwise returns Right s where s indicates the size and fibrancy.

sort :: Sort -> Type

ssort :: Level -> Type

telVars :: Int -> Telescope -> [Arg DeBruijnPattern]

telView' :: Type -> TelView

Takes off all exposed function domains from the given type. This means that it does not reduce to expose Pi-types.

telView'UpTo :: Int -> Type -> TelView

telView'UpTo n t takes off the first n exposed function types of t. Takes off all (exposed ones) if n < 0.

teleLam :: Telescope -> Term -> Term

Abstract over a telescope in a term, producing lambdas. Dumb abstraction: Always produces Abs, never NoAbs.

The implementation is sound because Telescope does not use NoAbs.

telePi :: Telescope -> Type -> Type

Uses free variable analysis to introduce NoAbs bindings.

telePi' :: (Abs Type -> Abs Type) -> Telescope -> Type -> Type

telePiVisible :: Telescope -> Type -> Type

Only abstract the visible components of the telescope, and all that bind variables. Everything will be an Abs! Caution: quadratic time!

telePi_ :: Telescope -> Type -> Type

Everything will be an Abs.

typeArgsWithTel :: Telescope -> [Term] -> Maybe [Dom Type]

Given arguments vs : tel (vector typing), extract their individual types. Returns Nothing is tel is not long enough.

univSort :: Sort -> Sort

univSort' :: Sort -> Either Blocker Sort

univSort' univInf s gets the next higher sort of s, if it is known (i.e. it is not just UnivSort s).

Precondition: s is reduced

unlamView :: [Arg ArgName] -> Term -> Term

usePatOrigin :: PatOrigin -> Pattern' a -> Pattern' a

usePatternInfo :: PatternInfo -> Pattern' a -> Pattern' a

data SizeOfSort

A sort can either be small (Set l, Prop l, Size, ...) or large (Setω n).

Constructors

SizeOfSort

Fields

szSortUniv :: Univ

szSortSize :: Integer

data TelV a

Constructors

TelV

Fields

theTel :: Tele (Dom a)

theCore :: a

type TelView = TelV Type

class TeleNoAbs a where

Performs void (noAbs) abstraction over telescope.

Methods

teleNoAbs :: a -> Term -> Term

module Agda.TypeChecking.Substitute.Class

module Agda.TypeChecking.Substitute.DeBruijn

data Substitution' a

Substitutions.

Constructors

IdS
Identity substitution. Γ ⊢ IdS : Γ

EmptyS Impossible
Empty substitution, lifts from the empty context. First argument is IMPOSSIBLE. Apply this to closed terms you want to use in a non-empty context. Γ ⊢ EmptyS : ()

a :# (Substitution' a) infixr 4
Substitution extension, `cons'. Γ ⊢ u : Aρ Γ ⊢ ρ : Δ ---------------------- Γ ⊢ u :# ρ : Δ, A

Strengthen Impossible !Int (Substitution' a)
Strengthening substitution. First argument is IMPOSSIBLE. In 'Strengthen err n ρ the number n must be non-negative. This substitution should only be applied to values t for which none of the variables 0 up to n - 1 are free in t[ρ], and in that case n is subtracted from all free de Bruijn indices in t[ρ]. Γ ⊢ ρ : Δ |Θ| = n --------------------------- Γ ⊢ Strengthen n ρ : Δ, Θ @

Wk !Int (Substitution' a)
Weakening substitution, lifts to an extended context. Γ ⊢ ρ : Δ ------------------- Γ, Ψ ⊢ Wk |Ψ| ρ : Δ

Lift !Int (Substitution' a)
Lifting substitution. Use this to go under a binder. Lift 1 ρ == var 0 :# Wk 1 ρ. Γ ⊢ ρ : Δ ------------------------- Γ, Ψρ ⊢ Lift |Ψ| ρ : Δ, Ψ

type Substitution = Substitution' Term

Orphan instances

Abstract Clause

Instance details

Methods

abstract :: Telescope -> Clause -> Clause

Abstract Sort

Instance details

Methods

abstract :: Telescope -> Sort -> Sort

Abstract Telescope

Instance details

Methods

abstract :: Telescope -> Telescope -> Telescope

Abstract Term

Instance details

Methods

abstract :: Telescope -> Term -> Term

Abstract Type

Instance details

Methods

abstract :: Telescope -> Type -> Type

Abstract CompiledClauses

Instance details

Methods

abstract :: Telescope -> CompiledClauses -> CompiledClauses

Abstract Definition

Instance details

Methods

abstract :: Telescope -> Definition -> Definition

Abstract Defn

Instance details

Methods

abstract :: Telescope -> Defn -> Defn

Abstract FunctionInverse

Instance details

Methods

abstract :: Telescope -> FunctionInverse -> FunctionInverse

Abstract NumGeneralizableArgs

Instance details

Methods

abstract :: Telescope -> NumGeneralizableArgs -> NumGeneralizableArgs

Abstract PrimFun

Instance details

Methods

abstract :: Telescope -> PrimFun -> PrimFun

Abstract ProjLams

Instance details

Methods

abstract :: Telescope -> ProjLams -> ProjLams

Abstract Projection

Instance details

Methods

abstract :: Telescope -> Projection -> Projection

Abstract RewriteRule
tel ⊢ (Γ ⊢ lhs ↦ rhs : t) becomes tel, Γ ⊢ lhs ↦ rhs : t) we do not need to change lhs, rhs, and t since they live in Γ. See 'Abstract Clause'.

Instance details

Methods

abstract :: Telescope -> RewriteRule -> RewriteRule

Abstract System

Instance details

Methods

abstract :: Telescope -> System -> System

Abstract Permutation

Instance details

Methods

abstract :: Telescope -> Permutation -> Permutation

Apply BraveTerm

Instance details

Methods

apply :: BraveTerm -> Args -> BraveTerm

applyE :: BraveTerm -> Elims -> BraveTerm

Apply Clause

Instance details

Methods

apply :: Clause -> Args -> Clause

applyE :: Clause -> Elims -> Clause

Apply Sort

Instance details

Methods

apply :: Sort -> Args -> Sort

applyE :: Sort -> Elims -> Sort

Apply Term

Instance details

Methods

apply :: Term -> Args -> Term

applyE :: Term -> Elims -> Term

Apply CompiledClauses

Instance details

Methods

apply :: CompiledClauses -> Args -> CompiledClauses

applyE :: CompiledClauses -> Elims -> CompiledClauses

Apply Definition

Instance details

Methods

apply :: Definition -> Args -> Definition

applyE :: Definition -> Elims -> Definition

Apply Defn

Instance details

Methods

apply :: Defn -> Args -> Defn

applyE :: Defn -> Elims -> Defn

Apply DisplayTerm

Instance details

Methods

apply :: DisplayTerm -> Args -> DisplayTerm

applyE :: DisplayTerm -> Elims -> DisplayTerm

Apply ExtLamInfo

Instance details

Methods

apply :: ExtLamInfo -> Args -> ExtLamInfo

applyE :: ExtLamInfo -> Elims -> ExtLamInfo

Apply FunctionInverse

Instance details

Methods

apply :: FunctionInverse -> Args -> FunctionInverse

applyE :: FunctionInverse -> Elims -> FunctionInverse

Apply NumGeneralizableArgs

Instance details

Methods

apply :: NumGeneralizableArgs -> Args -> NumGeneralizableArgs

applyE :: NumGeneralizableArgs -> Elims -> NumGeneralizableArgs

Apply PrimFun

Instance details

Methods

apply :: PrimFun -> Args -> PrimFun

applyE :: PrimFun -> Elims -> PrimFun

Apply ProjLams

Instance details

Methods

apply :: ProjLams -> Args -> ProjLams

applyE :: ProjLams -> Elims -> ProjLams

Apply Projection

Instance details

Methods

apply :: Projection -> Args -> Projection

applyE :: Projection -> Elims -> Projection

Apply RewriteRule

Instance details

Methods

apply :: RewriteRule -> Args -> RewriteRule

applyE :: RewriteRule -> Elims -> RewriteRule

Apply System

Instance details

Methods

apply :: System -> Args -> System

applyE :: System -> Elims -> System

Apply Permutation

Instance details

Methods

apply :: Permutation -> Args -> Permutation

applyE :: Permutation -> Elims -> Permutation

Subst ProblemEq

Instance details

Associated Types

type SubstArg ProblemEq

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg ProblemEq = Term

Methods

applySubst :: Substitution' (SubstArg ProblemEq) -> ProblemEq -> ProblemEq

Subst Name

Instance details

Associated Types

type SubstArg Name

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg Name = Term

Methods

applySubst :: Substitution' (SubstArg Name) -> Name -> Name

Subst BraveTerm

Instance details

Associated Types

type SubstArg BraveTerm

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg BraveTerm = BraveTerm

Methods

applySubst :: Substitution' (SubstArg BraveTerm) -> BraveTerm -> BraveTerm

Subst ConPatternInfo

Instance details

Associated Types

type SubstArg ConPatternInfo

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg ConPatternInfo = Term

Methods

applySubst :: Substitution' (SubstArg ConPatternInfo) -> ConPatternInfo -> ConPatternInfo

Subst DeBruijnPattern

Instance details

Associated Types

type SubstArg DeBruijnPattern

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg DeBruijnPattern = DeBruijnPattern

Methods

applySubst :: Substitution' (SubstArg DeBruijnPattern) -> DeBruijnPattern -> DeBruijnPattern

Subst EqualityTypeData

Instance details

Associated Types

type SubstArg EqualityTypeData

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg EqualityTypeData = Term

Methods

applySubst :: Substitution' (SubstArg EqualityTypeData) -> EqualityTypeData -> EqualityTypeData

Subst EqualityView

Instance details

Associated Types

type SubstArg EqualityView

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg EqualityView = Term

Methods

applySubst :: Substitution' (SubstArg EqualityView) -> EqualityView -> EqualityView

Subst Pattern

Instance details

Associated Types

type SubstArg Pattern

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg Pattern = Term

Methods

applySubst :: Substitution' (SubstArg Pattern) -> Pattern -> Pattern

Subst Term

Instance details

Associated Types

type SubstArg Term

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg Term = Term

Methods

applySubst :: Substitution' (SubstArg Term) -> Term -> Term

Subst Range

Instance details

Associated Types

type SubstArg Range

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg Range = Term

Methods

applySubst :: Substitution' (SubstArg Range) -> Range -> Range

Subst Candidate

Instance details

Associated Types

type SubstArg Candidate

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg Candidate = Term

Methods

applySubst :: Substitution' (SubstArg Candidate) -> Candidate -> Candidate

Subst CompareAs

Instance details

Associated Types

type SubstArg CompareAs

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg CompareAs = Term

Methods

applySubst :: Substitution' (SubstArg CompareAs) -> CompareAs -> CompareAs

Subst Constraint

Instance details

Associated Types

type SubstArg Constraint

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg Constraint = Term

Methods

applySubst :: Substitution' (SubstArg Constraint) -> Constraint -> Constraint

Subst DisplayForm

Instance details

Associated Types

type SubstArg DisplayForm

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg DisplayForm = Term

Methods

applySubst :: Substitution' (SubstArg DisplayForm) -> DisplayForm -> DisplayForm

Subst DisplayTerm

Instance details

Associated Types

type SubstArg DisplayTerm

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg DisplayTerm = Term

Methods

applySubst :: Substitution' (SubstArg DisplayTerm) -> DisplayTerm -> DisplayTerm

Subst LetBinding

Instance details

Associated Types

type SubstArg LetBinding

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg LetBinding = Term

Methods

applySubst :: Substitution' (SubstArg LetBinding) -> LetBinding -> LetBinding

Subst NLPSort

Instance details

Associated Types

type SubstArg NLPSort

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg NLPSort = NLPat

Methods

applySubst :: Substitution' (SubstArg NLPSort) -> NLPSort -> NLPSort

Subst NLPType

Instance details

Associated Types

type SubstArg NLPType

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg NLPType = NLPat

Methods

applySubst :: Substitution' (SubstArg NLPType) -> NLPType -> NLPType

Subst NLPat

Instance details

Associated Types

type SubstArg NLPat

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg NLPat = NLPat

Methods

applySubst :: Substitution' (SubstArg NLPat) -> NLPat -> NLPat

Subst RewriteRule

Instance details

Associated Types

type SubstArg RewriteRule

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg RewriteRule = NLPat

Methods

applySubst :: Substitution' (SubstArg RewriteRule) -> RewriteRule -> RewriteRule

Subst ContextEntry

Instance details

Associated Types

type SubstArg ContextEntry

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg ContextEntry = Term

Methods

applySubst :: Substitution' (SubstArg ContextEntry) -> ContextEntry -> ContextEntry

Subst ()

Instance details

Associated Types

type SubstArg ()

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg () = Term

Methods

applySubst :: Substitution' (SubstArg ()) -> () -> ()

DeBruijn BraveTerm

Instance details

Methods

deBruijnVar :: Int -> BraveTerm

deBruijnNamedVar :: String -> Int -> BraveTerm

deBruijnView :: BraveTerm -> Maybe Int

DeBruijn NLPat

Instance details

Methods

deBruijnVar :: Int -> NLPat

deBruijnNamedVar :: String -> Int -> NLPat

deBruijnView :: NLPat -> Maybe Int

Eq Level

Instance details

Methods

(==) :: Level -> Level -> Bool

(/=) :: Level -> Level -> Bool

Eq NotBlocked

Instance details

Methods

(==) :: NotBlocked -> NotBlocked -> Bool

(/=) :: NotBlocked -> NotBlocked -> Bool

Eq PlusLevel

Instance details

Methods

(==) :: PlusLevel -> PlusLevel -> Bool

(/=) :: PlusLevel -> PlusLevel -> Bool

Eq Sort

Instance details

Methods

(==) :: Sort -> Sort -> Bool

(/=) :: Sort -> Sort -> Bool

Eq Substitution

Instance details

Methods

(==) :: Substitution -> Substitution -> Bool

(/=) :: Substitution -> Substitution -> Bool

Eq Term
Syntactic Term equality, ignores stuff below DontCare and sharing.

Instance details

Methods

(==) :: Term -> Term -> Bool

(/=) :: Term -> Term -> Bool

Eq Candidate

Instance details

Methods

(==) :: Candidate -> Candidate -> Bool

(/=) :: Candidate -> Candidate -> Bool

Eq CandidateKind

Instance details

Methods

(==) :: CandidateKind -> CandidateKind -> Bool

(/=) :: CandidateKind -> CandidateKind -> Bool

Eq Section

Instance details

Methods

(==) :: Section -> Section -> Bool

(/=) :: Section -> Section -> Bool

Ord Level

Instance details

Methods

compare :: Level -> Level -> Ordering

(<) :: Level -> Level -> Bool

(<=) :: Level -> Level -> Bool

(>) :: Level -> Level -> Bool

(>=) :: Level -> Level -> Bool

max :: Level -> Level -> Level

min :: Level -> Level -> Level

Ord PlusLevel

Instance details

Methods

compare :: PlusLevel -> PlusLevel -> Ordering

(<) :: PlusLevel -> PlusLevel -> Bool

(<=) :: PlusLevel -> PlusLevel -> Bool

(>) :: PlusLevel -> PlusLevel -> Bool

(>=) :: PlusLevel -> PlusLevel -> Bool

max :: PlusLevel -> PlusLevel -> PlusLevel

min :: PlusLevel -> PlusLevel -> PlusLevel

Ord Sort

Instance details

Methods

compare :: Sort -> Sort -> Ordering

(<) :: Sort -> Sort -> Bool

(<=) :: Sort -> Sort -> Bool

(>) :: Sort -> Sort -> Bool

(>=) :: Sort -> Sort -> Bool

max :: Sort -> Sort -> Sort

min :: Sort -> Sort -> Sort

Ord Substitution

Instance details

Methods

compare :: Substitution -> Substitution -> Ordering

(<) :: Substitution -> Substitution -> Bool

(<=) :: Substitution -> Substitution -> Bool

(>) :: Substitution -> Substitution -> Bool

(>=) :: Substitution -> Substitution -> Bool

max :: Substitution -> Substitution -> Substitution

min :: Substitution -> Substitution -> Substitution

Ord Term

Instance details

Methods

compare :: Term -> Term -> Ordering

(<) :: Term -> Term -> Bool

(<=) :: Term -> Term -> Bool

(>) :: Term -> Term -> Bool

(>=) :: Term -> Term -> Bool

max :: Term -> Term -> Term

min :: Term -> Term -> Term

Ord Candidate

Instance details

Methods

compare :: Candidate -> Candidate -> Ordering

(<) :: Candidate -> Candidate -> Bool

(<=) :: Candidate -> Candidate -> Bool

(>) :: Candidate -> Candidate -> Bool

(>=) :: Candidate -> Candidate -> Bool

max :: Candidate -> Candidate -> Candidate

min :: Candidate -> Candidate -> Candidate

Ord CandidateKind

Instance details

Methods

compare :: CandidateKind -> CandidateKind -> Ordering

(<) :: CandidateKind -> CandidateKind -> Bool

(<=) :: CandidateKind -> CandidateKind -> Bool

(>) :: CandidateKind -> CandidateKind -> Bool

(>=) :: CandidateKind -> CandidateKind -> Bool

max :: CandidateKind -> CandidateKind -> CandidateKind

min :: CandidateKind -> CandidateKind -> CandidateKind

Abstract a => Abstract (Case a)

Instance details

Methods

abstract :: Telescope -> Case a -> Case a

Abstract a => Abstract (WithArity a)

Instance details

Methods

abstract :: Telescope -> WithArity a -> WithArity a

DoDrop a => Abstract (Drop a)

Instance details

Methods

abstract :: Telescope -> Drop a -> Drop a

Abstract t => Abstract (Maybe t)

Instance details

Methods

abstract :: Telescope -> Maybe t -> Maybe t

Abstract [Polarity]

Instance details

Methods

abstract :: Telescope -> [Polarity] -> [Polarity]

Abstract [Occurrence]

Instance details

Methods

abstract :: Telescope -> [Occurrence] -> [Occurrence]

Abstract t => Abstract [t]

Instance details

Methods

abstract :: Telescope -> [t] -> [t]

Apply t => Apply (Blocked t)

Instance details

Methods

apply :: Blocked t -> Args -> Blocked t

applyE :: Blocked t -> Elims -> Blocked t

TermSubst a => Apply (Tele a)

Instance details

Methods

apply :: Tele a -> Args -> Tele a

applyE :: Tele a -> Elims -> Tele a

Apply a => Apply (Case a)

Instance details

Methods

apply :: Case a -> Args -> Case a

applyE :: Case a -> Elims -> Case a

Apply a => Apply (WithArity a)

Instance details

Methods

apply :: WithArity a -> Args -> WithArity a

applyE :: WithArity a -> Elims -> WithArity a

DoDrop a => Apply (Drop a)

Instance details

Methods

apply :: Drop a -> Args -> Drop a

applyE :: Drop a -> Elims -> Drop a

Apply t => Apply (Maybe t)

Instance details

Methods

apply :: Maybe t -> Args -> Maybe t

applyE :: Maybe t -> Elims -> Maybe t

Apply t => Apply (Maybe t)

Instance details

Methods

apply :: Maybe t -> Args -> Maybe t

applyE :: Maybe t -> Elims -> Maybe t

Apply [NamedArg (Pattern' a)]
Make sure we only drop variable patterns.

Instance details

Methods

apply :: [NamedArg (Pattern' a)] -> Args -> [NamedArg (Pattern' a)]

applyE :: [NamedArg (Pattern' a)] -> Elims -> [NamedArg (Pattern' a)]

Apply [Polarity]

Instance details

Methods

apply :: [Polarity] -> Args -> [Polarity]

applyE :: [Polarity] -> Elims -> [Polarity]

Apply [Occurrence]

Instance details

Methods

apply :: [Occurrence] -> Args -> [Occurrence]

applyE :: [Occurrence] -> Elims -> [Occurrence]

Apply t => Apply [t]

Instance details

Methods

apply :: [t] -> Args -> [t]

applyE :: [t] -> Elims -> [t]

Subst a => Subst (Arg a)

Instance details

Associated Types

type SubstArg (Arg a)

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg (Arg a) = SubstArg a

Methods

applySubst :: Substitution' (SubstArg (Arg a)) -> Arg a -> Arg a

Subst a => Subst (Ranged a)

Instance details

Associated Types

type SubstArg (Ranged a)

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg (Ranged a) = SubstArg a

Methods

applySubst :: Substitution' (SubstArg (Ranged a)) -> Ranged a -> Ranged a

Subst a => Subst (WithHiding a)

Instance details

Associated Types

type SubstArg (WithHiding a)

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg (WithHiding a) = SubstArg a

Methods

applySubst :: Substitution' (SubstArg (WithHiding a)) -> WithHiding a -> WithHiding a

Subst a => Subst (Abs a)

Instance details

Associated Types

type SubstArg (Abs a)

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg (Abs a) = SubstArg a

Methods

applySubst :: Substitution' (SubstArg (Abs a)) -> Abs a -> Abs a

Subst a => Subst (Blocked a)

Instance details

Associated Types

type SubstArg (Blocked a)

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg (Blocked a) = SubstArg a

Methods

applySubst :: Substitution' (SubstArg (Blocked a)) -> Blocked a -> Blocked a

Subst a => Subst (Level' a)

Instance details

Associated Types

type SubstArg (Level' a)

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg (Level' a) = SubstArg a

Methods

applySubst :: Substitution' (SubstArg (Level' a)) -> Level' a -> Level' a

Subst a => Subst (PlusLevel' a)

Instance details

Associated Types

type SubstArg (PlusLevel' a)

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg (PlusLevel' a) = SubstArg a

Methods

applySubst :: Substitution' (SubstArg (PlusLevel' a)) -> PlusLevel' a -> PlusLevel' a

(Coercible a Term, Subst a) => Subst (Sort' a)

Instance details

Associated Types

type SubstArg (Sort' a)

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg (Sort' a) = SubstArg a

Methods

applySubst :: Substitution' (SubstArg (Sort' a)) -> Sort' a -> Sort' a

EndoSubst a => Subst (Substitution' a)

Instance details

Associated Types

type SubstArg (Substitution' a)

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg (Substitution' a) = a

Methods

applySubst :: Substitution' (SubstArg (Substitution' a)) -> Substitution' a -> Substitution' a

Subst a => Subst (Tele a)

Instance details

Associated Types

type SubstArg (Tele a)

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg (Tele a) = SubstArg a

Methods

applySubst :: Substitution' (SubstArg (Tele a)) -> Tele a -> Tele a

Subst a => Subst (Elim' a)

Instance details

Associated Types

type SubstArg (Elim' a)

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg (Elim' a) = SubstArg a

Methods

applySubst :: Substitution' (SubstArg (Elim' a)) -> Elim' a -> Elim' a

Subst a => Subst (List1 a)

Instance details

Associated Types

type SubstArg (List1 a)

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg (List1 a) = SubstArg a

Methods

applySubst :: Substitution' (SubstArg (List1 a)) -> List1 a -> List1 a

Subst a => Subst (Maybe a)

Instance details

Associated Types

type SubstArg (Maybe a)

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg (Maybe a) = SubstArg a

Methods

applySubst :: Substitution' (SubstArg (Maybe a)) -> Maybe a -> Maybe a

Subst a => Subst [a]

Instance details

Associated Types

type SubstArg [a]

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg [a] = SubstArg a

Methods

applySubst :: Substitution' (SubstArg [a]) -> [a] -> [a]

DeBruijn a => DeBruijn (Pattern' a)

Instance details

Methods

deBruijnVar :: Int -> Pattern' a

deBruijnNamedVar :: String -> Int -> Pattern' a

deBruijnView :: Pattern' a -> Maybe Int

(Subst a, Eq a) => Eq (Abs a)
Equality of binders relies on weakening which is a special case of renaming which is a special case of substitution.

Instance details

Methods

(==) :: Abs a -> Abs a -> Bool

(/=) :: Abs a -> Abs a -> Bool

Eq t => Eq (Blocked t)

Instance details

Methods

(==) :: Blocked t -> Blocked t -> Bool

(/=) :: Blocked t -> Blocked t -> Bool

Eq a => Eq (Pattern' a)

Instance details

Methods

(==) :: Pattern' a -> Pattern' a -> Bool

(/=) :: Pattern' a -> Pattern' a -> Bool

(Subst a, Eq a) => Eq (Tele a)

Instance details

Methods

(==) :: Tele a -> Tele a -> Bool

(/=) :: Tele a -> Tele a -> Bool

Eq a => Eq (Type' a)
Syntactic Type equality, ignores sort annotations.

Instance details

Methods

(==) :: Type' a -> Type' a -> Bool

(/=) :: Type' a -> Type' a -> Bool

(Subst a, Eq a) => Eq (Elim' a)

Instance details

Methods

(==) :: Elim' a -> Elim' a -> Bool

(/=) :: Elim' a -> Elim' a -> Bool

(Subst a, Ord a) => Ord (Abs a)

Instance details

Methods

compare :: Abs a -> Abs a -> Ordering

(<) :: Abs a -> Abs a -> Bool

(<=) :: Abs a -> Abs a -> Bool

(>) :: Abs a -> Abs a -> Bool

(>=) :: Abs a -> Abs a -> Bool

max :: Abs a -> Abs a -> Abs a

min :: Abs a -> Abs a -> Abs a

Ord a => Ord (Dom a)

Instance details

Methods

compare :: Dom a -> Dom a -> Ordering

(<) :: Dom a -> Dom a -> Bool

(<=) :: Dom a -> Dom a -> Bool

(>) :: Dom a -> Dom a -> Bool

(>=) :: Dom a -> Dom a -> Bool

max :: Dom a -> Dom a -> Dom a

min :: Dom a -> Dom a -> Dom a

(Subst a, Ord a) => Ord (Tele a)

Instance details

Methods

compare :: Tele a -> Tele a -> Ordering

(<) :: Tele a -> Tele a -> Bool

(<=) :: Tele a -> Tele a -> Bool

(>) :: Tele a -> Tele a -> Bool

(>=) :: Tele a -> Tele a -> Bool

max :: Tele a -> Tele a -> Tele a

min :: Tele a -> Tele a -> Tele a

Ord a => Ord (Type' a)

Instance details

Methods

compare :: Type' a -> Type' a -> Ordering

(<) :: Type' a -> Type' a -> Bool

(<=) :: Type' a -> Type' a -> Bool

(>) :: Type' a -> Type' a -> Bool

(>=) :: Type' a -> Type' a -> Bool

max :: Type' a -> Type' a -> Type' a

min :: Type' a -> Type' a -> Type' a

(Subst a, Ord a) => Ord (Elim' a)

Instance details

Methods

compare :: Elim' a -> Elim' a -> Ordering

(<) :: Elim' a -> Elim' a -> Bool

(<=) :: Elim' a -> Elim' a -> Bool

(>) :: Elim' a -> Elim' a -> Bool

(>=) :: Elim' a -> Elim' a -> Bool

max :: Elim' a -> Elim' a -> Elim' a

min :: Elim' a -> Elim' a -> Elim' a

Abstract v => Abstract (Map k v)

Instance details

Methods

abstract :: Telescope -> Map k v -> Map k v

Abstract v => Abstract (HashMap k v)

Instance details

Methods

abstract :: Telescope -> HashMap k v -> HashMap k v

Apply v => Apply (Map k v)

Instance details

Methods

apply :: Map k v -> Args -> Map k v

applyE :: Map k v -> Elims -> Map k v

Apply v => Apply (HashMap k v)

Instance details

Methods

apply :: HashMap k v -> Args -> HashMap k v

applyE :: HashMap k v -> Elims -> HashMap k v

(Apply a, Apply b) => Apply (a, b)

Instance details

Methods

apply :: (a, b) -> Args -> (a, b)

applyE :: (a, b) -> Elims -> (a, b)

Subst a => Subst (Named name a)

Instance details

Associated Types

type SubstArg (Named name a)

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg (Named name a) = SubstArg a

Methods

applySubst :: Substitution' (SubstArg (Named name a)) -> Named name a -> Named name a

(Subst a, Subst b, SubstArg a ~ SubstArg b) => Subst (Dom' a b)

Instance details

Associated Types

type SubstArg (Dom' a b)

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg (Dom' a b) = SubstArg a

Methods

applySubst :: Substitution' (SubstArg (Dom' a b)) -> Dom' a b -> Dom' a b

(Coercible a Term, Subst a, Subst b, SubstArg a ~ SubstArg b) => Subst (Type'' a b)

Instance details

Associated Types

type SubstArg (Type'' a b)

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg (Type'' a b) = SubstArg a

Methods

applySubst :: Substitution' (SubstArg (Type'' a b)) -> Type'' a b -> Type'' a b

(Ord k, Subst a) => Subst (Map k a)

Instance details

Associated Types

type SubstArg (Map k a)

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg (Map k a) = SubstArg a

Methods

applySubst :: Substitution' (SubstArg (Map k a)) -> Map k a -> Map k a

(Subst a, Subst b, SubstArg a ~ SubstArg b) => Subst (a, b)

Instance details

Associated Types

type SubstArg (a, b)

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg (a, b) = SubstArg a

Methods

applySubst :: Substitution' (SubstArg (a, b)) -> (a, b) -> (a, b)

(Apply a, Apply b, Apply c) => Apply (a, b, c)

Instance details

Methods

apply :: (a, b, c) -> Args -> (a, b, c)

applyE :: (a, b, c) -> Elims -> (a, b, c)

(Subst a, Subst b, Subst c, SubstArg a ~ SubstArg b, SubstArg b ~ SubstArg c) => Subst (a, b, c)

Instance details

Associated Types

type SubstArg (a, b, c)

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg (a, b, c) = SubstArg a

Methods

applySubst :: Substitution' (SubstArg (a, b, c)) -> (a, b, c) -> (a, b, c)

(Subst a, Subst b, Subst c, Subst d, SubstArg a ~ SubstArg b, SubstArg b ~ SubstArg c, SubstArg c ~ SubstArg d) => Subst (a, b, c, d)

Instance details

Associated Types

type SubstArg (a, b, c, d)

Instance details

Defined in Agda.TypeChecking.Substitute

type SubstArg (a, b, c, d) = SubstArg a

Methods

applySubst :: Substitution' (SubstArg (a, b, c, d)) -> (a, b, c, d) -> (a, b, c, d)