Agda.Utils.Benchmark

Description

Tools for benchmarking and accumulating results. Nothing Agda-specific in here.

Benchmark trie

type Account a = [a]

Account we can bill computation time to.

type CurrentAccount a = Maybe (Account a, CPUTime)

Record when we started billing the current account.

type Timings a = Trie a CPUTime

data BenchmarkOn a

Constructors

BenchmarkOff

BenchmarkOn

BenchmarkSome (Account a -> Bool)

isBenchmarkOn :: Account a -> BenchmarkOn a -> Bool

data Benchmark a

Benchmark structure is a trie, mapping accounts (phases and subphases) to CPU time spent on their performance.

Constructors

Benchmark

Fields

benchmarkOn :: !(BenchmarkOn a)

Are we benchmarking at all?

currentAccount :: !(CurrentAccount a)

What are we billing to currently?

timings :: !(Timings a)

The accounts and their accumulated timing bill.

mapBenchmarkOn :: (BenchmarkOn a -> BenchmarkOn a) -> Benchmark a -> Benchmark a

Semantic editor combinator.

mapCurrentAccount :: (CurrentAccount a -> CurrentAccount a) -> Benchmark a -> Benchmark a

Semantic editor combinator.

mapTimings :: (Timings a -> Timings a) -> Benchmark a -> Benchmark a

Semantic editor combinator.

addCPUTime :: Ord a => Account a -> CPUTime -> Benchmark a -> Benchmark a

Add to specified CPU time account.

Benchmarking monad.

class (Ord (BenchPhase m), Functor m, MonadIO m) => MonadBench (m :: Type -> Type) where

Monad with access to benchmarking data.

Minimal complete definition

getBenchmark, (putBenchmark | modifyBenchmark), finally

Associated Types

type BenchPhase (m :: Type -> Type)

Methods

getBenchmark :: m (Benchmark (BenchPhase m))

putBenchmark :: Benchmark (BenchPhase m) -> m ()

modifyBenchmark :: (Benchmark (BenchPhase m) -> Benchmark (BenchPhase m)) -> m ()

finally :: m b -> m c -> m b

We need to be able to terminate benchmarking in case of an exception.

getsBenchmark :: MonadBench m => (Benchmark (BenchPhase m) -> c) -> m c

setBenchmarking :: MonadBench m => BenchmarkOn (BenchPhase m) -> m ()

Turn benchmarking on/off.

switchBenchmarking

Arguments

:: MonadBench m

=> Maybe (Account (BenchPhase m))
Maybe new account.

-> m (Maybe (Account (BenchPhase m)))
Maybe old account.

Bill current account with time up to now. Switch to new account. Return old account (if any).

reset :: MonadBench m => m ()

Resets the account and the timing information.

billTo :: MonadBench m => Account (BenchPhase m) -> m c -> m c

Bill a computation to a specific account. Works even if the computation is aborted by an exception.

billToCPS :: MonadBench m => Account (BenchPhase m) -> ((b -> m c) -> m c) -> (b -> m c) -> m c

Bill a CPS function to an account. Can't handle exceptions.

billPureTo :: MonadBench m => Account (BenchPhase m) -> c -> m c

Bill a pure computation to a specific account.