Agda.Utils.FileName

Description

Operations on file names.

Documentation

newtype AbsolutePath

Paths which are known to be absolute.

Note that the Eq and Ord instances do not check if different paths point to the same files or directories.

Constructors

AbsolutePath Text

filePath :: AbsolutePath -> FilePath

Extract the AbsolutePath to be used as FilePath.

mkAbsolute :: FilePath -> AbsolutePath

Constructs AbsolutePaths.

Precondition: The path must be absolute and valid.

absolute :: FilePath -> IO AbsolutePath

Makes the path absolute.

This function may raise an __IMPOSSIBLE__ error if canonicalizePath does not return an absolute path.

canonicalizeAbsolutePath :: AbsolutePath -> IO AbsolutePath

Resolve symlinks etc. Preserves sameFile.

sameFile :: AbsolutePath -> AbsolutePath -> IO Bool

Tries to establish if the two file paths point to the same file (or directory). False negatives may be returned.

doesFileExistCaseSensitive :: FilePath -> IO Bool

Case-sensitive doesFileExist for Windows.

This is case-sensitive only on the file name part, not on the directory part. (Ideally, path components coming from module name components should be checked case-sensitively and the other path components should be checked case insensitively.)

isNewerThan :: FilePath -> FilePath -> IO Bool

True if the first file is newer than the second file. If a file doesn't exist it is considered to be infinitely old.

relativizeAbsolutePath

Arguments

:: AbsolutePath
The absolute path we seek to relativize.

-> AbsolutePath
The root for relativization.

-> Maybe FilePath
The relative path, if any.

A partial version of makeRelative with flipped arguments, returning Nothing if the given path cannot be relativized to the given root.

makeRelativeCanonical :: FilePath -> FilePath -> IO FilePath

Makes a path relative to a root without assuming that either path is canonical.

stripAnyOfExtensions :: [String] -> FilePath -> Maybe FilePath

Generalizes stripExtension.