Agda.Utils.HashTable

Description

Hash tables.

Documentation

data HashTable k v

Hash tables.

empty :: IO (HashTable k v)

An empty hash table.

insert :: (Eq k, Hashable k) => HashTable k v -> k -> v -> IO ()

Inserts the key and the corresponding value into the hash table.

lookup :: (Eq k, Hashable k) => HashTable k v -> k -> IO (Maybe v)

Tries to find a value corresponding to the key in the hash table.

toList :: (Eq k, Hashable k) => HashTable k v -> IO [(k, v)]

Converts the hash table to a list.

The order of the elements in the list is unspecified.

keySet :: Ord k => HashTable k v -> IO (Set k)