Agda.Utils.Lens

Description

A cut-down implementation of lenses, with names taken from Edward Kmett's lens package.

Documentation

(%%=) :: MonadState o m => Lens' o i -> (i -> m (i, r)) -> m r infix 4

Modify a part of the state monadically, and return some result.

(%=) :: MonadState o m => Lens' o i -> (i -> i) -> m () infix 4

Modify a part of the state.

(%==) :: MonadState o m => Lens' o i -> (i -> m i) -> m () infix 4

Modify a part of the state monadically.

(.=) :: MonadState o m => Lens' o i -> i -> m () infix 4

Write a part of the state.

(^.) :: o -> Lens' o i -> i infixl 8

Get inner part i of structure o as designated by Lens' o i.

contains :: Ord k => k -> Lens' (Set k) Bool

Focus on given element in a set.

focus :: forall (m :: Type -> Type) o i a. Monad m => Lens' o i -> StateT i m a -> StateT o m a

Focus on a part of the state for a stateful computation.

iso :: (o -> i) -> (i -> o) -> Lens' o i

Build a lens out of an isomorphism.

key :: Ord k => k -> Lens' (Map k v) (Maybe v)

Access a map value at a given key.

lFst :: forall a b f. Functor f => (a -> f a) -> (a, b) -> f (a, b)

lSnd :: forall a b f. Functor f => (b -> f b) -> (a, b) -> f (a, b)

lens :: LensGet o i -> LensSet o i -> Lens' o i

Build a lens from a getter and a setter.

lensProduct :: Lens' s a -> Lens' s b -> Lens' s (a, b)

Only sound if the lenses are disjoint!

locally :: MonadReader o m => Lens' o i -> (i -> i) -> m a -> m a

Modify a part of the state in a subcomputation.

locally' :: ((o -> o) -> m a -> m a) -> Lens' o i -> (i -> i) -> m a -> m a

locallyState :: MonadState o m => Lens' o i -> (i -> i) -> m r -> m r

Modify a part of the state locally.

over :: Lens' o i -> LensMap o i

Modify inner part i of structure o using a function i -> i.

set :: Lens' o i -> LensSet o i

Set inner part i of structure o as designated by Lens' o i.

use :: MonadState o m => Lens' o i -> m i

Read a part of the state.

view :: MonadReader o m => Lens' o i -> m i

Ask for part of read-only state.

type Lens' o i = forall (f :: Type -> Type). Functor f => (i -> f i) -> o -> f o

Van Laarhoven style homogeneous lenses. Mnemoic: "Lens outer inner", same type argument order as 'get :: o -> i'.

type LensGet o i = o -> i

type LensMap o i = (i -> i) -> o -> o

type LensSet o i = i -> o -> o

(<&>) :: Functor f => f a -> (a -> b) -> f b infixl 1

Flipped version of <$>.

(<&>) = flip fmap

Examples

Expand

Apply (+1) to a list, a Just and a Right:

>>> Just 2 <&> (+1)
Just 3

>>> [1,2,3] <&> (+1)
[2,3,4]

>>> Right 3 <&> (+1)
Right 4

Since: base-4.11.0.0

(&&&) :: Arrow a => a b c -> a b c' -> a b (c, c') infixr 3

Fanout: send the input to both argument arrows and combine their output.

The default definition may be overridden with a more efficient version if desired.

    ╭───────╮ c
  b │ ┌─ f ─┼───>
>───┼─┤     │
    │ └─ g ─┼───>
    ╰───────╯ c'