Agda.Utils.List

Description

Utility functions for lists.

Documentation

(!!) :: HasCallStack => [a] -> Int -> a

A variant of !! that might provide more informative error messages if the index is out of bounds.

Precondition: The index should not be out of bounds.

(!!!) :: [a] -> Int -> Maybe a

Lookup function (safe). O(min n index).

align :: [a] -> [b] -> [These a b]

Analogous to zip, combines two lists by taking the union using These (strict).

allConsecutive :: (a -> a -> Bool) -> [a] -> Bool

Check whether all consecutive elements of a list satisfy the given relation. O(n).

allDuplicates :: Ord a => [a] -> [a]

Remove the first representative for each list element. Thus, returns all duplicate copies. O(n log n).

allDuplicates xs == sort $ xs \ nub xs.

allEqual :: Eq a => [a] -> Bool

Checks if all the elements in the list are equal. Assumes that the Eq instance stands for an equivalence relation. O(n).

asum :: Alternative m => [m a] -> m a

A version of asum that avoids a final empty. It is right-folding just like asum.

Precondition: the right-unit law holds, i.e. m | A.empty = m.

asum1 :: Alternative m => m a -> [m a] -> m a

A right-folding asum for nonempty lists, never producing empty.

breakAfter :: (a -> Bool) -> [a] -> ([a], [a])

Breaks a list just after an element satisfying the predicate is found.

>>> breakAfter even [1,3,5,2,4,7,8]
([1,3,5,2],[4,7,8])

breakAfter1 :: (a -> Bool) -> a -> [a] -> (List1 a, [a])

Breaks a list just after an element satisfying the predicate is found.

>>> breakAfter1 even 1 [3,5,2,4,7,8]
(1 :| [3,5,2],[4,7,8])

breakJust :: (a -> Maybe b) -> [a] -> (Prefix a, (b, Suffix a))

Break a list when the given predicate returns Just b and place b as pivot between the prefix (where the predicate returns Nothing) and the suffix.

Crashes when the predicate holds nowhere.

caseList :: [a] -> b -> (a -> [a] -> b) -> b

Case distinction for lists, with list first. O(1).

Cf. ifNull.

caseListM :: Monad m => m [a] -> m b -> (a -> [a] -> m b) -> m b

Case distinction for lists, with list first. O(1).

Cf. ifNull.

chop :: Int -> [a] -> [[a]]

Chop up a list in chunks of a given length. O(n).

chopWhen :: (a -> Bool) -> [a] -> [[a]]

Chop a list at the positions when the predicate holds. Contrary to wordsBy, consecutive separator elements will result in an empty segment in the result. O(n).

intercalate [x] (chopWhen (== x) xs) == xs

commonPrefix :: Eq a => [a] -> [a] -> Prefix a

Compute the common prefix of two lists. O(min n m).

commonSuffix :: Eq a => [a] -> [a] -> Suffix a

Compute the common suffix of two lists. O(n + m).

distinct :: Eq a => [a] -> Bool

Check whether all elements in a list are distinct from each other. Assumes that the Eq instance stands for an equivalence relation.

O(n²) in the worst case distinct xs == True.

downFrom :: Integral a => a -> [a]

downFrom n = [n-1,..1,0]. O(n).

dropCommon :: [a] -> [b] -> (Suffix a, Suffix b)

Drops from both lists simultaneously until one list is empty. O(min n m).

dropEnd :: Int -> [a] -> Prefix a

Drop from the end of a list. O(length).

dropEnd n = reverse . drop n . reverse

Forces the whole list even for n==0.

dropFrom :: Eq a => List1 a -> [a] -> [a]

dropFrom marker xs drops everything from xs starting with (and including) marker.

If the marker does not appear, the string is returned unchanged.

The following two properties hold provided marker has no overlap with xs:

  dropFrom marker (xs ++ marker ++ ys) == xs
  dropFrom marker xs == xs

duplicates :: Ord a => [a] -> [a]

Returns an (arbitrary) representative for each list element that occurs more than once. O(n log n).

editDistance :: Eq a => [a] -> [a] -> Int

Implemented using dynamic programming and Data.Array. O(n*m).

editDistanceSpec :: Eq a => [a] -> [a] -> Int

Implemented using tree recursion, don't run me at home! O(3^(min n m)).

fastDistinct :: Ord a => [a] -> Bool

An optimised version of distinct. O(n log n).

Precondition: The list's length must fit in an Int.

filterAndRest :: (a -> Bool) -> [a] -> ([a], Suffix a)

Like filter, but additionally return the last partition of the list where the predicate is False everywhere. O(n).

findOverlap :: Eq a => [a] -> [a] -> (Int, Int)

Find the longest suffix of the first string xs that is a prefix of the second string ys. So, basically, find the overlap where the strings can be glued together. Returns the index where the overlap starts and the length of the overlap. The length of the overlap plus the index is the length of the first string. Note that in the worst case, the empty overlap (length xs,0) is returned.

Worst-case time complexity is quadratic: O(min(n,m)²) where n = length xs and m = length ys.

There might be asymptotically better implementations following Knuth-Morris-Pratt (KMP), but for rather short lists this is good enough.

findWithIndex :: (a -> Bool) -> [a] -> Maybe (a, Int)

Find an element satisfying a predicate and return it with its index. O(n) in the worst case, e.g. findWithIndex f xs = Nothing.

TODO: more efficient implementation!?

genericElemIndex :: (Eq a, Integral i) => a -> [a] -> Maybe i

A generalised variant of elemIndex. O(n).

hasElem :: Ord a => [a] -> a -> Bool

Check membership for the same list often. Use partially applied to create membership predicate hasElem xs :: a -> Bool.

First time: O(n log n) in the worst case.

Subsequently: O(log n).

Specification: hasElem xs == (elem xs).

headWithDefault :: a -> [a] -> a

Head function (safe). Returns a default value on empty lists. O(1).

headWithDefault 42 []      = 42
headWithDefault 42 [1,2,3] = 1

holes :: [a] -> [(a, [a])]

All ways of removing one element from a list. O(n²).

indexWithDefault :: a -> [a] -> Int -> a

Lookup function with default value for index out of range. O(min n index).

The name is chosen akin to genericIndex.

init1 :: a -> [a] -> [a]

init of non-empty list, safe. O(n). init1 a as = init (a:as)

initLast :: [a] -> Maybe ([a], a)

init and last in one go, safe. O(n).

initLast1 :: a -> [a] -> ([a], a)

init and last of non-empty list, safe. O(n). initLast1 a as = (init (a:as), last (a:as)

initMaybe :: [a] -> Maybe [a]

init, safe. O(n).

initWithDefault :: [a] -> [a] -> [a]

init, safe. O(n).

last1 :: a -> [a] -> a

Last element of non-empty list (safe). O(n). last1 a as = last (a : as)

last2 :: [a] -> Maybe (a, a)

Last two elements (safe). O(n).

last2' :: a -> a -> [a] -> (a, a)

last2' x y zs computes the last two elements of x:y:zs. O(n).

lastMaybe :: [a] -> Maybe a

Last element (safe). O(n).

lastWithDefault :: a -> [a] -> a

Last element (safe). Returns a default list on empty lists. O(n).

listCase :: b -> (a -> [a] -> b) -> [a] -> b

Case distinction for lists, with list last. O(1).

mapMaybeAndRest :: (a -> Maybe b) -> [a] -> ([b], Suffix a)

Like mapMaybe, but additionally return the last partition of the list where the function always returns Nothing. O(n).

mcons :: Maybe a -> [a] -> [a]

Maybe cons. O(1). mcons ma as = maybeToList ma ++ as

mergeStrictlyOrderedBy :: (a -> a -> Bool) -> [a] -> [a] -> Maybe [a]

nubAndDuplicatesOn :: Ord b => (a -> b) -> [a] -> ([a], [a])

Partition a list into first and later occurrences of elements (modulo some quotient given by a representation function).

Time: O(n log n).

Specification:

nubAndDuplicatesOn f xs = (ys, xs List.\\ ys)
  where ys = nubOn f xs

nubFavouriteOn

Arguments

:: forall a b c. (Ord b, Eq c, Hashable c)

=> (a -> b)
The values returned by this function are used to determine which element from a group of equal elements that is returned: the smallest one is chosen (and if two elements are equally small, then the first one is chosen).

-> (a -> c)
Two elements are treated as equal if this function returns the same value for both elements.

-> [a]

-> [a]

A variant of nubOn that is parametrised by a function that is used to select which element from a group of equal elements that is returned. The returned elements keep the order that they had in the input list.

Precondition: The length of the input list must be at most maxBound :: Int.

nubM :: Monad m => (a -> a -> m Bool) -> [a] -> m [a]

Non-efficient, monadic nub. O(n²).

nubOn :: Ord b => (a -> b) -> [a] -> [a]

Efficient variant of nubBy for lists, using a set to store already seen elements. O(n log n)

Specification:

nubOn f xs == 'nubBy' ((==) `'on'` f) xs.

partitionMaybe :: (a -> Maybe b) -> [a] -> ([a], [b])

Partition a list into Nothings and Justs. O(n).

partitionMaybe f = partitionEithers . map ( a -> maybe (Left a) Right (f a))

Note: mapMaybe f = snd . partitionMaybe f.

snoc :: [a] -> a -> [a]

Append a single element at the end. Time: O(length); use only on small lists.

sorted :: Ord a => [a] -> Bool

Check whether a list is sorted. O(n).

Assumes that the Ord instance implements a partial order.

spanEnd :: (a -> Bool) -> [a] -> (Prefix a, Suffix a)

Split off the largest suffix whose elements satisfy a predicate. O(n).

spanEnd p xs = (ys, zs) where xs = ys ++ zs and all p zs and maybe True (not . p) (lastMaybe yz).

spanJust :: (a -> Maybe b) -> [a] -> (Prefix b, Suffix a)

A generalized version of span. O(length . fst . spanJust f).

splitExactlyAt :: Integral n => n -> [a] -> Maybe (Prefix a, Suffix a)

splitExactlyAt n xs = Just (ys, zs) iff xs = ys ++ zs and genericLength ys = n.

stripPrefixBy :: (a -> a -> Bool) -> Prefix a -> [a] -> Maybe (Suffix a)

Check if a list has a given prefix. If so, return the list minus the prefix. O(length prefix).

stripReversedSuffix :: Eq a => ReversedSuffix a -> [a] -> Maybe (Prefix a)

stripReversedSuffix rsuf xs = Just pre iff xs = pre ++ reverse suf. O(n).

stripSuffix :: Eq a => Suffix a -> [a] -> Maybe (Prefix a)

stripSuffix suf xs = Just pre iff xs = pre ++ suf. O(n).

suffixesSatisfying :: (a -> Bool) -> [a] -> [Bool]

Returns a list with one boolean for each non-empty suffix of the list, starting with the longest suffix (the entire list). Each boolean is True exactly when every element in the corresponding suffix satisfies the predicate.

An example: suffixesSatisfying isLower AbCde = [False, False, False, True, True]

For total predicates p and finite and total lists xs the following holds: suffixesSatisfying p xs = map (all p) (init (tails xs))

tailMaybe :: [a] -> Maybe [a]

Tail function (safe). O(1).

tailWithDefault :: [a] -> [a] -> [a]

Tail function (safe). Returns a default list on empty lists. O(1).

takeExactly :: forall a n. Integral n => a -> n -> [a] -> [a]

takeExactly a n as == take n (as ++ repeat a)

takeWhileJust :: (a -> Maybe b) -> [a] -> Prefix b

A generalized version of takeWhile. (Cf. mapMaybe vs. filter). @O(length . takeWhileJust f).

takeWhileJust f = fst . spanJust f.

uniqOn :: Ord b => (a -> b) -> [a] -> [a]

Efficient variant of nubBy for finite lists. O(n log n).

uniqOn f == 'List.sortBy' (compare `'on'` f) . 'nubBy' ((==) `'on'` f)

If there are several elements with the same f-representative, the first of these is kept.

unzipWith :: (a -> (b, c)) -> [a] -> ([b], [c])

updateAt :: Int -> (a -> a) -> [a] -> [a]

Update nth element of a list, if it exists. O(min index n).

Precondition: the index is >= 0.

updateHead :: (a -> a) -> [a] -> [a]

Update the first element of a list, if it exists. O(1).

updateLast :: (a -> a) -> [a] -> [a]

Update the last element of a list, if it exists. O(n).

zipWith' :: (a -> b -> c) -> [a] -> [b] -> Maybe [c]

Requires both lists to have the same length. O(n).

Otherwise, Nothing is returned.

zipWithKeepRest :: (a -> b -> b) -> [a] -> [b] -> [b]

Like zipWith but keep the rest of the second list as-is (in case the second list is longer). O(n).

  zipWithKeepRest f as bs == zipWith f as bs ++ drop (length as) bs

type Prefix a

Arguments

= [a]
The list before the split point.

type ReversedSuffix a = [a]

data StrSufSt a

Internal state for stripping suffix.

Constructors

SSSMismatch
Error.

SSSStrip (ReversedSuffix a)
"Negative string" to remove from end. List may be empty.

SSSResult [a]
"Positive string" (result). Non-empty list.

type Suffix a

Arguments

= [a]
The list after the split point.

uncons :: [a] -> Maybe (a, [a])

\(\mathcal{O}(1)\). Decompose a list into its head and tail.

If the list is empty, returns Nothing.

If the list is non-empty, returns Just (x, xs), where x is the head of the list and xs its tail.

Examples

Expand

>>> uncons []
Nothing

>>> uncons [1]
Just (1,[])

>>> uncons [1, 2, 3]
Just (1,[2,3])

Since: base-4.8.0.0