Agda.Utils.Monad

Documentation

finally :: MonadError e m => m a -> m () -> m a

Finally for the Error class. Errors in the finally part take precedence over prior errors.

bracket_

Arguments

:: Monad m

=> m a
Acquires resource. Run first.

-> (a -> m ())
Releases resource. Run last.

-> m b
Computes result. Run in-between.

-> m b

Bracket without failure. Typically used to preserve state.

andM :: (Foldable f, Monad m) => f (m Bool) -> m Bool

mapMaybeM :: Monad m => (a -> m (Maybe b)) -> [a] -> m [b]

A monadic version of mapMaybe :: (a -> Maybe b) -> [a] -> [b].

dropWhileM :: Monad m => (a -> m Bool) -> [a] -> m [a]

A monadic version of dropWhile :: (a -> Bool) -> [a] -> [a].

concatMapM :: Monad m => (a -> m [b]) -> [a] -> m [b]

ifM :: Monad m => m Bool -> m a -> m a -> m a

Monadic if-then-else.

unlessM :: Monad m => m Bool -> m () -> m ()

guardWithError :: MonadError e m => e -> Bool -> m ()

Like guard, but raise given error when condition fails.

tryMaybe :: (MonadError e m, Functor m) => m a -> m (Maybe a)

Try a computation, return Nothing if an Error occurs.

tell1 :: (Monoid ws, Singleton w ws, MonadWriter ws m) => w -> m ()

Output a single value.

(==<<) :: Monad m => (a -> b -> m c) -> (m a, m b) -> m c

Binary bind.

whenM :: Monad m => m Bool -> m () -> m ()

guardM :: (Monad m, MonadPlus m) => m Bool -> m ()

Monadic guard.

ifNotM :: Monad m => m Bool -> m a -> m a -> m a

ifNotM mc = ifM (not $ mc)

and2M :: Monad m => m Bool -> m Bool -> m Bool

Lazy monadic conjunction.

allM :: (Foldable f, Monad m) => f a -> (a -> m Bool) -> m Bool

or2M :: Monad m => m Bool -> m Bool -> m Bool

Lazy monadic disjunction.

orM :: (Foldable f, Monad m) => f (m Bool) -> m Bool

anyM :: (Foldable f, Monad m) => f a -> (a -> m Bool) -> m Bool

altM1 :: Monad m => (a -> m (Either err b)) -> [a] -> m (Either err b)

Lazy monadic disjunction with Either truth values. Returns the last error message if all fail.

orEitherM :: (Monoid e, Monad m, Functor m) => [m (Either e b)] -> m (Either e b)

Lazy monadic disjunction with accumulation of errors in a monoid. Errors are discarded if we succeed.

mapM' :: (Foldable t, Applicative m, Monoid b) => (a -> m b) -> t a -> m b

Generalized version of traverse_ :: Applicative m => (a -> m ()) -> [a] -> m () Executes effects and collects results in left-to-right order. Works best with left-associative monoids.

Note that there is an alternative

mapM' f t = foldr mappend mempty $ mapM f t

that collects results in right-to-left order (effects still left-to-right). It might be preferable for right associative monoids.

forM' :: (Foldable t, Applicative m, Monoid b) => t a -> (a -> m b) -> m b

Generalized version of for_ :: Applicative m => [a] -> (a -> m ()) -> m ()

mapMM :: (Traversable t, Monad m) => (a -> m b) -> m (t a) -> m (t b)

forMM :: (Traversable t, Monad m) => m (t a) -> (a -> m b) -> m (t b)

mapMM_ :: (Foldable t, Monad m) => (a -> m ()) -> m (t a) -> m ()

forMM_ :: (Foldable t, Monad m) => m (t a) -> (a -> m ()) -> m ()

mapMaybeMM :: Monad m => (a -> m (Maybe b)) -> m [a] -> m [b]

A version of mapMaybeM with a computation for the input list.

forMaybeM :: Monad m => [a] -> (a -> m (Maybe b)) -> m [b]

The for version of mapMaybeM.

forMaybeMM :: Monad m => m [a] -> (a -> m (Maybe b)) -> m [b]

The for version of mapMaybeMM.

dropWhileEndM :: Monad m => (a -> m Bool) -> [a] -> m [a]

A monadic version of dropWhileEnd :: (a -> Bool) -> [a] -> m [a]. Effects happen starting at the end of the list until p becomes false.

partitionM :: (Functor m, Applicative m) => (a -> m Bool) -> [a] -> m ([a], [a])

A `monadic' version of @partition :: (a -> Bool) -> [a] -> ([a],[a])

fromMaybeMP :: MonadPlus m => Maybe a -> m a

Translates Maybe to MonadPlus.

catMaybesMP :: MonadPlus m => m (Maybe a) -> m a

Generalises the catMaybes function from lists to an arbitrary MonadPlus.

scatterMP :: (MonadPlus m, Foldable t) => m (t a) -> m a

Branch over elements of a monadic Foldable data structure.

tryCatch :: (MonadError e m, Functor m) => m () -> m (Maybe e)

Run a command, catch the exception and return it.

localState :: MonadState s m => m a -> m a

Restore state after computation.

embedWriter :: forall w (m :: Type -> Type) a. (Monoid w, Monad m) => Writer w a -> WriterT w m a

when :: Applicative f => Bool -> f () -> f ()

Conditional execution of Applicative expressions. For example,

when debug (putStrLn "Debugging")

will output the string Debugging if the Boolean value debug is True, and otherwise do nothing.

unless :: Applicative f => Bool -> f () -> f ()

The reverse of when.

class (Alternative m, Monad m) => MonadPlus (m :: Type -> Type) where

Monads that also support choice and failure.

Minimal complete definition

Nothing

Methods

mzero :: m a

The identity of mplus. It should also satisfy the equations

mzero >>= f  =  mzero
v >> mzero   =  mzero

The default definition is

mzero = empty

mplus :: m a -> m a -> m a

An associative operation. The default definition is

mplus = (<|>)

(<$>) :: Functor f => (a -> b) -> f a -> f b infixl 4

An infix synonym for fmap.

The name of this operator is an allusion to $. Note the similarities between their types:

 ($)  ::              (a -> b) ->   a ->   b
(<$>) :: Functor f => (a -> b) -> f a -> f b

Whereas $ is function application, <$> is function application lifted over a Functor.

Examples

Expand

Convert from a Maybe Int to a Maybe String using show:

>>> show <$> Nothing
Nothing
>>> show <$> Just 3
Just "3"

Convert from an Either Int Int to an Either Int String using show:

>>> show <$> Left 17
Left 17
>>> show <$> Right 17
Right "17"

Double each element of a list:

>>> (*2) <$> [1,2,3]
[2,4,6]

Apply even to the second element of a pair:

>>> even <$> (2,2)
(2,True)

(<*>) :: Applicative f => f (a -> b) -> f a -> f b infixl 4

Sequential application.

A few functors support an implementation of <*> that is more efficient than the default one.

Example

Expand

Used in combination with (<$>), (<*>) can be used to build a record.

>>> data MyState = MyState {arg1 :: Foo, arg2 :: Bar, arg3 :: Baz}

>>> produceFoo :: Applicative f => f Foo

>>> produceBar :: Applicative f => f Bar
>>> produceBaz :: Applicative f => f Baz

>>> mkState :: Applicative f => f MyState
>>> mkState = MyState <$> produceFoo <*> produceBar <*> produceBaz

(<$) :: Functor f => a -> f b -> f a infixl 4

Replace all locations in the input with the same value. The default definition is fmap . const, but this may be overridden with a more efficient version.

Examples

Expand

Perform a computation with Maybe and replace the result with a constant value if it is Just:

>>> 'a' <$ Just 2
Just 'a'
>>> 'a' <$ Nothing
Nothing