Agda.Utils.Monad

Documentation

(==<<) :: Monad m => (a -> b -> m c) -> (m a, m b) -> m c

Binary bind.

whenM :: Monad m => m Bool -> m () -> m ()

unlessM :: Monad m => m Bool -> m () -> m ()

guardM :: (Monad m, MonadPlus m) => m Bool -> m ()

Monadic guard.

ifM :: Monad m => m Bool -> m a -> m a -> m a

Monadic if-then-else.

ifNotM :: Monad m => m Bool -> m a -> m a -> m a

ifNotM mc = ifM (not $ mc)

and2M :: Monad m => m Bool -> m Bool -> m Bool

Lazy monadic conjunction.

andM :: (Foldable f, Monad m) => f (m Bool) -> m Bool

allM :: (Functor f, Foldable f, Monad m) => f a -> (a -> m Bool) -> m Bool

or2M :: Monad m => m Bool -> m Bool -> m Bool

Lazy monadic disjunction.

orM :: (Foldable f, Monad m) => f (m Bool) -> m Bool

anyM :: (Functor f, Foldable f, Monad m) => f a -> (a -> m Bool) -> m Bool

altM1 :: Monad m => (a -> m (Either err b)) -> [a] -> m (Either err b)

Lazy monadic disjunction with Either truth values. Returns the last error message if all fail.

orEitherM :: (Monoid e, Monad m, Functor m) => [m (Either e b)] -> m (Either e b)

Lazy monadic disjunction with accumulation of errors in a monoid. Errors are discarded if we succeed.

mapM' :: (Foldable t, Applicative m, Monoid b) => (a -> m b) -> t a -> m b

Generalized version of traverse_ :: Applicative m => (a -> m ()) -> [a] -> m () Executes effects and collects results in left-to-right order. Works best with left-associative monoids.

Note that there is an alternative

mapM' f t = foldr mappend mempty $ mapM f t

that collects results in right-to-left order (effects still left-to-right). It might be preferable for right associative monoids.

forM' :: (Foldable t, Applicative m, Monoid b) => t a -> (a -> m b) -> m b

Generalized version of for_ :: Applicative m => [a] -> (a -> m ()) -> m ()

mapMM :: (Traversable t, Monad m) => (a -> m b) -> m (t a) -> m (t b)

forMM :: (Traversable t, Monad m) => m (t a) -> (a -> m b) -> m (t b)

mapMM_ :: (Foldable t, Monad m) => (a -> m ()) -> m (t a) -> m ()

forMM_ :: (Foldable t, Monad m) => m (t a) -> (a -> m ()) -> m ()

mapMaybeM :: Monad m => (a -> m (Maybe b)) -> [a] -> m [b]

A monadic version of mapMaybe :: (a -> Maybe b) -> [a] -> [b].

mapMaybeMM :: Monad m => (a -> m (Maybe b)) -> m [a] -> m [b]

A version of mapMaybeM with a computation for the input list.

forMaybeM :: Monad m => [a] -> (a -> m (Maybe b)) -> m [b]

The for version of mapMaybeM.

forMaybeMM :: Monad m => m [a] -> (a -> m (Maybe b)) -> m [b]

The for version of mapMaybeMM.

dropWhileM :: Monad m => (a -> m Bool) -> [a] -> m [a]

A monadic version of dropWhile :: (a -> Bool) -> [a] -> [a].

dropWhileEndM :: Monad m => (a -> m Bool) -> [a] -> m [a]

A monadic version of dropWhileEnd :: (a -> Bool) -> [a] -> m [a]. Effects happen starting at the end of the list until p becomes false.

partitionM :: (Functor m, Applicative m) => (a -> m Bool) -> [a] -> m ([a], [a])

A `monadic' version of @partition :: (a -> Bool) -> [a] -> ([a],[a])

fromMaybeMP :: MonadPlus m => Maybe a -> m a

Translates Maybe to MonadPlus.

catMaybesMP :: MonadPlus m => m (Maybe a) -> m a

Generalises the catMaybes function from lists to an arbitrary MonadPlus.

scatterMP :: (MonadPlus m, Foldable t) => m (t a) -> m a

Branch over elements of a monadic Foldable data structure.

finally :: MonadError e m => m a -> m () -> m a

Finally for the Error class. Errors in the finally part take precedence over prior errors.

tryMaybe :: (MonadError e m, Functor m) => m a -> m (Maybe a)

Try a computation, return Nothing if an Error occurs.

tryCatch :: (MonadError e m, Functor m) => m () -> m (Maybe e)

Run a command, catch the exception and return it.

guardWithError :: MonadError e m => e -> Bool -> m ()

Like guard, but raise given error when condition fails.

bracket_

Arguments

:: Monad m

=> m a
Acquires resource. Run first.

-> (a -> m ())
Releases resource. Run last.

-> m b
Computes result. Run in-between.

-> m b

Bracket without failure. Typically used to preserve state.

localState :: MonadState s m => m a -> m a

Restore state after computation.

when :: Applicative f => Bool -> f () -> f ()

unless :: Applicative f => Bool -> f () -> f ()

class (Alternative m, Monad m) => MonadPlus (m :: Type -> Type) where

Minimal complete definition

Nothing

Methods

mzero :: m a

mplus :: m a -> m a -> m a

(<$>) :: Functor f => (a -> b) -> f a -> f b

(<*>) :: Applicative f => f (a -> b) -> f a -> f b

(<$) :: Functor f => a -> f b -> f a