Agda.Utils.Pointer

Documentation

data Ptr a

newPtr :: a -> Ptr a

derefPtr :: Ptr a -> a

setPtr :: a -> Ptr a -> Ptr a

updatePtr :: (a -> a) -> Ptr a -> Ptr a

updatePtrM :: Functor f => (a -> f a) -> Ptr a -> f (Ptr a)

If f a contains many copies of a they will all be the same pointer in the result. If the function is well-behaved (i.e. preserves the implicit equivalence, this shouldn't matter).

Orphan instances

Typeable a => Data (IORef a)

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> IORef a -> c (IORef a)

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c (IORef a)

toConstr :: IORef a -> Constr

dataTypeOf :: IORef a -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c (IORef a))

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c (IORef a))

gmapT :: (forall b. Data b => b -> b) -> IORef a -> IORef a

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> IORef a -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> IORef a -> r

gmapQ :: (forall d. Data d => d -> u) -> IORef a -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> IORef a -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> IORef a -> m (IORef a)

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> IORef a -> m (IORef a)

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> IORef a -> m (IORef a)