Agda.Utils.SmallSet

Description

Small sets represented as a bitmask for fast membership checking.

With the exception of converting to/from lists, all operations are O(1).

Mimics the interface of Set.

Import as: import qualified Agda.Utils.SmallSet as SmallSet import Agda.Utils.SmallSet (SmallSet)

Documentation

newtype SmallSet a

Constructors

SmallSet

Fields

theSmallSet :: Word64

class (Bounded a, Ix a) => SmallSetElement a

An element in a small set.

This must implement Bounded and Ix, and contain at most 64 values.

class Ord a => Ix a

The Ix class is used to map a contiguous subrange of values in a type onto integers. It is used primarily for array indexing (see the array package).

The first argument (l,u) of each of these operations is a pair specifying the lower and upper bounds of a contiguous subrange of values.

An implementation is entitled to assume the following laws about these operations:

inRange (l,u) i == elem i (range (l,u))

range (l,u) !! index (l,u) i == i, when inRange (l,u) i

map (index (l,u)) (range (l,u))) == [0..rangeSize (l,u)-1]

rangeSize (l,u) == length (range (l,u))

Minimal complete definition

range, (index | unsafeIndex), inRange

(\\) :: SmallSetElement a => SmallSet a -> SmallSet a -> SmallSet a

Time O(1).

complement :: SmallSetElement a => SmallSet a -> SmallSet a

Time O(n).

delete :: SmallSetElement a => a -> SmallSet a -> SmallSet a

Time O(1).

difference :: SmallSetElement a => SmallSet a -> SmallSet a -> SmallSet a

Time O(1).

elems :: SmallSetElement a => SmallSet a -> [a]

Time O(n).

empty :: SmallSetElement a => SmallSet a

The empty set. Time O(1).

fromList :: SmallSetElement a => [a] -> SmallSet a

Time O(n).

fromAscList :: SmallSetElement a => [a] -> SmallSet a

Time O(n).

fromDistinctAscList :: SmallSetElement a => [a] -> SmallSet a

Time O(n).

insert :: SmallSetElement a => a -> SmallSet a -> SmallSet a

Time O(1).

intersection :: SmallSetElement a => SmallSet a -> SmallSet a -> SmallSet a

Time O(1).

member :: SmallSetElement a => a -> SmallSet a -> Bool

Time O(1).

notMember :: SmallSetElement a => a -> SmallSet a -> Bool

not . member a. Time O(1).

null :: SmallSetElement a => SmallSet a -> Bool

Time O(1).

singleton :: SmallSetElement a => a -> SmallSet a

A singleton set. Time O(1).

toList :: SmallSetElement a => SmallSet a -> [a]

Time O(n).

toAscList :: SmallSetElement a => SmallSet a -> [a]

Time O(n).

total :: SmallSetElement a => SmallSet a

The full set. Time O(1).

union :: SmallSetElement a => SmallSet a -> SmallSet a -> SmallSet a

Time O(n).