Agda.Utils.SmallSet

Description

Small sets represented as immutable bit arrays for fast membership checking.

Membership checking is O(1), but all other operations are O(n) where n is the size of the element type. The element type needs to implement Bounded and Ix.

Mimics the interface of Set.

Import as: import qualified Agda.Utils.SmallSet as SmallSet import Agda.Utils.SmallSet (SmallSet)

Documentation

data SmallSet a

class Ord a => Ix a

Minimal complete definition

range, (index | unsafeIndex), inRange

(\\) :: SmallSetElement a => SmallSet a -> SmallSet a -> SmallSet a

Time O(n).

complement :: SmallSetElement a => SmallSet a -> SmallSet a

Time O(n).

delete :: SmallSetElement a => a -> SmallSet a -> SmallSet a

Time O(n).

difference :: SmallSetElement a => SmallSet a -> SmallSet a -> SmallSet a

Time O(n).

elems :: SmallSetElement a => SmallSet a -> [a]

Time O(n).

empty :: SmallSetElement a => SmallSet a

The empty set. Time O(n).

fromList :: SmallSetElement a => [a] -> SmallSet a

Time O(n).

fromAscList :: SmallSetElement a => [a] -> SmallSet a

Time O(n).

fromDistinctAscList :: SmallSetElement a => [a] -> SmallSet a

Time O(n).

insert :: SmallSetElement a => a -> SmallSet a -> SmallSet a

Time O(n).

intersection :: SmallSetElement a => SmallSet a -> SmallSet a -> SmallSet a

Time O(n).

isSubsetOf :: SmallSetElement a => SmallSet a -> SmallSet a -> Bool

Time O(n).

mapMemberShip :: SmallSetElement a => (Bool -> Bool) -> SmallSet a -> SmallSet a

Time O(n).

member :: SmallSetElement a => a -> SmallSet a -> Bool

Time O(1).

notMember :: SmallSetElement a => a -> SmallSet a -> Bool

not . member a. Time O(1).

null :: SmallSetElement a => SmallSet a -> Bool

Time O(n)!

singleton :: SmallSetElement a => a -> SmallSet a

A singleton set. Time O(n).

toList :: SmallSetElement a => SmallSet a -> [a]

Time O(n).

toAscList :: SmallSetElement a => SmallSet a -> [a]

Time O(n).

total :: SmallSetElement a => SmallSet a

The full set. Time O(n).

union :: SmallSetElement a => SmallSet a -> SmallSet a -> SmallSet a

Time O(n).

zipMemberShipWith :: SmallSetElement a => (Bool -> Bool -> Bool) -> SmallSet a -> SmallSet a -> SmallSet a

Time O(n).