Agda.Utils.StrictState

Description

This is a plain strict state monad, where state update, monadic binding and return are all strict. If you only need a single strict state effect, use this module.

Do not use Control.Monad.State.Strict for the same purpose; it's not even strict in state updates and is much less amenable to GHC optimizations than this module.

Documentation

newtype State s a

Constructors

State

Fields

runState# :: s -> (# a, s #)

put :: s -> State s ()

get :: State s s

gets :: (s -> a) -> State s a

modify :: (s -> s) -> State s ()

execState :: State s a -> s -> s

runState :: State s a -> s -> (a, s)

evalState :: State s a -> s -> a