Agda.Utils.String

Documentation

quote :: String -> String

quote adds double quotes around the string, replaces newline characters with n, and escapes double quotes and backslashes within the string. This is different from the behaviour of show:

> putStrLn $ show "\x2200"
"\8704"
> putStrLn $ quote "\x2200"
"∀"

(The code examples above have been tested using version 4.2.0.0 of the base library.)

haskellStringLiteral :: String -> String

Turns the string into a Haskell string literal, avoiding escape codes.

delimiter :: String -> String

Adds hyphens around the given string

>>> putStrLn $ delimiter "Title"
———— Title —————————————————————————————————————————————————

addFinalNewLine :: String -> String

Adds a final newline if there is not already one.

indent :: Integral i => i -> String -> String

Indents every line the given number of steps.

unwords1 :: [String] -> String

unwords, but remove empty words first.

showThousandSep :: Show a => a -> String

Show a number using comma to separate powers of 1,000.

ltrim :: String -> String

Remove leading whitespace.

rtrim :: String -> String

Remove trailing whitespace.

trim :: String -> String

Remove leading and trailing whitesapce.

Orphan instances

(IsString (m a), Monad m) => IsString (ReaderT r m a)

Instance details

Methods

fromString :: String -> ReaderT r m a

(IsString (m a), Monad m) => IsString (StateT s m a)

Instance details

Methods

fromString :: String -> StateT s m a